[roll] Roll fuchsia [kernel][arm] fix a typo in the boot allocator
The code is trying to compute a bitmask for a page size, but
accidentally uses -2 instead of -1, which would result in a mask that
has the bottom bit unset. It should mean it only matches on every other
large page, so it's been mostly harmless except it doesn't utilize large
pages as well as it should have been able to.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043712
Original-Revision: 21c473df26f2acce5daf8aa7c7fc8d4b627d13cc
GitOrigin-RevId: 23867dc85dcfe564d6ae41ed03f043826989561f
Change-Id: Icd556c00254920d21b9133d927cf45dbdd6bdeba