[roll] Roll fuchsia [power] Link to plan to move lease to eventpair
People keep rediscovering that leases should be eventpairs. This CL adds
a TODO comment that links to the bug for that change so that we are less
likely to spend energy rediscovering that property.

At the moment, moving to an eventpair isn't blocking anything, so
actually executing this change isn't urgent.

Original-Bug: b/339474151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045157
Original-Revision: 4ab1966cc1e8d645a27f8c90708d6d10bf7d9667
GitOrigin-RevId: ad673e970b7b36ac15b570001a231e8d6e65e70b
Change-Id: I838e7773b56672fdea720088997d67d535c8966f