[roll] Roll fuchsia [power][sag] Create FakeSAG AIB
This change splits out the system-activity-governor out of the power
framework shard, so that we can have two variants of it based on which
AIB gets picked from the configuration. The testing variant is the new
fake SAG, and we also allow it to be accessed from the console launcher
so that bootfs tests can access its protocols. This change also allows
power framework to be included in the bootstrap feature set. Finally it
allows the fake-suspend driver to bind to the same node the real suspend
driver can bind to.

Original-Bug: 335526421
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042045
Original-Revision: d881929116a1dd76e9b951e539c49f56e488b092
GitOrigin-RevId: 1a094df26969ba486bae3913dc09d599a04733c4
Change-Id: Ia5b4f54ead81db3b0c34f181bb3b661a89a396b1