[roll] Roll fuchsia [fdf] Pass properties from composite node spec
Node properties of composite nodes contained all the properties of the
parent. This is different from the properties used for matching against
bind rules. This CL publishes only the properties of the parents exposed
from the composite node spec. We may have to follow up this change to
merge the properties of the parent as well.

Original-Bug: 338667323
Test: composite_node_spec_v2_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042464
Original-Revision: 946f7497b52f3e0c9732d7810a629e0af94e1711
GitOrigin-RevId: 4723810ec6f941be9b013b9993178fd2684ad34b
Change-Id: I3282a7b11ca830e9a7d73416cf748dbf72e55c2d