[roll] Roll fuchsia [sdk][fidl] add comment to fuchsia.hardware.power.SagElement
When searching for a list of SAG elements, I was confused by the
`SagElement` enum. I thought the enum specified all of the
SAG elements.

That turns out not to be the purpose of the `SagElement` enum.
Rather, the enum is meant to be used by clients of the power
support library. (In retrospect, `fuchsia.hardware.power`
does suggest something driver-specific.)

To help my future self, add a comment explaining the purpose
of `SagElement`.

Alternatively: we could add the currently missing element to
the enum. But that has a couple of problems:

1. It invites driver authors to use the enum value, although
   that value isn't supported by the power support library.
2. It creates a maintenance burden, if/when we add additional
   elements to SAG.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046092
Original-Revision: 616382c1ac3bedb1665c9f778d4bb087d220ead7
GitOrigin-RevId: 21832843e6e2926a90c322743d6530d9b7a1b763
Change-Id: I8fc826c4bd1d5d182487b33fb73df9629e095046