[roll] Roll fuchsia [fxfs][fsck] Fix spurious fsck error
Running fsck on a filesystem with an empty volume was erroneously
failing.  The allocator reports (correctly) that 0 bytes are allocated
to the volume, but Fsck fails to filter this out and complains that the
owner_allocated_bytes is different than what is observed.

Original-Bug: 306749272
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045220
Original-Revision: a562ae6dcb54fb4feaa8a67913b4e46affbf8319
GitOrigin-RevId: 845d40b7b9cb8364c0c40aacbb4d9167b120e6e0
Change-Id: Ib307f0b9a5884555cb3c8cc008cd68e935c9b31c