[roll] Roll fuchsia [fidl][power] Mark statecontrol protocols as server=platform
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or extenal components as such.

This marks all of the fuchsia.hardware.power.statecontrol protocols as
allowing clients in platform or external components but only allowing
servers in platform components.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046056
Original-Revision: 6d5c99433f419ca21406dac56552068efcaf9ad4
GitOrigin-RevId: ba5d3bdf06ca0725355536d205c50175f4b61a9d
Change-Id: I4af50b6b8abb6e70cec8d805cbe9c4887462aa06