[roll] Roll fuchsia [fidl][castauth] Mark protocols as server=platform
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or extenal components as such.

This marks all of the fuchsia.castauth protocols as allowing clients in
platform or external components but only allowing servers in platform
components.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046055
Original-Revision: ec2ba24a7d1800a647a53d24403ca3877ec6c02d
GitOrigin-RevId: 86599426618844c7c970a76a0055ef0c5e034d05
Change-Id: Ib60b8f1b3925f0133e93cc05293924e41d9ebced