[roll] Roll fuchsia [sdio][fidl] Correct usage of MAX_TRAN_SPEED CIS field
MAX_TRAN_SPEED is set in the function 0 CIS tuple but applies to all
functions on the card. Providing it only in the function 0 info makes it
inaccessible, as no node is added for function 0. Move it to the
top-level device info and update the core driver and tests accordingly.

Original-Bug: 333726427
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045601
Original-Revision: 65e8baa4e43bd1648e9c637171de0142c38fe13a
GitOrigin-RevId: bf4c26638a22087b895c0aa813538f752f957780
Change-Id: Ie63b27b3721c49727c6b90e525863532f7734b80