[roll] Roll fuchsia [audio][drivers] move to started/bypassed
Migrate drivers and driver clients away from signalprocessing
property `enabled`, to `started` and/or `bypassed`.

Field enabled is already in the process of being deprecated.
Also a coming CL creates a subset of ElementState that will
be used for SetElementState, and it does not include enabled,
so this CL is early preparation for that change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046066
Original-Revision: c723aacec01261c11c3b78d55422b876e6038ca3
GitOrigin-RevId: afd4b7a82067954fef35bdc28a9ab9c93ea47086
Change-Id: I6e9a4311aa4ce5eaa7c1737658998dc02a887c72