[roll] Roll llvm-project/libc [libc] Fix GPU handling for unsupported backends (#92271)
Summary:
If the user does not have the selected backend enabled, we should still
be able to build the LLVM-IR an ddistribute it. This patch makes logic
to suppress tests if the backend can't build it, as well as removing a
flag for the building that's only present int he NVPTX backend.
GitOrigin-RevId: e5a190cf1567a4a62e0bcc6d79dcdaa5a9866308
Original-Revision: 64cf380a67fa8e8c844d52e4d1f22b9db5fe9832
Change-Id: Id4926e39a402e20b770b38864933c266ed0666f7