[roll] Roll fuchsia [block][sdhci] Refactor PendingRequest
Move PendingRequest into an optional. This will be helpful when buffered
IO is implemented, as it enables the lifetime of an fzl::VmoMapper to be
tied to the PendingRequest object.

Also simplify the PendingRequest status fields somewhat, and use them to
enforce the expected interrupt order (i.e. it is acceptable for the
command and transfer complete interrupts to be received at the same
time, but not for the latter to be received before the former).

Test: sdhci-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048376
Original-Revision: 7d483ea6264adffdd24cdb17f6807fd8eef0e32d
GitOrigin-RevId: 9407c2fb3f5e24b77353db7f4496edaa0ee34d6c
Change-Id: Id7ed56bfd053f6aa9e0a1c1ddd6eab71614d952a