[roll] Roll fuchsia [kernel][vm] Remove page request batching
The batching system, both internal and external, are vestigial and from
a time when every page offset was looked up individually for a request.
Current lookup cursors and methods now find and request their entire
range at once, and do not need to append to an existing range. The
current batching logic is, therefore, completely unused and the code
paths are not triggered.

Although batching itself is unused, a page request having been created
with a different batch state changed how failed ranges worked. This
behavior does not make sense, and the comment justifying it is wrong,
since the external callers are in fact explicitly requesting the range
that is being failed, and so will not be surprised. Changing this
behavior does impact some tests that were explicitly testing for it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045573
Original-Revision: af8d32d2c9763ba0b5f1ecf818675a0ce8230b00
GitOrigin-RevId: de3189b3f4e4ff7b1b52f2bf42f95dd62719baea
Change-Id: Ib742469d77dd13d00583db9719b416d7c56cdc5a