[roll] Roll fuchsia [cmc] Clear source_dictionary when source is voided
Fix a bug where, if `source_availability: unknown` is used and the
source is a dictionary exposed by a child, the source is set to Void but
the source_dictionary is left set.

Not clearing source_dictionary may break assumptions about state (in the
validator for example).

Original-Fixed: 340624184

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048374
Original-Revision: a401a604d50fcacf2c025e3ebe3b2b5733e048c1
GitOrigin-RevId: 46833a6ce0c70bc679afc123dd63c22e15084930
Change-Id: Ie15aab7a36d781cc7e31734a4e2fd7327cbca7e1