[roll] Roll fuchsia [zircon][scheduler] Move the concept of "active" from mp -> Scheduler.
Prior to this change, the `mp` layer of the system was maintaining 3
different masks of per-cpu state in order to reflect 3 different
concepts.

1) The set of CPUs which are currently "online".
2) The set of CPUs/Schedulers which are currently "active".
3) The set of CPUs/Schedulers which are currently "idle".

This CL applies feedback from
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/1056093 to move
concept #2 out of the `mp` layer and into the `class Scheduler` layer.

This reflects the fact that "active" is really a Scheduler concept
(scheduler is active and accepting threads) and not so much a CPU
concept like "online" is ("online" CPUs are powered on and not
halted).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064308
Original-Revision: 0d1be1111e4f9d68d1ee40823505a4895303f8b1
GitOrigin-RevId: 480dcf06a84a346661aac239a92df2e4fbd187af
Change-Id: Idf4aab593b4e3a04662d59ae525d17c35d58946b