[roll] Roll fuchsia [power] Element levels can be supplied in non-increasing order
Power elements are sometimes contracts between multiple repositories.
The most common example is the platform (fuchsia.git) can declare a
power element that a driver (out-of-tree) must implement. The platform
will take dependencies on the levels of that power element in order to
control the behavior of the driver.

In these situations, the repositories must be able to evolve separately
while maintaining backwards compatibility. We accomplish this by
allowing for the platform to introduce new power levels and insert them
in between previous levels. For example, assume we have these levels to
start with:

1. LOW_POWER
2. HIGH_POWER

The platform wants to introduce a MID_POWER level, but must not re-use
the number 2 or else the driver repository could misinterpret the
meaning. The platform should change the power levels to:

1. LOW_POWER
3. MID_POWER
2. HIGH_POWER

This requires that we forego the assumption that the value of the levels
declares the order of increasing power consumption. The order of
increasing power consumption is now determined by the order the levels
were supplied in AddElement(valid_levels=[1, 3, 2]).

Original-Bug: b/340673244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056572
Original-Revision: 8b21587dce47e7006de9b679379a09beb52b99c8
GitOrigin-RevId: 17ba3be44f6a7d8bd6922b0ae3acaa7092a5e679
Change-Id: I64a1c998d490431cdc20c409312616a24f9e53b8