[roll] Roll fuchsia [starnix] Test that futex_wait can be interrupted.
If a thread waiting on a futex receives a signal that triggers a signal
handler, and the signal disposition didn't set `SA_RESTART`, the futex
syscall should return with a `EINTR` error number.

This CL adds a test for that. The main thread sets up a signal handler
for `SIGUSR1` that doesn't do anything, spawns the second thread and
waits on a futex. The second thread blocks `SIGUSR1` on that thread, and
sends a process-directed `SIGUSR1` every 100ms.

The expected result is that the main thread is interrupted while waiting
for the futex.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065248
Original-Revision: 4adee895372f6e78cf076528be5abf28699f9deb
GitOrigin-RevId: 0a504b229a5e58f72a7b9ab034fb1f3e03970d69
Change-Id: I03d6c0c22475e696d9c573fae566b2baa7a05c52