[roll] Roll fuchsia [brcmfmac] Fix SDIO teardown on Start failures
If initialization of the SDIO bus failed for any reason the bus object
would not be stored and the driver would fail to start. However, because
netdevice was initialized before the bus there could be netdev requests
in flight that would then be served by a non-existent bus. This would
then attempt to access a null pointer and crash.

Fix this by initializing netdevice after the bus has been successfully
initialized. Also do a complete shutdown of the device if bus
initialization fails to ensure that any resources that were initialized
are properly disposed of.

Original-Bug: 344662647
Test: Manual testing with missing firmware files

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065872
Original-Revision: 8969affbbecbc818063ed7290911258a35dc1ad2
GitOrigin-RevId: cbd3032ae70ac098bf4d5638aec9d7fb831b9d3f
Change-Id: I87ee6b7ef2ec79e7e1d97f2d387db8298c701ee9