[roll] Roll fuchsia [netstack3] Carry errors with serializers
Introduce ErrorWithSerializer and an alias to it, SendFrameError. This
encodes the pattern of including errors alongside serializers when
errors are observed on the send path.

Bubble all of it up through to the IP layer.

A follow-up CL will use the now better-placed errors to introduce
IpSendFrameError to scope errors to the IP layer, like invalid loopback
addresses on the wire.

Original-Bug: 346585534
Original-Fixed: 42167236

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065958
Original-Revision: 512ed6d2c89741cd05530ac593dbd32e0ba58a84
GitOrigin-RevId: 6cd293bc2b0dea87cb5b44e06953184765aafd55
Change-Id: I365ea61b051c180d3d299af253340017b7ec8b68