[roll] Roll fuchsia [fx] Fix list-packages script
The list-packages tool can be used to output a list of packages
available in the current Fuchsia build. The list is currently sourced
from several build artifacts that are not guaranteed to be available
depending on how the build was assembled, and should instead be sourced
from all_package_manifests.list which is always present.

Changes:
- Package info for list-packages is now sourced from
  all_package_manifests.list
  - Note that this eliminates the ability to list packages based on
    whether they were base, universe, or cache packages; that ability
    relied on the filename of the build artifact the list was sourced
    from and is not contained within the actual manifest files for each
    package (which are now used to source package names)

Original-Bug: b/337957623

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052925
Original-Revision: e2e9a2aedb901a68bc164cec0458428cd0e7a30c
GitOrigin-RevId: 16d250126c2ec3aa46c454b3a881f5fe74a7ff42
Change-Id: Ic90deacd4b2457b876e11df21640e1c9d3005321