[roll] Roll fuchsia [kernel] Convert platform_set_oneshot_timer to take a raw ticks value
Modifies the platform_set_oneshot_timer function to take a raw ticks
value as the deadline instead of a monotonic time. This will make it
easier to reuse this function when scheduling boottime timers.

This required changing the pdev_timer_ops struct (and the RISC-V
bindings that use it) to also operate on raw ticks values.

Functionally, I expect this to be a no-op. I'm not changing the
computations performed, I'm just moving them up a layer in the stack.
The monotonic time -> raw ticks transformation that previously occurred
at the platform layer is now happening in the platform-agnostic code in
kernel/timer.cc.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063733
Original-Revision: 8b91b5de4b7a4c49010f59ec00ed2979bcc15187
GitOrigin-RevId: b2568329e78786fa4a8dfa9f70dfc5b8ae64abc0
Change-Id: I5e56f7edc048be01196397de2d54c475202aaa23