[roll] Roll fuchsia [build] Only include aml drivers & tests in arm64 builds
These drivers are for an arm64 SoC, and not useful in builds of
other architectures, but GN doesn't know that, and so includes
them, triggering the rebuilding of many artifacts in the arm64
toolchain.

This should save multiple minutes on a core.x64 build that
includes the buildbot tests.

NOTE: this doesn't move all the aml drivers, just a number of them, more
remain and need to be moved.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065706
Original-Revision: 666cd72b634444587d77d563b1d46d32c2322880
GitOrigin-RevId: 8ee9d43b885ee69cd52ff98a19ca439cb9dab5dc
Change-Id: I234db6f981d172b6dc163cbb8d90864431ca3ccb