[roll] Roll fuchsia [fidl][abi_compat] Allow protocol openness mismatch
It's okay to change protocols between open, ajar and closed. The
problems are when you add and remove things when one of the peers thinks
that's not allowed, which is handled elsewhere.

This adds a small test for this case which required a bunch of changes
to support testing protocol-level skew.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065801
Original-Revision: 03765ca3e3e502b81aef0b00bacbfa2ed55c20f7
GitOrigin-RevId: c62a7e2c819266cf853d2be226ca2bba673ad4f8
Change-Id: I659b8c8b718ca28f5ef968a81336f79bdc7e1de7