[roll] Roll fuchsia [test][core] Clean test phys vmo
Performs a cache clean of the test physical VMO before returning it to
the calling test. This ensures it's in a known state and no previous
test has left data in the cache that might get written out as a cache
bomb if the current test is using the VMO uncached.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069592
Original-Revision: 47a074bf4e823112f7d64df7830fb6d2768f482e
GitOrigin-RevId: 1b0999a88b30eaeb543918a578ab6e4d0d0b9185
Change-Id: If212259a7ce2b282be8b98eb90dd8a99cc2d280c