[roll] Roll fuchsia [sysmem] Assembly config for sysmem memory sizes
This deprecates sysmem driver metadata for this purpose. Kernel
cmdline for this is already gone as of a previous CL.

This allows setting PlatformSysmemConfig fields in board info and/or
in platform config. The fields in platform config override any fields
set in board info.

This also changes the contiguous_guard_pages_unused default to false.

BUG: b/322009732
BUG: b/344065609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063881
Original-Revision: 1f4329a5547b61469fbc42048432a45999ca7954
GitOrigin-RevId: a4c5d08f786f6738e048ea00f6ce7eae4d71ed53
Change-Id: I50993bd3258e2b583c754f7ae382aad30d2d30c7