[roll] Roll fuchsia [lowpan] update discovery proxy workaround
The Fuchsia mdns do not support subscribe service via a specific network
interface. To make lowpan work with that, there was a workaround to keep
track of a list of addresses that both presented in the OT and
lowpan0. One case missed is that netstack added the address which
already presented in the OT, and that caused missing addresses in the
list and resulting discovery proxy may unable to work with locally
published services.

Original-Bug: b/341965925
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069779
Original-Revision: f6b60f36762f8debda022c91da653a85afc0085f
GitOrigin-RevId: 1dc3f379a30d0d788fc6e80c7642bfb5a72d5f5e
Change-Id: Ib0bfaebdca36a38dbeef31b9183589283b150ec9