[roll] Roll fuchsia [ld] Provide ubsan-custom runtime for test modules
The test modules can be built in the standalone ubsan variant,
while fancier sanitizer variants like asan are already excluded.
It just needs a custom runtime.  On POSIX, this can be complete.
On Fuchsia, it has no way to print but it formats a message on
the stack and crashes in a way where the pointer to the message
buffer can be found in a known register with a debugger.

Original-Bug: 348678677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066593
Original-Revision: 17c7c5d3adf4c44ed987d7301a4be94f50211f38
GitOrigin-RevId: 2d6d72c06da5bd3197665859910460baf06619d2
Change-Id: Ibe42d4e65e6fe0269ec092248aae310226b4883b