[roll] Roll fuchsia [netstack3] Remove datagram sharing update race
The form of `update_sharing` in the datagram API allowed for potentially
racy behavior in the UDP implementation. This is not really an
observable race because bindings linearizes all the operations on the
socket, but this new form makes ite clearer that we shouldn't have races
here.

Original-Fixed: 345740748
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069783
Original-Revision: e323c2308c485e75d1e5598fc9f452fabd0fdc08
GitOrigin-RevId: b22672a3302ae221af995fef47d12b80c1b605f4
Change-Id: I7e4ab0cbaadaf14e73ddc5dbf4e7ef4c06e6a733