[roll] Roll fuchsia [build] Support FUCHSIA_BAZEL_DISK_CACHE environment variable.
Allow the user to set an environment variable to point to an
absolute directory path that will be used to store build
artifacts out of the Ninja build directory.

This is useful when full Bazel remoting is not enabled,
(i.e. when `bazel_enable_rbe = true` it not set in args.gn),
when performing builds on different checkouts of the same
machine, or when doing repeated clean builds, since the
Fuchsia platform build does not use the standard output_base
locations.

While the Bazel documentation states that using --disk_cache
is compatible with remote caching, the RBE documentation
states that this flag 'does not mix' with RBE and thus
should be removed from the `.bazelrc` file, so avoid using
the flag when `--config=remote` is used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070574
Original-Revision: 20de61c2f976907b4b95d09241a4d27d0a35871c
GitOrigin-RevId: 10d4b0f4f42a5a9bf82c02f78a8e39b4c4c752d6
Change-Id: If3aab29daadcebad4410729b6853c8d01f3d9f8c