[roll] Roll fuchsia [starnix] Report restricted VMAR of starnix programs
In Ifbb5614f450c3ac80a9a3a6005f0019524e61618, we taught starnix_kernel
to report which programs are run in the starnix container, but did not
include information about how much memory is devoted to each program.
In this CL, we teach starnix_kernel to report how much memory by
reporting the KOID of the restricted VMAR of each program.

There are a few options to identify memory used for each program. In
theory, starnix_kernel could send over VMO handles as soon as a VMO is
mapped into the starnix program. However, the QPS of that operation is
likely too high to afford in the critical path of mmap. A typical
program is measured to have 50-250 backing VMOs.

Therefore the approach in this CL is to report the restricted VMAR KOID.
Since memory_monitor already knows all the VMAR of Zircon processes, and
which VMOs are mapped into them, memory_monitor can easily derive which
VMOs are mapped in a particular reported VMAR.

The discussion linked in fxbug.dev/341981291 has more details.

Original-Fixed: b/341981291
Original-Fixed: b/337865227

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068120
Original-Revision: 8ff3b132da9ac475dfea844842ee2021aa757d67
GitOrigin-RevId: 40119fd321dd093438cf0f23f37204646a8325f7
Change-Id: I1eaadb7bc9307031028141781bbb2bcc62f12d03