[roll] Roll fuchsia [Honeydew] Replace subprocess.* with host_shell.*
* subprocess.check_output() has been replaced with host_shell.run()
* subprocess.check_call() has been repalced with host_shell.run()
* subprocess.Popen() has been repalced with host_shell.popen()

We have many places in Honeydew that performs subprocess oeprations and
handles the exceptions. All of this code is being repeated in multiple
places. So created host_shell module with all subprocess related
operations and replaced all of subprocess.* calls in Honeydew with
host_shell* calls.

As part of this, also removed 'exceptions_to_skip' argument used in all
ffx.run() and fastboot.run() as it is not being used anywhere (except
for one usecase which can be handled differently) and is making the code
less readable.

BUG: b/347778716, b/332538317

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065893
Original-Revision: 4573fbcb4c65c1330a4bf852cdf379e6a1c7494a
GitOrigin-RevId: 4b78fca97e81691c1f56189767ecd2b0e9f6ced8
Change-Id: Ie6f528cb35d59514ba8077726e62bcf1bc57a060