[roll] Roll fuchsia [time][rtc] Adds a crude retry for RTC detection
The RTC drivers seems to have a race issue where driver init may happen
after a would-be user attempts to read the device.

Since the init is fairly quick, we put in a crude retry to work around
this issue temporarily, until the underlying problem is resolved.

This change can be reverted safely in case it causes startup issues on
devices.

Original-Bug: 348530594
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069936
Original-Revision: d3aef7491e66983cf1990796344f3923025061c0
GitOrigin-RevId: 2b4de81fb248504d4b20ddedc81438e33dae58b0
Change-Id: I48619ee5f08e214ed3a75144d1c7a68ad61693b2