[roll] Roll fuchsia [netstack3] Fix deferred TCP socket destruction
The logic for marking a socket for DOA was inverted: if a socket is *not
yet* marked for destruction it must be inserted into the sockets set.

Added details to loom test to cover this, which required exposing
deferred destruction through `FakeBindingsCtx`.

Original-Bug: 349404597
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071839
Original-Revision: b2799f0fa7ea8ce3629e96cc59bd73ea62fa9002
GitOrigin-RevId: f533869cce0742e7500670940d5ca1dae316e3c9
Change-Id: Ia4a0a406315e14d5ff89bec0be281f4b5bb173d8