[roll] Roll fuchsia [kernel] Give a simpler handling of kernel.memory-limit-mb
We do not need the complexities of libmemory_limit to handle this boot
option. With a normalized accounting of RAM, we can more simply
determine a RAM cut-off address and then truncate all mention of RAM
passed that point.

Removing the use of libmemory_limit, which parameterizes repeated calls
to pmm_add_arena() is a first step toward refactoring how the arenas are
initialized (which should happen all at once, and not an arena at a
time).

The code in this CL is a switchback, tweaking the legacy zbi_mem_range_t
ranges - as described by the bootloader - that are handed off to the
kernel. When we are ready to hand off the normalized ranges, we'll be
able to remove this special handling entirely, relying on the truncation
of the normalized ranges that now occurs at the very beginning of phys
memory intialization.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1067512
Original-Revision: a7dfb01dcd54c8d8e5f3ec24d86b7c4eee735b34
GitOrigin-RevId: 9843cf7e29f8efd8bb53baddb9b78ba39b90a1a1
Change-Id: I7b346078a97fdd1a7abeca8605a238fcdd0b93be