[roll] Roll fuchsia [zircon] Clarify state of FIDL library `zx`
Attempt to make it clearer that the library is not experimental and is
in a stable but interim state.

In addition, clarify the difference between the two `zx` libraries,
including adding a README for the `zx/` subdirectory.

Also update and fix links in the existing README.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051240
Original-Revision: d62f56ba486d942a224232c55445cccbf17f8e87
GitOrigin-RevId: ece7f70bafd61ed5deeb95fbda23faa570aa2772
Change-Id: Ifcc3cdbdde4f41b36575bac9ee2bb8534b64b95d