[roll] Roll fuchsia [fusb302] Fix logic error in handling SourceCapabilities in Ready state.
The USB PD sink state machine incorrectly handles SourceCapabilities
messages in ProcessMessagesInReady(). The method is expected to return a
state that the state machine will be migrated to. The handling code
calls ForceStateTransition() and falls through, which causes the fusb302
driver to reply with a NotSupported message.

This CL fixes the logic error, so the hardware replies to
SourceCapabilities messages whenever we receive them.

Original-Bug: 42182911
Test: Booted core.vim3 connected to Apple A1719 charger
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071842
Original-Revision: abedb90e86d6ecb1bcef900c52ed30e5f48f6155
GitOrigin-RevId: 2c9c80a23b059878832cd492a5b776faecd8f0f2
Change-Id: I26f1251dea268ceb88e8b27dc92262b7088120fc