[roll] Roll fuchsia [iob][tests] Remove flake from Iob::GetInfoProcessMaps
Remove a cause of non-deterministic flake in the
Iob::GetInfoProcessMaps.  Prior to this, the test would count the
total number of VMOs in the test process, then create two IOB
endpoints, each with 3 regions, and assume that the total number of
VMOs in the test process should have gone up by exactly 6.  Then it
would close an endpoint and assume that the number of VMOs should go
down by 3, and finally close the last endpoint and assume that the
number of VMOs should have returned to its original value.

The problem with this approach is that it was counting the total
number of VMOs in the process, not all of which are under the control
of the test.  For example, other VMOs could come and go as threads are
created and destroyed by the runtime, creating and destroying VMOs
used for things like stacks for thread local storage and storage in
the process.  So, we could close an IOB endpoint and expect the total
number of VMOs to go down by 3, then observe it only go down by 1
because two new unrelated VMOs suddenly showed up.

Avoid this issue by being more explicit.  The test already had code to
enumerate the VMOs and find the specific VMOs it thinks it should have
created in the first place.  Reuse this code so that when we close the
first endpoint of the IOB, we can confirm that the specific VMOs
associated with that endpoint go away, while the VMOs associated with
the other endpoint stick around.  Other VMOs outside the control of
the test are free to come and go as we no longer care about the
absolute count.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071838
Original-Revision: c74e2cddde8fcf9842eca8c69544e802783816f5
GitOrigin-RevId: e379c454971500434598d8853b541d85ce347e22
Change-Id: Ibbe10269831422b310e4262769ee86cf767d03c6