[roll] Roll fuchsia [session] make cobalt unit tests more robust
Configure the unit tests to panic if the executor stalls.

This makes the tests fail-fast in some conditions where they would
otherwise wait forever.

For example: suppose we introduce a bug in `log_session_launch_time()`,
which causes that function to believe that the `end_time` in the
`test_log_session_launch_time_swap_start_end_time` test is greater
than the `start_time` in that test.

With the previous code, the test would hang forever. This is because
`log_session_launch_time()` would be waiting for a reply to its
call to `log_integer()`, but the test does not have any affordance
to return a reply. (And the test _shouldn't_ have such an
affordance, because the test is not expecting such a call to be
made.)

Test: fx test session_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069534
Original-Revision: 3b0bc50e6feb164a0de1d7af1cbb275d8f13bbc5
GitOrigin-RevId: e5fb42d897f9989208fc106bafc591d001040316
Change-Id: I5cadb7f304be615b6d4a269cbd7077fd4e822e95