[roll] Roll fuchsia [iob][tests] Fix flaky test
Fixes a test that attempts to query all the mappings in a process, but
allocates between determining the number of mappings, and performing the
query, potentially increasing the number of mappings due to heap growth.
If the heap was placed, due to aslr, below where the IOB mapping was
placed, and the IOB mapping is virtually the last mapping in the
process, this will cause the IOB mapping to not be in the list.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073112
Original-Revision: 299e7fc26cf724384b9694b5bf91982d6ed59efc
GitOrigin-RevId: 010b601c57dc076b0d9a241897430f895886275b
Change-Id: I350e2d2457687277326a43652c12155dae00120d