[roll] Roll fuchsia [starnix][vfs] Fix crash when prerun fails
When the prerun function fails, we were crashing when unwinding the
thread in some situations because the task_complete closure was holding
some releasables. When we dropped those objects, they would register
themselves in the RELEASABLES thread local but never actually get
released.

Now, we finalize the DelayedReleaser to catch those objects when they
try to register themselves instead of catching them when we tear down
the thread.

Also, we drop the task_complete closure in the error codepath to drop
the releasables while the thread local is still able to accept them.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073480
Original-Revision: e7ab8b1cdb41790b48cc52d8e712dda0ced839bc
GitOrigin-RevId: ac73e33889d1642c219dec829720eeab8dae2cfa
Change-Id: Ic8b25df31e6ff4740e6fbabbf2681224f14088af