[roll] Roll fuchsia [routing] move with_policy_check from component manager
The `with_policy_check` function will need to be invoked by
`cm_fidl_analyzer`, so this commit moves it to the routing library in a
new trait.

This function requires the ability to get a moniker from a
`sandbox::WeakComponentToken`, which was not possible to do before in a
generic way due to the type erasure perfomed by `WeakComponentToken`. To
address this, a new field `moniker` has been added to the struct.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1059752
Original-Revision: d8dbdbaea4d98a306a9a6c5905ce3eea95fe37e3
GitOrigin-RevId: 9730c5d77ebca87fbc23b81c1c519d5b1a8ae45c
Change-Id: I63dac294b188c9068eee5fb5194edbbff83ac9c0