[roll] Roll fuchsia [tefmocheck] Create new tefmo for seriallistener timeout failures.
The Boot-test-successful check is to capture cases where the boot test
passed but we failed to detect it and thus failed the test. However, if
a boot test legitimately times out because it doesn't see the success
string, it will print out that it fails to find the success string
Boot-test-successful in the swarming output. The original check tracks
both of these types of failures together, but we should separate the two
since they are two different types of failures.

NO_IFTTT=just added the check to keep the strings in sync
Original-Bug: 42067738
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064833
Original-Revision: 6ab7e9e7e4c599cdaa0dc5cff2ef83c77823ab59
GitOrigin-RevId: f9ee67d50b9ee9cc6a7c582a884f839308b91c7c
Change-Id: I2b71c596da689c7a4a36e42c8419c182a5ccdeb9