[roll] Roll fuchsia [starnix][tests] Run reboot test in debian container
This CL migrates the reboot test to run in the debian container. To make
this work, we need to write our own client for the reboot syscall
instead of relying on one from the container.

I tried using the `/sbin/reboot` binary in the Alpine container, but
that binary just sends a signal to `init` rather than calling the reboot
syscall.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073372
Original-Revision: 76df5aee3459d7fb8a0a5d5441863dfd9d00f188
GitOrigin-RevId: a5b97d9f629343f226d6d41d3c63f22cd2bae509
Change-Id: I4cd8d0677dd5e9e25be405fd254a33825e20a6e7