[roll] Roll fuchsia [components] accept args in LifecycleController.StartInstance
Adds a new function to fuchsia.sys2/LifecycleController named
StartInstanceWithArguments. This is identical to the StartInstance call,
but it has an additional arguments of fuchsia.component/StartChildArgs.
The provided arguments are used when starting the child, just as if the
child had been started with fuchsia.component/Controller.

Also modifies `ffx component run` such that the `--connect-stdio` uses
the newly added approach to give stdio handles to the run component,
instead of relying on `CreateChildArgs`. This removes the limitation
that `--connect-stdio` can only be used in single-run collections.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/998691
Original-Revision: 45a80bd3a5a4b0e974d3771d03f4ce7074cf1a2e
GitOrigin-RevId: 02caf43e720e4f71d4e7d17148f112826e5efefd
Change-Id: I5c9111eaaba8438e2ddc9d36f7375c2546c2a68c