[roll] Roll fuchsia [debugger] Add TestFailureStackMatcher with tests
This cleans up the code that detects when we are running embedded in fx
test and allows ConsoleContext to have enough external context to tell
when we can jump straight into the user's code from a test failure.

The tests use some real stack frames from failed gTest and Rust tests
and show that we can identify the top frame of boilerplate. This is done
by mimicking the namespace hierarchy with fake symbols to create the
same symbol names for the test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074257
Original-Revision: 237084134fa10ce0877096faf3ef56ba19cc020a
GitOrigin-RevId: abe0d214a74a3faf15aa8f6b4f577bfa48f6b36c
Change-Id: I0f9093db231e7008d69b17785b07192e5977d22f