[roll] Roll fuchsia [rbe] Rename to 'legacy_default'
This makes it a little clearer that 'legacy_default' (resembling the old
default from the Goma-only days) is not a choice most users want,
even if it is what they are used to.
We want to encourage users to pick a baseline mode that
best matches their development environment.

Original-Bug: 344673187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074465
Original-Revision: 5ddf9cecba372b46047b08e523750f3334b1fb5a
GitOrigin-RevId: fdb8f0c3ae79dad004783064ae7bbb0bcce837c6
Change-Id: Ib7a9df4ed2b86846635ed5f6b95d5df12148d2fd