[roll] Roll fuchsia [covargs] Fix bug where multiple embedded build ids causes hang.
This was introduced in fxrev.dev/1074518 because we create a channel
that only has the size = number of profiles, but if there are multiple
embedded build ids in a profile, we'd exceed that size and hang trying
to send more entries to that channel. Instead, we'll use a mutex to just
directly append to the slice of entries.

Original-Bug: 42075455
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075933
Original-Revision: 288b6bd2ae124b88366b8eb55ae8af65fc2fca63
GitOrigin-RevId: 498e7fab09f7e1b15a794b9111becb2f1e5a13e7
Change-Id: Ie2691d441700171eda852c2b7fc3c58155070008