[roll] Roll fuchsia [wlan] Adds logging to WLAN state checks in tests
The WLAN policy layer requires that clients acknowledge
receipt of past state updates before serving new updates.
We are seeing some test flakes that suggest that we may not
be synchronizing on this correctly.  This change adds
logging to report the state reports that we are rejecting.

Original-Bug: 350536751
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076235
Original-Revision: 86d2cfa0cc987a7953cfa4506f9774ed31ba5902
GitOrigin-RevId: 9679f3d007074a9332dccbf7a991378fd31fed1d
Change-Id: I1ad6289e1d87b27d1a53bde03c31434ae170bcc2