[roll] Roll fuchsia [botanist] Temporarily increase max attempts for capturing syslog.
We have a max attempts for capturing the syslog within a minute in case
`ffx log` is broken, in which case there's no use continuing trying.
Recently, we've been seeing `ffx log` take a few seconds to start
capturing logs and it fails more than 5 times before that, causing us to
get a "failed to capture syslog" flake. Although this is always
considered a flake at the moment and thus won't fail the build, it does
cause botanist to stop trying to capture the syslog, so we should
temporarily mitigate this by increasing the max attempts until ffx log
can be fixed to not fail so immediately. I also added a new tefmocheck
to fail for the case where we suspect ffx log is broken so we don't miss
it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076936
Original-Revision: 3c8181cc7412aedf102a115c71c6012ff1fe2551
GitOrigin-RevId: 34114e5b31af87b2ff159865af9c62fbf2cda8e0
Change-Id: I8ec83d60562897541c9955c8aea678f273ced001