[roll] Roll fuchsia [starnix][memory] Report the VMAR range instead of VMAR KOID
Previously we had starnix_kernel report the restricted VMAR KOID to the
attribution protocol, such that memory_monitor could find which VMOs are
mapped into that VMAR and hence relate them to the running starnix
programs. That turned out insufficient because the zx_object_get_info
syscall does not report the KOIDs of VMARs mapped in a process.

There are some upcoming investigations to design faster syscalls for
memory_monitor. We might be able to add the KOID as part of that
evolution. In the meantime, we switch the protocol to report the process
KOID and a (base, len) tuple of the VMAR, which will let memory_monitor
find out which VMOs are mapped in this VMAR by checking ranges in the
process-maps syscall information. We can remove this dedicated VMAR case
when the KOID is reported by zx_object_get_info or other syscalls.

Original-Fixed: b/349116395

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073212
Original-Revision: 5801a266f3d97e8e44d9d116eb005bfff63817c2
GitOrigin-RevId: cf2d11bfeacc944a765317928a036f1cba264c3c
Change-Id: Ib64cb6551fd3b72556b728b42668d3ec4a40a36d