[roll] Roll fuchsia [kernel] Add timer support for pausing the monotonic clock
Adds functions to the timer API to support pausing and unpausing the
monotonic clock. We do this by reusing the `mono_ticks_offset`, now
renamed the `mono_ticks_modifier` to store the value of the monotonic
clock when it was paused during a system suspension. The logic goes like
this:
1. If the value of mono_ticks_modifier is negative (its most significant
   bit is 1), then it contains the offset between the raw ticks counter
   and the monotonic ticks. In this case,
   `timer_current_mono_ticks_synchronized` will return the sum of this
   offset and the raw ticks value.
2. If the value of mono_ticks_modifier is positive (its most significant
   bit is 0), then it contains the value of the monotonic clock when it
   was paused. In this case, we return the value of this variable.
Note that this only works if the raw ticks counter does not make use of
the 63rd bit. On x86 and ARM, it is extremely unlikely for this to
happen, as the counter registers start at zero and would take > 10 years
to tick past the 63rd bit. RISC-V does not make the guarantee that the
register will start at 0, but it does mention that it is unlikely for
the counter to overflow, so we're probably fine there.

In order to catch cases where the ticks counter may use the 63rd bit, we
emit a kernel oops from the timer initialization code when the counter
is deemed to be "too close" to a 64 bit value.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064932
Original-Revision: 0e8af64a09103e8d7f252ceb1ef90f1bd9351dec
GitOrigin-RevId: f4a75f1ae00a648238e830a7a295e3a9a986463d
Change-Id: I107d78013ec74e69536e704c48ce7378cf683fc6