[roll] Roll fuchsia [doc-checker] Adding support for reference docs
This adds the feature of being able to check the contents of a reference
documentation directory which is independent of the docs root folder.

This allows generating reference docs and then checking they do not
contain missing or broken links and that they are connected to the
_toc.yaml graph so all reference documentation is reachable from the
table of contents.

Original-Bug: b/344621741
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076852
Original-Revision: aa0aeb2ab8c33a0a6d0975b31b6da953ffe3e6b2
GitOrigin-RevId: d4e3664de20cfcda9b71716ad2393f6488cf7fcc
Change-Id: I0423b49ab86f61f903ffca50e021b21ed8a77cd2