[roll] Roll fuchsia [fx] Make `fx format-code --files` print an error on nonexistent files
Previously it would silently skip nonexistent files. This is the correct
behavior when the list of files to format is retrieved from Git, since
Git reports deleted files, but causes `fx format-code` to misleadingly
skip files passed via the `--files` flag that don't exist, which implies
to the user that the files *do* exist and are already correctly
formatted. This is a user error that should yield a helpful message on
what went wrong.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077752
Original-Revision: b1a70d857ab884b91c16b7447811e102889a4507
GitOrigin-RevId: 9a6007d8be74f607ea339041a75368eb9955ca5c
Change-Id: If8682198540d9e758b8461de0cdff50f1a9fb3c9