[roll] Roll infra/recipes [git_checkout] Always rebase merge commits
The default behavior of `git rebase` is to skip merge commits.

Only if the `--rebase-merges` flag is set will it respect merge commits,
so it's necessary for this flag to be set when trying to check out a
Gerrit change that is a merge commit.

Merge commits are uncommon in normal workflows but are often done for
third-party projects, so I see no reason not to set `--rebase-merges` by
default – the old default behavior of skipping merge commits seems
obviously incorrect.

Original-Bug: 349990455
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1078713
Original-Revision: c9c41a77206cae36a70211f0d74aaf5d0fe45470
GitOrigin-RevId: 5ed51dd967ea3a51e947dae27b077f48d2728641
Change-Id: I7b0dd9c46cc152e2afd0a4e71300737c00c3c85a