[roll] Roll fuchsia [driver_runtime] Arena & dispatcher rust bindings
This adds implementations of safe rust bindings for the arena and
dispatcher parts of the driver runtime api surface. This allows for the
following in the public api of the crate:

- Creating and destroying arenas
- allocating arena lifetime-constrained memory from an arena that drop
  appropriately when they go out of scope
- posting tasks to a running dispatcher

It also allows for creating a top level driver runtime dispatcher to run
tests under in the crate-internal api, and uses that to exercise the
driver runtime async executor driving rust futures to completion.

Multiply: fdf_rust_tests
Original-Fixed: b/338249591
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070133
Original-Revision: a1e70dc6a3db01cc500682a71425ddc687e482cf
GitOrigin-RevId: 7538feff43d4f1971775afe874f7c795de65b8fd
Change-Id: Ie10d084ae475637b0336f4054a6dcc4a7df79952