[roll] Roll fuchsia [fuchsia.hardware.network] Deprecate DeviceClass
... in favor of a new flexible enum `PortClass`. This will ensure the
FIDL types can evolve as Fuchsia adds support for new classes of ports.

All internal APIs have been migrated to use the new `PortClass`. The
fuchsia.net.interfaces API, however, is published to the SDK, so its
`Properties` struct contains both `device_class` (old) and `port_class`
(new). Once support for `DeviceClass` is officially retired, this can
be cleaned up. It is therefore the responsibility of the netstack to
populate both fields; an integration test covering this has been added.

The majority of the "real changes" in this CL is introducing error
handling in the components that interact with the `PortClass`, and now
need to handle the unknown variant gracefully. The largest changes from
this occurred in Netstack/Netcfg.

While here, introduce the `Lowpan` port variant, which was the
motivating case for making the enum flexible. The lowpan component
now specifies this class when creating the interfaces. Notably, a TODO
in MDNS could be cleaned up, as they were previously searching for the
lowpan interface by name.

Original-Bug: 42157740
Original-Fixed: 349436251
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073292
Original-Revision: 7e411bcf092cd7680dcae41a38ab15e287693f2a
GitOrigin-RevId: 6e12ea9dc20a1ff73c69ea8b0f2e675beddc910d
Change-Id: If3a5fd7fae989081e0083bd60b445b2598f3450e