[roll] Roll fuchsia [kernel][elfldltl] Complicate static_assert to avoid tautological test
GCC 14's "tautological test" warning triggers on trivial
expressions like "X != X" even when X has a template-dependent
value and the whole thing is never evaluated because it's in an
uninstantiated template or an `if constexpr` branch not taken.
Making the expression for constant false more nontrivially
template-dependent avoids this.  The easy mechanism is to use the
non-type template parameter as a parameter in another template
class and test that with `std::is_void`; it will always be false
since no template class instantiation can be the void type, but
the compiler won't really clock that because there could be an
explicit specialization for some `std::is_void<T>` where T is
template-dependent, and it doesn't know there won't ever be one.

Original-Bug: 351904809
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078587
Original-Revision: c09a38cc9e533b55d5e060dc0b751e4c209fe8b5
GitOrigin-RevId: 1a2f5973f52e0904c84561cee3388ea3a998f2dd
Change-Id: I89b4eca6c3359184df58cc57b4c323b9a7255272