[roll] Roll fuchsia [starnix] Remove default impl of FileOps::sync.
Having no-op default impls makes it easy for implementors to
accidentally overlook the need to implement operations. Move this
default impl to a macro that allows implementors to opt-in to a
noop.

This caught missing real impls for loop devices, overlay files,
and FUSE.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079373
Original-Revision: 4817ff798ad89fabdb22388e43f76f176a2234ea
GitOrigin-RevId: 60167a374b7c159f90304b4646ad69d4023b864b
Change-Id: I680f9a6545ca65b9340c5853a39820f43e7dfc3f