[roll] Roll fuchsia [async_utils] Support long lines in command output
It turns out the internal buffer for StreamReader is only 64KiB, and
when a single line is longer than the limit an error is raised. Since we
are running in a task using asyncio.wait(), the exception is not
propagated and the program hangs.

This change manually processes outputs from the StreamReader without
using the internal buffer, and it additionally ensures that exceptions
in any read or write tasks are propagated up.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080074
Original-Revision: 022382c69317931ceaadb66ed77caa4714a07174
GitOrigin-RevId: 89bb1dcb62496999e2cc263195e915f4d1513499
Change-Id: Ic5c01bdc7b7a97e919746bd97604da50c6a414a7