[roll] Roll fuchsia [netstack3] Expose new many tx allocation fn
A previous CL introduced a function to allocate more than one TX buffer
at once, but failed to expose it on the crate's public API.

In order to allow a borrowed method from Session, it must carry a strong
reference to the inner state, which defeats the Task not running
mechanism. It turns out that mechanism is very racy and not offering
great protection, so we remove it altogether.

Original-Bug: 42057204
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080634
Original-Revision: 12d1cda2514ba5bc65145fe7f4497a0359995fd3
GitOrigin-RevId: a812f2fe42a5574098fc0d0e8c920bca03b3ab42
Change-Id: I109c29cdb6426b7c821ba87ff14fb96678377235