[roll] Roll fuchsia [driver_runtime] Remove rust `Task::cancel` for now
We don't actually need this yet and there are problems with the way it
works making it hard to properly test. Removing this without completely
removing the mechanisms that help make it possible so that it can be
re-added in the future when a more resilliant design can be made.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080433
Original-Revision: 422cd69804e255e7cbedf7fdf4aa3c2268858113
GitOrigin-RevId: ed9019dc0f17816eef374f933ee137ae7d4ac8f3
Change-Id: Ifaab21a881f7f3e7d7a033a77bac54b3015f7610