[roll] Roll fuchsia [fxfs] Use bloom filters for queries
Adds a bloom filter to the persistent layer file format which allows us
to quickly determine whether a layer file needs to be consulted for a
given search.

To make use of the bloom filters, the LSM tree needs to know what the
goal of the caller is when performing an `LsmTree::seek`.  Added a
`Query` type which expresses the goal and allows us to use the bloom
filter appropriately; see lsm_tree/merge.rs for details.

Original-Bug: 333401205
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017312
Original-Revision: 35915801279312e1c8fa70d1078999e04db69828
GitOrigin-RevId: 7f988c0d2dbc9fda79adebe650629618cef5d7e5
Change-Id: I33d97048bc32952e486fc760614d842d3675f9ed