[roll] Roll fuchsia [idk][build] Remove sdk_molecule's host_tools_deps
https://fxrev.dev/943927 introduced the `host_tools_deps` argument to
the `sdk_molecule()` template as a means to remove the global
`sdk_no_host_tools` arg. However, not all targets assigned to it are
host tools, it is only used once per IDK, and it hides a decision
about what is included in the IDK build.

Instead, let the IDKs specify when to include such dependencies. In the
future, we will be able to use other GN variables and remove the global
GN arg `sdk_inside_sub_build`.

Original-Bug: 347094435
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074468
Original-Revision: a97ebe118170a9d74d201c5af0985f6f7343bf47
GitOrigin-RevId: f96b697d36f1c102d5e79e5330954ba7689fa98d
Change-Id: Ic6eca22d5b58cb13912da8107bc73b46c1e3ae37