[roll] Roll fuchsia [netstack3] Don't spawn a TX task for loopback
We're spawning this task but the loopback interface is not actually
configured to use any queues. This is just wasting memory and increasing
complexity. We can reintroduce the TX task when we add queueing
discipline for loopback.

To make the lack of support clear, don't create a tx notifier for
loopback, which will make the stack panic loudly if we ever configure
loopback to have a tx queue.

Right now this is just getting in the way of a better tx queue
implementation for netdevice-backed interfaces.

Original-Bug: 42057204

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080726
Original-Revision: 795b935ae6ea78132a90fe7fefcfc051c6756194
GitOrigin-RevId: e007ec7b5f97dc5b6883f055d9cb7de529c838ba
Change-Id: I8680088b86b4958cdd88b7ee59163c7172d0b348