[roll] Roll fuchsia [kernel][pi] Fix an old copy-pasta bug
Fix an old bug in an accessor used to access a thread's TID via a
thread adapter class in the scheduler PI code.

The `tid()` method was supposed to return the thread's TID, but
instead it was being used to return the thread's state instead.  The
method was being used only in one place, and it was for a
DEBUG_ASSERT_MSG which (by design) has not been firing, so we have
just never noticed.

It's fixed now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081976
Original-Revision: ad0925953d9fd12883f54d3163e49aaf7021d24c
GitOrigin-RevId: c6ca3827ce8eca8f633c2edfd19fe5d2796c7d74
Change-Id: I757298133eff4f7d44723ee450d9a62d7f8f9af2