[roll] Roll fuchsia [memalloc] s/UpdateFreeRamSubranges/UpdateRamSubranges
When UpdateFreeRamSubranges() was introduced, we wished to effectively
reserve a special range (which could span both free RAM and existing
allocations), but wanted to do so in a non-destructive way: the result
was just to update the free RAM subranges leaving the existing
allocations in place. However, with allocations either 'surviving' into
the kernel proper or being freed at hand-off, this approach is no longer
suitable. We want the whole range with the intended allocated type to
be present in the kernel proper, and not with "free RAM" carveouts
resulting from the subranges that were originally occupied when we
called this method.

Accordingly, we destructively update the method to reallocate all RAM
ranges. These types are mostly for diagnostic purposes, and even though
such a destructive operation might cloud that observability, it should
remain clear enough what it going on in practice around any of these
callsites.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079555
Original-Revision: 8f1955e2a54e3700c08f61f47c659a2fca0d08f4
GitOrigin-RevId: 5a6f30d634abf6068fc1ee1e671da9e95c4be857
Change-Id: Ifda1a29cc53d63e24608bf83e9ee1e1aa1f45d2b