[roll] Roll fuchsia [starnix][iptables] Separate namespace ID from table name
When creating a fuchsia.net.filter/Namespace to correspond to an
iptables table, starnix adds a prefix indicating the IP version ("ipv4-"
or "ipv6-") so that IPv4 tables don't conflict with IPv6 tables.
Refactor to clarify the distinction between the NamespaceId and the
table name. This fixes a bug in the logic that was using the prefixed
NamespaceId to determine routine type (IP or NAT), which meant that all
routines were determined to be IP routines.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082712
Original-Revision: 434cde9ae76360fa2f20b5ebe76244ae7b1a7e26
GitOrigin-RevId: ea711b1b83761faa0b8d7dd6a6e3398da71a04d4
Change-Id: Ib49b981b5d48dc6ad75b3d0319c3ba08e9a3e383