[roll] Roll fuchsia [bt][a2dp] Prevent lock over await
Perform release tasks on StreamEndpoint in the background after
responding to the peer.  Pass a future out of the handle_request
function which holds a lock on PeerInner, so we can await on the result,
instead of holding the lock while waiting.

This prevents the possibility of double-locking the peer if a reservation
was released while awaiting during a Close operation or while suspending
in response to a requested Start that could not be satisfied due to
a lack of resources.

Fixes: fxbug.dev/326585243
Test: updated unit tests
Test: fx test bt-a2dp-lib-tests bt-a2dp-tests bt-avdtp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078733
Original-Revision: c7774e8b81a102cf17a99cff0efa4c74c29b8d5a
GitOrigin-RevId: 9e69242c34623b6b8517ff06d6b09a6f12c96c83
Change-Id: I283b213fad2c86c7c0412c85ef656f39f9feb798