[roll] Roll fuchsia [libc] Improve debugdata test for failure cases
Preparing for a new implementation of __sanitizer_publish_data,
improve its tests.  The failure case was not quite tested
explicitly, only that it didn't make the helper crash.  Now the
helper knows whether it's a success test or a failure test and it
fails if the __sanitizer_publish_data appears to succeed when it
should have failed.  In addition to the test for no /svc in the
namespace, also add a test for an /svc whose channel peer is
closed so the channel send fails on a valid channel handle.

Original-Bug: 342469121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080354
Original-Revision: 137caf798537b5e46eb2b88e4e85cd54a5c9059e
GitOrigin-RevId: 2b338bf22e64dd620e4e042abe2d0c9cdc1de249
Change-Id: Ie63caba11ed7d4ea47939f407778969f17bcdee2