[roll] Roll fuchsia [msd-arm-mali] Keep lease while suspended
Currently, once the GPU is no longer active due to suspend, the power
lease is quickly discarded (because the current_gpu_powerdown_timeout_ is
set to 30 ms in the future). Once the system is resumed, the GPU will
remain suspended until some new GPU work triggers
PowerOnGpuForRunnableAtoms to re-acquire a lease.

To enable the GPU to restart work immediately when the system is
resumed, the lease must be held as long as there is runnable work in the
scheduler, even if the GPU isn't currently active.

Original-Fixed: 351904457
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081932
Original-Revision: 11eeccb1db98453c14d2da644682168ab59f6d32
GitOrigin-RevId: dc7855b8fa45a2eed452f8cabeb081de4136c7d3
Change-Id: Iada75c034ac4425a0c29e357cb3529439c44f44c