[roll] Roll fuchsia [kernel][trace] Kernel generated channel message flows events.
Add kernel-generated flow events to the ChannelMessage trace events that
emit the FIDL ordinal for a channel message. This allows convenient
visualization of the connection between the source and destination of a
channel message in Perfetto and works in conjunction with the command
`ffx trace symbolize --fxt <filename>` to product trace files that
provide both the FIDL method names and flows connecting senders and
receivers.

Original-Bug: b/340540026
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006055
Original-Revision: c482a09026f80485c37359cba51791393ec1ef78
GitOrigin-RevId: 5818e0ab4ff9f62de63ba7b5334285f44e676b97
Change-Id: Id02c4fa1cd003a588b425085bb3fa45b0761c4a4