[roll] Roll fuchsia [kernel][vm] Remove redundant trim
Due to the way looking up multiple pages in a VmObject use to work,
that is the LookupPagesLocked interface, the requested pages from a
page fault was trimmed to prevent more than the actual faulting page
from being marked dirty. The current LookupCursor interface uses an
explicit IfExistPages lookup, which is not allowed to perform dirty
transitions, and so the requested length does not need specific
trimming.

This does not improve performance or change behavior, but it improves
code clarity as the existing comment, check and logic is now confusing
given the current LookupCursor behavior.

There are existing core-tests that explicitly validate that page faults
do not cause additional dirty transitions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048655
Original-Revision: 78d34c2413834be2238307e499b73e898920e49c
GitOrigin-RevId: ec53a237a2eb210c153f5d280671cc1d15dcf075
Change-Id: I4dddd2a8b224f9acb567d9c5dda746cb4f7defd1