[roll] Roll fuchsia [components][sandbox] New interface for dictionaries
This change adds a new sandbox interface which reflects a fairly
comprehensive revamp of the bedrock dictionary APIs. It introduces a new
framework protocol, [CapabilityStore], that represents a container for
bedrock capabilities which are known to component_manager.
[CapabilityStore] presents an interface that implements the
"client-assigned ids" fidl pattern:
https://fuchsia.dev/fuchsia-src/development/api/fidl#eventpair_correlation.
This design is the outcome of a discussion on an internal mailing list.

[CapabilityStore] has a few responsibilities:

- Provide methods to operate on [Capability]s that are opaque types.
  These are now used to manipulate [Capability::Dictionary]s instead of
  the [Dictionary] protocol methods. (Later we'll integrate this with
  [Connector])
- Expose [Export] and [Import] methods to take capabilities in and out
  of a store. This is useful for transferring capabilities between
  processes.
- Support resource management for [Capability]s. When a
  [CapabilityStore] connection is dropped, all the capabilities inside
  the store are released as well. Also, there is a [Drop] method to
  release individual capabilities.

Note that, to reduce the size of the change, Dictionary/Drain and
Enumerate APIs (currently unused) have been deleted. We will re-add them
in a followup change.

Finally, for compatibility reasons we keep a pared down version of the
`Dictionary` protocol, which is still used by Test Realm Factory
APIs, but we hope to delete this eventually.

This change introduces only the interface. The subsequent change
includes the implementation.

Original-Bug: 339027363
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081982
Original-Revision: 9d0c45db644d10d520b09c0989a20d0c5d1f3cde
GitOrigin-RevId: 817de870fa14f494bbbf3b1aa131cd4a294b8628
Change-Id: Iede12828fd4aad45eb9757e0a2d4b2cb062b3818