[roll] Roll fuchsia [dl] Add DlClose to System Test Fixture
This will call dlclose at the end of tests so that glibc can unload the
dlopen-ed file.

Musl's dlclose is a no-op, and will require tests that rely on a "clean
state" to use unique test modules and shlibs so that modules previously
loaded by other tests do not leak into multiple tests.

A TODO is added to implement dlclose in libdl.

Original-Bug: 342028933, 342483491
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052414
Original-Revision: 550245718d400911789f4399ba54c81009044f7f
GitOrigin-RevId: d766984d0d3d360cc3ccaece82611d3724553d59
Change-Id: I05727a897b95ef3eb4b4e8f91a4d1ea39fbb0cd2