[roll] Roll fuchsia [driver_manager] Refactor for driver dynamic linking.
Pass in loader and bootstrap receiver to driver host runner.

We want the option to specify which loader is used for which
driver host.

The bootstrap receiver zx::channel will be used to communicate
with the driver host. Eventually this will become a proper protocol.

Original-Bug: 341998660,49831408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1085000
Original-Revision: 625074c30f002796270cfd71384adda7a5f4b473
GitOrigin-RevId: aebd7d6f29989981e5836709e88356844313a3dc
Change-Id: I216d58afa1200314a33cf9ba403e12af5b8ccbfe