[roll] Roll fuchsia [display][fake-coordinator-connector] Create a single FakeDisplayCoordinatorConnector instance.
Previously the fake-coordinator-connector binary created two separate
FakeDisplayCoordinatorConnector instances to serve /svc and /dev
directories. This didn't reflect the actual use case in display clients
and it was not compatible with DFv2 driver backends.

This change removes the CreateAndPublishService method and replaces it
with the outgoing directory's "AddUnmanagedProtocolAt" method which
allows a service instance to be served at multiple paths.

Test: fake-display-coordinator-connector-integration-test
Test: //src/ui:tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1085287
Original-Revision: c657e7111649771712ad5b5c544636e2cf34bb98
GitOrigin-RevId: 7002fe98ecfa2b18bb6d89c360289744d33629d7
Change-Id: I65aba095da13bbb64455f6e40c16c92dd3397fd6