[roll] Roll fuchsia [codec][vaapi] Add const to copy assignment
https://github.com/llvm/llvm-project/pull/89652 made
std::pair trivially copyable if its members are, and
it caused the following issue: "error: the parameter
for this explicitly-defaulted copy assignment operator
is const, but a member or base requires it to be
non-const". This patch adds const to copy assignment
operator to fix the issue.

Original-Bug: 354627057
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086537
Original-Revision: 3d310ae50b6ea4982450cdb6f817fd2acfb02ead
GitOrigin-RevId: 25df2142c8e3b35418b44dc7ffa7afe524d0dd41
Change-Id: I616f86982e64af8efac78d200a74f1fb6484d2f7