[roll] Roll fuchsia [camera][ge2d] Add const to copy assignment
https://github.com/llvm/llvm-project/pull/89652 made
std::pair trivially copyable if its members are, and
it caused the following issue: "error: the parameter
for this explicitly-defaulted copy assignment operator
is const, but a member or base requires it to be
non-const". This patch adds const to copy assignment
operator to fix the issue.

Original-Bug: 354627057
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086536
Original-Revision: a2fef69a6c032a50ac2c62ead4e7212a488ad238
GitOrigin-RevId: 18c0605dc1feadcd7020552002db782a3b26fd18
Change-Id: Iba5488f6e2d16d8f4383f10f1d77f003301d7ca8