[roll] Roll fuchsia [boot-shim] Introduce DevicetreeGenericWatchdogItem
This item will generate a generic watchdog item when any of the
registered watchdogs provides a matching string and is able to
successfully fill the payload.

A 'Watchdog' type must satisfy the API such that the generic item does
the matching and delegates the filling and internal characteristics of
the item.

Fixes: 352391159
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056873
Original-Revision: 87d1ae8041ddb7ad1b5907f3344336ef89754e72
GitOrigin-RevId: a94b9f9839a1a2759060911b8bf3480e5ee7e5ee
Change-Id: Ib2968336bd74474ad6bca00f7dddb1c0224411dd