[roll] Roll fuchsia [merge_power] Normalize Data before Correlating
Suppose we have a signal and feature
```
signal: [5, 0, 0, 0, 0, 2, 2, 2, 2, 2]
feature: [10, 5, 5, 5, 5]
```

The feature has a high, followed by relative lows, so it intuitively
make sense to align it at the beginning.

But if we do the correlations
```
[5, 0, 0, 0, 0] o [10, 5, 5, 5, 5] = 50 + 0+ 0 + 0 + 0 = 50
```
And the other way:
```
[2,2,2,2,2] o [10, 5,5,5,5] = 20 + 10 + 10 + 10 + 10 = 60
```

It's better to align it with all the twos -- the medium values beat out the matching high values.

However, if we normalize first, we get
```
[1, 0, 0, 0, 0, .4, .4, .4, .4]
[1, 0, 0, 0 ,0]
````

Here, the feature now aligns correctly with the beginning of the
feature.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086957
Original-Revision: 1495cc7313e7d12800c41d94d4506304263c518c
GitOrigin-RevId: 68132342763e97b5b5f9db88b6938ca3d535ba46
Change-Id: I9e3edb8e5d10c6a051f6b1c83c587890aa81317e