[roll] Roll fuchsia [ffx] [pkg] store repo from CLI not daemon
The daemon should not make config changes. This changes one occurrence,
the setting of the repository.repositories config, so that it is updated
from the `ffx repository add/add-from-pm` plugins instead.

Also, because we are now writing the config from two places (the CLI and
the daemon), we need to make sure that the daemon picks up any changes
in the CLI by invalidating its cache. Otherwise, it may not see changes
made by the CLI, and it could  any updates. Otherwise the daemon can
overwrite the changes made by the CLI.

Original-Bug: b/343293737
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060975
Original-Revision: 7fa6e2c587ef45456f5108f344b68027b25f38c9
GitOrigin-RevId: 61d653463e9fc7a94b0599278d7a32c5ac2e6448
Change-Id: I4c141820100726d83fa018f0bcd83613696f5255