[roll] Roll fuchsia [ffx] client-side discovery must set ssh port
When using client-side discovery, the port was not being set in
the daemon, meaning that it could only talk to a target with ssh
port 22. (In addition, the `ffx self-test` test
`test_get_ssh_address_includes_port()` assumed the port was always
22.)

Also, because this triggers an `ffx self-test` run, we need to
remove a test that had stopped working when client-side discovery
was introduced. (The test, `test_get_ssh_address_timeout()`, is
based on likely-obsolete functionality, but rather than also dealing
with the changed behavior with get-ssh-address timeouts in this CL,
I'm just removing it.)

Original-Bug: 354273937
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086912
Original-Revision: 94bb1b801db1e86d06043d3d3e3666ff4fb6052b
GitOrigin-RevId: 2cbaa81162c4d4c16cff10a41a6839f58e5c6528
Change-Id: I6f4efeae161d06d9cb32452a82483c71eebde77a