[roll] Roll fuchsia [scenic] Use zx_cache_flush
Replaces usages of ZX_VMO_OP_CACHE_* operations with zx_cache_flush.
With RFC-0252 these VMO cache operations are incorrect as the VMO can
be a physical VMO. In most cases there is a mapping available and the
preferred zx_cache_flush can be used instead. For one case where a
mapping is not available a fallback to create one is added if a
physical VMO is detected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054132
Original-Revision: 1d2eb5aca7d98c7b0039d32d54c53c41baa09455
GitOrigin-RevId: f94cb575cfdfe490a377cc3a8332a4a68ef1f254
Change-Id: I4f8660b395130416601e67ea7374d63e82013a39