[roll] Roll fuchsia [wlan-frame-writer] Add support for writing to padded buffers
The write_frame_with_fixed_slice!() macro now supports filling the
beginning of the buffer with zeroes so a frame can be written with no
space at the end of the slice. This is being added in preparation for
network device buffers which will require this functionality.

Additionally write_frame_with_fixed_buffer was renamed to
write_frame_with_fixed_slice so that an arbitrary amount of space at
the back of a buffer could be excluded from a write by specifying
a slice.

Test: fx test wlan-frame-writer-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087178
Original-Revision: 48f79ef1c8869ad57aa1e8fdf106ea874c1d98c6
GitOrigin-RevId: 9e69013894b9dbfd8d4bac226034d51678167a6a
Change-Id: I3107baab2e893178a2b51da81de869041047fd45