[roll] Roll fuchsia [kernel][x86] Only scan PTs on failed mapping
The RemoveMapping code attempts to handle the scenario where the
AddMapping had been installing PTs, but got an error part way through
the process. In this case AddMapping would call RemoveMapping to clean
up, but since no mappings were found to remove RemoveMapping would
assume it should leave the PT there. The any_pages bool fixes this case
by ensuring that the PT is still scanned.

This boolean is overly aggressive though and, in the case of doing a
user requested Unmap, if no entry is found to unmap we would still do a
PT scan on the first walk back up. In this case we know we did not
install any page tables and so the extra scan is wasteful.

In this change a flag is added to track whether existing empty page
tables are expected or not.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084273
Original-Revision: 1c32821081579a07d1a163fafe1722080a8a7624
GitOrigin-RevId: a22960b769b020ca9a6736cc5eaf3bea5dacd313
Change-Id: Iacf2482d7b263f87a5da46ab243234636d248db1