[roll] Roll fuchsia [lacewing] Ensure test terminates within 10sec on SIGTERM
On receiving the SIGTERM from infra, infra gives about 10sec for the lacewing test to handle the SIGTERM before sending SIGKILL to kill the test. So make changes to ensure there is a better chance that test termination is completed within 10sec on receiving SIGTERM.

1)
Mobly Driver on receiving the SIGTERM from infra does the following in 10sec:
1 Terminates the test and wait for the test to terminate
2 Runs mobly driver's teardown
Currently we are giving only 3 sec for test to terminate and remaining 7 sec for mobly driver teardown. Increase the timeout related to waiting for test to terminate so that there is a better change for test to properly terminate.

2)
Honeydew has some code where it does retry for in some APIs (such as sending HTTP requests) for any exceptions it is received. In case of SIGTERM, infra sends it to Mobly driver which sends it to Mobly which sends TestAbortAll to test case. So at the time test receives TestAbortAll, if it is in middle of running the retry API, then it wont honor TestAbortAll as these retry based APIs currently retry for any Exception. So update Honeydew retry based APIs to not retry for any Exception but only for specific exceptions related to API (in case of HTTP requests, retry only for ConnectionError and urllib.error.URLError)

3)
In continuation to #2, remove the places in Honeydew where it is
listening on any exception ("except Exception:") as much as possible. So
that Mobly's TestAbort* erros are not caught by Honeydew.
Note: There are still some places where Honeydew is catching broad
exception. Esecially ones in
https://fuchsia.googlesource.com/fuchsia/+/refs/heads/main/src/testing/end_to_end/honeydew/honeydew/utils/common.py
may cause intervention with the SIGTERM handling

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087732
Original-Revision: 0c9f3747802b4e0457ee5af326221ec8113828fe
GitOrigin-RevId: 940c059f40e44b71454bbb3b77ac348e5747179b
Change-Id: If736af12a2762b8293f440c112bbd836fbe26187