[roll] Roll fuchsia [cm_fidl_analyzer] include capability source in route results
The detailed route segments is going to be removed by future commits as
part of component framework's bedrock workstream. The most important
thing that these route segments are used to confirm is the source of a
route. This commit adds the capability source to route results, so that
tests can assert that the source of a route matches expectations without
using the route segments.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084412
Original-Revision: ce5334d12825730ab67f0c482d9acfaac2317dd4
GitOrigin-RevId: bfe397a2f1332012edcfea41e674d2d6f0aeb6c3
Change-Id: Ic908b142cfe9ab4ecbcc5aa8c6dbca79901e46f5