[roll] Roll fuchsia [banjo][fidl][display] Remove the layer z-index concept from the API.
In the API between the display coordinator and display engine drivers,
layers currently have a z_index property. Display engine drivers are
supposed to use the property to merge the layer list in a display
configuration passed via ApplyConfiguration() with the layer list in the
current configuration.

The z-index concept is under-documented and under-tested, and only
supported on the intel-i915 driver. This CL removes the concept, in
order to reduce cognitive load and the cost of driver migration efforts.

Original-Bug: 316631158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089297
Original-Revision: 92f074eee5e62152af43e955f5c76f53163bd406
GitOrigin-RevId: 4277a80289b26a0fa2c1e3024b168819d339801f
Change-Id: I74056b26ac0a47d2780602d88b16d7ffd04d7ff8