[roll] Roll fuchsia [kernel][phys] Remove kArchHandoffGenerateUartPeripheralRanges logic
This is used to synthesize a PERIPHERAL range to cover UART MMIO;
however, this is now redundant with similar logic added to the arm64
Linux devicetree boot shim and first-party bootloaders being required to
supply it to cover all kernel MMIO access. This also is applied to the
construction of `PhysHandoff::mem_config`, which is slated to be
replaced in the context of fxbug.dev/347766366, so it seems prudent to
remove it instead of look to add new support for it in the new codepath.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089237
Original-Revision: c1b7b709feb660d306a98edcaf9beccefd94a7f6
GitOrigin-RevId: fa9115ced2df0fe838984c4effb3af5e3efc931e
Change-Id: I772eaa0e2e37f64ed34f8b284b8cc54793eee9c2