[roll] Roll fuchsia [starnix] Only run gvisor tests on hardware.
There are a number of tests we inherit from gvisor which use
sleep() to "control" the ordering of the test logic between multiple
processes. It seems that these tests sometimes flake on emulators
because the underlying host has not scheduled the emulator recently
enough.

Move these tests to only running on hardware to get more predictable
timing.

In the future, we may be able to work with the gVisor team to migrate
to other methods of synchronization but for now this should reduce
the flake rate.

Needs internal CL 864671.

Original-Bug: 355311252
Original-Bug: 353614027
Original-Fixed: 354249283
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084721
Original-Revision: bc4e0a2edf2395d8c63d58d5b29c8d5645744175
GitOrigin-RevId: 60e53b96e86b0a6c07c220963ca4e0c1462b19d9
Change-Id: Idec3452ad8178d03aef6a0f1cd6fa75011be7912