[roll] Roll fuchsia [kernel][pc] Remove x86 kernel PCI support
Hasn't been used/supported in a long while. Plus, some of this code
would need to be updated in the context of fxbug.dev/347766366, so it
seems prudent at this point to remove the code (and revive it later if
need be).

Original-Bug: 42108122
Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089236
Original-Revision: 4bfe4883873d5c89ed02a270fbf861baf41cf757
GitOrigin-RevId: 4a5ee466f981bc40765e43d654934770045434ca
Change-Id: I4fd4108c62fc0c332e5d4b00019060752d2b387f