[roll] Roll fuchsia [kernel][phys] Hack PhysLog to leak memory
While trampoline booting is in effect in the x86 codepath, care needs to
be taken with `Allocation`s made before the fixed-address image
recharacterization that TrampolineBoot does. In particular, we do not
want to absent-mindedly free a region (e.g., due to Allocation::Resize()
or Allocation::~Allocation()`) that was recharacterized as being  a part
of a fixed-address image. Until TrampolineBoot is removed from kernel
boot, defensively leak log buffer allocations when trying to resize or
destruct an Allocation.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088739
Original-Revision: a89ffed590554d722774bc7c2ad9422141c0f1c8
GitOrigin-RevId: 253654c4e8dc6630865b67383ae05507796ef7f0
Change-Id: I80cfb55b7df11827ac6080971ca0717cb31ab9e4