[roll] Roll fuchsia [starnix][vfs] Prevent umount when files are open
If there's a file open inside a mount, we are supposed to prevent the
mount from being removed by returning EBUSY from umount. To count the
number of clients, we add a marker to the Mount object that we reference
from each NamespaceNode. It's important to hold this reference from
NamespaceNode objects rather than FileObjects because the cwd and root
fields on the FsContext need to keep the mount from being removed and
those fields are NamespaceNode objects rather than FileObjects.

Test: umount02
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088114
Original-Revision: 2c1d445c66e30d15300a44068467ebbb9b158601
GitOrigin-RevId: 95f91ba94bdb4ae43d9f7c45c6fd1dc94e384d99
Change-Id: I21b2dcc1dfd7518069cea8881832734927c77864