[roll] Roll fuchsia [shac] Settle the age-old question: _ or -?
The answer is: do whatever files near you (with the same extension as
you) do. This is a SHAC check that attempts to codify that answer.

It only ever emits a warning, which we could also downgrade to a notice.
I manually looked through exists files that are out-of-compliance, and
they largely seem to fit the category of "files that would ideally be
differently named."

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082053
Original-Revision: 84001d2b18b3973e4efcf284c0927742fc2628d1
GitOrigin-RevId: 96953be2ec30cb64b37b5c3a0f15d8867a1f044d
Change-Id: Ida75ca48ad5691df3049d050a2a1ae519df2a27c