[roll] Roll fuchsia [starnix] Add diagnostic checks for TheadGroup children
This forces `remove` to takes a OwnedRef<Task> to ensure that the task
is still alive when it is removed, and this add a debug only check on
ThreadGroup that the children map is empty when the ThreadGroup is
removed.

The goal is to get better information on why there is case where some
children cannot be upgraded.

Original-Bug: b/354754795
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089834
Original-Revision: 159273d3ebd895c6274d9850f3fdc71aa440b50a
GitOrigin-RevId: 8a0eebd21da0ef1c121eb36f7b11b581ba9e60c7
Change-Id: I0c21aadb92f80585816db0e9674a1aaa77eb5c3f