[roll] Roll fuchsia [msd-arm-mali] Add inspect for slow atoms from connections
slow_atom_count and slow_semaphore_set_atom_count will count the number
of atoms that the MSD receives that it takes > 5 seconds to mark as
resolved. This will help detect if the MSD is hanging, or if there's
some other reason semaphore take a while to be signaled.

Original-Bug: 355485837
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089875
Original-Revision: 1427b14789bfd31a447fa052562cfe1f814296e1
GitOrigin-RevId: b9423e2e82de9c0641083b19a165a3be59f6e9b2
Change-Id: If5eb866afc2b29ce5dde68c68bbdcbfef40f4172