[roll] Roll fuchsia [netstack3] Rework multicast route witness types
Rework the witness types of source addresses for multicast routes. The
source addresses must be unicast, but `UnicastAddr<Ipv4Addr>` is not a
safely constructable type (because unicastness of an IPv4 address can
only be determined in the context of a subnet).

For IPv4, rather than using `UnicastAddr`, use a custom witness type
that ensures "specified and not multicast".

Original-Bug: 353329136
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089036
Original-Revision: 5e7cf7f96f6a81af24cca4d6a4373526c6d663a8
GitOrigin-RevId: ca53206fcb43b1d7ff678ea6712812d39ef7d12e
Change-Id: I8c77976b1b4653ccd81bd4afca0c825394c76c86