[roll] Roll fuchsia [fake-display-stack] Add SysmemServiceProvider interface.
Previously, `FakeDisplayStack` created a mock-ddk-based device root
and used the `GenericSysmemWrapper` template class to bind the sysmem
"devices" to the root. This caused a lot of confusion in the fake
display stack:

(1) the SysmemProxyDevice doesn't need to be bound to the fake root;
(2) the template class requires the SysmemProxyDevice and
    sysmem_driver::Device to have similar interfaces, which is also
    not necessary;
(3) SysmemProxyDevice implements the DriverConnector protocol, which
    is not usable by any clients.

This change is the first step to refactor the FakeDisplayStack.

This change defines a new interface type `SysmemServiceProvider`. Its
only responsibility is to provides the fuchsia.hardware.sysmem/Service
to an outgoing directory, consumed by display Coordinator and the
fake-display library.

The `SysmemServiceProvider` doesn't place any constraints on the source
of the sysmem service protocols. It could be from the component's
incoming directory, or from a self-owned fake sysmem device.

Test: no-op
Original-Bug: 354836088
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087592
Original-Revision: f0319b9a8c2a769092d3d87b176f626d662ce1aa
GitOrigin-RevId: 0e96697f966495c5e51ef787f818c571190e1604
Change-Id: Ie20de5b583da1ce3bfccfb2b995ffd56874f8404