[roll] Roll fuchsia [netstack3/conntrack/tcp] SynSent
Since transitioning to the next state requires knowing all of the data
it will contain, this also adds the Peer struct, which will be core to
all the upcoming states (minus TimeWait).

Implementing this state took a bit of refactoring to plumb the new
update status result into the right places, but no more should be
required for the remaining states.

Original-Bug: 328064736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088372
Original-Revision: 7188e41695b59200bcda6572f8f8e90f18336359
GitOrigin-RevId: 67d099480fded50e11f83cfbe237c6cc4b4e6ec9
Change-Id: Ied7517694e0767bc1f41fc69973815425216a14d