[roll] Roll fuchsia [ffx][build] Adding --no-environment to ffx_action()
This adds the `--no-environment` switch to the
ffx_action() template by default.

This is done to improve hermeticity of running ffx during
the build.

The option is exposed as a parameter in the template if there
is some edge-case that requires the configuration to be read.

Original-Bug: b/355672251
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090795
Original-Revision: 785ef6f0098051230bfdb1656bb496fd007fdf35
GitOrigin-RevId: 8251d73ba54a372e719404eea447e6826579766c
Change-Id: I3b9fb4a64847092eed89dd05d53151641029fef1