[roll] Roll fuchsia [fx][rust3p] Call ninja directly instead of `fx build`
This removes the warning from `fx build` that directs users to use GN
targets instead of Ninja outputs.

I opt to not use GN target for build in this particular case because
it doesn't improve the script. In fact it makes it slightly worse
because moving the GN target would requiring changing this script, which
building the Ninja output directly does not, and we'll need to hardcode
the path to the Ninja output anyways to invoke it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089714
Original-Revision: d65f2563711ab51cc715315bd4b4a35a345a13ae
GitOrigin-RevId: 27c6c0aedcefb6af8f1a05e34b3e028bf5f7ec95
Change-Id: Ia32513450ea28034baa6d06eb4b72d9524b7ae31