[roll] Roll fuchsia [msd-arm-mali] Add magma info dump to inspect
Having information about the state of the driver is useful in hang
scenarios. A lot of the scheduler code is on a latency-sensitive path,
and some information is read on demand through hardware registers, so we
can't update it as changes happens.

Because of that, it makes the most sense to create a LazyNode that grabs
the GPU dump whenever it's requested.

Original-Bug: 355485837
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090032
Original-Revision: 9a14fcf1bf3e8b4344bb025ec099fdd4b3256c4a
GitOrigin-RevId: fe9d4f4672a0511e0a7718d3dff7e33f66434d9c
Change-Id: I85fb3d131835d190b6815794ecfbec39c78a3c0a