[roll] Roll fuchsia [ffx][metrics] Refactoring analytics dir calculation
This refactors the initialization of the metrics library
to make the metrics folder be passed in.

This aligns the metrics folder with the rest of ffx configuration
logic, so it now works correctly with isolation directories and
then `--no-environment` flag.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089473
Original-Revision: 61a03be9f47344c0a6b827d145cb41bcd3919c2a
GitOrigin-RevId: e671bd1e95be9d974e3e78f595da9c045e7c1b86
Change-Id: I9bcf2d47e5c7cf6d0eac9e1e5b5278a557879f9c