[roll] Roll fuchsia [netstack3] De-macro "process_fragment"
Convert the "process_fragment" macro into a function. Originally it was
a macro to circumvent problems with the borrow checker: It was taking in
both `buffer` and `packet` as arguments, which as a function call would
"move" `buffer` and invalidate the reference held by `packet`.

Here, I move error handling to outside of the `process_fragment` call,
which means it no longer needs to accept `buffer` as an argument, and we
can avoid reference invalidation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091573
Original-Revision: d7b2f403d3b0b8979be1b62ea93d062356b1326a
GitOrigin-RevId: 9088a86a788ad481046c11cddf50655f1baf85a8
Change-Id: I87561325f4e8855da7a8b6994d2226ea4b74c151