[roll] Roll fuchsia [starnix] Trace the start and completion of suspend
This was requested in comment #12 of b/348021898. The motivation is to
understand the performance of these operations.

Test: I ran

  ffx trace start --categories "starnix_runner" --buffer-size 64 --duration 10

on one terminal, and

  ffx starnix suspend
  ffx starnix resume

on another terminal. I was able to find suspend durations in the
resulting `fxt` file. The duration slices may be very hard to spot if
your container is not doing much.

Original-Bug: b/348021898
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091338
Original-Revision: 1c6635dd89d64a25306a0645d2f3ec89eb57eb83
GitOrigin-RevId: 83af87a803449f73b8f461f1e9301552aa347849
Change-Id: Ie39cb1d69d18a51d10ab2a914661241cf2e60ad0