[roll] Roll fuchsia [fake-display-stack] Add FakeSysmemDeviceHierarchy.
This change adds FakeSysmemDeviceHierarchy which is an
implementation of the `SysmemServiceProvider` interface.
It creates a mock device hierarchy with a fake sysmem
Device.

This is similar to the GenericSysmemWrapper with a
fake "sysmem_driver::Device". The difference is that the
FakeSysmemDeviceHierarchy contains all the dependency
devices and doesn't depend on any components from the
FakeDisplayStack.

Test: fake-sysmem-device-hierarchy-tests
Original-Bug: 354836088
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087594
Original-Revision: 7b772783641e5042199db8e451f9234af61512fe
GitOrigin-RevId: ee17bfe9b08318d585db27980146a960e6904673
Change-Id: I1149d9678ee9a275be2b897fff2a01799ffeca1b