[roll] Roll fuchsia [kernel] Remove last use of boot_reserve_foreach()
This was used when mapping EFI regions to double-check that they are
reserved (i.e., not marked as RAM). Rather than look for some PMM
replacement for this logic for this (currently) single callsite, it
seems more prudent to find an alternate, specific means of doing this
for now. In particular, this is straightforwardly done in this change by
consulting the normalized hand-off memory description and checking for
overlaps with RAM.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093074
Original-Revision: 7e60ebf856ed637f91f476b4ce66933df834ef5d
GitOrigin-RevId: a24a937729ce24d6e121bb8d7585f849cace18b2
Change-Id: If266987604d35f9da0faee7a1e0bb1801226856e