[roll] Roll fuchsia [fxfs] Get or insert keys to use None instead of NotFound errors.
Lets the caller decide how to handle a missing encryption key. In the
case of prefetch-keys it is best effort and the object may be deleted by
the time the prefetch arrives. In most other cases where a key is
expected they can turn it into an Inconsistent error.

Notably this handling change will prevent the error logged failure in
`UnwrapResult::Set` when the key is missing from the tree, as only the
caller can decide if this is an issue.

Original-Bug: b/356626753
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093072
Original-Revision: 930ea81ce78ae33b13faa611e9aea6ac70861a99
GitOrigin-RevId: 93273d79f576f624a9d96d893f352a4ff3473b3c
Change-Id: I0c98881cde1c26f735bfb50c5788ead232bf2cd3