[roll] Roll fuchsia [routing] simplify AllowedSourcesBuilder.framework
The optional framework field in AllowedSourcesBuilder held a closure to
convert a Name to an InternalCapability. In practice this closure only
ever converted a name into the same InternalCapability as the
CapabilityTypeName declared on the AllowedSourcesBuilder.

To simplify things this closure has been replaced with a boolean field.
When set to true, the framework_sources function will return an
InternalCapability matching the CapabilityNameType. There is no behavior
change due to this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091933
Original-Revision: f7d0159c021087cca330a41b520d995ac116b75f
GitOrigin-RevId: 5e0c7e207244b41ee346f8532b99039852718156
Change-Id: Id96b035555f2c08dadd52e4163b63358aed1ead6