[roll] Roll fuchsia [net-cli] Add integration test
To facilitate writing more meaningful tests of net-cli (rather than
change-detector tests that mainly unit-test the FIDL bindings), adds
an integration test that uses netemul and network-test-realm to be able
to test net-cli invocations against a network-test-realm-owned netstack.

In particular, this will make it easier to test the route FIDL API
modernization changes for net-cli; the added test case checks that
`net route add` and `net route del` work.

For simply integration testing net-cli, network-test-realm would be
unnecessary -- netemul is sufficient and running network-test-realm via
netemul is redundant. However, because there are out-of-tree end-to-end
test usages of net-cli within network-test-realm, it's helpful to
exercise net-cli-within-network-test-realm in this way in order to help
validate that the correct capabilities are available in-tree. (This is
noted as a file-level comment in the integration test as well.)

Original-Bug: 42081105
Test: net-cli-integration
Multiply: net-cli-integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091315
Original-Revision: 3209928f436b25c07992f1a42317eb26c10f3f9d
GitOrigin-RevId: a31762c9680172a8526a6e011a7866ecd07a0dfe
Change-Id: I5a5f3ca08031d00bc63b91addb6a486fef859723