[roll] Roll fuchsia [audio][test] VirtualAudio VMO rights
Update virtual_audio driver (Composite, Dai, StreamConfig)
to apply the appropriate handle rights to the VMO that it
returns in response to `RingBuffer/GetVmo`.

In the process of doing this, also fixed the Composite
implementation of `WatchTopology` (which tests use to
determine whether ZX_RIGHT_WRITE is required) so that
driver signalprocessing state is correctly reset when the
Composite protocol channel is closed.

Test: audio_driver_tests
Test: audio-core-api-pipeline-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093753
Original-Revision: ee7e5dba2b95858ee1f8f176b851ebbd1e9ee61d
GitOrigin-RevId: f22a617b5f74232bb50e6185e245be7bea55e199
Change-Id: Ia6bc81e30aa5bd30ed49d475e5395d9ef27727f4