[roll] Roll fuchsia [rbe][cxx] Workaround for newer clang toolchain
Newer versions of clang drive the compiler in a manner that
is sensitive to the existence of target libdirs.
re-client isn't aware of this yet, and only includes the correct headers
as remote inputs, without any linker directories.

This temporarily works around the issue by adding one file per critical
target libdir as a remote input to force those directories to
exist in the remote environment.  Since this wrapper is only used for
compilation and not linking, these extra remote inputs are not actually
used.

Original-Bug: 354016617
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092316
Original-Revision: 6cecde067cd647de032bb9ecb95827bdd4f8d982
GitOrigin-RevId: 3c792c718b2c6636ad679ac4e833ef644724e1d0
Change-Id: Iae866c9f6d5d25a49848556f2a96e0c1b39317d3