[roll] Roll fuchsia [netstack3] MulticastForwardingDeviceContext
Move the check for whether multicast forwarding is enabled on a device
to its own context trait. This moves in the direction of narrowly scoped
context traits, which reduces the amount of boilerplate required in
unit tests.

This is the follow up to a comment left on https://fxrev.dev/1089037

Original-Bug: 353329136
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093173
Original-Revision: 001d96d94f8cc0c70d03e01470aa8d69fcec3fe1
GitOrigin-RevId: 2565ecda2eee2d230c58be5f685dd4516d2721ca
Change-Id: I658458350453d9b0427b57aeed5db76aafdc1fde