[roll] Roll fuchsia [kernel][mmu] Remove DropPAddrs
The MappingCursor::DropPAddrs method exists to allow a MappingCursor to
switch to a mode where it only tracks virtual addresses, allowing it to
consume a non-page sized portion for error purposes. Consuming such a
large portion of virtual address is excessive for the outcome needed,
which is to ensure that the table is checked for emptiness, and a
single PAGE_SIZE slot being consumed is sufficient.

Removing the DropPAddrs method will allow for splitting the
MappingCursor into type safe variants that are either always tracking
paddrs, or always tracking just a virtual range.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093473
Original-Revision: 6762b1503d7d393c41236a069f1c67ec15cf9567
GitOrigin-RevId: dc25618134e1389e1e6ccd3c19dd0d6a9b0b927e
Change-Id: I9b81b7e610c5da4386f5258c03f6c6579a029829