[roll] Roll fuchsia [kernel][mmu] Separate VirtualAddressCursor
Splits the mapping cursor into a MappingCursor that always tracks
physical addresses to map, and a VirtualAddressCursor that only tracks
a virtual range.

This provides a safer set object as previously there were methods on
the MappingCursor that were only valid to be called if it was in the
correct mode and this had to be asserted. Now these methods are just
on their respective specific implementation and mistakes cannot be
made.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093474
Original-Revision: 2b439189738dba659d235adf30b6af4034173b68
GitOrigin-RevId: 73f09b4441bca75a91f8bf2916d87e81ed08799e
Change-Id: I768a498862adde70adcd3267d562743fb132a024