[roll] Roll fuchsia [aml-rawnand] Wait for command queue on reads
Wait for the command queue to be empty when reading before checking the
status bits.

The nand cycle time for the Micron chip is 24ns and the page size is
1024. This results in sleeping for 24.5us. Waking up, checking the bit,
and sleeping again can take ~5us. 4k is read from the device at a time
which should result in sleeping 4 times and taking roughly 118us. From
looking at traces, a 5th sleep is frequently required. 5 sleeps * 5us =
25us of CPU time spent per 4k read. Blobfs typically reads in 6 pages at
once, 25us * 6 = 150us. That's half of the nand-worker's deadline
scheduler budget just sleeping.

Multiplying the number of ecc_pages into the sleep would only sleep for
98us and more polling with a shorter interval would be necessary.

Testing different sleep durations shows that 130us works almost all the
time. 130us is also the command queue flush initial sleep.

This change goes back to the previous implementation which waited for
the command queue to be empty. That single 130us sleep worked 99.99% of
the time. To account for the 0.01%, the status bits are still checked
with a short polling interval.

Original-Bug: b/350020539
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092194
Original-Revision: 4d36ef68d699ee90c33f4047a07103d0cd9a48ea
GitOrigin-RevId: fb0afed2d11a25423d5dbc2cc1343d18d4441c0b
Change-Id: Ie51bc39256d2c959714465df8f2a6703f05368d5