[roll] Roll fuchsia [connectivity] Remove duplicate routes
Routing LogSink manually is not needed,
as this is done implicitly through the client shard.
Future CLs will update the routing used in the shard,
so it's necessary to cleanup these duplicate routes
in order to update the shards.

Original-Bug: b/345827642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095733
Original-Revision: 3ebd66b48465f628501d47839fc4872d645a8d25
GitOrigin-RevId: a4f12401ab808c9f75ead172103e8ab1bf7d7b35
Change-Id: I9848eb3f95702ec918c6d7ca9b72f158a323c751