[roll] Roll fuchsia Reland [ffx][emu] Refactoring disk hashing to emu_instance
> Reason for revert: Broke mac builders
> Failed builders: go/bbid/8740400844815955649, go/bbid/8740397533543846081

The original was reverted because of a build error on mac. This has been
fixed.

This refactors the API used to hash the disk images to be
in emulator_instance so it is easier to use this when the
configuration is using EFI vs. ZBI to start the emulator.

> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094269

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094835
Original-Revision: 2e1a3db79b8056ede6715c385a6dc0eb8350e351
GitOrigin-RevId: e4e39a905a5692af9de7f0875fd61a4873e72e18
Change-Id: Iac8ebd77c3048a76685e30deedd51caab4b4965e