[roll] Roll fuchsia [debug] Remove duplicate routes
Routing LogSink manually is not needed,
as this is done implicitly through the client shard.
Future CLs will update the routing used in the shard,
so it's necessary to cleanup these duplicate routes
in order to update the shards.

Original-Bug: b/345827642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095735
Original-Revision: 36167b6900ecf7426da226ff1325c020d7b08ae4
GitOrigin-RevId: dc17ed65fdd2357130d5df53c1ebe9273b2ee105
Change-Id: I4b3529dd577265cb7bb064191485a2c4a6b58337