[roll] Roll fuchsia [kernel][x86] Protect physmap
Unmaps the non-RAM portions of the physmap when possible on x86,
instead of protecting. Unmapping is possible if not using a MMIO based
UART, and as these are not very common this should result in unmap
being the common operation.

Original-Bug: 42124648
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053097
Original-Revision: af317f4bf49ae602e70c0db63ef58c6d06cd3f76
GitOrigin-RevId: 32d718c4f8afe2f500ba741aa95e4d014d901494
Change-Id: I72067d648bf1ebf6a12ce632308c661b8c68ff8f