[roll] Roll fuchsia [vulkan] Remove duplicate routes
Routing LogSink manually is not needed,
as this is done implicitly through the client shard.
Future CLs will update the routing used in the shard,
so it's necessary to cleanup these duplicate routes
in order to update the shards.

Original-Bug: b/345827642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096901
Original-Revision: d432615672f335f7e660f806e06ee8a889624182
GitOrigin-RevId: 8d910d692d39fb69aa37d2fc8ab3e546f4b70cd5
Change-Id: I576fb4d9a8dac6837f91e8ab6a43135a60e04d14