[roll] Roll fuchsia [abi-compat] More optimizations
This changes the compare::Path type to be much simpler, just carrying
the version and a string representing the path (rather than a Vec of
path elements). The code to generate path strings moves to
convert::Context and is more efficient - calculating the length of the
string before allocating the space all at once and filling it in.

This cleans up the compare::ProblemPaths type to be simpler and hangs
onto values used for display and comparison rather than fetching or
calculating them.

This is an incremental 33% improvement. Overall relative to baseline
this is a 56% improvement.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096172
Original-Revision: 727264c3bea4ed7e51ab167074dcd5ced82171ba
GitOrigin-RevId: 8e1bf22bd93e06aa2fcc5d92ae874eca135628e2
Change-Id: I39c02b54a2952d4bbef121dc2e1c7ee8f3c8f768