[roll] Roll fuchsia [abi-compat] Clean up & optimize error reporting
This includes a bunch of simplifications and optimizations in error
reporting. Originally the API levels and paths of conflicting
declarations made up part of the error message (CompatibilityProblem)
type, but as the tool developed the API levels, ended up in error
messages too, and we only ended up referring to common paths. So this
removes API levels and individual paths from CompatibilityProblem and
also the types we use for tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096173
Original-Revision: 10e8cdb9c993c4b72be9883c7653df3ffe6b59fa
GitOrigin-RevId: 37406998b979135339c3f0234e46b9c57b23bbe1
Change-Id: I4bc9a5fcc62a6b8f41f97f4d6d332791b38190b8