[roll] Roll fuchsia [components] Support component exit codes
We add a `OnStopInfo` event to the `ComponentController` channel. The
first usage is an optional int64 exit code.

The ELF runner reports the Zircon process return code as the exit code.

Component manager will relay this in the StopPayload component event.

This will improve diagnostics experience and paves way to only reboot
the system if a critical component exits with an error later, which is
useful for escrow.

Original-Bug: b/42081082
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094214
Original-Revision: b07a4843eca072fc018407d0252ca9fbf7d27d3e
GitOrigin-RevId: 30a02356eedf8e8b82fb7781f74ac3dc0fa01b68
Change-Id: Ie7f699b7c08cdd2395d2f37d3674123e7d1db5dd