[roll] Roll fuchsia [libc] Temporarily disable UBSan on __dls3 for riscv64
The bringup.riscv64-asan builder was crashing very early during bringup
with a stack overflow in the next clang roll.
-Rpass-analysis=stack-frame-layout showed a bump in stack usage for
__dls3 (almost 56 extra bytes). Disabling UBSan for this particular
function on RISCV results in the same stack usage before and after the
clang roll. It's likely that upstream changes to UBSan resulted in extra
stack usage. We've experienced this type of issue before regarding UBSan
and bringup.riscv64-asan. We have plans to address this in the long run,
but for now we can just suppress UBSan for just this one function to
unblock the clang roll.

Original-Bug: 42080808, 357634850
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097105
Original-Revision: df7e4d4624304567f016982a0742059426084898
GitOrigin-RevId: c6684ac3b17c9df3dc3ed26d2e1fe9ce8e6c5c33
Change-Id: I869197fbc5e16c705212a86193c031e01d43cd63