[roll] Roll fuchsia [ffx][emu]Changing --dry-run to run after staging
This changes the emulator setup flow to stage the emulator instance
before printing the command line when the `--dry-run`  option is
provided.

This is being done to make the `--dry-run` output match the actual
command line that would be run to start the emulator.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097796
Original-Revision: 6b9faa436a964dc37e6afe66ab77a40a67236e01
GitOrigin-RevId: e64defa86e4708d6c3302094eefd59dbaca17054
Change-Id: Ib873f037180347bffb557c9c50f31e8ea510af07