[roll] Roll fuchsia [netstack3/conntrack/tcp] Move self in State::update
...and return a new state to replace the old one.

Most of the upcoming helpers use this pattern, and pushing it all the
way up to the highest level means it can be used in more places.

As part of this, we add a "Closed" terminal state, which is used to
signal that a connection has been closed and should be removed from the
conntrack table.

With all of this simplification, the return value of the update function
has become (State, bool), where the boolean is whether the segment is
valid.

Original-Bug: 328064736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092877
Original-Revision: 9c4bf8976a5117688fabe463365533a59b208a7a
GitOrigin-RevId: 6b1b0a8dfa2305b8545d2440c812529be9348a3c
Change-Id: Ifba0b7fed7d637ec9b04531042573f0e65139326