[roll] Roll scudo [scudo] Added test fixture for cache tests. (#102230)
The test fixture simplifies some of the logic for allocations and
mmap-based allocations
are separated from the cache to allow for more direct cache tests.
Additionally, a couple
of end to end tests for the cache and the LRU algorithm are added.

GitOrigin-RevId: 87d9b29e2996edff083375d22495473d1b3f09e7
Original-Revision: 72353a410ab866d268bd27d00f8f63f2c9d4d56b
Change-Id: I15809dfc2e8d41a6ea98893b963305c7c81ba51e