[roll] Roll fuchsia [kernel] Select PMM arenas algorithmically
This change updates PMM initialization to be driven by the normalized
accounting of hand-off-described memory, running that through the new
selection algorithm in <vm/phys/arena.h>.

See the commit message of fxrev.dev/1074514 for the limited arena
selection differences for supported boards.

In the interest of incrementality, we continue to use bootreserve to
manage reserved regions. This will be removed in a follow-up.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076373
Original-Revision: f5ec26f2585411067b4d95795a0584ea3cca3a5d
GitOrigin-RevId: c061733db55032f1fbf445f20cee20eb3d77d6cb
Change-Id: I3ecb1abc6c2d73a58f953d28da03222889964d35