[roll] Roll fuchsia [ffx] Logging of daemon connection
There are a lot of "failed to connect to RemoteControlProxy" flakes,
and the only information is that we find the daemon link, and then
never connect to it. This CL adds some debugging so we can track
the failure a bit further.

Original-Bug: 350541791
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098020
Original-Revision: 889496995972f70aec53a760925b883e1bf8db68
GitOrigin-RevId: 0187b2694ce4357acdf519d64d6131f90023fb92
Change-Id: I48bf17a5b6ed7bfb4559c0e585c85253218ceb99