[roll] Roll fuchsia [routing] every RoutingError variant has a moniker
In order to preserve UX in scrutiny, every routing error variant should
include a moniker indicating where the routing error happened. This is
necessary because today this information is guaranteed by the presence
of list of route segments in the scrutiny output, and in a future change
those route segments will be removed. This error type thus must carry
all the necessary information for diagnosing and resolving the issue it
signifies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094672
Original-Revision: 9bfca821527adac0753b2cfaefd1884599976eb9
GitOrigin-RevId: 2e3f7aa28bc69e27d1abd92f1c1a9373230171de
Change-Id: I1ca31c7a6982b403853ac2d7dbd8a52d5ef65d45