[roll] Roll fuchsia [starnix][task] Check for X during exec()
Previously, we were checking for the R throughout the interpreter chain
and then checking again for X at the end of the interpreter chain. We
now check for X (and not R) throughout the interpreter chain.

We still fail SuidTest.SuidBinaryBecomesRoot, but we get much further.
For example, we're now able to actually exec the test binary.

Original-Bug: b/329142824
Test: SuidTest.SuidBinaryBecomesRoot
Test: InterpreterScriptTest.BinaryNoExecute
Test: prctl06
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097555
Original-Revision: ab2cfd74ca865ba5ea12563ea14866ec7b7b5478
GitOrigin-RevId: dccdbc29662dd125ddc0bbbd21056cefd592be55
Change-Id: I81b72cb8e684a586bdc895af1eb19b3a3a2af6a6