[roll] Roll fuchsia [build] Don't use -Wc++20-compat
This switch is meant to be used in C++ 17 mode to catch things
that would need to change for 20.  However, we were using it
instead only in 20 mode, which is not useful.  There are too many
warnings in 17 mode for things that are already conditionalized
appropriately for 20 mode so it's not useful at all.

Original-Bug: 42064981
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068733
Original-Revision: 02a680054af9c9618a6846db9d0409e2cca439ef
GitOrigin-RevId: 5a0277527ccbd53f487ec5f56f6508f44cd05721
Change-Id: I871bc4183eda8087fa0a9b65e003a83af6ae7cd3