[roll] Roll fuchsia [kernel] Teach IdlePowerThread how to pause the monotonic clock
Adds the functionality needed to pause the monotonic clock to the
IdlePowerThread. All of this code is checked in behind a feature flag
that is disabled, so this CL should have no functional impacts. It will
give us a chance, however, to review the code and get it checked in so
that we can flip the flag as soon as the boot timeline RFC is accepted.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093967
Original-Revision: c607e04b1675936e19c1a372ab33dc35c6aee518
GitOrigin-RevId: 264d037d1c4602c8df383ae579ee01c9fedc8cb6
Change-Id: I7d7488a855fc4fe7659df48210d0824ecbaeb8e6