[roll] Roll fuchsia [bazel] Also ignore @internal_platforms_do_not_use
A previous CL added @host_platform to the ignore-list of
bazel_action.py, which was enough to remove build errors when
the script was checking inputs. However, this was not enough
and tqr/872950 failed to roll Bazel 7.2 due to Ninja no-op
failures after a successful build.

The root cause is rather ... esoteric, but the work-around
is to ignore the special @internal_platforms_do_not_use
repository as well, as it now appears in Bazel queries
that return the list of inputs. It's unsure at this point
why this repository name appears in query results at all,
but this should be enough to roll the new Bazel binary.

This was tested locally by verifying that a clean build
+ incremental fx build invocation returned "no work to do".

Original-Bug: 359151210
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099174
Original-Revision: 85bd378418814856c4b2d676359d4fe12bddb927
GitOrigin-RevId: 0d79c6b515ca08642cd94969743d25bb95baedc5
Change-Id: Icd439cd10ff3f6e84a494481b4ae72a4fd6360d1