[roll] Roll fuchsia [reboot] Ignore failure during reboot process
Previously the reboot process attempted to background a reboot  and
immediately exit with a return code of 0. This doesn't seem to work,
however, and a return code of 255 was returned. This change ensures
that no matter the result of the shell command, the script's execution
will continue to eventually get to a successful exit.

Original-Fixed: 322226825
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099512
Original-Revision: 434a90df887221517094274bbf3950af547b50e3
GitOrigin-RevId: b7ff0caf96a10c29cf540f217f056231bac53554
Change-Id: I0571a8e92be571dfe03f32c5c6ee09ce4e4f2c49