[roll] Roll fuchsia [kernel][vm] Avoid SupplyPagseLocked assert misfiring
The assert at the end of VmCowPages::SupplyPagesLocked is overly
aggressive and can spuriously trigger if AddPageLocked fails with
ZX_ERR_NO_MEMORY. This failure causes the loop to 'break' before
incrementing `offset`, leading to a discrepency in the number of
pages pulled from the VmSpliceList, and the number of offsets we
have successfully supplied.

As supplied_len is not used if any other error kind is returned, this
change redefines the function to state this, and relaxes the assertion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099052
Original-Revision: 8f1ea0501ac1be1febca56f41c158efc346eebd1
GitOrigin-RevId: d88afb28478f2834041d92c0ed0f56119999d7ab
Change-Id: I29aeba178ffc367bf5dc4630f41f427aa7c80896