[roll] Roll fuchsia [ffx][log] Remove dependency on FFX daemon_protocol
Refactors the reconnection logic in ffx log to use APIs provided by FFX
to remove the usage of daemon_protocol. This was identified as a blocker
towards the removal of the ffx daemon.

The core change is reasonably simple and was done in Patch#1, the rest
of the change is about refactoring the existing tests to write them on
top of existing ffx APIs that integrate well with the newly introduced
usage of fho::Connector.

I'll make a few additional improvements to the tests in a follow-up.

Multiply: ffx_log_lib_test
Original-Fixed: 299981807
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098792
Original-Revision: 62200b2e2b3817bd6c6836ba2f4d1408099585e9
GitOrigin-RevId: 31f5a4dd41ba76e3dbb1e5852640f92a11321a3e
Change-Id: Ib77a405507b4b375905db20901767da17c8b9062