[roll] Roll fuchsia [ffx][repo]Correctly handle daemon server in list
This adds calling the FIDL proxy to determine  the state of the
repository server when running as part of the ffx daemon. The previous
logic relied on the process status to determine if it is running or not.
The server in the daemon can be stopped without changing the process
status, which causes the list command to incorrectly report the server
is running.

Original-Bug: b/359290928
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099633
Original-Revision: 688425ca467ef1001fc24c364408b912f6b6f81f
GitOrigin-RevId: 79ce38d70592d7b9c42b640f6e66905b2668bc56
Change-Id: If0b68261fea79f20a98e08db02bef894d0b30942