[roll] Roll fuchsia [driver_runtime][rust] support for "messages"
This implements a "message" object that can be used to hold an arena and
accompanying data, and then allows unpacking that data in various ways.
It also adds a handle type for managing driver runtime handles, which
are stored the same as zircon handles, but are slightly different.

Original-Fixed: 359226657
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087878
Original-Revision: 168b59fc0afcd678f508df4cdc6a2cd2e39da7b8
GitOrigin-RevId: 00f181ddb7cf5b08aa8ec660d5cace21369ccf09
Change-Id: I523939638b33cd9d32eaa65e6971d164d0b24730