[roll] Roll fuchsia [ffx][logging] Allow subtool override of logging file.
This adds a method in FfxMain trait to allow the subtool
to override the basename of the log file to use.  This
then takes effect just prior to running FfxMain::main().

The basename is not changed if the log file is specified on
the command line.

This change also changes `ffx doctor --record` to capture all
.log files in the log dir.

Original-Bug: b/358404932
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099613
Original-Revision: 41fade892c0c31e0bf93e52e14f598c2c5f2aad6
GitOrigin-RevId: e70b9a39a556b0fd60ca316799564991db03fffe
Change-Id: I539f61f4207ce0d3b38c5963de771e2c34092a1e