[roll] Roll fuchsia [bazel] pass copy_debug_symbols in bazel_action
The debug symbols for product bundles were not being copied into their
final location because we were not actually passing along the
copy_debug_symbols value for the directory_outputs scope. This also was
causing us to not track all values that were being passed in to the
tracked_files field when we did not specify copy_debug_symbols.

This CL does not fully solve the problem of product bundles not copying
debug symbols but that will be fixed in a later CL since it has to do
with how we are exposing the symbols in the bazel build rule.

BUG: 354039207
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100414
Original-Revision: 4e82bd1b61263f8b7c90fdf65874dab2820bcfa6
GitOrigin-RevId: 9e37002284e55921ae942f0c728fe4d6a3d963a9
Change-Id: I2bd1b02a6e7f47d63f614f7bcf3b89226dc64fb3