[roll] Roll fuchsia [components] Fold ComponentRunner epitaph into OnStop
https://fxrev.dev/1094214 added an OnStopInfo event to ComponentRunner,
which lets a runner report a component's exit code to component manager.
(The exit code is the status returned by the _program_ for example the
value passed to exit() in a C program.)

The channel's epitaph, renamed to the component's "termination status", is
used for the status code reported by the _runner_, as documented in the
ComponentController fidl doc. Right now the termination status and the
exit code are reported in two different ways: one by an epitaph, the
other by an event.

This changes the behavior so that both the exit code and the termination
status are reported in the OnStopInfo event (renamed to OnStop). If
OnStop is used, the channel epitaph is ignored.

In the future, we could update all runners to adopt OnStop and deprecate
the epitaph, although no effort is made to do so in this patch.

Original-Fixed: 358182549
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099940
Original-Revision: 194fca0cb647948b22bdeca820774b5420f8211f
GitOrigin-RevId: e4b2da92256ba63a472383c84fafdb04193ee6b0
Change-Id: I5c0bfa3a8a398c0bd3a19fb52885eb5c97c74788