[roll] Roll fuchsia [dwmac] Implement NetworkDevice protocol
Directly implements the NetworkDevice protocol in the dwmac driver
instead of relying on the netdevice-migration adapter.

This is almost universally an improvement in benchmarks, with the
exception of udp/send having higher CPU usage. However, this higher
CPU usage is corrolated with increases in throughput, so it's not
clear this is the driver's fault.

Original-Bug: 42068759
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089712
Original-Revision: d6d1869682796d41eeae5e673576669c082ef1a6
GitOrigin-RevId: 2ede24ea8236016ae4034a0908b00dc85bf938f8
Change-Id: Id4b84c0a6c55e71d0904c9b23970546cadecaedc