[roll] Roll fuchsia [bazel_sdk] propagate debug symbols to product
This CL will run an aspect to collect all of the FuchsiaDebugSymbolsInfo
providers in a product and expose them as a top level
FuchsiaDebugSymbolsInfo.

The build-id directories are then exposed as an OutputGroupInfo so they
can be copied into the root build-id directory when building in
fuchsia.git.

BUG: 354039207
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101273
Original-Revision: 93da3874924090e00022a2dc5f7a6f7063d682af
GitOrigin-RevId: 2691215de6d34243467ee2db10abe4ceb52b83f7
Change-Id: I6bf4509cb1542bfb7e78f07f0d857424461138d0