[roll] Roll fuchsia [kernel][vm] Correct old refptr upgrade comments
The refptr upgrade use to be able to fail when pages were cleaned out
of a VmCowPages in the destructor. This is no longer the case and these
couple of upgrade sites have a stale comment that could lead to
confusion.

For additional clarity a check in the evictor for a valid cow is
changed to an assert with an updated comment.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099879
Original-Revision: 60850f6d8a2be06898e111a2f68f48d85dc23f63
GitOrigin-RevId: 31c27a84863d2f50bfc751ebb354d520f76bab13
Change-Id: I5fe53a5b039d7ea84107a00a45b46b7b09e5b66c