[roll] Roll fuchsia [kernel][arm64] Align .bss size and fixup code
This fixes two separate but related alignment nits in the kernel
image layout.  The fixup code needs to be properly aligned for
instructions, even though the raw kernel image itself might not
have an aligned size (and it doesn't need to).  The .bss size
needs to be aligned to 16 so that the zeroing code can store 16
bytes at a time without extra logic for an odd tail size.

Original-Fixed: 360404817
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102213
Original-Revision: 8723e340af659c77939cc8d49eb20342f3669f0f
GitOrigin-RevId: e0194ffb42977b1a113f73de7caf3cb85f78c196
Change-Id: I887ac5d948a8cfdf1b61f414d4396c8da00fe7c2