[roll] Roll fuchsia [power] Add benchmark test for Topological Test Daemon
The Power Topology Testing fidl calls are tested for performances using
criterion. An integration test was created to verify the correctness of
the functions acquiring the fidl proxies, and actually toggling lease
states.

The location of this bench is moved from the System Activity Governor
to src/power. The metric names are revised as well.

Original-Bug: b/357445279
Test: fx set terminal.vim3 --with //src/tests/end_to_end/perf:test \
    --with-test //src/power:tests --release
  fx test --e2e power_framework_microbenchmarks -o
  fx test -o power-framework-bench-integration-tests
    --test-filter=*test_topologytestdaemon_toggle -- --repeat 5000
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098517
Original-Revision: 676c648be43d19cb162569390666ba170b171ab0
GitOrigin-RevId: 2fa32481777bd2cd665cf827408800131eb87f6b
Change-Id: I7282fda457e3316fb56e1e2de9129fdb82469c8a