[roll] Roll fuchsia [bt-transport-uart] Fix crash
A crash happened because a receive packet came before HciTransport
connection is established. This was fixed in the test with a
libsync::Completion, but it also observed on a real device now. Fixing
it with a new trick -- delay the start of read until HciTransport is
available.

This change also fixes the flake in unittest which results from a late
establishment of ScoConnection protocol.

Test: Unittest and on device test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097832
Original-Revision: f7da0b025aececfaaf5822aeda23ba8da93913a2
GitOrigin-RevId: e7b44338f45801431c676825285d566f58f58b5d
Change-Id: Id5516552fd129386e18c2b1638bb638dd9a95010