[roll] Roll fuchsia [starnix][input] Use unique Fuchsia VID and PIDs
It turns out that implausible USB VID or PID prevent correct
device configuration in Android. Fix that by adopting
plausible-seeming identifiers.

The system is already configured to use the correct files for
keyboard. For touch, this is not the case; but the current state
of touch configuration is broken anyways, so I don't expect
things to change after this change goes in. It does give us
a chance to right the config on Android side, though.

Original-Bug: 353765537
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102435
Original-Revision: ca237edefe8fde6e3bb7216479628db80eb5893a
GitOrigin-RevId: dd212b2ca1a17c3bf99537e6f1b1b1e2b109185f
Change-Id: Ia02e011360ecf23a7b5c4f3165a5ea2a89c58a08