[roll] Roll fuchsia [clidoc] Making subcommand anchors unique
This makes the anchor for subcommands to be built from
the full command line. This avoids the problem of having and
anchor like `#start` for multiple subcommands, so the URL
only can address the first command with a subcommand of `start`.

Original-Bug: b/360893569
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104313
Original-Revision: e94776d95fdcb94618d1e67667917aafeb2e0802
GitOrigin-RevId: 4c65067d0c414252195f866f29247e98e48c75a0
Change-Id: If5ec259736ee0f4fc2a9e35e79e38044a7bbf731