[roll] Roll fuchsia [kernel][sched] Move location of previous_thread_
There is a member variable used during context switching by the
scheduler called "previous_thread_".  The (new) current thread needs
to know what the previous thread was in order to hand off locks
properly, but has no way of immediately knowing from its (now
restored) context.

So, prior to this change, the SchedulerState structure held the
previous thread member (and there is one instance of SchedulerState
per thread in the system).  So, the scheduler would record the
previous thread in this member variable of what is about to be the
current thread, and after the context switch, the new current thread
uses the information then zeros it out again.

Moving forward, I'm in the process of trying to be a bit more formal
about expressing the locking requirements for various
scheduler-related shared variables.  Members of the SchedulerState
structure in the post GTL world are actually protected by a more
complex set of locks and protocols because of requirements for being
able to coherently observe them without needing to hold all of the
locks which protect the variables.

From a locking requirement perspective, the "previous_thread" member
is _not_ one of these variables.  It is used in one *very* specific
place, and if one had to define an owner for it, it would probably be
either the percpu data structure for the current CPU, or the
scheduler for the current CPU.  It does not ever need to be accessed
concurrently, and requires no locking protection.

So:  Move the `previous_thread_` out of the `SchedulerState` instance
stored in the `Thread`, and put it into the high level `percpu` data
structure instead.  Later on, when the `scheduler_state` member of a
`Thread` has been properly annotated, we will not be subjecting this
field to more locking-protocol requirements than are actually needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099994
Original-Revision: 70e634894e138b1a9682fd6881fd0760bd8c7902
GitOrigin-RevId: e6a4fac027d9e638718ade1b0089078f8c18048f
Change-Id: I18bf0c30f0e05b80c8298a67c078f0c69c7022a1