[roll] Roll fuchsia [time] Allow force-starting from RTC
Force-starting the clock was so far done only when RTC is not present.

This change does the same if RTC is present, but its reading is known
bad. In which case, we start UTC from backstop, and leave the faulty
state in RTC until the next regular time update.  If the device is not
network-capable or network-connected, that RTC update may never come.

In normal operation this behavior is relevant only when RTC is
unpowered by a battery, or pristine. Both are very unlikely scenarios in
general. We do deal with unpowered RTCs in some devices today, but this
is supposed to be a temporary state of affairs.

Original-Bug: 355293758
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100898
Original-Revision: dbf7dd096cc20c9ece0207b4d2b265ece8edd023
GitOrigin-RevId: ceaa0237d54914ef1621afca8016859ac95925fb
Change-Id: I8ac619d0acb3128330dfa5f437488d359c0b440c