[roll] Roll fuchsia [dl] Rename ModuleHandle & LoadModule
This CL renames ModuleHandle to RuntimeModule and LoadModule to
SessionModule.

The RuntimeModule is ultimately managed by the RuntimeDynamicLinker, as
the SessionModule will only live as long as a single LinkingSession. The
LinkingSession data structure will be introduced in a forthcoming CL and
the SessionModule will become private to that class.

Original-Bug: 341473244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104316
Original-Revision: cae00f9610fade6f8b2e52d53ffd4fafd90052f9
GitOrigin-RevId: 5dd88912c82c4cb458d2cc88cb2ac5a54c706ca9
Change-Id: Idf370f7fa1ef0495d2c3583d7cfe45bf9ccfac4e