[roll] Roll fuchsia [assembly] Specifying 'update_epoch' without 'update_version_file' is an error
Previously, this silently did nothing, but also didn't create an update
package.  Now, providing a value for 'update_epoch', but not a value
for 'update_version_file' is an error.

The default value for 'update_epoch' is still '1', but it's moved from
being a default of the rule, to a default of how the rule invokes
the creation of the update package.

This corrects in-tree usages to not cause errors on this corner case.

Downstream repos may experience new errors with the following message:
"'update_version_file' must be supplied in order to build an update package."

Original-Bug: b/361328909
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105314
Original-Revision: 02d658e33fa2333197490cb948f813c0e59c2b46
GitOrigin-RevId: 6d0b05c0fa54d184b4366d59aff492324be4d23a
Change-Id: Ia36987957ca06387bce457c30de87c78dccb9b87