[roll] Roll fuchsia [kernel][top] fix bug in error case when allocating idle threads
An error in the logic when handling an error case when allocating idle
threads would cause it to go ahead and mark all of the idle threads as
created, even if it stopped early. It's not really a condition that is
good to have and survive properly, but is still more graceful than
starting additional secondary cpus that have no idle thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100505
Original-Revision: c4f1499cdef81f7c5f39c94ca2e6b17243bb9b19
GitOrigin-RevId: 1afab4817c5c1bb7d5f8c92e6e75b399c624e5fd
Change-Id: I63a3d673c6ed48e1056ee930288228fc75554de6