[roll] Roll fuchsia [ffx] Support direct connections with --core flag.
This updates the ffx behavior for handling connections and providing
FIDL protocols for developers. When passing the `--core` flag to a
command, non-daemon behavior will be put to use where possible.

For this initial pass, only `RemoteControlProxy` will be provided without
the daemon when the `--core` flag is given.

This defines a device connector trait that is implemented by a default
(ssh-only at the moment) connector that takes a target spec and resolves
it to an ssh-able address.

After completion of a command, if errors are encountered AND the
`--core` flag has been passed, the underlying ssh connection will be
inspected for errors, and they will be reported to the user to aid
in debugging.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097107
Original-Revision: 755d34896675516c06d93882f3c7e460699670f1
GitOrigin-RevId: a51a99a2eb7dbb95efa2d7a217e7d5a0f474ab26
Change-Id: I98903d877765794fdfb6907d59952ce44ee97e54