[roll] Roll fuchsia [kernel][vm] Correct partial success case in commit
It is possible for the loop in CommitRangeInternal to successfully
commit all the pages in the VmCowPages, i.e. have a successful
CommitRangeLocked call, but fail to process the whole range due to
ReplacePagesWithNonLoanedLocked failing to allocate a page. In this
case the assert at the bottom of the loop would execute and fail, since
it's checking commit_status, which was ZX_OK, but we failed to process
the whole range.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106553
Original-Revision: 4635f3e5d195c4ce1f0333fa0d330076c4a19a1f
GitOrigin-RevId: 95c2440cc69142d21a1772e0f49e7dba5f3674f5
Change-Id: I22850caa253370f5f7fa899038fbb2ed98438549