[roll] Roll fuchsia [ffx][config] Fixing build-dir level config generation
This fixes the root cause of the rollback of using the
build generated config file.

The target to generate the file was moved to be a dependency
of a target that is built by default when building the
default configuration from `fx set product.board`.

There also was a double substitution going on when the
variable had a relative build directory.

Original-Bug: b/360197942
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105703
Original-Revision: ff8a547a9674c56347065bc57d33be20c631f69f
GitOrigin-RevId: dc30db3a02f2880273830e5a9064f899759d384b
Change-Id: Ie737c414e24984b32fa008b039d66db23a738772