[roll] Roll fuchsia [fuchsia-lockfile] Fix flake with unfair locking
fuchsia-lockfile doesn't guarantee fair locks, so it's possible for one
or more threads to not grab the lock. This changes the library to expose
if the lock attempt timed out, and to change the test to loop again if
it times out a few times to avoid flakes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103612
Original-Revision: e4572c417d7fbd6058541bac9470643529d2a9bd
GitOrigin-RevId: 3d3fc86b45b88843350f9216441230119fddc4b0
Change-Id: I450f8475a230c2b9d8eff73be2a58d22a25d24f3