[roll] Roll fuchsia [starnix] Centralize file system creation
Previously, a different set of file systems were available on the
CurrentTask versus the Kernel object. Now, we create all the file
systems via the CurrentTask.

However, when we boot the kernel, we need to create the root file system
without a CurrentTask because we need the root file system in order to
create the initial task. This CL restricts the set of file systems that
can be used for the root mount to remote_bundle, remotefs, and tmpfs,
all of which can be created without a task.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105168
Original-Revision: e1f7f0fdbbd62e917c0b6317321e74d4a4bd4009
GitOrigin-RevId: 6e4358ecd5e06fd295ff6a0f0e3722fb8425bcb7
Change-Id: I8a509f5b45cbe07cdf67e027c5494652eb060ca5