[roll] Roll fuchsia [test-manager] Add pkg resolver protocol
This change routes the pkg resolver protocol to the test manager, so
that it can provide a hermetic package resolver implementation for
tests. This will be used by the driver test realm to implement support
for subpackaged drivers, as that requires it to use the package resolver
to query for subpackages.

Original-Bug: 330189874
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100698
Original-Revision: 3cf887fd542c2396b13a5541e0fab11757f071e0
GitOrigin-RevId: b601485f4efc201cd58aaa7eb26ba2c449946f42
Change-Id: Ie827de93a2488ecd67223b60ce6ea454ac2b0b19