[roll] Roll fuchsia [bazel_sdk] add rule to ensure cml files are compiled
Adds a rule which is used in the fuchsia_component macros which makes
sure we are always passing in a compiled manifest when building the
component.

Previously, we were checking the string that was passed in but this is
not always correct and is fragile. This change makes sure that the end
result is a compiled manifest.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105166
Original-Revision: ea91ca1f1418b3cd8288e0fc5d220bca1f789ed3
GitOrigin-RevId: 8d8403a5216c981329e8468525b6b8d2e120020c
Change-Id: If25368d528b7c66107fabb0da488e2da871d6610