[roll] Roll fuchsia [dtr] Provide resolved component to startargs
Provide the resolved test component information in the driver test realm
start args so that it can be used to discover pkg drivers. With this
change we can support loading drivers from the test package directly as
well as its subpackages. This also removes the need to manually provide
a pkg handle, because the resolved component already contains that as
well.

There are a few things that need to be improved still:
 - Should have a CF introspection for getting own resolved component

Tests: basic-driver-test
Original-Bug: 330189874

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075272
Original-Revision: 817b7a6fc7cb06aa6c01af496c1631ef1051a5d5
GitOrigin-RevId: c9efaf65217686d64cc31bb00507ce234dd48df3
Change-Id: I69b1b95fe1d4da6391a185c29d5a786531d322fd