[roll] Roll fuchsia [blackout] destroy test component after test for suspend/resume
The continuous load generation is causing timing issues when it is
generating load while the storage driver is suspended. To prevent this,
destroy the test component right before we suspend, so the load
generation is stopped.

This somewhat degrades the usefulness of the test, since the load
generation won't go right up to the point of the suspend and get
interrupted like it will for the reboot scenarios, but consistency in
this case is certainly something we still want so it's not completely
useless.

Original-Bug: b/355511475
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106873
Original-Revision: 93a3324af076a4e87be8f726a332471919e10cf7
GitOrigin-RevId: e69ca11751bd8ad6544f25c67ad867cbbaed98c2
Change-Id: I459f4f430d9bc31e6617ef4990989f292764b7b4