[roll] Roll fuchsia [netstack3] Refactor IP Forwarding
Move the IPv4/IPv6 forwarding logic into a helper function. This will
enable a follow up CL to call the function when performing multicast
forwarding.

Regrettably, forwarding requires both the packet, and the
underlying buffer from which the packet is parsed. Moving the buffer
(such as by passing it as an argument to a function) invalidates the
reference to it in the packet. Thus, the approach in this CL is to
split forwarding up into two steps:
* Determine the forwarding action (using the parsed packet), and
* Forward the packet (used the underlying buffer).

While here, I opted to write my helper functions in an IP Generic
context, rather than the original separate IPv4/IPv6 implementations.

Original-Bug: 353329136

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106255
Original-Revision: aff25c8e249063f238bb3d40cf4af0ded0d3a1a5
GitOrigin-RevId: b1f8bfa6cb97c924e5903266b9c34a85c0ff15dd
Change-Id: I67a0506ef2922524e6350bd0408093194bdc7173