[roll] Roll fuchsia [kernel][phys] Fix 'false hand-off memory' bug (once and for all)
It got complicated and this was an oversight in the previous change that
attempted to ensure that only the one range (backed by the one true
data zbi) actually made it across of type kDataZbi. This did not cover
the case where the orignal the data ZBI - not actually passed to
TrampolineBoot - was being relocated due to an intersection with
the fixed address load image: the part outside the overlap was not freed
(and this was justified due to faulty logic). In this change we update
the freeing logic to avoid the overlap.

We will soon have a regression stop in place in the form of the PMM at
the end of hand-off asserting that all "phys VMO"-type memory was
ended up unwired. These false data zbi regions would be 'unowned' by a
PhysVmo and result in this assertion firing (which is actually how I
discovered that this bug remained).

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105772
Original-Revision: 70bc10d192e4b9637dda6a48f069ab5a97b4295b
GitOrigin-RevId: 89f52f4b6427cbf23b6bcff7807d0c5c75aaa591
Change-Id: I857223a06c77da90cacace1739474f4b36716a81