[roll] Roll fuchsia [bt][hfp] Implement CodecAudioControl
Forward requests to start and stop audio to the event stream, and
respond when the ScoConnection is setup, as well as tear down the
ScoConnection when it is finished.

Update to indicate that the responders from CodecRequest are Send.

Original-Bug: 341114499
Test: added unit tests, fx test bt-hfp-audio-gateway-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1057872
Original-Revision: 292c34731bcf47fb6b578b185e35870f1b8a71ff
GitOrigin-RevId: 7d7d6bba06c0e00d1ec89895db45f720642a60e8
Change-Id: I98292615cec5688085586e6a09f675cd77de041e