[roll] Roll fuchsia [network/test] Test SO_REUSEADDR listener conflicts
A bug in this logic in Netstack3 was fixed in
Ide63b3767e2e256222d8b63e6128cbffabdaa797, but it is also useful to have
a test that runs on Linux as well to validate that we match its behavior
for this case. Netstack2 is currently expected to fail a subset of these
tests.

Original-Bug: 361402347
Test: netstack3_streamsocket_test.cm
      netstack2_syncudp_streamsocket_test.cm
      host_x64/netstack_streamsocket_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108584
Original-Revision: a745deafa1204d5d5d6e1e4ca25c3c40e00a083a
GitOrigin-RevId: 1283a43f1b0cc403d6630c3e8d97e38d1d603af0
Change-Id: I746ce96ecf6a3240057a5ae892b5a9ecf4f27b59