[roll] Roll fuchsia [geni][uart] Fix RX and RFR (ready-to-receive) watermark config.
Previously RX and RFR registers were configured with values
that were incorrect (RX watermark was written to the RFR register)
and inconsistent with the bootloader configuration.

This caused a failure where the serial engine hung up and lost the
last 4 bytes received when the host wrote multiple bytes continuously
to the UART pins (for example, when the host pastes texts to the tty).

This change removes the RX and RFR config statements so that they will
not conflict with the pre-existing configs. The bootloader is fully
responsible for setting them prior to Fuchsia boot.

Test: geni-tests
Test: See bug.
Original-Bug: 361345380
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108146
Original-Revision: d15b027ec85f6c2834d1af7e12db4c7926075085
GitOrigin-RevId: 913389c964651bc072837d7597a0e5ea6a3e9b52
Change-Id: I059a6537e1ee352c7c5ec6670625fc92a1d38396