[roll] Roll fuchsia [recovery] Remove unused recovery images
This removes all recovery images other than the recovery-fdr
image, which is the only one which has active tests in place.

This is the first of several CLs that are cleaning up this
file, they are being done in a stack to make the changes
more clear.

Run-All-Tests: true

Original-Bug: b/362336812
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107976
Original-Revision: d488d090fcba055bcdd0b4046ad8fd837ee70659
GitOrigin-RevId: 320f001d40d1e248f58341e30211676166017407
Change-Id: Ife7000935413166a4afafafd42c7ad37b2c1a78b