[roll] Roll fuchsia [tools][refresh-bazel-compdb] Fix directory paths for in-tree bazel packages.
In-tree bazel packages may start with "internal_sdk" or start with  a
Fuchsia tree path. We need to rewrite the path and treat them as local
files for all such cases.

Previously we used a heuristic that only covered "//sdk". This change
adds "//src", "//zircon" and "//vendor" to that list. These covers
most of the cases for in-tree development. This is not an exhaustive
list and in the futurewe should look for a better solution to
distinguish all these in-tree directories.

Test: fx refresh-bazel-compdb --dir //src/devices/board/drivers/vim3-devicetree
All <lib/ddk/..> directories (located in //src/lib/ddk/... are parsed
correctly in compile_commands.json. clangd didn't report any error
when opening C++ sources in //src/devices/board/drivers/
vim3-devicetree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108940
Original-Revision: 6b36a429733b1a42e8a69541b55f43e409b4575a
GitOrigin-RevId: 557322f0ae4b55eff55f0b6e6138bbfe13b1843f
Change-Id: I2c68dee180778edbacf85c7fc7cbd4187f1b7921