[roll] Roll fuchsia [tee] Export xtest expectations from fuchsia.git
This change introduces //src/tee/tests/xtest-expectations.json5 to
define the current expectations of pass/failure of xtest test cases. The
intention is that this will be consumed by the xtest test suite
definition in the appropriate vendor repo, allowing us to enable test
cases in the same change in which we implement the associated TEE
runtime features (sparing us a bunch of cross-repo roundtrips in the
process).

Original-Bug: 348474435
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108935
Original-Revision: eb3275557a3754872a3f0d01582185a15fbfbe0a
GitOrigin-RevId: add0186517a0fd09e0efec601fb4fe81ffa79b68
Change-Id: I5482b4be99bfa9651ead42e5f33b06e90ecea7b4