[roll] Roll fuchsia [netstack3] Make tcp::SendBuffer's payload generic
Now that we have GAT we can put tcp::SendBuffer's payload type behind a
GAT and reduce the reliance of non test code on the concrete
`SendPayload` type, which is only useful in the context of ring buffers.

For now, all the buffers use `SendPayload` as their impl. A follow-up CL
will replace `SendPayload` with an array of slices.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108832
Original-Revision: 73515f8dc6d63b2b229fa47bc6333bd67579a5dd
GitOrigin-RevId: 05fb72bc3286629470178966e056a953299d317c
Change-Id: Iea701eb97c921e7c098026477f019f0aca2e34ea