[roll] Roll fuchsia [build][cpp] Disable -Walloc-size-larger-than
This is not a particularly useful error: we don't specify a particular
value for this warning, and so it defaults to PTRDIFF_MAX, and it seems
unlikely that we'd be in a situation where we're making statically-known
allocations that could exceed this value. Moreover, it's buggy.

Original-Bug: 362618077
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103674
Original-Revision: e35d67fbf9f3ca0115908897fc38908a2ef43594
GitOrigin-RevId: ab519ad79708d64eb60d73ef3f943a7a8f2ab089
Change-Id: Iffba6b1e0cbb3fb0265db0fd69f5035804c210e5