[roll] Roll fuchsia [starnix][modules] Move selinuxfs into a module
This code does not have a circular dependency with starnix_core and can
be moved into a module. Also, most of our file systems are now
implemented as modules.

Because "selinuxfs" interacts with the SELinux LSM, which is not yet
modularized, some APIs are added to the "security" hooks solely for use
by this filesystem.

Original-Bug: 362917997, 335397745
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109332
Original-Revision: 77fc88739423022787afab4c16ebb293ce8506e7
GitOrigin-RevId: 38e6705bbcc5e01915dc9f1b74f6ea758a8a379e
Change-Id: Idc49bce7783b0b9ca6774c0896bb6cbd8f85a7b4