[roll] Roll fuchsia [abi-compat] Make @discoverable contradictions non-fatal by default
If a request in a @discoverable(client="external") protocol A contains a
client_end for a @discoverable(server="platform") protocol B then that
sort of implies that there's an external server for the that protocol.
But in practice, in a few places we pass through protocols whose servers
are in the platform through external components.

This makes the this situation a warning unless you pass
--fatal-contradictions to abi-compat.

Original-Bug: 362970530
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106813
Original-Revision: e8af3da88dae1349642e2dbb369f279508672da8
GitOrigin-RevId: db5cbd2f27dd2798d69863c492e5930dc76fa995
Change-Id: I6936c4bf7d85980f7ce8c8cd8e175d38def322ce