[roll] Roll fuchsia [bt][avrcp] GetItemAttributes is supported for Target role
Currently, when the remote CT sends GetItemAttributes command, we
respond with an error since it's not implemented on TG side.

According to the AVRCP v1.6.2 section 6.10.4.3, When the TG supports
GetItemAttributes, the CT shall use GetItemAttributes and not
GetElementAttribute. Since our AVRCP implementation supports the
browsing, if the browse channel is successfully established, we would
expect the remote CT to send the GetItemAttributes command instead of
GetElementAttributes. Therefore, we should handle the command
appropriately so that the media attributes are returned.

Original-Bug: b/306416869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940273
Original-Revision: 349b38fe78140178fef1cd86ffd91e23728bc4cd
GitOrigin-RevId: 3fa5694f5d7e5ff2c18d0d058e0ce1e01530e037
Change-Id: I654ef695fae764d3a816a9fad1334a71bc1839e7