[roll] Roll fuchsia [windowed-stats] Correct Simple8bRleRingBuffer num blocks output
PackedU4VecDeque, which is used to hold the selectors of
Simple8bRleRingBuffer, doesn't decrement the len when the oldest
selector is requested to be evicted. This commit fixes this by
decrementing the len.

Also, the unit test case for rotating out the oldest value doesn't have
the #[test] tag on it. Here we add the test and also fix the test.

Original-Bug: 337086207
Test: Fixed unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107919
Original-Revision: 28e4e24ed03ff20e30358616bf49a4429f4989cd
GitOrigin-RevId: 324cad9e26fb23983fcbeb3ce280ed69ac8422d6
Change-Id: Ib72d66fb29e8391eb1b4a0f528e360002f5e65a7