[roll] Roll fuchsia [kernel][phys] Realize PhysVmo machinery
This change makes good on long-standing plans to have PhysVmo represent
physical memory as-is to be converted directly into a VMO (rather than
with temporary contents to be copied out into a new VMO). This was
previously blocked on being able to convey these regions to the kernel
as regions to wire and avoid when placing PMM arena bookkeeping.

PhysVMO memory in the hand-off is simply identified as being one of
several special types, abstracted behind PhysHandoff::IsPhysVmoType().
As guardrails against this memory not ending up unwired at the end of
hand-off (especially if it does not have an associated description
elsewhere in the hand-off on how to convert it into a VMO), we have the
PMM put this memory into a dedicated page list and then assert that it
is empty in PmmNode::EndHandoff().

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104559
Original-Revision: 14bec424c27601d61d7d3ed22ee4fa420e564c97
GitOrigin-RevId: 5a83545bb16f2b18584ba011fe00bf61acdd273b
Change-Id: I0580d39229a084bfc4c44fe3de74812b8f2567b5