[roll] Roll fuchsia [wlan][policy] Snapshot on successful recovery.
wlancfg will intervene when it detects that WLAN is not
behaving as expected.  When this intervention is successful,
upload a snapshot to try to capture the state that the
system was in that led to WLAN being non-functional.

Original-Bug: 355290936
Test: fx test wlancfg-tests
Test: fx test triage
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109979
Original-Revision: 126918b6c0631c3f86a8b78ceb2f5065c060d6ba
GitOrigin-RevId: 516c8cecbf0cb0ec187f607359f04c78a3d0424f
Change-Id: I8bc983f1e668f56307c80af423aae70bf7ea8e40