[roll] Roll fuchsia [dl] Store dep_names_ in SessionModule
This CL will store the decoded Sonames corresponding to DT_NEEDED
offsets when they are decoded from the loaded ELF file. This set of
names will later be used to look up the modules needed to include in a
resolution list.

Since dep_names_ is stored in the class and scoping is not a factor for
as it was for the local variable, the function returns early if the
reserve operation fails; this makes the code a little more readable.

Original-Bug: 338232267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110341
Original-Revision: 21f6677f2600134dd041d4e4a41d8c0c3f66cec4
GitOrigin-RevId: 0d02c01233de5c54e23bc9a58d700852219e818c
Change-Id: I6e6f83860de518dc4b9215e26eb0c318fd07ffa4