[roll] Roll fuchsia [kernel][vm] Correct page queues reclaim check
Corrects a debug method for testing if a page is in a reclaimable queue
to not be limited to evictable (i.e. pager backed) VMOs. This
restriction is from a time when the reclaimable queues were actually
called the pager backed queues, but when that was changed this method
was not updated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110233
Original-Revision: f8cd25e9a7365d2f34ccf3a4407145113334c297
GitOrigin-RevId: 88551c13cd1d9789d81ad564545cc0f23f5b43a5
Change-Id: I098769f1aacb413e2f108a5668aa8cd4d2fa2a80