[roll] Roll fuchsia [kernel][vm] Refactor loan cancellation
Refactors how the the PhysicalPageProvider performs loan cancellation
and tracks down loaned pages. As the PhysicalPageProvider is now the
only caller of GetCowWithReplaceablePage this method is re-designed
into GetCowForLoanedPage with a narrower contract. This narrower
contract allows for combining the stack ownership wait that use to be
in that method and the PhysicalPageProvider into a single wait in the
PhysicalPageProvider, and allows for the functionality of
GetCowForLoanedPage to be logically simpler.

Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107993
Original-Revision: 2d8bf23cf5535919e08f7c6f3b8413bb659722f8
GitOrigin-RevId: ef32d696b290fb70400bf01763e40a3a27103048
Change-Id: I4b673ca133fbb0e04bc9e021e56e306a38d121f0