[roll] Roll fuchsia [bazel_sdk] remove dep on fuchsia_sdk_common
The fuchsia_sdk_common repository was needed for us to share the
sanitizer features but this requires external repos to call the
bazel_rules_deps macro in their workspace files. This required a load of
the fuchsia rules which would trigger a download. In order to move away
from requiring loads in WORKSPACE files we needed to break this
dependency.

Original-BUG: b/346416385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112315
Original-Revision: ba06992040226eb7605bdf80d9f35582a08355b5
GitOrigin-RevId: d2a15408ddcc7b50790524014f0de68376f9cae1
Change-Id: I0ee47325575f539ff036524245f7248bdd81850e