[roll] Roll fuchsia [components] Clarify process launcher error for dynamic linker timeouts
When a new process's job is killed while the process is being launched,
it's possible this causes the attempt to retrieve the dynamic linker to
timeout. In this situation, the error is expected, so use an INFO log
instead of ERROR, and make the error message more clear.

This is a sequel to https://fxrev.dev/1090961. If we discover more
errors that are like this we can follow the same pattern.

Original-Fixed: 358166703
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113312
Original-Revision: fc693ba074dc389b88dfe2329c44949ab7d04b59
GitOrigin-RevId: 279ce6fe81e4eabf59214b1a128b9e430bd40a48
Change-Id: Icfeb3b62b99e6e2d487cfbb20e9d64c6d0b14dd3