[roll] Roll fuchsia [zircon] Remove the flash stress component
The flash stress component was dependent on FVM which has been
superseded by Fxfs as the default storage architecture, and
as far as I know, we're not actively using it. If and when we
need something to stress test the flash, we can ressurect this
or develop something new.

Original-Bug: 357467643
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110034
Original-Revision: f59f9febcbf066e7b0f2f0e1ab9354a3afd695fc
GitOrigin-RevId: 84046647511bdb3740994fa1501bf28c81e6805e
Change-Id: I45a3afbfdd7af1796dfaec43aa0bb5f10193dc15