[roll] Roll fuchsia [netstack3] IpDeviceAddr witnesses "non multicast"
Ensure instances of IpDeviceAddr are not multicast addresses. This
allows us to remove a check for multicast in source address selection.

To facilitate this, introduce a new Witness type to the net-types lib:
`NonMulticastAddr`. Note that `UnicastAddr` was not suitable because the
semantics of "unicastness" are complicated for IPv4.

This also increases the "depth" of the nested witnesses held by
`IpDeviceAddr` to 3. Thus this CL adds support to `impl_nested_witness`
in the net-types lib for this additional witness layer.

Original-Fixed: 1112292

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112092
Original-Revision: 5f266b291cdf0b7a926fcfc37c7056e8b529e23f
GitOrigin-RevId: 54fe5735e8f53c7ced84c8864bebae4ccb58e292
Change-Id: I57a302c62b36ef8507ca0aaf0dcd7e0bc003e414