[roll] Roll fuchsia [power][trace] Power element trace events
Added trace to lease request and level trasnsitions in power broker.

On the client side, added trace to the Daemon so that in the future
the overhead from client calling CurrentLevel::Update to the power
broker starting to handle the fidl can be computed.

Trace is enabled in the src/power/bench integration test and routed
to the daemon and the components created in the test.

Original-Bug: 322026959
Test: CQ and manual test
  fx test power-broker-unittests \
    --test-filter=*test_initialize_current_and_broker_status
  fx test -o power-framework-bench-integration-tests \
    --test-filter=*test_topologytestdaemon_toggle -- --repeat 100000
  while the integration test runs, collect trace (a few seconds) with
  ffx trace start --buffer-size 200 \
    --categories '#default,kernel:ipc,power-broker'

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090133
Original-Revision: ec027b194bb38940afc242b43d1890577ca39c90
GitOrigin-RevId: cdf18954f49f970588bec54550f8da70fc63de58
Change-Id: I126e040f2effe41a348f0b86193ea04b0ddc6ce0