[roll] Roll fuchsia [sshutil] Log timed out message for context deadline exceeded.
A time out trying to connect to SSH can either show up as a net.Error or
as a context deadline, but we were only logging the TimedOutConnecting
message for the net error, so the context deadlines were not getting
caught by a tefmocheck and thus showing up as the generic task_failure.
This should group these failures with the ssh connection timeout
failures.

Original-Bug: 356029657
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114120
Original-Revision: 7ed49f225dc0a94418fcb4175edb66cf13bab986
GitOrigin-RevId: 070fdc4f3f4dafdc921f05a2158c2b37bf923961
Change-Id: I22ea9f356fd2a1e587d3a05faff88bc49837d3ac