[roll] Roll fuchsia [starnix] Add wake leases to runner
This change updates the starnix runner to acquire wake leases for all
the containers it runs. These are dropped when the container is
suspended, and reacquired before the container is resumed.

The change is hidden behind a feature that is currently disabled.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113914
Original-Revision: 010d28036fcefc6e5ca0ab6f081ec4bec42adf38
GitOrigin-RevId: e790c4880cd900e4772c015299c084269486b599
Change-Id: I872a47cb89becea196e41b429ea53fd4635e1a91