[roll] Roll fuchsia [paver] expose boot attempts
Expose a new paver FIDL API that returns the slot status as well as the
number of times it has attempted boot.

Eventually we may want to deprecate the API that only returns slot
status and use this everywhere, but that can be a follow-on effort if we
decide it's worth the time.

Original-Bug: b/349617985
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110009
Original-Revision: adc6e781221280fb57974451d6c4edbc882a712b
GitOrigin-RevId: 960bf211b1f80891eee7fae330b64238c9009af5
Change-Id: If6378e1860ca6ca779f654cc12f48eed92b728ce