[roll] Roll fuchsia [vsock] Remove SendVmo functionality
This removes the ability to sent entire VMOs through a virtio socket in
the both the service and driver layer. This functionality is currently
unused and untested - the data path used by sockets is a Zircon socket.

In an earlier version of the system this data path was used to more
efficiently send large amounts of data from a file into a socket. If and
when this use case arises again we can either revive this logic or find
a new way to optimize the stack as many other parts of the system behave
differently from when this analysis was last performed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114036
Original-Revision: 05d3eeea342ab3907def655dd5a36c2bdc5fb4ae
GitOrigin-RevId: b841299028a9a66a44313e480d4b2d8040e7f51f
Change-Id: Ia1af470b6fbeb25a1c9024559d6c4219dee9bdd5