[roll] Roll fuchsia [bazel] define sysroot filegroups in fuchsia_sdk
This CL defines the sysroot filegroups in the fuchsias_sdk repository
instead of the fuchsia_clang repository. This allows us to remove the
symlink from the fuchsia_clang repository.

If we keep a symlink in the fuchsia_clang repository then that symlink
will materialize when we ship the fuchsia_clang repository. This would
put the sysroot in the clang repository where it does not belong.

Original-BUG: b/346416385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113545
Original-Revision: 40f6f4b42c060687a1ae932be434bbf87fe68b3c
GitOrigin-RevId: f4c499ac94c1131aafdb6811fd52aa2da7099b81
Change-Id: Ie626812231fd5e85276ff29265ab5ced60bda1b5