[roll] Roll fuchsia [cobalt][fidl] Delete removed FIDL elements
This FIDL library is not in the SDK, so it only needs to support HEAD.
This also correctly marking the entire library `@available(added=HEAD)`.

`@available(added=7)` was added in https://fxrev.dev/664072 and not
removed when the category was changed in https://fxrev.dev/792622.

Original-Fixed: 315496857
Original-Bug: 364628320
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114041
Original-Revision: 44d76f1cbfd701f60359ecb23ac28b0b38af25fc
GitOrigin-RevId: 1fe444bae48746c2623381f2e10cc234788d91ae
Change-Id: I2fbd7b5054821cd912833c83c1b8bdfb148c5635