[roll] Roll fuchsia [recovery] Add recovery to Assembly
This allows Assembly to be able to include the recovery-fdr components
in the assembled image.

As part of that, it removes a number of unused GN args that changed
the contents of the recovery packages.  The next step in this process
is to move configuration to a domain config package and config-
capabilities.

Original-Bug: b/362365879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108586
Original-Revision: cc0f9ab30ea7d05378a6b57d7a1eb706623f78f0
GitOrigin-RevId: abf714aa01911929adfea56f4de747a80f4351b0
Change-Id: Ie373293347aeb629cd9e1dcd859a785d5e768cc6