[roll] Roll fuchsia [debugger] Treat non-symbol index servers as public
With the introduction of DEBUGINFOD_URLS support in symbolizer we should
stop assuming that servers coming from anywhere other than the symbol
index have authentication.

If we have a need to specify servers that require authentication on the
command line then we can add a new argument for those. Debuginfod was
never intended to be used with non-public servers and so these will
always be assumed to be public.

This also fixes a crash in zxdb when trying to initialize a
std::filesystem::path with a null pointer (specifically when "HOME" is
not set in the environment).

Original-Bug: 365193885
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115593
Original-Revision: ed70ec298adb75c778e3c6f7952e46ef4d20e5e1
GitOrigin-RevId: 85c61e10a900b9bd42f11d93574d8655181e8af5
Change-Id: I82de594a1f7717c222e50cd466058eee13b00384