[roll] Roll fuchsia [honeydew] Add Tracing.is_active()
Exposes the tracing module's notion of whether there is an ongoing
trace. Some users of this module are doing their own bookkeeping of this
state, and all that does is create the opportunity for inconsistencies
to arise and create bugs.

Original-Bug: None
Test: fx test //src/testing/end_to_end/honeydew/tests/unit_tests --host --output
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110526
Original-Revision: eb31146cf8fb2751ef06d0dac91447fc37dec13f
GitOrigin-RevId: a27c3c802995f7cabe4e796e410d87d6e376bd6b
Change-Id: I6c8a55fc7c84d5d25c4f66df018c8210da33a98c