[roll] Roll fuchsia [tools][artifactory] Only upload artifacts related to the last build.
This is a reland of [1], after the fix in [2], for a bug that justified
the revert.

Ensure that `artifactory` only tries to upload Ninja artifacts that
correspond to the last successful build. This avois uploading stale
artifacts from previous incremental builds, or artifactory
complaining that it does not find breakpad symbol files matching
(stale) debug files in the build directory.

- Add new `lastBuildOnly` argument to NewBuildAPIClient()
  function is added, which creates a BuildAPIClient
  instance to return filtered results.

- Add `NewModulesForLastBuildOnly()` function to return
  a `Modules` instance corresponding to these filtered
  modules.

- Use `NewModulesForLastBuildOnly()` in the artifactory
  `up` command implementation.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1113733
[2] https://fuchsia-review.googlesource.com/c/fuchsia/+/1114412

Original-Bug: 42175326
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114414
Original-Revision: d60763ae01655d3f37c94a4bfc771e7c735ec781
GitOrigin-RevId: 9cfacc596b62d5323fb8b55bef58afa2341601be
Change-Id: I78d6ca8870d76abae9c5dd9a3733b1ec97847069