[roll] Roll fuchsia [ffx] Remove redundant ssh stderr read.
If there's an error setting up SSH and running a compatibility check,
then there is no need to read the stderr a second time in the midst of
encountering an error. This will remove the original failure message,
causing the error reported to the user to be an empty string, which is
reported as "unknown" and then tells the user to report an issue, while
hiding the original error.

Original-Bug: 362976369
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116852
Original-Revision: 7fe537129ca1a3486aabbab34b70cc59e47c54c0
GitOrigin-RevId: 60cfb617a5019edb1e776107b8994bebbacda671
Change-Id: Ie1de3e0a6437ced43a6a2623fe683c185bf3e66f