[roll] Roll fuchsia [assembly] default assembled_system() to use eng platform AIBs
All in-tree uses of assembled_system() are for eng build-type
assemblies, so make that the default to reduce the common user
error of not setting this option correctly.

Bazel assemblies are left as-is, because they can't accidentally
use the eng build-type AIBs when not setting the flag correctly,
as their action sandboxes won't include the eng build-type AIBs
at all.

Original-Bug: b/365667289
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116873
Original-Revision: adf4da89739c948f122e151e8f36ce01691cf46e
GitOrigin-RevId: af1124e80d8b8a756316ceaf6e2d537f21f965ad
Change-Id: I1eb7ef2520502e1fc5ccd829aa9006f82e73f24e