[roll] Roll fuchsia [bus][pci] Attempt to use INTA if InterruptPin reads 0x00
According to the PCI spec, if the Interrupt Pin register in the
configuration header reads 0x00 it means the device function does not
have an interrupt. Crosvm does not appear to set this pin correctly
presently, so for now we'll try pin 1 (INTA) for the purposes of
checking the interrupt routing table we've built. This should not cause
a problem on other platforms because the interrupt configuration is
hardcoded, or does not have routing information for that device and pin
combo if InterruptPin was zero. In fact, it's generally rare to not have
a legacy interrupt pin at all.

Original-Bug: b/365837900
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116894
Original-Revision: d1b8d90b737842adb64d1114f6ea5b4559ea6175
GitOrigin-RevId: be9d684df95d02112c9a7dd621ec325f4ec29e28
Change-Id: Id88e146709388c49db378d48b3cb8f5c2fe65a0e