[roll] Roll fuchsia [rfc] Monotonic clock suspension and the boot timeline
Adds an RFC to further elaborate on the montonic clock suspension
described in RFC-0230. This RFC also introduces a new timeline, called
the boot timeline, which programs that need to be aware of time spent
suspended can switch to before we pause the clock.

This is a high level RFC intended to describe the conceptual changes to
the system and the areas of the system that need to be modified before
we can pause the clock. More concrete designs and implementation details
will follow in future RFCs or design documents.

Original-Bug: 341785588, 328306129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102437
Original-Revision: 0e7bb067ea6e80dc85beade653b7b4882db4bb52
GitOrigin-RevId: db93f65421d4754ebf8c3c8568b0fbda399d100b
Change-Id: Ib2b0bf2719457c19e9b3215a0de8b9e8e2d6d2b6