[roll] Roll fuchsia [ffx] Use absolute paths in isolate dirs
When the daemon is autostarted, there is no guarantee it starts from
the same directory as the invoking CLI. This means that if an isolate
dir is passed in as a relative path (e.g. `ffx --isolate-dir foo ...`)
the daemon may end up using a different isolate dir. The user-visible
behavior is that CLI cannot make a connection to the daemon (because the
ascendd socket ends up in a different location).

To fix this, when a isolate dir is specified as a relative path, we
canonicalize it into an absolute path.

Original-Fixed: 362704093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116679
Original-Revision: 794b8c93147dd2466eb0625435504344c650b4bf
GitOrigin-RevId: ca5b5d398376c32c2f95e407d7da07604737b223
Change-Id: Idffba22106365be75546d0d6f6a6f7e2f0a8b15c