[roll] Roll fuchsia [ffx] when getting config objects, merge levels
The default behavior of `ffx config get` is to stop when it reaches
the first level with an existing value. Unfortunately, if the
value in question is an "object", then that behavior is not
helpful. E.g. if at the user level, we have `obj: { foo=bar }`
and at the run-time level we have `obj: { x=y }`, then when querying
`obj.foo` or obj.x` we get the respective values. But if we just
query `obj`, we only get `obj.x` (since that is the first one found).

Instead, when we find an object, we should merge in from all the
levels, ignoring keys that already exist, to match the semantics
of individual field queries.

Original-Fixed: 359934352
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117175
Original-Revision: 8cb7e1dd173f413273315f43285a66a9e4f10aa2
GitOrigin-RevId: c1515a5aad29ef03e7265eb4f9901d9b942394ec
Change-Id: I7ec81aa76aa34f124b53c80ca2e3b422bbdd8b3b