[roll] Roll fuchsia [gn] Enable no_stamp_files option.
Now that fxr/1115414 has been submitted, and the corresponding
Ninja binary rolled into the Fuchsia tree, the new GN
`no_stamp_files` option in `//.gn` can be turned on.

Its effect is to replace stamp files in the Ninja build plan
(which are actual files written to disk during the build)
with phony aliases (which are paths like `phony/lib/foo`
which are never written to disk).

This is necessary because GN is going to remove support for
stamp files entirely in the near future, as requested by the
Chromium team for improve their Siso tool.

Original-Fixed: 363930634
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115537
Original-Revision: cee2566e056b884a50ca1621be17f814eb6999d3
GitOrigin-RevId: 1dc76da328a7f060dbe98d4c4b3fa18007eff7ae
Change-Id: I6af677704e05f18206840720e940db4a1db4556b