[roll] Roll fuchsia [usb][fidl] Make FIDL libraries versioned
Add `@available(added=HEAD)` to libraries in the "fuchsia" namespace
that are included in an SDK [category] but do not currently specify
`@available`. This will allow the `fidl()` template to verify all
libraries in the "fuchsia" namespace in a subsequent CL.

In four libraries, `added=7` is replaced with `added=HEAD`. The
`@available(added=HEAD)` was added in https://fxrev.dev/664072 but not
removed when the category was fixed in https://fxrev.dev/679350.

Original-Bug: 364628320
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114042
Original-Revision: da2778d6bdf5e82aaf084ec652edf1dbfa6e501d
GitOrigin-RevId: 7a349773542f5ce689d63223874e4ab22fa141b6
Change-Id: Ib977ad0bfda0bf8f8efc88c4dca8dad0f2939af3