[roll] Roll fuchsia [display] Convert Amlogic and vim3 display drivers to new GPIO calls
Call ConfigureInterrupt()/GetInterrupt2() instead of GetInterrupt(),
and SetBufferMode() instead of ConfigIn() and ConfigOut(). The pull
configuration that was previously passed to ConfigIn() is handled
differently depending on the board:

On vim3, the pull configuration for LCD_RESET is dropped as its GPIO
controller does not support pull-ups/-downs. The configuration for
HDMI_HOTPLUG_DETECT is moved to the devicetree and a corresponding
gpio-init composite node is added to the amlogic-display driver bind
rules.

On Astro/Nelson/Sherlock the pull configuration is dropped because none
of the supported panels use kPowerOpAwaitGpio (which called ConfigIn()).

Original-Bug: 42082459
Test: fx test //src/graphics/display/drivers/amlogic-display
Test: Verified that the display works on Nelson
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117174
Original-Revision: 7515717e311815d046b5204a2fc7c15eaec2fa75
GitOrigin-RevId: 379fc14c1483d5c71d32ed2d814bed5102e08900
Change-Id: I6765e6c3d1651b14cdb3ef868656b7e4b16e13fc