[roll] Roll fuchsia [trace_processing] revise computation of suspend metrics
In some conditions, it's possible for the trace to contain
a slice that began before the test itself.

When such a slice is present, the previous method of computing
`total_time` would misrepresent the duration of the test,
and thereby cause the `UnsuspendedTime` to be high, and the
`SuspendPercentage` to be low.

To resolve this, we compute the trace start time based on the
first scheduler slice. That slice should be generated very
soon after the test enables tracing. So it's a good
representation of when the test started.

Original-Bug: 363067753
Test: manually verified that a trace from a recent tryjob achieves
Test: ~60% suspend time.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117574
Original-Revision: 377d239e4866534fc35762e07a18ec52abfff3cd
GitOrigin-RevId: cd2d1a78254f5c11492ae05a872de0bbc064af0e
Change-Id: I2b40dd2d26d09bf7bd1c2a7cc864768fab96e8ab