[roll] Roll fuchsia [fidl] Correct added=7 for libraries in internal SDK category
Since FIDL libraries in the "internal" SDK category are not in an SDK or
stable, they should be `@available(added=HEAD)`.

`@available(added=7)` was added in https://fxrev.dev/664072 but not
removed when the category was fixed in https://fxrev.dev/679350.

Original-Bug: 364628320
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114044
Original-Revision: 0aa000e8bb48809c20359e9cabb7e9f697f901ba
GitOrigin-RevId: 591f7c704f4179fbc837db9cdcc428eac2b4dd7a
Change-Id: I8965eb39f9b21efe6a7cda96928f6b1c2a7306f6