[roll] Roll fuchsia [rust-3p] Re-land "Update rustls to 0.21.7""
This patch re-lands the rustls update, which was reverted in
http://fxrev.dev/1089232, or 04c91aa757acc91cd6a83b1216f7460db9c3f422.

The issue we ran into was in `src/lib/fuchsia-hyper/src/not_fuchsia.rs`.
The old version would error out if `rustl_native_certs` failed to parse
any certificates, but didn't actually error out if it found no
certificates to try to parse. The new version treated this as an error.

The fix here then is to first check that we actually have certificates
to add before checking if we ran into any problems trying to parse them.

Note that this does not remove rustls 0.19.7, we'll remove that in a
follow on CL when this sticks.

Original-Bug: 364892934

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112055
Original-Revision: c4e5cd5b3e74211f00147887603aa464f11ecfc3
GitOrigin-RevId: b1d6d55ffbc1ec02998f88ccbac38a7e20b11da1
Change-Id: I36aaa943e6075cf2deaa4c09c22d8defb1741e97