[roll] Roll fuchsia [bazel_sdk] Override host tool paths when invoking ffx.
This is a reland of [1], which should work thanks to [2].

Ensure that `ffx` and its plugins do not rely on the SDK manifest
to locate the host tool they may invoke at runtime, by passing their
location directly on the command line.

New functions are added to `ffx_tool.bzl` to retrieve the start
of the command-line invocation, that includes the necessary
`--config "sdk.override.<tool1>=<path1>,sdk.override.<tool2>=<path2>"`
argument.

This is required to allow moving the generation of the IDK to the
Bazel workspace. In this scenario, host tools will be built on-demand
as Bazel targets, and their location cannot be predicted before the
build, making it difficult to write them to the SDK manifest in advance.

(The SDK manifest will instead contain Bazel labels such as
`@fuchsia_idk//tools/x64:cmc` that `ffx` definitely cannot convert to
a proper host path on its own).

[1]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114415
[2]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119632

Original-Bug: 367274537
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115539
Original-Revision: 957e159e5467bb29f5994ad1a109605cc8bda913
GitOrigin-RevId: faa99204b49ce09152e4c252576e4c182799a925
Change-Id: I6e24cc26de1c79e1ae6f5ecdcfd5a56460d72a49