[roll] Roll fuchsia [elfldltl] Use C++20 constexpr std::string
This replaces a lot of constexpr shenanigans with just using
std::string, which is now possible in constexpr functions.  Some
new shenanigans are required at the end to persist a std::string
from constexpr context into something approximating a literal
that can be used in constexpr variables.  The ConstString type is
still around but now just reifies a std::string that was computed
in consteval into a constexpr string_view/C-string equivalent.
This requires constructing ConstString with a captureless lambda
of no arguments that returns a std::string.  That lambda type can
be recreated and its return value *used in* constant expressions,
just not saved *as* a constant expression; that makes it possible
to get its returned std::string's `.size()` into a constant
expression, which in turns makes it possible to initialize a
constexpr array from the std::string contents.  Aggregates as
template arguments make it straightforward to approximate a
string literal a little better by merging identical persisted
strings, not just uses of the identical lambda to construct it.

The Diagnostics printf format generation is also cleaned up a bit
by using std::string in consteval functions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118992
Original-Revision: 442c1413b3ef044f6c650888151b1523c3344b08
GitOrigin-RevId: 5cbb685ed7d4a9a9feaf80fee7e8a08f5caa0bd3
Change-Id: Ia3cc6277bfcc4d9750d9e470bb62afcebb6a0d7a