[roll] Roll fuchsia [sestarnix] Eagerly initialize FileSystem labeling details
Previously labeling information for FileSystems was resolved on-
demand, to simplify handling of the pre-policy-load state.

FileSystems are now eagerly labeled, either at construction, or
when a policy is first loaded. This requires splitting the
FileSystem security state initialization across several hooks, each
with dependencies that can be accommodated during Starnix boot:

1. Initialization of SELinux mount-options, based on the FileSystem
   mount options.
2. Resolution of the FileSystem labeling information, based on the
   loaded policy, and SELinux mount options for the FileSystem.

The kernel-global SELinux state is extended to accommodate a queue
of FileSystems used to track those constructed prior to policy load
so that the SELinux subsystem can enumerate and resolve labels for
them as soon as a policy is first loaded.

Original-Bug: 366405587, 366405530, 367309046
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117992
Original-Revision: 6c9c585431263a89a5bfb17c2fec323c59cfb019
GitOrigin-RevId: 5e12e2e7510b0e13f2ac37323ae155e2709dba90
Change-Id: If9607d431b31cadfc37a24c013dbf67d6d06fbb1