[roll] Roll fuchsia [fxfs] Fix race when deleting volumes
We need to hold the flush lock while deleting an object store, because
deletion can't run concurrently with a flush.  Flushing is the only
operation that can mutate a store when it is not currently open, so it's
the only operation that could race with deleting the volume (which
removes it from ObjectManager, leading to a crash when we attempt to
apply a transaction to the store).

Note that there is no need to check if the store is deleted in
ObjectStore::flush, because this already happens implicitly due to the
call to `ObjectManager::needs_flush` (which returns false for a deleted
store).  However, a small change to ObjectManager::flush was necessary
to avoid a race between getting the list of object IDs to flush, and
actually going to flush them.

Original-Bug: 365601359
Multiply: fxfs-platform-test.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118493
Original-Revision: 377ecce90ef0540c79af4a54bbba05431b8c14f4
GitOrigin-RevId: a78d3ed6436fa20ff52acaab7de083732ec1c4a3
Change-Id: I851103d6873b88ed27912e0edeedb1b95994a3f3