[roll] Roll fuchsia [starnix][sysfs] Store KObject in device registry
We now store the KObject for each registered device. Storing this object
will let us implement /proc/misc and /proc/devices in a future CL.

In order to store the KObject for a device, we need to add the device to
the KernelObjects store before registering the device, which is the
opposite order from what we did previously.

This CL also cleans up the implementation of the device registry to use
more type names and to avoid storing the DeviceMode redundantly inside
each of the MajorDeviceRegistry objects.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120213
Original-Revision: d12cf42000051f40bc778a40faeec766d8682ec6
GitOrigin-RevId: 313437105007462b874e08b5ee2cb4bef0602b4d
Change-Id: I93216856d9665cd4b4facd190ff8af28d8d1a231