[roll] Roll fuchsia [kernel] Account for save_state_list during CPU offlining
The MigrateUnpinnedThreads function, which moves threads off a CPU as it
goes offline, did not correctly account for the save_state_list. This
fixes the issue by migrating all of the threads in that list along with
all of the threads in the RunQueue.

Original-Bug: 362353680
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121248
Original-Revision: 5be5834370b936ad9599b619f4d4ba618aa5dfe0
GitOrigin-RevId: d51f66638a6fef467a8ddfd91a40c0a7b09eed01
Change-Id: I773095eb60bc39d62e6b665510178e32810ccfde