[roll] Roll fuchsia [ffx] Use absolute paths in isolate dirs (reland)
When the daemon is autostarted, there is no guarantee it starts from
the same directory as the invoking CLI. This means that if an isolate
dir is passed in as a relative path (e.g. `ffx --isolate-dir foo ...`)
the daemon may end up using a different isolate dir. The user-visible
behavior is that CLI cannot make a connection to the daemon (because the
ascendd socket ends up in a different location).

To fix this, when a isolate dir is specified as a relative path, we
turn it into an absolute path. (Note: we cannot fully "canonicalize"
it, resolving sym links, because that doesn't play well with
Google3 build environments.)

Original-Fixed: 362704093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117582
Original-Revision: 5d8726ef1fad22e4f3b3efbe9e870b568c913a93
GitOrigin-RevId: fd9cc1b8d075a54e168f8c2ddedffca9520b91aa
Change-Id: Ib28099b791b62c18eead0bd37c1b75b6af2ceaf3