[roll] Roll fuchsia [wlanix] Add integration tests for iface name
This change includes the new WifiStaIfaceBaseTestClass that creates an
iface for use by all tests in the subclass. The sole test in
IfaceCorrectnessAndConsistencyTest checks that the name of the iface
reported by WifiChip is consistent with the name the iface reports for
itself.

Various wlanix TODO bugs are also added to wlanix code.

Original-Bug: 367836470
Test: fx test --e2e --output vim3_vg_wlanix
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118695
Original-Revision: 10925e970cbd6c5c3c7ea9dc9f8e4fd4167263cd
GitOrigin-RevId: 6f7fd7687df0fb7a291eb39d4d52f3153dc8f994
Change-Id: I2e649c85e53a06d212a576cc6d7d62a68ab96929