[roll] Roll fuchsia [ffx] SSH parsing does not interpret debugging as err.
In doing some testing, I was able to get an empty error message by
attempting to connect to a nonexistent device. It was reporting that
using an address was an error, because the line was not something the
parser would catch, given it started with `debug2:`.

Original-Bug: 366238654
Original-Bug: 362976369
Original-Bug: 365638219
Original-Bug: 357958365
Original-Bug: 361140479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121661
Original-Revision: b1e6cb583ef2a3fe61a8eca4277324658c8ef6e7
GitOrigin-RevId: 3ef08fe97fe7c57c8159e56b84c8a1fad22413d0
Change-Id: I5cc5a9c2bcf168b0e23e969176b747ce62de5d19