[roll] Roll fuchsia [Lacewing][Suspend-Resume]Add an option to disable timer duration check
idle-suspend-auto-resume based PowerBAT tests are failing because of
b/355511475, until this issue is resolved, we are introducing an option
to disable the timer duration check. Please note that PowerBAT users are
still expected to have a test with timer duration check enabled as this
is how we found this bug in the first place. However users now have an
option to also run with timer duration check disable as a separate test

Original-BUG: b/355511475

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121660
Original-Revision: 389ff18c22d69824f2c01e86196417d04f6bc1a5
GitOrigin-RevId: 83f67ded46529674b51b9239ffa96085f2f3b625
Change-Id: I99535b22031a76daa3011f286f7cc747b288f5bc