[roll] Roll fuchsia [build][sdk] Disallow use of "public" SDK category
There is no consumer of the "public" SDK category, and compatibility
enforcement is bypassed when using it, so disallow specifying this
category until it is needed and properly implemented.

Add checks to ensure that only valid and supported SDK categories are
used in atoms in the SDK. (These checks may not be executed if one of the
more specific `sdk_()` templates does not generate an `sdk_atom()`
for a target.) Also update the documentation.

Change a few existing instances of `[sdk_]category = "public"` to
"partner", remove `sdk_category from two FIDL libraries that aren't
intended to be in an SDK, update an example, and fix an invalid
category.

Original-Bug: 366221510
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120677
Original-Revision: 777522e22a578c3dcb3636b8bd2e902885fc994f
GitOrigin-RevId: 53cdc524ed8f9fa0d969d9208d42375ed51364ee
Change-Id: Ieabefa8b831c4306cb5fb4bb1d822d9018060a11