[roll] Roll fuchsia [netstack3] Define Masquerade NAT action
Define the Masquerade action, a special type of source NAT that rewrites
the source address of an outgoing packet to an address that is assigned
to the outgoing interface (optionally rewriting the source port as
well).

This change defines the fuchsia.net.filter/Action.Masquerade union
variant and adds logic in Netstack3 to validate Masquerade rules, but
does not add an SNAT implementation to Netstack3 Core, so any installed
rules are effectively no-ops. The rest of the implementation will come
in future changes.

Original-Bug: 341771631
Test: netstack3-filter-test
      netstack-filter-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120542
Original-Revision: 7229c18cdf1e58d56c06ae28dd47244ba39ecc54
GitOrigin-RevId: 7303b101672f84973babb65fec6e51f58448f308
Change-Id: I9a3826b0e48e2ed54578f29859f12449c44b5547