[roll] Roll fuchsia [sandbox] Router interface admits null request
Update the interface of Router to allow a request to be `None`. If the
request is `None`, the router may supply a default request (or it may
choose to return an error and require an explicit request).

This commit does not actually change the behavior of any router to
accept a null request. That will come in a followup.

Note that `debug` (not present in fidl) is still required, because we
wish to require the caller to choose whether they are performing a debug
route.

Original-Bug: 366458428

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112031
Original-Revision: 659e83562bc0260431e6e6ef00ffcbfc9c9662b1
GitOrigin-RevId: 4941205fe31a2a186d96bd73ff83b8c1d3c028af
Change-Id: Iacd7f6b1a875bdf1fe76040db76011f613aba0c2