[roll] Roll fuchsia [vsock] Add cid to listen call
This allows us to start redirecting listen calls to different backends
as necessary, such as a loopback when VMADDR_CID_LOCAL is supplied.

This change also starts enforcing that the only valid CID which can be
supplied today is VMADDR_CID_HOST as that's really the only one we
support.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120733
Original-Revision: c1c833c51a0635411f540787a20dc9334d4208fb
GitOrigin-RevId: e3a7c071e4d1bbec0cffae411cd1604fa0adfe9a
Change-Id: I714aa3386fb3eaafa734ad1ce58bfa19cf6d46d2