[roll] Roll fuchsia [network/tests] Test ping forwarding
Add a test verifying that a client on Network A can reach the router's
address on Network B, if the router is configured to forward between the
two networks.

Essentially, this verifies that the Netstack operates as a Weak Host
when configured as a router. Currently it passes against NS2 but fails
against NS3.

While here, refactor the original "forwarding" test so that the setup
can be shared.

Original-Bug: 368343873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122153
Original-Revision: 2a56fd74be20cc62de774359ae28c5b66bbad2be
GitOrigin-RevId: 94fd90de3309b5552d22f04a1e8e4127eb6f3f65
Change-Id: Ib0f063a42c36378df1b8435f5d1f5d8aac00aa79