[roll] Roll fuchsia [ffx] Add target disconnect request
When a test causes the target to reboot, the daemon can be left
with a stale connection. While that connection will eventually
be detected, the detection is based on timeouts that are changeable.
If the test wants a predictable result, it needs to inform the daemon
that the connection it has is stale, i.e. that the daemon should drop
the existing connection.

By itself, this CL does nothing since there are no clients for the
new Target.Disconnect() FIDL request. But there should be a follow-up
change to add the method to Fuchsia Controller. And possible a
new ffx subtool, `ffx daemon disconnect`, to give a similar ability
to non-FC clients.

Original-Bug: 367453247
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122213
Original-Revision: 0cbf9b7caa83a6458401ede0e5f497ab0f5b7320
GitOrigin-RevId: 93461842bdb9aa33bf2a6e6eac201ed5ad4a4f94
Change-Id: Ica8e686a9827ea978324ae348281fb28f51fbf9a