[roll] Roll fuchsia [ffx] Add "ffx daemon disconnect"
Add a command to tell the daemon to disconnect from a target.
This is useful in a reboot test, which will know that the connection
the daemon is maintaining is no longer valid. It can be used to
be explicit, rather than relying on the unknown timing of the daemon
to detect that the target has gone away.

Original-Bug: 367453247
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122219
Original-Revision: 3147bbf2f620b371efffe8c48e859dbf4062158b
GitOrigin-RevId: 4481941ef80425cadec82b0850aa859da7b54ee6
Change-Id: I7aba488fffe1192c4daeaa9b2b2e41c7f651c48c