[roll] Roll fuchsia [bazel_sdk] add ability for bt-host to not include c++
The bt-host wants to statically link c++ to test its impact on size.
However, we don't currently have a way to not bundle libc++ libraries in
a package if it statically links in c++. This CL adds a hack which
allows them to remove the libraries from the package. When we build out
better logic for determining if libc++ should be included we will remove
this hack.

Original-BUG: 368434718
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122452
Original-Revision: 858bcdf4f1a4c7f93dfa3229cf941dcb45d5796e
GitOrigin-RevId: 344e65e5d65779bbfd120c2d1626de869bd21ae7
Change-Id: I4acf4780845569f8726f70ed5ff552f653509abe