[roll] Roll fuchsia [starnix] Split overloaded FsNode::socket field
This field had two jobs: For FDs created by syscalls like `socket()`, it
held the underlying socket so socket syscalls could access it through
the FD. For unix domain socket nodes, it held the socket bound to that
node. It turns out that for the first type of FD the socket is already
held by the FileOps, and nothing needs to access it through the FsNode.
So switch the first type of field accessors to go through the FileOps,
and rename the field to `bound_socket` for use for bound unix domain
sockets.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121712
Original-Revision: ffe0588d0f86e1c58937e7c217e328e17903062f
GitOrigin-RevId: 4b592cab67d8f6cd8892de18c2966146e738a88c
Change-Id: I93bf65bc6f69ae59bbb68bca144751337995e5d1