[roll] Roll infra/recipes [checkout] Make release branch resolution more robust
Previously we wouldn't emit a `release_branch` output property if the
`ref` in the triggering `GitilesCommit` was a tag rather than a branch.

Now if the ref is not a branch, we'll fall back to using `git branch
--contains` to determine which branch the commit is on.

This is necessary due to a recent change (http://tqr/876812) to builder
triggers that updated some release builders to trigger on tags rather
than on branch heads.

Original-Bug: 361799038
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1121533
Original-Revision: 62defb8441af6c6af3fa745765521318f7d08847
GitOrigin-RevId: cef98e9f88511939619d28a4b047a973fb3870b8
Change-Id: I95d859978eee8d7ae0683f8d5030bf762c375f77