[roll] Roll fuchsia [kernel][userboot] Construct RoDso from a VmObject
This change refactors RoDso to be constructible from an explicitly
provided VMO, factoring out the "embedded" logic (which we'll soon
invalidate) and the kernel handle logic (which is not appropriate for
the userboot image case). This eases the upcoming removal of the
embeddings of userboot and the vDSO from the kernel.

Original-Bug: 1097802
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121717
Original-Revision: a88444022bab9bd4f4a5f2ba24ca6d9d6a9cceb0
GitOrigin-RevId: 4f32ba03850b737125638293d79d950a78633f60
Change-Id: I5a0a942117ee5140b7648221454e6eef3a96ba7f