[roll] Roll fuchsia [netdevice] Fix race on TX in-flight counter
Currently in Inner::send a buffer is queued to be sent to the driver
before the number of TX buffers in flight is incremented. This means
that it's possible for the buffer to be sent to the driver and complete
a round trip back to the client before the in-flight counter is
incremented, at which point it's decremented while currently at 0,
causing a debug assertion to trip.

Fix this race by incrementing the in flight buffer counter before
actually queueing it to be sent to the driver.

Original-Fixed: 369180969, 368652664, 368689639
Multiply: tun_socket_benchmarks_with_netstack3_test.sh
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123853
Original-Revision: d72296153a7268090a15f8f394a7f822914f08b8
GitOrigin-RevId: f3cc43aa1ffaf964bd32be10ec16320c47386ec7
Change-Id: I5a54294fb9f01358c9bcf5c630a5bccbcaa7e065