[roll] Roll fuchsia [starnix] Add type for architectural invariants on futex addresses
Addresses used for futex operations in Linux including the futex family
of system calls and the robust futex list feature must always satisfy
two architectural invariants:

- The address must be aligned to a 4 byte boundary
- The address must be within the range accessible to user space

This adds a type in starnix_uapi that checks for these properties on
attempted conversion and generates an error if the value does not meet
them. The futex logic in the futex table and the robust list
implementation use this type to perform these checks exactly once on
each incoming value. The memory manager's atomic operations routines
(which are used exclusively for futex operations) accept this type to
verify that the caller has checked that the value is architecturally
valid before attempting the operation.

This has the side effect of removing several redundant alignment checks
from the futex table and adding missing range checks for the value.

Addresses below the base address value (currently 2MB) are
architecturally guaranteed to have nothing mapped, so that value does
not need to be explicitly checked before access is attempted.

This also moves the restricted address space constants to starnix_uapi
so that they can be used in other logic as they are architecturally
defined constants that are applicable to all of Starnix and not specific
to the memory manager now that we always use the same address space
model.

Original-Fixed: 369214563
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123925
Original-Revision: c09e8c32d89bec02d0ff09466fc1ef40dcf8569f
GitOrigin-RevId: 9ae8227aef0eadcf44cbcfc54c89d6c6d95b787f
Change-Id: I641f464dede95951d2a26e5eda107ae66495b82a