[roll] Roll fuchsia [tee] Stop skipping xtest regression case 1040
It used to hang, but now successfully fails! The test opens up a number
of connections in succession, so the recent change to make app endpoint
access more thread-safe likely addressed the past hang.

The case is actually currently commented out in optee_test.git, so this
is the first of two changes to get it running again.

Original-Bug: 348473568
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123929
Original-Revision: f54e97c2f64e59971320b233d2be8796961163d6
GitOrigin-RevId: b6093a100aaf13d4294230615fb5d49df5e99883
Change-Id: Ia6e2c59910ca8c5617e39a0c17968d10c947b4b3