[roll] Roll fuchsia [superproject] Roll third_party/pigweed pw_allocator: Store TestHarness::Allocation inline
This module include a TestHarness class which can generate requests for
an allocator to allocate, reallocate and deallocate regions of memory.
Previously, this harness took a vector of Allocation structs that was
used to stored the details of these requests. This allowed, for example,
the test harness to deallocate memory that had previously been
allocated.

Requiring this vector could result in significant overhead if a
substantial number of allocations were to be tested at once. This
overhead is unecessary though, as there is another place to store the
request details: in the allocations themselves!

This change sets a minimum size and alignment on requests. It makes the
Allocation struct an IntrusiveList item, and stores the Allocations in
the allocated memory.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/230894
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Original-Revision: a4dab0be18a649c79759880308ed051b0a2576a4
GitOrigin-RevId: 547308eef8f304a60f5b5d535fcf7c4c5ec683d6
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124718
Original-Revision: d593f913e8546f7200fa7e1885bab9b04c6d729c
Change-Id: I989938ef73b52b17b603d616b8b5a1f73037ffda