[roll] Roll fuchsia [cml] Revise syntax for dynamic dictionaries
Instead of:

```
capabilities: [
  {
    dictionary: "dict",
    extends: "program/svc/fuchsia.component.sandbox.Router",
  },
]
```

The new syntax is:

```
capabilities: [
  {
    dictionary: "dict",
    path: "/svc/fuchsia.component.sandbox.Router",
  },
]
```

This harmonizes better with other capability types. The semantics of
`extending from program` diverge from `extending from
parent/self/child`, and it's a good idea to reflect this in the API.

In a followup, I will land a change to decouple dynamic dictionaries
from the extension codepath entirely.

In addition, disallow `extends` and `offer to self/dict` in this mode -
they no longer make sense as in the followup we're no longer creating an
intermediate dictionary. If users want this behavior, they can always
create a second dictionary and extend the dynamic one.

Original-Bug: 368155371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121718
Original-Revision: 64d7522aa87ebf7f9876338abc52062919937f26
GitOrigin-RevId: de67026d667261fed3938657a53384a66fce6a9c
Change-Id: I78f8496a330b3dc4004cdeb630bf4c06b6e95e75