[roll] Roll fuchsia [components][dictionaries] Simplify dynamic dictionaries implementation
Today, we implement dynamic dictionaries by creating a new dictionary
that _extends_ the DictionaryRef obtained from the Router protocol
served by the program.

This extension is an unnecessary step and carries the limitation that if
the program adds new items to the dictionary after it is routed for the
first time, they won't appear in subsequent routes. This is because the
extension operation takes a 'snapshot' of the source dictionary.

Original-Fixed: 368155371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121832
Original-Revision: cb3005de3158c6fa60375537e212b5793269a8b5
GitOrigin-RevId: 8aa38b90332e441d63b201daef8b9fd682d9e422
Change-Id: I47dc9bf3c9475c807cec65977e7330bb0464504a