[roll] Roll fuchsia [dl] Move LinkingSession::TraverseDeps to RuntimeModule
This CL is a verbatim code move only, to make review easier. The
`TraverseDeps` function always operates on the `RuntimeModule` that will
serve as the root of its constructed graph.

This function fills out the runtime module's `module_tree` when dlopen
is called on the module. Currently, it is only called on the root module
when it is loaded. This will eventually also need to be called to fill
out the module tree for a module that was loaded as a dependency of
another module, when dlopen is called on that loaded dependency.

As this will always operate on a root `RuntimeModule`, moving this
method to the `RuntimeModule` class will make it easier to invoke for
both scenarios.

Original-Bug: 354786114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124680
Original-Revision: 9f69e84c8c11143139c8b32cf9efe0f39039f6e6
GitOrigin-RevId: f27d6eef4133f193d36911f9fef4dd8767e2c36f
Change-Id: Ia788974cc307f7b23eca8b8e13c56ce10ddd9f5c