[roll] Roll fuchsia [kernel][vm] Correct SOLPI assert
Due to how chainlock transactions work and the need to perform backoff
it is possible that a thread sets is_ready_for_waiter_, and yet does
not actually end up blocking. In this case the thread will have
observed that the stack ownership interval has ended, and so there is
no problem with this case, except for this assert spuriously
triggering.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115313
Original-Revision: b2ccac880498da3c9e30fd2a91be4844ebddc2d8
GitOrigin-RevId: 2bbd06b16b34bbac64fd14113b411fac22dddd8f
Change-Id: Ifd143e9d15eeafa005df2c28852882473c821eca