[roll] Roll fuchsia [storage][power] Remove wake-on-request support.
In the process, we remove the dependencies of hardware power
elements on Execution State. This disables storage power management
altogether, as the drivers' leases on these elements will be
unconditionally fulfilled.

Storage power management will be re-enabled shortly using a
simpler scheme.

power-system-integration-storage-test is dependent on this
integration, so it must be disabled. It will be re-enabled with the
revised powermanagement scheme.

Original-Bug: b/368409324
Original-Bug: b/368636358

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109447
Original-Revision: 53783868a18571aaf19cb92a6c2a0e17355e670a
GitOrigin-RevId: 87e1cdae94dfa848dc06ddfa3ce6ad855be9cc12
Change-Id: I4964fed6f440360159b4983ab821373a2a47ea92