[roll] Roll fuchsia [wlan][policy] Remove connection selection sender error log
The connection selector logs an error when it can't send its selection
to the receiver. This is expected to occur when connection selections
are cancelled and the receiving end of the channel is dropped.

This change removes that error log, as the connection selector should
not care if the receiver is dropped. True errors passed through the
channel will be logged by the receiver.

Original-Bug: 369662954
Test: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125768
Original-Revision: bfcb5b953dcceca40c003e50b2273e623e1fa39d
GitOrigin-RevId: 05851348c011c2e937c74ec2b4352d380790977f
Change-Id: I5cadd682a6f0d63b30b94b63ea0272dde0936f59