[roll] Roll fuchsia [netstack3] Use weak host semantics when a router
By default, NS3 acts as a strong host. However with this change, it
begins acting as a weak host when forwarding is enabled (i.e. when
configured as a router).

The newly permitted weak-host behavior is:
  * When sending packets, allow a src addr that is not assigned to the
    output device, provided it is assigned to another device that has
    forwarding enabled.
  * When receiving packets, allow a dst addr that is not assigned to the
    input device, provided it is assigned to another device and the
    input device has forwarding enabled.

In both cases, the packet passes through the forwarding hook of the
filtering engine, because conceptually we're "internally forwarding" the
packet from device A to device B. Note that this CL also updates the
integration tests to prove this.

Note, the send path requires checking if forwarding is enabled on a
device during route lookup. To accommodate this, I adjusted the lock
ordering so that the device configuration can be locked after the
routing table has already been locked.

Multiply: netstack-forwarding-integration-test

Original-Bug: 368343873

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115635
Original-Revision: 113fc28b40a9b5879529f4c8acb517a30bdf6c44
GitOrigin-RevId: f32e7a15edcd26427952cee4a69a36ecceadfdd0
Change-Id: Icce47c351c2f3bad6f418b34dc30658d04ea38ce