[roll] Roll fuchsia [build][rust] Support rustc_cdylib loadable_module
Adds support for building rust cdylibs that use `loadable_module` under
the hood instead of `shared_library` and behave in a similar way as if
you were compiling a C++ library with `loadable_module`.

Also implements allowing the caller to override the output prefix in
general, with correct behaviour for when it's set to use loadable_module
(which requires inverting the sense of the override flag since the
rustc_cdylib tool only expects to build shared_libraries).

Original-Bug: 364901670
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126633
Original-Revision: 0fa033949e435412b43179376139c7f8882e7426
GitOrigin-RevId: a9151e68bb4c2f472ef424df6a9b1eb1cada2905
Change-Id: I9a4e8a77355d16ad4c6484dc6028d9753a69495e