[roll] Roll fuchsia [owners] Remove glamrock@google.com from src/diagnostics/config/fire/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/glamrock@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118447
Original-Revision: 0d8adc46a3c4a71cbf1ab540994e414fd42123db
GitOrigin-RevId: 1d576226e3b538f16626d10cac88bc2370c72b3c
Change-Id: If6218fa8d7a7248e2dd36112a112d9a4098afc7d