[roll] Roll fuchsia [f2fs] Improve read performance
This CL tracks recently used pages, and excludes
them from readahead I/Os on read requests. Also,
it creates a VMO as large as needed to supply
pages to the area of a pager's VMO where pages
are not likely to be present. This way
effectively avoids unnecessary readahead I/Os
and the oversupply of VMOs to the area of
a pager's VMO to which it has already supplied
pages before. It uses the hint of the recently
used pages only when memory pressure is low.
On high memory pressure, it assumes that all
pages are not present on the read request area
of a pager's VMO, and does not readahead. Also,
it clears clean pages on FileCache of
pager-backed vnodes during writeback to speed
up lookup for dirty pages on fsync().

Original-Fixed: b/368139435

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116913
Original-Revision: 8be4fbf36f2f2d1d28c8b321111c42b4ad1c07d6
GitOrigin-RevId: cd61a2beedbd7acad5a3c131f57d687b4963e170
Change-Id: Id535e9a50b1ead8ddca8936ca35d7005165ed214