[roll] Roll fuchsia [owners] Remove kevinwells@google.com from src/sys/lib/fuchsia_backoff/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/kevinwells@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127729
Original-Revision: 420d1a91acc3e61e7de27868cd461749de9aa209
GitOrigin-RevId: 374e023f5d7c71f3c855a377c9409412d31f3231
Change-Id: Ifada85ff472e55346a18ff78ca99539fd10a8512