[roll] Roll fuchsia [zircon][docs] Remove mention of ZX_CPRNG_DRAW_MAX_LEN.
This constant is an implementation detail of the vdso and syscall
interface, and should not significantly affect the behavior of callers.
There are a few instances where the name seems to have suggested to
callers that they can't pass a buffer larger than this value, so
we should just not mention it to avoid confusion.

Original-Bug: 369851484
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126632
Original-Revision: 70bf5be6749f0ee3be35b3066fdb2c27334b5db1
GitOrigin-RevId: 6133711296ee34fdb35eeed49efad31d6760e9da
Change-Id: I8200652871bd74974ffee50f4a2e9ff720a7519b