[roll] Roll fuchsia [fidl][rust] Decouple resources from core library
This allows the same FIDL bindings to be used on host and target, even
when the host side does not support handles. This also lays the
foundations for other resource types in FIDL, such as driver handles and
FDomain handles.

With this change, it is possible to use the core FIDL library and
bindings without pulling in any Fuchsia-specific dependencies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125893
Original-Revision: 00bcdade28a391d1748bc55dd2537c5cad316945
GitOrigin-RevId: da28814714db0e7556d57d0ee9c455164e05d505
Change-Id: I0d64f29bba0e99ea7260f65550b99d6b3185909b