[roll] Roll fuchsia [owners] Remove kevinwells@google.com from src/sys/pkg/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/kevinwells@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127730
Original-Revision: baafc24914c1a46f22c166f60c370b701c4fa436
GitOrigin-RevId: ed6d3f54040cbbe301c4df3cc9003b06aa761425
Change-Id: I639579298b466f7ef6a30ae48f703310f90ada79