[roll] Roll fuchsia [package-tool] Avoid overwriting TUF repo data inside a product bundle
This will copy the TUF metadate into a user-provided empty directory
which serves as the (mutable) TUF metadata store for the blobs in the
product bundle. This ensures the original TUF metadata in the product
bundle is not modified when the blobs are served via the repo server.

Original-Bug: b/335008631
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128015
Original-Revision: cd599313851f3bc10e9f164119d3671e5fc01ca5
GitOrigin-RevId: a9d878d34de3be82c8a85f022205f8aa918183e3
Change-Id: I6e6f524c07ebd88f4af3fad2f0ec78280cf4520e