[roll] Roll third_party/github.com/google/emboss Restructure ACTION and GOTO tables. (#194)
With this change, the `Parser.action` and `.goto` tables are now 2-layer
hash tables, using states as the keys for the outer tables and symbols
as the keys for the inner tables.  Functionally, this means that
`action[state, symbol]` becomes `action[state][symbol]`, and likewise
for `goto`.

This also allows the `expected` table to be eliminated, as it can be
quickly computed from `action[state]` when needed.
Original-Revision: 886cfbb659401db1c6911493f264c6586a37f5c8
GitOrigin-RevId: 89e2564a8ffe804725b1502c85d98d5ebe9126b2
Change-Id: Idba5e1589e01868e66aa6c40a11c62dd3561ebe6