[roll] Roll fuchsia [superproject] Roll third_party/github.com/google/emboss Restructure ACTION and GOTO tables. (#194)
With this change, the `Parser.action` and `.goto` tables are now 2-layer
hash tables, using states as the keys for the outer tables and symbols
as the keys for the inner tables.  Functionally, this means that
`action[state, symbol]` becomes `action[state][symbol]`, and likewise
for `goto`.

This also allows the `expected` table to be eliminated, as it can be
quickly computed from `action[state]` when needed.
Original-Revision: 886cfbb659401db1c6911493f264c6586a37f5c8
GitOrigin-RevId: 0043eda87257ed784a3a4bb89dd45252b1e5baa6
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129758
Original-Revision: 602595daa6df5f0b7562382c0718bcd2301735cb
Change-Id: I4b29b3d73c81fa6ae690a63251cac1f309990fbe