[roll] Roll fuchsia [ffx] remove BuildOverride
Based on code inspection and some testing, BuildOverride doesn't seem
to do anything for queries. (In particular, regardless of the setting
of the query's "build" parameter, the same config is returned.)  This
suggests that the functionality is not being used, and since it adds a
lot of complexity to the ffx config code, it should be removed.

Another use of BuildOverride is used to implement the setting of
build files in Environment. But in this CL it is simplified down to
an Option<Path>.

Original-Bug: 342038349
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126673
Original-Revision: 781610dd3dcdf11ea7b6b09613d94257f1d8f8c7
GitOrigin-RevId: 8afba478defccab31d97e69d7dfb5be8a0ea22e8
Change-Id: I6064320cc30a10037b29aa36d0a70195fd202d4b