[roll] Roll fuchsia [spmi] Switch to debug service in spmi-ctl and update routing
The --target option is now mandatory and accepts an integer ID of the
SPMI target device to connect to. An optional --controller option is
added to specify the name of the controller to connect to: if
unspecified, the first entry in the service directory is used. The
ability to connect using a hexadecimal service name has been removed.

All routes that previously referred to TargetService and
SubTargetService have been changed to DebugService to reflect the
changes to spmi-ctl.

Test: spmi-ctl-test
Test: Used spmi-ctl to read registers on Kola
Original-Fixed: 368117439
Original-Bug: 361853516
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128274
Original-Revision: 485af07906f30dc9e3c46ef4b537e5105bfbde75
GitOrigin-RevId: 763c8c58a614adcb7b8519a67fe1f4b4f6f44d4c
Change-Id: Id8794ef5715dc499d28eac88a17645d477dffbc2