[roll] Roll fuchsia [logs] Use random per-boot integer for detecting reboots instead of UTC
The UTC to monotonic offset can change frequently and is
not guaranteed to stay the same within the same boot.
Instead of using this to detect reboots, use a per-
boot random number that is generated from the monotonic
clock when Overnet starts on the device.

Original-Fixed: b/370548387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128476
Original-Revision: 9c2c2664a793b4827f8f31fb4fb0b7216fb6cce0
GitOrigin-RevId: 42906e5224faf9028e264bfc3b32eb20ae9d0f36
Change-Id: I8e468efacb51adc625152364e58dbf232bfde43d