[roll] Roll fuchsia [build][fidl] Limit unstable to internal and partner categories
Only allow FIDL libraries to be unstable when the SDK category is
"internal", "partner", or not specified. This specifically disallows
unstable libraries in the "partner_internal" category, which as noted in
https://fxbug.dev/364727245, does not make sense.

Move `fuchsia.wm` to the "internal" SDK category to meet this condition.

Allow exceptions for known exceptions in "partner_internal" due to ffx
build requirements.

Original-Bug: 325669391, 364727245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125780
Original-Revision: 8521352712d32c24afb200ce2ad7b211fb6ff531
GitOrigin-RevId: f15c6aec2923edabed4758bc25690a993888125e
Change-Id: I826de1c52a959fdbad670dab5b25dd06556bdc71