[roll] Roll fuchsia [ffx][repo] Fixing product-bundle server registration
This fixes 2 issues related to package serving with a product
bundle as the repo source.

First, it constructs a repository name based on the "base repo name",
which is either the `--repository` option, the default configured repo,
or devhost. Then appends `.` and then first alias for each repository
configured in the product bundle. This avoids having multiple servers
named "fuchsia.com" and serving different packages.

Second, this adds registering the aliases configured with the repository
when registering the repo with the device. Previously, since the name of
the repo was the first alias, it would work for 1 product bundle with
repositories that had 1 alias each.  Now it works for multiple product
bundles.

Original-Bug: 370594742
Original-Bug: 370089652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128280
Original-Revision: 6deb7ca9a08a03c7510f2b8fffe66d7ecd11840e
GitOrigin-RevId: 0b2e1b0befe7af399ac4fe7d329af6f4262a4aad
Change-Id: Ia2816c4c56dfb8bd3cf2d511f529f3bb801e4f84