[roll] Roll fuchsia [netemul] Disambiguate protocol names using source
Currently, netemul installs all capabilities provided a realm's children
to the root node. This introduces panics when multiple components expose
the same protocol.

This change adds a suffix to each exposed protocol to ensure that
clients can specify the name of the protocol they would like to connect
to, or leave it unspecified to connect to the first component offering
that protocol.

Adds a `counter-alternative` component for testing to demonstrate
which protocol from which child will be used when they both offer the
same protocol. One component has program_args specified for the starting
value so that we can determine which one is being used.

There will be a follow-up CL to allow specifying child name in the
client netemul API.

Original-Bug: 42151461

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126658
Original-Revision: 670699acdcdef00f1ec389f81444c573e5190ac5
GitOrigin-RevId: f6f10c1af1a917073d483deaff2916a5e18414c3
Change-Id: Ia2a9f5c5e482b4d87090ea7bd03c013670c43c8b