[roll] Roll fuchsia [netemul] Add endpoint for protocol child name
Users of the netemul TestRealm API can specify a particular child to use
when connecting to a protocol.

Add a test to ensure that specifying the child component's name when
connecting to a protocol results in creating a proxy that returns the
expected results.

Original-Fixed: 42151461
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129975
Original-Revision: 06abdeeb2c028095f32ac68783e15f114fd9391a
GitOrigin-RevId: 36fdd5241e77373db2886196d84364d0b634bab0
Change-Id: I89b0fe469957f7160591ba87be0ef97f82d6956b