[roll] Roll fuchsia [bt][fidl] Add generic UUID to BR/EDR service search params
Currently, the BR/EDR Profile Search parameters require a mandatory
profile/service identifier. This identifier is actually a UUID and the
Profile FIDL API provides a list of known IDs for profiles & services. In
some cases, it is desirable to search for a custom UUID instead.

Add an optional field, `full_uuid` to the service search parameters. If
specified, the `bredr.Profile` server will use this UUID as the search
filter. It is invalid to specify both a `full_uuid` and `service_uuid`
in the parameters - an error will be returned in this case.

Original-Bug: 355487849
Test: API only - fx build

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110654
Original-Revision: 4fbed5b475282a2ce94bb13fa052a4bbdc5302ec
GitOrigin-RevId: b2a438c1eb0962ac2a3480b4aa378600e4c9e33f
Change-Id: I74d408bd278c4fb4def7d18e9a81bfd2adf5412a