[roll] Roll fuchsia [driver_runtime][rust] Add driver server
Implements a `DriverServer` that interfaces between the driver host and
the rust driver object, including managing its async lifetime. This is
analogous to the C++ `DriverServer`.

Also adds the basic outline of a `Driver` trait that is kind of like
C++'s `DriverBase`, except that instead of holding the context in a base
class it is given a `DriverContext` with that information.

As part of this, this also breaks up the main `fdf_rust` library so that
`fdf_sys` is its own (visibility-gated) library that both `fdf_rust` and
`fdf_rust_server` depend on. This avoids a dependency loop since the
server depends on generated driver transport fidl, which in turn depends
on `fdf_rust` for things like Channels.

Original-Fixed: 359228261
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090314
Original-Revision: 40c310645b9e2c2501b11efb8ed28cd63c27de1b
GitOrigin-RevId: 4321a28ed06e24656fd3f9ee9c49a96d3693f438
Change-Id: I74870003f7133c769f704118991ae07ed4a09a41