[roll] Roll fuchsia [netlink] Route table state outside routes workers
The route table state will need to be shared with the rules workers. To
prepare for this, move it out of the routes worker to make it clearer
that this is the case (rather than having the rules worker take &mut
routes worker).

Test: netlink-lib-test
Original-Bug: 358649849
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1131278
Original-Revision: 3dfe6d8a09cc331e4d686a52b7d952a62be5aa23
GitOrigin-RevId: b4d3a9d2b6ac4e6ee19554ec5479f3ec8d7a9b56
Change-Id: Id0f828f3a7fa6ac88b5adcd5e30d63210334b0a4