[roll] Roll fuchsia [gpt-rust] Restore primary metadata from backup
This is required by the UEFI spec (see 5.3.2).  It also ensures that we
are resilient to sequences of partially completed updates; for example,
if the following sequence happens, primary restore is necessary to
ensure the GPT remains consistent:

1. Metadata is updated.  Only the backup metadata is fully written, and
   the primary metadata is partially written.
2. The system boots and uses the backup metadata.
3. (This is when primary ought to be restored.)
4. Metadata is updated, but this time, the backup metadata is partially
   written.

Without restoring the primary metadata in step (3), the GPT would be
corrupt after this sequence, as both copies would be corrupt.

Original-Bug: 371060853
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1131492
Original-Revision: 7ef21cd28382e3ace467fe513feff53dfac801c5
GitOrigin-RevId: 4c218ff97d4f5beeeccde700a28d2b3530113b5e
Change-Id: I444e6697d3f958c89d653650a18cfab5be2d1f34