[roll] Roll fuchsia [components][sandbox] Support DirEntry on host
Support on host to prepare for typed routers, including DirEntryRouter.

The host version of this type is basically empty, but it allows code
that uses this type to compile on host. In practice, the host code will
use debug routing so it doesn't need DirEntry to be functional.

Original-Bug: 360221168

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107927
Original-Revision: ed573bd221464cf0a63291f724b2473a8756eaf4
GitOrigin-RevId: 340828a5fb6627b776c59630777c66b6d187794b
Change-Id: Ibcef40172dc2f11f3a7efa27bd4769bac05f13d9