[roll] Roll fuchsia [kernel][crashlog] Print boot time in crashlog
As dictated by RFC-0260, use the current_boot_time() to indicate system
uptime in the crashlog. However, the user may also want to know how long
the system was up not including periods of suspension, so add a line
that prints current_time() as well; this quantity is now called
"RUNTIME" to reflect the fact that it is the amount of time the system
was awake and running.

This change does not add RUNTIME to the RAM mappable crashlog; that is
left to a future change.

Original-Bug: 328306129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130234
Original-Revision: 013a6fd2c3164cc4781f227e0ca4b92b5976c89f
GitOrigin-RevId: 7342b69c03b84cd4b81e0dec3e872e0e43e0d146
Change-Id: I48661324932501dbd03f213c92022d59f2edccc7