[roll] Roll fuchsia [starnix][aio] Refactor internals of AioContext
This CL moves a bunch of logic from AioContext into simpler objects. For
example, the work of converting between the UAPI opcode and the Rust
enum for the OpType is now part of the OpType implementation.

This change makes it easier to understand the main control flow of the
asynchronous IO operations because all the type conversion is moved out
into leaf types.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133872
Original-Revision: b96f56794b8f23774a293e0f96c07f8098577b00
GitOrigin-RevId: 4d224d17881aab887524a9415a846144404da48a
Change-Id: I6296e48767274a870f932127cde676a0edb4d6ed