[roll] Roll fuchsia [owners] Remove dworsham@google.com from tools/devshell/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134675
Original-Revision: f0f2555b9219f703fb2e6e4302e7bef0abc753d2
GitOrigin-RevId: 4126150a7b4ed0af0173936c7effa4ac70a16acf
Change-Id: I08c72c6cb4a496c9ff6d56784ea12766f3cd3b07