[roll] Roll fuchsia [owners] Remove dworsham@google.com from zircon/kernel/vm/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134597
Original-Revision: c1ef59632f7e3ce78610868a03ef835b49f5997b
GitOrigin-RevId: 9c20e7c799888464f8f11e7ba34a4ec1d2e91578
Change-Id: I5db362467f00d32ed03b8d02a6b8876d389ab446