[roll] Roll fuchsia [kernel][vm] Back split bits w/ a share count
Add a share count field to pages and back split bits with it.

Also add some opaque metadata storage to the compression system. The
caller is able to update the metadata for a page anytime, including
while compression is in-progress or the page is compressed and bound
to a reference. When compression produces a page for the caller (via
Decompress, MoveReference, or Compress failure) it will return the
most updated version of the caller-supplied metadata.

VmCowPages uses this metadata storage to preserve the share count across
Page<->Reference conversions.

No functional change.

Test: compression_unittests
Original-Bug: 42058561
Original-Bug: 338300808
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068495
Original-Revision: c3fb81b6fe9ceecedc00dfaef2a6f6a6b438f6b4
GitOrigin-RevId: 904e5b33a84bf92138ffacb660ca0e4b1fcdd3d0
Change-Id: I0e18cdc31da0881943ad2f26ee9984a226241550