[roll] Roll fuchsia [fidldoc] Render vectors and arrays recursively.
For whatever reason, when printing a `vector<T>` in fidldoc, the type
`T` wasn't being rendered using the same logic that renders types in
general. It would call a different template that prints a simplified
kind of type.

I don't see why that would be preferable, so in the absence of push-back
I'll change it to work the way I thought it would work :)

Also, I've changed the bounds on arrays and vectors to more closely
resemble the modern FIDL syntax.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128892
Original-Revision: f84f74c1c1d59dde06f1dde0bcc20b84b7749034
GitOrigin-RevId: 72caedd5d4d961ed1002e3a5af5ae1fb5e439cff
Change-Id: Ibe0a5a677b4afd73b2265471186818374c293fcf