[roll] Roll fuchsia [debugger] Add explicit options for public/private servers
Before this, there was a single "symbol-server" command line argument to
all the debug tooling. This is inherently ambiguous, because it's
unclear if the server should go through the authentication flow or not.
If a public server is specified this way, and never responds to the
authentication flow, we'll never actually use that server.

When servers are added via the symbol-index, this is typically never an
issue because the ambiguity is solved there. The symbol-index is not
available in all contexts, however, so adding explicit parameters is the
way to go.

Servers specified with the DEBUGINFOD_URLS environment variable are
always treated as if they were added as public servers.

The existing --symbol-server argument is unchanged, and means "private" servers. This is to maintain compatibility with existing users (particularly for symbolizer). We can change this later to "--private-symbol-servers" to be consistent in the future.

Original-BUG: 365193885
Original-BUG: 359660815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134732
Original-Revision: 80001d91f93f4b6181a1a841cdd8cb030962255b
GitOrigin-RevId: 9f51b0d59bc27103fded4d137e1d7cbf6b850ef2
Change-Id: Ieb157141382c4ef32e22c096a6050bcec48da510