[roll] Roll fuchsia [netstack3] Enforce device MTU at IP layer
Wrap the serializer with MTU-limits in the IP layer instead of the
device layer. That allows loopback and pure IP to actually start
observing MTU limits, which was hidden away in the Ethernet
implementation.

This sets the stage for IP layer to fragment packets when needed.

Original-Fixed: 371593049
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134814
Original-Revision: 99598c498a969cb84f4670e1022bb1413fca1f3f
GitOrigin-RevId: d43ee5c9bcf27596aef89625247ff7c2ac7a8c8c
Change-Id: I350e3748d4e08e7e63723b499a3819dedd49ce4e