[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/lib/ui/scenic/cpp/tests/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134581
Original-Revision: b50795a713298558a5b0dce965adc77c427b8c32
GitOrigin-RevId: 5b33620e7abceed0b63ac459f9602e759afdc14f
Change-Id: Ie9400fa88d7983b6895b176438fd45411bafa1c3