[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.observation.geometry/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134576
Original-Revision: 5d8ef03bf64e29327477f80d2cbc1ab5671bc26e
GitOrigin-RevId: a6f98867f7a18c560850954c057283ab20eb9daa
Change-Id: I10b024b1132d46df4d5d4c1e77eaac2987e0d460