[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.observation.test/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134594
Original-Revision: c0a2c8e393baa6884549a75c4e30c563a7cf423c
GitOrigin-RevId: adfc74ecaef7ac8efc4a5e0744dd479274c9fbac
Change-Id: Ib4aab901b9369bffeb2ea5fdd2065ad1ec21edb6