[roll] Roll fuchsia [bazel_sdk] Do not depend on C++ toolchain for preprocess_file()
Before this CL, the preprocess_file() function required the rule
to depend on the current C++ toolchain, which made it impossible
to run in the host build configuration when no such toolchain
is defined.

This CL changes the function to use @fuchsia_clang//:bin/clang-cpp
directly instead,

On the other hand, the dependencies of fuchsia_devicetree_source()
and fuchsia_devicetree_fragment() must be reachable through a
transition to a Fuchsia-compatible build configuration, as these
can be cc_librrary() targets that need a valid C++ toolchain
definition.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135772
Original-Revision: b47462f4e7694353b18139851f348f664d31791c
GitOrigin-RevId: 369561cab0942934ace2399a83a6f23812988d64
Change-Id: I689a6e77a209d0b2a644e1ada35afa6b4f47f726