[roll] Roll fuchsia [ffx][repo] Use correct repo name with product bundles
This uses the generated name from a server that is serving
a product bundle. Before this change, the target register
command could not find the repo server instance that is
serving a product bundle since the name is prefixed with
the repo_name option, and uses the alias as the suffix of
the name.

Original-Bug: 372026735
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133672
Original-Revision: 0efe16c5874444de31a8fc4892bc95de896d9be7
GitOrigin-RevId: 832ee4ee44760d7aa6ab618afc0652ff20be6d8b
Change-Id: I8b3e588a190e7fe8bb50b841ef61c65d02df2c69