[roll] Roll fuchsia [sdk][driver] Add ScopedGlobalLogger to partner
Adds the ScopedGlobalLogger to the sdk. This library is separate from
the regular testing library as it is meant to be used in tests that
don't require setting up the entire driver, which is what the regular
library helps with.

Original-Bug: 369120118

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127472
Original-Revision: 4a4acd035e0aca7994cc380d7a682aa057ea6137
GitOrigin-RevId: 9c7ef04f0ffa97f7178d451efce1e48f41a82985
Change-Id: Ic39f4339a409020d44227f4eedde638c6d01a5e4