[roll] Roll fuchsia [fuchsia-async] Fix timer memory issue
Prior to this CL, defunct timers wouldn't get removed from the heap
until the original time expired which meant it was possible to end up
with excessive memory use under certain usage patterns.

This change now uses a heap where elements can be removed.  There are
also a number of other optimisations and changes:

1. Timers will now usually avoid heap allocations. A heap allocation
   might be required if the heap needs to grow in capacity, but that
   should not happen very often.  A side-effect of this is that timers
   are no longer Unpin, and timers aren't registered until they're first
   polled.

2. Timer::reset now works no matter what state the timer is in.

3. Timer now implements FusedFuture.

Note that an earlier version of this CL tried to use BTreeSet, but that
caused significant code size increases.

I made some of the callsites Box::pin where it appeared to be difficult
to pin to the stack (which won't be a regression because prior to this
CL all timers caused a heap allocation).  Owners might want to revisit
this to remove the boxing.

Original-Bug: 370598401
Original-Bug: 371152025
Original-Bug: 371434567
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130312
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: cf3e8d891968ab17dd6c8235c4e402ae12c224ab
GitOrigin-RevId: 567cf7afd4361ac01e6d0d2cc605eb013814a35c
Change-Id: I5102b671bca6cb7576627cdbf452cb8d66cd3bea