[roll] Roll fuchsia [fshost] Launch storage-host dynamically
This is necessary to support multiple instances of storage-host, which
can happen when a ramdisk image is present in recovery mode (and looking
forward could happen if multiple GPT-formatted devices are present).

There is still a singleton storage-host instance associated with
FshostEnvironment which is the one that actually publishes its
partitions in fshost's outgoing directory.

Original-Bug: 339491886
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124776
Original-Revision: 118d03ca696da5a8606935369f40023b06b59d40
GitOrigin-RevId: 93f714a54efb4030629c8c5310565f63aa86b275
Change-Id: I0e8897b74c09c97c409bfc0bef8e1a4932ac4826