[roll] Roll fuchsia [inspect][ctf] Use default publish settings in Inspect tests
This makes the CTF Inspect publisher test use default PublishOptions.

Sometime ago it seemed better to me to not encode our defaults into CTF
tests. However, with the addition of name-filter-lists in selectors, I
now think it is better to have CTF ensure that if you
  1) Publish using the default Inspect tree name
you will always be able to
  2) Select from Archivist without specifying a name filter list.

Before the use of name-filter-lists in selectors, the tree name was
mostly an internal detail. Now it is part of the selector API, so
ensuring these behaviors stay consistent is important.

Original-Bug: 355732696
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129876
Original-Revision: abd02cd9c0389ae3d3169a5c92c17cad14ea491f
GitOrigin-RevId: ec460d3159bef752d7d1cc620371c06e3bd98366
Change-Id: I419c5d0a753f154a2a4c2afc52a0074043907712