[roll] Roll fuchsia [superproject] Roll antlion WLAN Policy-initiated roam test suite.
Tests that a DUT roams (or does not roam) as expected between two BSSs.
This suite connects the DUT to a BSS, then starts a second BSS, then
slowly attenuates the radio power on the AP for the original BSS. A DUT
that is configured to roam will notice that the original BSS is no
longer the strongest signal and will scan for a roam candidate. If the
roam candidate (the second BSS) has a stronger signal and a compatible
security configuration, the DUT should roam to it; if the roam candidate
has an incompatible security configuration, the DUT must not roam to it.

This suite does not test all permutations of incompatible security
configurations, but it is a decent first pass. It should also be
adjusted to use a hardware atttenuator instead of attenuating via
software. And finally, the suite endeavors to detect situations where
Policy disconnects and connects instead of attempting a roam, but
without a solid mechanism to detect intermediate connection states,
Antlion can't currently detect whether the DUT roamed or if the DUT
disconnected and connected. In its current state, this suite is still
quite useful for bench testing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1130812
Original-Revision: f80275bf31d552c7808ca13738f978751007c1a4
GitOrigin-RevId: 39b443841afb1da45c9c708736b2bdb5659c08d4
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136307
Original-Revision: 46df808bf86754c15dcefab2765bdf14fe137db1
Change-Id: I0b6995678327ed98071d4ee9584da8245eb6cbb7