[roll] Roll fuchsia [tee][api] Base handle types on u64
Since it will be convenient to model object handle's in FIDL (for the
purposes of a persistent storage service), it is more convenient to base
handle types on u64 over usize (as there is no FIDL integral type with a
Rust binding of usize). This change is cosmetic in practice, as usize
and u64 have the same ABI on all supported architectures - and the
ABI compatibility continues to be verified in the existing tests.

Original-Bug: 371213067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134731
Original-Revision: 4c5bd67993e9afade8b60ba45e334cd847fb41fe
GitOrigin-RevId: 153ced866c8b4f221fc0a1a3b2487dc35c21bfb7
Change-Id: Id6241f989031a5c844ae89a6b72eda1198cec4fa