[roll] Roll fuchsia [debugger] Move exception record to Exception.
Previously the function to get the debug_ipc::ExceptionRecord was on the
ThreadHandle. This matches the way the Zircon objects work where this is
a property of the "thread" kernel object.

But the exception record is conceptually more associated with the
exception object. This change may possibly reduce race conditions
because the exception record will be queried more closely in time to the
exception delivery. More relevant is that I plan in a followup to add
more POSIX exception information to the exception record which belongs
on the exception object.

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132532
Original-Revision: 2258a357724693e01528286891b7479236283db6
GitOrigin-RevId: 3d64f5264a463c9180f8ce5da83a1d38997484b8
Change-Id: I9a4b8c973b2782e81cced6bbd655e2a18b57a450