[roll] Roll infra/recipes [checkout] Fix changed files analysis on non-main branches
The old `changed_files` logic to determine files affected by a CL in
presubmit was incorrect. It compared against the `main` branch unless
the `is_release_version` was set, but that's only set on postsubmit
builds that run against tagged release commits, not on presubmit
builders.

The correct behavior is to determine the base commit by comparing
against the head of the target branch after rebasing the CL (and any of
its unsubmitted parent CLs) on top of the target branch.

The target branch can only be retrieved via the Gerrit API – we already
resolve it during the checkout, so I added some caching logic to ensure
we don't do a second Gerrit API request if we already have the data we
need.

Original-Bug: 373434901
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1137534
Original-Revision: cb2badcdde2620ae2fba2ca85f4041cc794bbf76
GitOrigin-RevId: 9a8ba96fad1e31f4740cbbc93616b9c92db4aa25
Change-Id: Ia94d9690df414957b59a31fd22a73e3edc85080b