[roll] Roll fuchsia [amlogic-display] Get interrupts using interrupt names.
This replaces GetInterruptById() pdev calls with GetInterruptByName()
throughout the amlogic-display driver, so that the driver won't rely
on the interrupt indices defined in the board driver or the devicetree.
This makes it easier for boards to add, remove or change pdev-provided
interrupts.

Test: On core.astro, core.sherlock, core.nelson, core.vim3:
      - "display-tool squares" worked.
      - "ffx inspect show" showed correct rdma stats.
      - "flatland_display_compositor_pixeltest" (using capture) passed.
Original-Bug: 372259382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133473
Original-Revision: d26eab099075c0a5ba0033bf72955783d303e8fa
GitOrigin-RevId: 8edb2188f6794444e574e8bb4b778e81a80b5e41
Change-Id: I1126ef2508b75c835d56f3ab7854dbea69495733