[roll] Roll fuchsia [owners] Remove jbauman@google.com from src/sysmem/tests/sysmem/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136395
Original-Revision: 58633ebdd3bf4258e8eb0d79e3cdbb8c9f93863b
GitOrigin-RevId: 61f5e3ffb8ab1f3e2fe94751616661611b850666
Change-Id: I879271c61fc599c8bd527fccc9c9e00128a5e4f6