[roll] Roll fuchsia [owners] Remove jbauman@google.com from src/sysmem/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134830
Original-Revision: 24f8cfa55b449a98ae6c13438c6061cce5ee9393
GitOrigin-RevId: 61bf8255744d50badb362002678ca8c57b7e33a5
Change-Id: I8d5a9b4fde71a785c455ba1025185fb1efb3ee09