[roll] Roll fuchsia [owners] Remove jbauman@google.com from sdk/fidl/fuchsia.sysmem/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134829
Original-Revision: 26de8cd59ce057c3ec5864737becae380b432b0b
GitOrigin-RevId: 49f653057350523e4ddc5a85097588cc2c866812
Change-Id: Ie7fc7a86acb93e7604f08fe0ac61ce2260b7cf1f