[roll] Roll fuchsia [fidl][style] Make FIDL request / response parameter documentation inline.
The previous FIDL Style guide required all request / response parameters
documented at the bottom of each method using the following format:

```
+ request `param1` <description>
+ request `param2` <description>
- response `param1` <description>
- response `param2` <description>
* error <description>
```

This was written in 2019, where both C and HLCPP bindings made requests
and responses as function arguments, and such documentation style was
helpful to map param names to function arguments.

Now that all the FIDL requests and responses are encapsulated in structs
or tables, and both the new C++ bindings and the Rust bindings put
requests and responses to structs / classes, the old documentation style
makes it difficult for readers to get parameter documentation in IDEs.

This change modifies the FIDL style guide so that request and response
parameters are documented inline in their corresponding structs. Errors
still needs to be documented in a separate subsection in FIDL methods.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093588
Original-Revision: c782fa2a7fcb17d5a7dbd45ed98853abda1c9613
GitOrigin-RevId: fcafab5772521f35928db9c66f7f5d68ae8216af
Change-Id: I0d8f6e8ec35f3e3feb5fc786dad15dec5eee82aa