[roll] Roll fuchsia [driver_manager] Parameterize DriverRunnerTest on using dynamic linking.
We want to run each test case on the dynamic linking path, and the
legacy path.

DriverRunnerTest2 is added as a temporary fixture so we can make the
fixture changes required over several CLs. This is as calling GetParam()
will crash if a test has not been migrated to TEST_P, but we don't
want to use the dynamic linking path if part of the test is not yet
supported.

Added the driver-runner-test.cml, so we can add the ability to
create processes that is required by the dynamic linker code.

MULTIPLY: driver-runner-test.cm : 50

Original-Bug: 363012744
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136298
Original-Revision: b2e35d1e78075a68a6e540bbd85008289c58db0a
GitOrigin-RevId: 26562f99101d61cbc46eb1501551d84e5eaf1cb0
Change-Id: I58a2b5419950689cfea593f76ffe042596737ef7