[roll] Roll fuchsia [starnix][ltp] Pass utsname03_clone
This test requires the CLONE_NEWUTS flag. We actually had all the logic
for this flag, but we didn't list the flag as supported, which meant we
returned ENOSYS and failed the test.

In addition to adding CLONE_NEWUTS to the list of supported flags, this
CL cleans up the implementation to match the normal naming conventions
for cloning and forking these kinds of objects.

Test: utsname03_clone
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137816
Original-Revision: bc5986fc6aecc4ec5be318d5ba177bcc4dc8add3
GitOrigin-RevId: 2e72959d9f20064ff1730b89455305cba55c4e2c
Change-Id: I3f92a4d7ab07787a0942549fe52950c0f69afe2f