[roll] Roll fuchsia [debugger] Use DebuggedJob for RootJob
This converts the handling of RootJob to be considered just another one
of the jobs that DebugAgent can attach to, and adds a few methods that
makes it easier to generally attach to jobs using the new DebuggedJob
object.

All notifications that we rely on coming from the root job should be
unchanged.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136252
Original-Revision: 07da5d71ecf97e777199bbfc8f053e06961a7ea9
GitOrigin-RevId: 5ea821038df656b559640ec31bcb29240b0a1b32
Change-Id: I3d5d2111dd283ea8aba8917e7058e5db8a31dcd5