[roll] Roll fuchsia [amlogic-display] Get MMIO regions using MMIO names.
This replaces GetMmioById() pdev calls with GetMmioByName()
throughout the amlogic-display driver, so that the driver won't rely
on the MMIO region indices defined in the board driver or the
devicetree. This makes it easier for boards to add, remove or change
pdev-provided MMIO resources.

Test: On core.astro, core.sherlock, core.nelson, core.vim3:
      - "display-tool squares" worked.
      - "ffx inspect show" showed correct rdma stats.
      - "flatland_display_compositor_pixeltest" (using capture) passed.
Original-Bug: 372259382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132519
Original-Revision: e756e7ccbed1ab5d1355654dfdef11e69b2ae48c
GitOrigin-RevId: 0292662835190ebfc597d8c0b4d0c556a370be40
Change-Id: Ia6da5d4cb3cf7efc947f44d74e4ff946aec31d03