[roll] Roll fuchsia [driver][sdk] Add ServiceValidator
Adds a new sdk class called ServiceValidator. This class is able to be
provided to the incoming namespace optionally. When provided, the
namespace will perform a validation of the service member connect
operation against the service validator. This can help terminate early
if a driver is connecting to a nonexistent optional node parent, or
connecting to the wrong parent for a specific service.

Original-Bug: 371393843

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132775
Original-Revision: 0d7c61a2a839c14c90b2c777ad93d5fc46ca56f4
GitOrigin-RevId: 9ae2c8a52389528cb2ae64d01567815fff833c05
Change-Id: Ifd47d3688250a12599ec7aace81c9dd43a5f6c45