[roll] Roll fuchsia [starnix] Fix peer-closed logic in proxy
Previously peer closed was not properly handled, and we could end up
with threads just spinning in the wait loop.

The proxy loop code was relying on the message forwarding to return an
error if the read channel was closed, but before reading we checked
whether or not the channel was readable. This meant that the peer closed
was never encountered.

Original-Bug: 371566166
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137133
Original-Revision: 0e3b6a5a6fb519ef13af882e3908f0a604ea35f9
GitOrigin-RevId: b22cb39420d951a47f7447b2719a40ab9d614b42
Change-Id: Ie3d3f96a25ebbf9155b37f527dfdd81b95f2cc81