[roll] Roll fuchsia [owners] Remove dworsham@google.com from third_party/flatbuffers/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134674
Original-Revision: 88b77d49ea61d274f6062b2ef48503b9b1f5d29a
GitOrigin-RevId: a422f2fe94a68621efbfc9c7458c2533c7385d89
Change-Id: I2fed00b7a900cd0101d2e91b8faf5d79e5404a81