[roll] Roll fuchsia [tee][api] Clarify undefined behavior of memory functions
The spec does not define the behaviour of TEE_MemMove, TEE_MemCompare,
or TEE_MemFill when the provided pointers are null (ditto for the
related libc functions). However, it is a precondition of the Rust
standard library functions that we implement these with that these
pointers are non-null. For clearer panicking and better interface
hygiene, we add non-nullilty asserts before these calls.

In the case of TEE_MemMove, we actually go further by treating a size of
zero as a no-op, even when either of the provided pointers is null! It
turns out that there are xtest cases that explicitly rely on being able
to do this! Given that, and the fact that this isn't against the letter
of the spec, we deem this behavior as sensible for our implementation.

Original-Bug: 371213067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138792
Original-Revision: 330b395363b7a8ad57fb7232b9ab85272e822c80
GitOrigin-RevId: b458ba9e4e96eb44216c8d2b7b72259b73ce6605
Change-Id: I49949341d59df865f7e2f18f529c7d2f41b341fc