[roll] Roll fuchsia [sysmem] add clarifying comment re. allocation
Add a comment clarifying why we don't wait for old collections to be
fully gone before attempting allocation of a new collection, and point
toward AttachLifetimeTracking in case a client wants to wait for any
old collections known to that client.

The docs are also separately updated to clarify this as well.

Original-BUG: b/42150211
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136614
Original-Revision: b14d4fd5081a17355290c59d8204360bdffe0298
GitOrigin-RevId: 2028a0e97152e5c5045554de60ae6e188d045fc7
Change-Id: I84fb1bea4f723a1b78ea085cd498fcf3f7a060b7