[roll] Roll fuchsia [sestarnix] Add Starnix kernel runner option for fs seclabels
Allow an SELinux Security Context to be specified in the manifest
of a Starnix component, for the Starnix container to apply to all
files in all filesystems mounted for that component.

The new "fsseclabel" argument may now be provided in the "program"
section, alongside the existing "seclabel" argument that controls
the label assigned to the component's initial task.

Original-Bug: 373819089
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139852
Original-Revision: c48f3d97a9e3e7a6bf72c2fe9d4d6c86e3696458
GitOrigin-RevId: 38a5e392114763c6de57861a5dbcd8110eb3570d
Change-Id: I331b1125e507e87e9b910e0373cf1869f5e8709e