[roll] Roll fuchsia [netstack3] Remove rules after table removal
This change prepares for installing rules in Core. In Core, a rule
can hold a strong reference to the table, so we need to make sure
when a table is removed, all rules that reference it should also
be removed.

Original-Bug: 357858471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139580
Original-Revision: a2ce2763cf9090631d873c2b952ddacc2e8ea9d4
GitOrigin-RevId: a3d6f7b55fddde0875c1d0ad4e80f80b71e72409
Change-Id: I9760b9e40e48d8d8e01c0551d03eb73248032c8f