[roll] Roll fuchsia [fdf] Create a quarantine node shutdown intent
This intent will put the node into a quarantine state after it fails to
start a driver, and stops+removes the failed-to-start driver.

This way disabling the driver that had failed will cause it to go
to the orphan list as part of the rebind, and then be available for new
drivers to match with.

This also makes our ffx tool print out "(quarantined)" with the driver if
the node is quarantined.

Tests: Created a new integration test, 'replacefaileddriver_test'

Original-Bug: 370048627

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127478
Original-Revision: d05d91e92cc16851d97c310e4cf74c02a195049a
GitOrigin-RevId: 789cf0998b38553a860f3cb66dc6277c8e67b9d5
Change-Id: I8fc92d6007fda54d07d58da58fe6d35d71ce3451