[roll] Roll fuchsia [profiler] Tolerate ZX_ERR_BAD_STATE when working with tasks.
This error code typically means that the task has already exited and
isn't able to be attached to the sampler or receive an exception
channel. It's common for this to happen with short-lived tasks
and is hit much more frequently when doing system-wide profiles.

Original-Bug: 359658592

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139576
Original-Revision: 81881f897147660bd0bcb6a7148b2ca9db62c653
GitOrigin-RevId: b81567b6f83d446e9892fddc160f7be23a380186
Change-Id: I776e4114349d2192aafdabcd488df03af2cfb260