[roll] Roll fuchsia [debugger] Add job-only flag to "attach"
This adds the ability to specify job-only filters from the "attach"
command. On Fuchsia, this will cause any matches for this filter to
attach directly to the job of any matches. Used in conjunction with the
"weak" option, this attaches to a job's Debugger exception channel. When
"weak" is not specified, this attaches to the standard exception
channel, putting us on the path to receive exception notifications
without being attached directly to the process that caused the
exception.

Specifying this option on linux will result in an error.

Original-BUG: 362240229
Original-BUG: 362240382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1139994
Original-Revision: f55990a85fd3e0c2a748c80886297c96a42332f9
GitOrigin-RevId: 0995e9ad5b52a02b51ea4cf4684457bb5a98976e
Change-Id: Ifc7f7b599a18c62abfd2c419fcf866971da325cf