[roll] Roll fuchsia [starnix] Always clear kernel signal on adb read
Normally for hanging gets we don't raise the signal until the next
hanging get has been scheduled. For adb reads we don't know when that
will come, and if the connection gets reset between reads the proxy will
get stuck waiting for the kernel signal.

The current code still mostly worked because the hang would not block
the new proxy that was created for the re-established connection.

However, if the connection was ever reset in this state it would cause
all future suspend requests to succeeed but immediately get woken up, as
if an adb command had come in.

Original-Bug: 373736681

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140792
Original-Revision: 4799890454edd749a2ef874653e5cc1b6d370987
GitOrigin-RevId: da1fada332ac8b23938e62184ae3233879c75683
Change-Id: I611226ccd847cd09f818726a9381defa710492a2