[roll] Roll fuchsia [drivers][rust] Improve incoming ergonomics
Implements library support for connecting to protocols and services in
the driver's incoming namespace more easily.

In the process, this adds some new functionality to `fuchsia-component`
libraries to connect to a service at an arbitrary root directory.

It also changes the `IterablePath` trait to use an impl return trait
instead of a boxed dyn in `cm_types` since that's possible now and
wasn't when it was originally written.

Original-Fixed: 373918951
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136552
Original-Revision: 05ecaa3913e34904e23717605390af7d21f2e3c0
GitOrigin-RevId: 863c3b0b6060ff7920f22595313c9539f51696b4
Change-Id: I4877c22cd5f692062b74bbf57bda5adcea7f0ed8