[roll] Roll fuchsia [thread_sampler] Handle attaching to Lifecycle:INITIAL threads
Right now, the kernel thread sampler will refuse to attach to threads
that are still initializing, dead, or dying, returning ZX_ERR_BAD_STATE
for all of them. This makes it difficult for a user to distinguish
between trying to attach to a thread that will not give samples, versus
attaching to a thread that will not give samples /yet/.

This allows the sampler to enable sampling on a thread that's not yet
running so that when it does start running, it's eligible to have
samples taken.

Original-Fixed: 360927412
Original-Fixed: 374109518
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140212
Original-Revision: aa2a2c2a283568f6d67cb3b74470b9bb94c2cb1a
GitOrigin-RevId: 55d69f73b13404645809d19ea7607de0450b8597
Change-Id: I2bc554b3065b0223e020b541d4b6bba33ab00280