[roll] Roll fuchsia [ffx][playground] Introduce DeferredError
This is just a little tool to aid error handling in the parser. It lets
you suppress error handling for one part of the parse, but then restore
the handled errors later if other paths fail. Suspect this will be
used more and more as we improve error handling.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098013
Original-Revision: 13478efd0c4376db915367f86e9e4e1f3d75313d
GitOrigin-RevId: ab26e1eebd78409646d8ec24b4745fe5e0be4507
Change-Id: I21c7afd1d2696c04eb0ecddfbb467aaa5d255a79