[roll] Roll fuchsia [flake][driver-host] Fix Start_ReturnError flake
This was flaking because in certain instances the driver's shutdown
handler where we quit the async loop, was running earlier than the
ResetQuit happening on the test side, where the quit there was to signal
completion of the start async call. This changes the start to use a sync
completion rather than the loop's quit so that we don't reset the quit
state being set from the shutdown side.

Multiply: driver-host2-test

Original-Bug: 374186610
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140643
Original-Revision: 5c724312d14896b876b73b60d5319b7fc9434e98
GitOrigin-RevId: b29299f4f07aec8f4831a08119f055e7c4029f1b
Change-Id: I1a12d0e0530b0b67c225399f1bfea54c385a3861