[roll] Roll fuchsia [fidl][build][sdk] Limit SDK categories in documentation
Previously, any FIDL library that specified an `sdk_category` would be
included in the generated and published documentation if its `..._sdk`
target appeared in the build tree. Rather than relying on appropriate
use of the targets, specify which categories should be published. This
should not change the set of libraries for which documentation is
published as long as the `..._sdk` targets are only depended on by
targets in `//sdk/fidl/BUILD.gn`.

This replaces a condition that was redundant / always true.

Original-Bug: 366221510
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140108
Original-Revision: 0f48b3009c9d53f1fe938fd40c416604ca1b3093
GitOrigin-RevId: fe7f01fdabacf582dad1dce0c0cdfbeb1f319415
Change-Id: Ib9399585707009dfe6715e642d4b4b9e49ac1457