[roll] Roll fuchsia [ffx][ota] Adding product bundle option to target update
This allows the specifying of a product bundle when running
`ffx target update check-now|force-install`.

By default, these commands communicate with the registered package
server, which may or maynot be serving the desired update package.

This feature allows the user to specify to use a product bundle to
provide the update package. This is either done with an explicit path
to a product bundle directory, or if the path is not specified, the
path is retrieved from the configuration using the key `product.path`.

The implementation of this feature then starts a package server,
registers it with the device, then performs the update.

Original-Bug: 364376140
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130267
Original-Revision: 9adfeefdf39f4170f3f588a4eaa21c2d0936665f
GitOrigin-RevId: 0335a7db8b52070c74a5d8a55ebb37e2f081174a
Change-Id: I151b61504fb9ac6381596abb36ae0b8eaf742b22