[roll] Roll fuchsia [kernel] Introduce power_management::ControllerDpc
This is a workaround for escaping scheduler context when requesting a
power level transition from the Cpu Driver.

The Scheduler provides the storage for both the Dpc task and the Timer
object, such that the objects are guaranteed to outlive the timer and
the DPC. This means that whenever a Dpc is queued, the scheduler wont be
able to request a new power level transition, thought It may choose to
just overwrite any existing pending transition.

When the Timer fires, it will queue a Dpc task, which when executed will
take ownership and reset the pending transition from the scheduler. This
means that until such moment happens, the scheduler is free to
update/change the pending transition or even cancel. The scheduler can
verify whether a transition was posted or not, if they have pending
transitions when it needs to rescheduler again.

The operation mode is when the scheduler decides to switch to a given
level it will:
+ Stash the transition provided by PowerState::RequestTransition
+ Call ControllerDpc::NotifyPendingUpdate

Original-Bug: 330349572, 347701043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137756
Original-Revision: 0721ff65e68227f939239751414e23a6c2dbc5ec
GitOrigin-RevId: aec9168bffd1a68c0c8ea71e42f196a44e390a8e
Change-Id: I7055b634cba0f931293eeeb727dce7c692a22e1f