[roll] Roll fuchsia [audio][adr] Wait for initial device detection
Ensure that preexisting devices are detected and initialized
before responding to `Registry/WatchDevicesAdded`.

In practice this is not a requirement today because the
ADR is eagerly started on products where it is included.
However, this will unblock certain testing scenarios, and
allows ADR to instead be started on demand if desired.

Original-Bug: b/359344835

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141292
Original-Revision: 32878c6d4a4043dd3d33bbc8d5c61142dfb37a71
GitOrigin-RevId: 545b6bc481ab38310d9a53a618e18ad63a01d449
Change-Id: If593f930bdaddee258a42a8f0d9d23f68934d91e