[roll] Roll fuchsia [netstack3] Retain conntrack entry across loopback
Currently all packets look up their conntrack entry when they hit the
INGRESS hook. This causes problems for packets that are NATed and then
looped back: they may no longer match the original or reply tuples of
their connection, so it cannot be retrieved, which results in a new
conntrack entry erroneously being created.

Instead, hold on to a weak reference to a packet's conntrack entry in
the loopback RX and TX queues' per-packet metadata. This allows
looped-back packets to hold onto their conntrack entry on the way out
and for the IP layer to retrieve it on ingress, and avoid allocating a
new connection. This also happens to allow us to drop packets when
finalizing a connection fails (https://fxbug.dev/372543386), something
that, without this change, is not possible to do without breaking NAT
over loopback.

Original-Fixed: 372545312
Test: netstack3-filter-test
      netstack3-core-ip-integration-test
      netstack-filter-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137754
Original-Revision: 0ee115aa26333643e1a51daa7d45aa2919a9f46f
GitOrigin-RevId: 50e754329f5a0c7047c238f98871f4084556a913
Change-Id: Ic44c5cac2869191098434b64836fff0f90587dde