[roll] Roll fuchsia [netcfg] Remove masquerade config on close
Ensure that when a `fuchsia.net.masquerade` controller disconnects, it's
masquerade config is removed. This upholds the documented behavior on
the API, that the "protocol is associated with the lifetime of the IP
masquerade".

While here, downgrade some `error!` calls to `warn!`. If disconnect
races with interface removal (which happens from time to time in our
integration test teardown), the disconnect logic will observe `NotFound`
when disabling the masquerade configuration. With `error!` this lead to
spurious test failures.

Multiply: policy-management-integration-test

Original-Fixed: 374285849
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140772
Original-Revision: 3096fdb04e46433d97dd2b4568204847293db837
GitOrigin-RevId: 1f59078d8ad57a5de361262ab0f7cb5cca2d0820
Change-Id: I3cc1aff4feba8f7b4a26f36f45fe904a2be7a7bd