[roll] Roll fuchsia [zxio] Remove attribute equality checks.
They had a few small bugs in them due to missing fields that have since
been added. It was only being used as part of attr_set to ensure that
extra "has" flags weren't being passed in. This deletes a long chunk of
human-managed code that we keep forgetting to update properly and
replaces it with an iteration over the struct of bools as a pointer
array.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135513
Original-Revision: 1195ec8dbbecc73948ed789d1d3a8e094dff9143
GitOrigin-RevId: bd12b231f3d87de999cf9be139683c54d01fedc3
Change-Id: I829faab892fc88ab4737377a4d36f04928d91cdd