[roll] Roll fuchsia [starnix][vfs] Restrict xattrs outside of user namespace.
We previously restricted access for non-file and non-dir types for the
user namespace, and restricted access to the security namespace. We did
not prevent others from using every other namespace which meant that
they could access them all.

The real intention here is that symlinks are used with permissions 777
which would allow anyone accessing a symlink to attach xattrs to it.
This is the reason the "user" namespace is only allowed on regular files
and directories, it presumes that all other namespaces require special
privileges.

This will notably prevent use for non-CAP_SYS_ADMIN users from accessing
the "system" and "trusted" namespaces that may often be used, but I
don't think we're supporting anything that does that right now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143212
Original-Revision: 32d64b9699268a427ff69ac49a7750d7997fd194
GitOrigin-RevId: d205cbd94ba9bde69c7f969b8e09e0ec2c085e3d
Change-Id: I7c566352b989306b95687e4a2de2a42548eb680b