[roll] Roll fuchsia [netstack3] Allow conflict of identical connections
Currently, when the netstack attempts to finalize a connection that
conflicts with an existing entry in the conntrack table, finalize fails
and the packet is dropped. Usually this is not a problem because packets
that are part of the same flow are typically handled sequentially.
However, it's possible for two packets from the same flow to race, e.g.
if they are arriving on different interfaces. In this case they can both
end up creating an exclusive conntrack connection and then whichever
packet loses the race to finalize will be dropped.

To mitigate this scenario, when detecting conflicts, only report a
conflict (and drop the packet) if there is a conflicting connection that
is not identical to the one being finalized. In the case that the
connection to be finalized is identical to the one in the table (in
terms of both its original and reply tuples, *and* its NAT
configuration), adopt the already-finalized connection from the table
and allow the packet through.

Original-Fixed: 375012282
Test: netstack3-filter-test
Multiply: netstack-socket-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143973
Original-Revision: 190555cb4a1f874a1f2422b41f09796ba28dc21a
GitOrigin-RevId: aa02b8e2357ae461126567e45fefb5d9c4dac883
Change-Id: I50510e0e1120133726ee680af6936ddb2ab8b204