[roll] Roll fuchsia [fdomain] Make futures disposable in client
This sets it so that most operations on handles that require an `await`
will still complete even if you throw the future away, so long as you're
polling the client's main future that was created when the client was
created.

We still require awaits for creating new sockets and channels and a few
other "global" operations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121665
Original-Revision: 1a88da5a979dcb8edbf16148b09b53f9d496457f
GitOrigin-RevId: bf6a6e60dd6e51f881763ee53170b9a3c4b63e12
Change-Id: Ieb1d705a4eac53db9c81731c1195822f8fd492f4