[roll] Roll fuchsia [wlanif] Remove trace log in destructor
We're seeing a race condition that results in a crash in unit tests
that seems to occur when we attempt to log on teardown. This log
statement isn't really useful, so we can remove it and see if it
fixes the test flakes.

Original-Bug: 373586951
Test: unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143912
Original-Revision: 68c53abab15e0c5340ceb39de405760d7f0bc88b
GitOrigin-RevId: d2946d487d9c25dde665d1cb3c663b830790350d
Change-Id: I6319be914f7c4f9bd9fcd60cbf4279c43896ecc6