[roll] Roll fuchsia [starnix] Add temporary component features for tests
These features will be present in the SELinux tests until we either
decouple the test and container cml files, or we add logic to the test
runner to modify the program block. This change does not do the latter
because that has caused a lot of confusion in the past.

Original-Bug: 371368307
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144872
Original-Revision: 98da53ff49c9ab0886c56aca9f6024d9e05b7d95
GitOrigin-RevId: 20a830bf5f5d8219543296ae7226fefd25c34c3d
Change-Id: I336de91fc600897142792ddf6fb096483e2aa46c