[roll] Roll fuchsia [test_list_tool] Fix handling file relative blobs
File relative blobs in package manifests are pre-processed by the
manifest parsing library to be relative to the working directory.
"working_dir" relative manifests are actually relative to the build
directory, which is assumed to be the working dir for package
operations. test_list_tool violates this assumption when run from fx
test because it is relative to the source directory.

While the relative path computation is technically correct, the
incorrect defaults were corrected by prepending the build directory in
loading. This change ensures that the build directory is not prepended
twice by skipping that operation if the prefix is already set.

One result was that any package using "file" relative blobs failed to
get the correct directory set when using fx --dir=<> test, which is now
fixed.

Original-Fixed: 375256332
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144841
Original-Revision: 9bfa3793a7d31955c9d95fbd631337831d3f2a4b
GitOrigin-RevId: bf4a4beec87c450d8f10599b2aa0fee5515fc6f6
Change-Id: I20f5375805613e415cf34e597b75a6a45984bdbd