[roll] Roll fuchsia [build] Work around std::__libcpp_verbose_abort signature change
This is the first of two changes to perform a soft transition.
The _LIBCPP_VERBOSE_ABORT_NOT_NOEXCEPT macro must be defined now
to keep the same signature the source code relies on after new
libc++ code rolls in with the next toolchain roll.  After the
toolchain roll, this workaround will be reverted and instead the
source code will change to use `noexcept` in the definitions.

Original-Bug: 374753100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143273
Original-Revision: fcea69a0f471979051ffe1b62a7215982a2885dc
GitOrigin-RevId: f84c72246c71808526cb155b270b66c3bcca8f58
Change-Id: Idd1fcf6269c65081b5b790616830a1f80c8efe6a