[roll] Roll fuchsia [zxio][netstack] Refactor UDP prelude constant sizes
The networking stack and client library (zxio) have to agree with each
other on the size of the UDP prelude messages passed in datagram socket
buffers. This constant must be available to C++ code in zxio and to Go
code in netstack. To arrange for this, there are constants defined in C
and exposed via CGo. This rearranges the definitions of these constants
into the following layers:

1. zxio defines the sizes that it uses as C++ inline constexpr variables
   in a udp_socket_private.h header. This header is restricted by visibility
   rules to be accessible only by zxio and parts of netstack.
2. src/connectivity/network/netstack/udp_serde relies on its C
   compatible definition for the routines exposed via CGo and contains
   static assertions that these constants match the values used by zxio.

As a result zxio no longer has a dependency on //src/connectivity/.. for
these particular constants and the zxio code can use constexpr variables
with the values available inline. This makes the values available for
use in constexpr contexts and provides a (minor) opportunity for further
toolchain optimizations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130269
Original-Revision: 32eb90cf69e283507622500c19139bdb4235161c
GitOrigin-RevId: b170f0aba825a28a58e6481f4110ed3e37946cfa
Change-Id: I0c15f5154a89272edff65f6d08e147fb47f55817