[roll] Roll fuchsia [starnix][remote-binder] Add ContainerPowerController
Add an API that a HAL can call to wake the Container.

The HAL has the ability to pass in a power baton which will not
be dropped until after the container has taken its power lease.

The HAL also has the ability to create a wake lock which holds
the container awake.

TEST: fx test starnix-tests
Original-Bug: b/372254780
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143313
Original-Revision: 780962651be77a02ab8e0ce5654ac93a09fb9436
GitOrigin-RevId: 47e0aa8908184c17562e8e3b79a8445b28e77e59
Change-Id: Ib8e91c91229ea3d8d6b5b3ed186127c9449027e1