[roll] Roll fuchsia [fxfs] handle multiple checksums per block in journal replay
When writing to a block on disk, checksums are emitted as journal
records regardless of whether that block has been written to before.
This means overwrite mode blocks that are written to multiple times
between journal flushes will have multiple checksums emitted for them.

In this case, journal replay should succeed if any of the checksums
match the data. This is so if later writes go through on that block, it
won't cause replays of earlier transactions to fail. While that would
technically be consistent, it would potentially lose a lot of data
because we would end replay earlier than we should have.

Original-Bug: 293943124
Test: fx test fxfs-tests fxfs-platform-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135552
Original-Revision: 36f16dfbafe057e9b6259696129d042e5fff54c5
GitOrigin-RevId: 492e3410dc80af731887277c0116f25765b326ca
Change-Id: Ieb46b348b6e1f3e1e206a5a813f877eb0be0c141