[roll] Roll fuchsia [kernel][arch] disable code coverage for early boot mmu code
For both arm64 and riscv64, disable generating code coverage on the
early boot mmu mapping code.

On arm64 in particular the compiler may generate unaligned stores to
the coverage array which will trigger an unaligned access since the
mmu is off and the data cache is disabled at this point.

For riscv64 it isn't an issue on current hardware, but a similar
condition may arise in the future so makes sense to disable it out
of an abundance of caution.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145432
Original-Revision: 7efa7ed5045b5f580f1790afb8659bbbca2ef349
GitOrigin-RevId: 472fc8622a89898371f327b037f62806a9758412
Change-Id: Iaad092dc30780b9b11cdfe3290f7b285f0e78e11