[roll] Roll fuchsia [remote-control] Host an FDomain inside RCS
This lets us connect to the toolbox via FDomain over a socket. It's only
available on device for now but it's a major part of getting us over the
network.

We precede each FDomain message in both directions with a 32-bit little
endian length for framing. This is documented in the Connector FIDL.

Original-Fixed: b/369883400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130656
Original-Revision: e0837ead5bfc4a34b8647b8fc4c751673234f33c
GitOrigin-RevId: 993d71d9390e52ca4c502d48630453cc2e3a600b
Change-Id: I941f4e71dd0991d97ab847710d9670f16abb67ae