[roll] Roll fuchsia [debugger] Fail loudly when symbol server auth fails
Previously, it was up to users to notice that their symbols were
missing, know to type `sym-server`, and also know how to interpret the
various states of the symbol servers.

With this change we remain silent in the good cases, but fail loudly
and give instructions on what to do in the common authentication failure
paths.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146874
Original-Revision: 4b6aa251001516afd69f9c4a3af21f00265182a1
GitOrigin-RevId: f52d2bf2b4c9b00bb83a5223aab74482d57c529b
Change-Id: I4e773a4f4873f35f230194c65bf5d29f709ed2ef