[roll] Roll fuchsia [userboot] Fix multi-process/boot entry tests.
When userboot launches a process it will use the 'test' entry result to
determine sucess. This is fine 99% of the time, except when we have test
that involve being launches as a boot test, `userboot.next`.

When this happens, if there is a stub (multi process test) or there is
just the boot program performing a test, the result will either always
be sucess (default int value) or the stubs return value (0).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143214
Original-Revision: 1eaf13fbc1624762b23fc672b321c7d93b379fc5
GitOrigin-RevId: 0399a41a927d8d0c0d4a6191298fab15212df5c3
Change-Id: I067dcd0c5a7a4b7c17f551e12cec0c3abbc2c815