[roll] Roll fuchsia [drivers][build] Only set assert_no_deps if needed
Previously, `assert_no_deps = []` was set regardless of whether a value
would be appended. The fact that it did not clear values forwarded from
the `invoker` is subtle and refactoring could change this. Use a safer
pattern closer to where it is needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146474
Original-Revision: 486e3329105566f38be8c75a255c66960e4af46e
GitOrigin-RevId: ae342dafff07f59e359441121f89dc5d958efac9
Change-Id: Ia6098fbb5ef295182b11c622e104259ed1095a4c