[roll] Roll fuchsia [fake-pdev] Use rvalue ref for SetConfig
This will help ensure better errors are output when the user accidently
supplies and lvalue instead. Normally we would just delete the copy
constructor and implement the move constructor, but that's not helpful
here as we lose aggregate initialization for the config in that case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145961
Original-Revision: 39160a7195053589a81f6954e6d962d534c6b649
GitOrigin-RevId: 9d0eb81bed12321be695eb05a4aa458b42c2369b
Change-Id: I61fb0bad174a68f0859d856b74362b9f7f11d41b