[roll] Roll fuchsia [kernel][rppm] Fix `zx_system_set_processor_power_state`.
The syscall itself is fine, but this expression here is wrong.
The bucket is INCLUSIVE, but the expression is exclusive. So we get to
the point where we need to set the power level and do nothing.

E.g.
With 10 CPUs for example, and a bucket size of 64, we get :
10/64 = 0. That is bucket 0 contains all CPUs, but the expression above
would prevent from ever visiting this bucket.

Original-Bug: 347701043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146753
Original-Revision: 163ce7190604a5b4f1d245b76dfc67cca35c347f
GitOrigin-RevId: 0d94628926c291e326cfb1e00d7dee7920b2d2e8
Change-Id: I778c1ee2dcecbe2024cb7cefdca72f2d26dbcfd4