[roll] Roll fuchsia [netstack3] Refactor NAT verdict types
Currently ConfigureNatResult is a struct that holds both a verdict
(typically either Accept or Drop) and a boolean `should_nat` value.
However, the latter is only meaningful in the context of an Accept
verdict; if the packet is to be dropped, its NAT configuration doesn't
matter. Invert the relationship between these two types such that
Verdict::Accept holds a value, which, in the case of NAT configuration,
is a NatConfigurationResult enum that indicates whether NAT should be
performed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145959
Original-Revision: c9b9aae19a6ef59f268ec17de7892184d71bffae
GitOrigin-RevId: 89eb79592d52ebfd66f45655a0bdf0a8ba789d01
Change-Id: I46e5013ab311f3d6a129f0bac25cf7422e2b41f9