[roll] Roll fuchsia [components][sandbox] Add DirConnector
A [DirConnector] is like a [Connector], but it's specific for
`fuchsia.io/Directory`.

We'll use this type to replace `directory` capabilities in the
namespace, replacing `Directory`. `Directory` has a flaw in its
implementation of `clone` that is not fixable. `clone` calls`
fuchsia.io/Directory.Clone`, but this has two problems:

- It's not a strictly in memory copy: it requires a channel write
  syscall, which we'd prefer to avoid
- If the original Directory channel ends up closed, it will not
  necessarily give the correct result. If the original channel was
  closed with an epitaph, the cloned channel will always have
  PEER_CLOSED. This is caught by component_manager routing unit tests.

[DirConnector] fixes this defect because it is a directory channel
_factory_. That makes it trivially cloneable, like [Connector]. It is up
to the client to call `DirConnectorOpen` (`send` in rust) to
obtain a channel.

Original-Bug: 360221168

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132663
Original-Revision: 8a003a178a4204ceac5ea5015c5befa962aeac64
GitOrigin-RevId: eec182485368e9a1bdbc0e4d000fed6cb3c09ab6
Change-Id: I8f1bb6baf657836d6e635325b674e61240915060