[roll] Roll fuchsia [build][bazel] Support FUCHSIA_DEBUG_BAZEL_SANDBOX=1 in environment.
Setting this environment variable forces all non-remote and non-local
 Bazel actions invoked by the build (i.e. from a bazel_build() GN
target) to use the `--sandbox_debug` option.

This makes Bazel much more chatty but also preserves the sandboxes
used by the Bazel actions, which is useful to look inside when
things go wrong (typically when input files do not appear because
they belong to missing dependencies of the target being built).

This doesn't not affect the build graph or how the result of build
actions, so an environment variable can be used safely here, without
the need for a new args.gn build argument to enable the feature.

NOTE: This does not work for remotely-built actions, i.e. if
      RBE is enabled. It is however possible to make a specific
      target non-remotable by adding the following to its
      Bazel target definition:

```
  tags = [ "no-remote" ]
```

Note also that Bazel targets that use the `local` tag (e.g.
assembly-related actions) are not sandboxed and will not benefit
from this either.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136013
Original-Revision: 9dab56648a0c8415f812982166bfa42cda73201d
GitOrigin-RevId: f74cd6f93bbf4d4b48248462d3065b3c2c034f64
Change-Id: I5da9f83c17bfb5b937a040b11fa3dc8f5c6dcbde