[roll] Roll infra/recipes [symbolizer] Don't provide DEBUGINFOD_URLS that require auth.
The symbolizer needs to know whether it has to authenticate to a
given symbol server, and the DEBUGINFOD_URLS variable is assumed
to be a public URL.

On public bots we continue providing DEBUGINFOD_URLS, since the
symbolizer can safely assume it's an unauthenticated storage bucket.

On private bots we remove DEBUGINFO_URLS and rely on
SYMBOL_INDEX_INCLUDE added in https://fxrev.dev/1146849.
Removing the public env var should be safe for existing flows that
use llvm-symbolizer since it does not seem able to authenticate to
GCS anyways.

Original-Bug: 359658592
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1142752
Original-Revision: d30d34365e17338f91565816a5f4407c263824d4
GitOrigin-RevId: f1429941a3a310b9c27dbc43157b92da7984c00b
Change-Id: I7ae8e877363f53dec77576dade5ec7c96e2d82ac