[roll] Roll fuchsia [ffx][config] Adding $SHARED_DATA placeholder
This adds $SHARED_DATA to the list of strings that are interpolated
when resolving configuration values. The $SHARED_DATA is used for
persisting data that is shared between isolates and ffx instances on
the current host machine for the current user.

This is intended to be used for data that is used to manage global
resources such as TCP/IP port usage.

Original-Bug: 374085877
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140397
Original-Revision: eabd1238908385fc3b85179ebf4c105cb690f306
GitOrigin-RevId: 42c10dd88a2dcc10f5c5410a0be61f707d01b6e2
Change-Id: I1739b338b4be5444dbce0897a49810e0f10a3bc7