[roll] Roll fuchsia [utest] Fix flake in Timer.Restart
The Timer.Restart core test was inherently racy and could fail if the
test thread was preempted for longer than the timer deadline in between
when the timer was set and when the wait_one was invoked. This CL fixes
the issue by setting the timer deadline to a year in the future; while a
race technicallys till exists, the test thread should never be preempted
for that long.

Multiply: core-tests-ticks_get_force_syscall
Original-Fixed: 375681700
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148133
Original-Revision: 9c3e8f10a6cdf02aad5c651c9685976f08a5bb06
GitOrigin-RevId: be9793aaf0b76e7ebb59e1b5d45ffec6f15da9e8
Change-Id: Ie1953683091f30768443f371e5e10bd501cb44a0