[roll] Roll fuchsia [kernel][vm] Destructor helper for lock annotations
Moves most of the VmObjectPaged destructor into a helper method so that
it checks lock annotations. This is useful because normally a
destructor is considered not to need analysis as it runs when it is
just being cleaned up. However, in this case the VmObjectPaged can
still be found from different lists and so the destructor must still
take part in correct locking behavior.

As part of moving to a helper method some missing lock assertions are
added.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135192
Original-Revision: 5f45535233a32011aabe133c309ec70763dcbcbc
GitOrigin-RevId: 292bb458581a5bfd77008be90e10506e5fb51753
Change-Id: Iaeb77e9d6a64dbec87aa1d0d144481c16cd50b76