[roll] Roll fuchsia [trace-provider] Silence warning about registration
On builds where trace_manager is not included, we're expected to fail to
register with trace manager. Let's not emit a warning in trace provider
so that we don't spam the log with every single trace provider on the
system warning it can't connect.

Original-Bug: 376703561

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148234
Original-Revision: 82e1c37c667c1274a161bd8e9c7f120a9ec9d72a
GitOrigin-RevId: 19134ef6fb3ccced94c2e36bb1937b9a0faa3518
Change-Id: Ibaf48ab5cdf122d5754e9cecd4f769b6ab0cd4eb