[roll] Roll fuchsia [kernel][vm] Remove unused ZeroPagesLocked optimization
This removes a specific optimization of the zero operation that made
zeroing the tail end of the clone of an anonymous VMO slightly more
efficient. The benefit is small, present users do not manage to hit
this case, and the usage of ReleaseCowParentPagesLocked here is
different compared to other use cases and complicates future hierarchy
changes.

Original-Bug: 42058561
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149013
Original-Revision: 1a77788b93d2ca24c87c0941fce682781c119d31
GitOrigin-RevId: 0937aa262fa50dfdc43d5372b3ffbd6b2298a148
Change-Id: I5de4a474018e3512801c87b21a9d0c16aa2319af