[roll] Roll fuchsia [test] Require filename when using ConstantMetricsProcessor to report freeform metrics
The filename under which freeform metrics are saved is part of the
publicly exposed information when the metrics are stored in our backend.
As such, require that callers using the ConstantMetricsProcessor to push
freeform metrics into the system pass an explicit identifier to be used
when publishing metrics.

Original-Bug: b/376922110
Test: trace_metrics_test cpu_breakdown_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149474
Original-Revision: 6df66e981cdff559df13f5cf71d30f02b98847de
GitOrigin-RevId: f0f3228fa72f3fbfa20989889791b75717d857cb
Change-Id: If60c8a2598b136c405bc7a706bc750729f5836f5