[roll] Roll fuchsia [Honeydew] re-enable input and session tests on internal vim3 builder
Session tests: For tests that require the session to be in a healthy
state before the next action, we use a timeout so that all
session-associated components can finish starting up or shutting down
fully.

Input tests: The screenshot data doesn't properly reflect the app
pixels, so we remove the assertion as was done for the screenshot test.
See b/320543407 for more details. It is still worth enabling this test
without the screenshot functionality so we can exercise touch device
creation and tap injection in the meantime.

Original-Fixed: 370481460
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137593
Original-Revision: 2197698c4f170b48ecd54206a8116469926b5671
GitOrigin-RevId: 088720ea71ca0a91e4a4410167ff88693b6d7647
Change-Id: Ia19c037fce99137c363d37492f178fa44fc40deb