[roll] Roll fuchsia [owners] Remove brettw@google.com from src/lib/line_input/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/brettw@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151472
Original-Revision: a969a83e4ad37b561166369d07068f0e2867b26f
GitOrigin-RevId: 2c8ce8d6f4c5a60994e681f27a6d79852d0e3602
Change-Id: I32387ac73c82a6c20ba246b703ea5026523fe1cf