[roll] Roll fuchsia [owners] Remove brettw@google.com from zircon/system/ulib/zxtest/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/brettw@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151267
Original-Revision: 17efe67fc1a8b9e39586f68a462c1e6055b98160
GitOrigin-RevId: 3fa8cfd29286898298af2f1286fb02faf8926209
Change-Id: I086551ed97864067e9ac24e055ce89c65fb79aa3