[roll] Roll fuchsia [network/bench] Shorten system metrics trace duration
The iPerf benchmarks rely on the system metrics daemon to collect CPU
usage stats and output them as trace events. In order to ensure that the
daemon is collecting data during the benchmark, the benchmark records a
short trace session before beginning and checks if it includes any CPU
usage trace events. The duration of this trace session is currently 10
seconds, which is longer than typically needed to capture some CPU usage
records. Shorten it to 1 second so that even if we have to retry
occasionally to get a CPU usage event, we don't always take at least 10
seconds to start the benchmark.

Original-Bug: 370786547
Test: netstack_iperf_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143532
Original-Revision: 43640ee63100dd210c34d2debafe3ad2a3a47f63
GitOrigin-RevId: d921d3a0d1b288804b4385f4f292d103b1b62713
Change-Id: Ia8103bf6b1dbb7780ce415b51fcb8f5cd497612c