[roll] Roll fuchsia [owners] Remove brettw@google.com from sdk/docs/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/brettw@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151263
Original-Revision: 2d4b23ec2be7c526f0b1b3b54c52185b6374637d
GitOrigin-RevId: ecaeb20298a71d67ae7bbf6b868a8861b44c66da
Change-Id: Ib811bb4f2887d0827a5816524b8f6f859213efe2