[roll] Roll fuchsia [netlink] Introduce IndexedRules data structure
Introduces an IndexedRules data structure that stores a set of rules
that all have the same priority. Internally, this is a
BTreeMap<u16, RuleMessage>, where the u16 is the rule's index among the
set of rules with the same priority.

This replaces what was previously a Vec<RuleMessage>, which will be
unsuitable once we start installing rules via the relevant FIDL
protocols, which expect the "index" of a given rule to remain the same
as long as the rule is installed.

Now, when a rule is removed, all the rules after it don't have their
indices changed.

Test: netlink-lib-test
Original-Bug: 358649849
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148072
Original-Revision: 24adf4cbe856c5bedc39acfe80513f11c79666cc
GitOrigin-RevId: d3ecc073a8cb719b5206f26f713c7f8619c21489
Change-Id: Idf8ed6b591b41e0ebccadf857918dffbb3f50e31