[roll] Roll infra/recipes [test-centric-dashboard] Create recipe to link dashboard to CL
This recipe is designed to create a link to the test centric dashboard
and create a comment including the link on the CL the builder was
triggered on. This will pass in the necessary parameters associated with
the CL to the dashboard to enable quick access to the relevant tests the
user will be looking for.

Original-Bug: b/377336699
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1148673
Original-Revision: bf46ecceb094cf1fb787044c4158a68ca265c2e4
GitOrigin-RevId: db8382fa9f5bd243057f25491333a1ad24fe4b8d
Change-Id: Ieef4a8e1ca9b5da47f2e03f547af9063382f2b67