[roll] Roll fuchsia [network/bench] Run iPerf for 5s instead of 10
Currently the iPerf benchmarks (both the host-driven ethernet and
hermetic loopback variants) run iPerf for the default 10-second
duration. The tests are currently taking up a lot of time to run in CI
and we can get results that are nearly as useful by running it for just
half the time, so reduce the runtime to 5 seconds.

From running perfcompare, the effects are this are likely to be:
- slightly lower CPU usage for multi-flow ethernet tests
- very slightly lower throughput on the ethernet receive tests
- fewer lost packets for UDP loopback netstack3 tests

For more analysis of why these changes are unconcerning, see
https://fxbug.dev/370786547#comment3.

Original-Bug: 370786547
Test: netstack_iperf_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143533
Original-Revision: daf731921b66ca46644f4ac6c864f3da1e6828be
GitOrigin-RevId: 70d0b00361acb5081da7204123ffd347428dec17
Change-Id: Ia14463f51ed08c057fda8bbf880b28b6f8ddd69f