[roll] Roll fuchsia [network/bench] Start iPerf server only once
Start the on-device iPerf server once at the beginning of the benchmark
and reuse it across test cases. This does have the downside of server
output not being saved separately for each test case, and a server
failure in one test case causing issues in a subsequent one, but it
saves a lot of overhead on server setup and teardown that currently
applies for every test case rather than once for the entire benchmark.

Also simplify the code now that we only run single-flow benchmarks over
ethernet; we no longer have to handle running multiple clients and
servers in parallel outside of the hermetic benchmark.

Original-Bug: 370786547
Test: netstack_iperf_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149536
Original-Revision: 6a96652a447e76271ad438ad1c42e43cd6d4447f
GitOrigin-RevId: 705db762cb25cf9ab43fb029a05947d177376c92
Change-Id: I3abebef5cd6c9f68d8f1d52b1fb25a7b40f442d5