[roll] Roll fuchsia [docsgen] Fixing clidoc dependencies
This uses the sdk_molecule manifest as input to determine
the list of tools to run to generate the reference docs.

This is done to make sure all the tools are dependencies of
the invoke_clidoc target to make sure the build is consistent.

Original-Bug: 376723774
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148896
Original-Revision: 87ebfa6dcee9f19e8a32e6d0ace1a4c40cd4f128
GitOrigin-RevId: 284501a8c2a1cdb7b573b9a66bd35bc48d24c97a
Change-Id: I7bc6fd66f1fcc09ef72de54b12612fd6ab89675b