[roll] Roll fuchsia [hrtimer][vim3] Add setup_event to StartAndWait
The `setup_event` event signals to a client that the driver has
setup the timer in the hardware. Hence for cases when the
system may need to wakeup from this timer, it is safe to allow the
system to suspend now that it is known that the timer is setup.

Original-Bug: 373928684
Test: fx test -o aml-hrtimer-test
Test: Exercise a timer on VIM3:
$ hrtimer-ctl -i 0 -w 1
Executing on device /dev/class/hrtimer/27021865
Starting timer...
Timer triggered
Keep alive released
Timer started

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147532
Original-Revision: 403ebbdc0e8d4c7d265f248298dbc9709c4937b1
GitOrigin-RevId: 68dbfbd3e23a835fd6d40a2ab7f1d7f4f32440c3
Change-Id: I9640eced1000a9430d15f69c9360f75a821998c1