[roll] Roll fuchsia [fxl][log] Expose CLI parsing without the syslog dep
The json_validator is used as part of the FIDL toolchain. It uses the
FXL command line parsing utilities. Syslog would like to be able to use
log levels defined in FIDL as the source of truth for log levels.
There's optional support in the FXL command line parsing utilities to
parse log levels. This causes a cycle.

This introduces a new cli_no_logging target to FXL that's just the
non-log-related parts of command-line parsing. This was already written
in separate files and is all that json_validator actually uses.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152516
Original-Revision: dc1a95e0a46a71aba1d2308ff155cc11c88426a1
GitOrigin-RevId: 59a7183f0ccd76a2caa62dd321e47d47d95b776a
Change-Id: I435a5cb57a95eb25ea7bdf073afe6fff02dcdfba