[roll] Roll fuchsia [netstack3] Add switch for sending test data
Many tests use the bind_listen_connect_accept_inner function to set up a
connected pair of sockets, but not all of them rely on the test data
that gets sent as part of that process. However, for an upcoming test
case, this test data breaks the test because it affects the advertised
windows.

Adding this flag required auditing the test cases to see whether they
actually relied on the data being sent, which for the most part they
don't.

Original-Bug: 372763809
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153712
Original-Revision: 95a87778cb1ba554d05eb4c3d0d808504f2eef10
GitOrigin-RevId: 8d70194342971c95025309183a16364314f09ccf
Change-Id: I8ee4cbdd0244f6a8911f1c10408b6c89a1f7e1f6