[roll] Roll fuchsia [elfldltl] Replace NewArrayFromFile with ContainerArrayFromFile
This eliminates a dependency on fbl::AllocChecker in the
widely-used <lib/elfldltl/memory.h> public header.

The only users of NewArrayFromFile are in test code and thus
don't need to be fully optimized.  The replacement for that class
is a helper function that adapts Container API object to the
Allocator API for ReadArrayFromFile in the Memory API.  This is
suboptimal because the container API has no way to allocate space
without default-constructing each object, whereas the old code
using `std::unique_ptr<T[]>` could allocate uninitialized space.

An optimized replacement would do something more like the old
NewArrayFromFile did, but could perhaps be separated into a
different public header to isolate the fbl dependency and/or
template-parameterized somehow to handle both plain new and
AllocChecker new.  But that can be revisited when a need like
that arises.  So far the only performance-critical uses of
ReadArrayFromFile (startup and libdl loaders for phdrs) just use
FixedArrayFromFile.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152874
Original-Revision: 3dbc6049e0d28282368ec91202e6b231d08d57d4
GitOrigin-RevId: 81ade71952607d350c9ccb7e61e7c5dc8113a6f6
Change-Id: If01ecd8e69cd6382134ec1e9925727a2607e3cd6