[roll] Roll fuchsia [ffx] Retry connections when we only connect to FDomain
Sometimes the Overnet connection fails when it *could* succeed, but the
FDomain connection succeeds first try. If FDomain didn't exist we'd
retry and get an Overnet connection. As-is we just proceed with FDomain,
which we don't support further up the stack so we end up panicking.

This makes an FDomain-only connection cause to retry.

Original-Fixed: b/377969139
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153341
Original-Revision: 63deec4f28f6c248afba0711e0c91a0718b6fbfe
GitOrigin-RevId: 6a1490459e6a8133040aed2ce40580bf5818c90c
Change-Id: Ib186937c1326f854a2131fc9e9a848d747dd8477