[roll] Roll fuchsia [kernel][aarch64] Distinguish captured faults
Not all kinds of data faults on arm can be resolved by the page fault
handler. When not capturing faults this distinction is performed, but
when capturing faults there is not enough information provided to the
caller to know if running the page fault handler can resolve it.

This change exposes additional information to the user_copy routine by
directly setting the return status to indicate whether the fault
information is useful or not.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153614
Original-Revision: efbb5ee18cf63d879cf334f5e5990bb50a1bcc42
GitOrigin-RevId: 5169fa27246199b56f96410d50abef8976e992bd
Change-Id: Iaa94784943f7904122d484b7676ecc5922000333