[roll] Roll fuchsia [power] Clarify FIDL protocol usage, update servers
The fuchsia.hardware.powersource/Source protocol was being used
a bit oddly by battery manager. In particular, battery manager
would reconnect to it to get information each time an event was
signaled. This requires far more FIDL interactions.

The protocol was also structured a bit oddly in that changes to
batter information would cause an update event, but the signal
could not be cleared by calling `GetBatteryInfo`. This change
fixes that in the ACPI battery driver and the fake driver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148912
Original-Revision: 5450c5ee1771179c9c97f6ae043d93419e66fd21
GitOrigin-RevId: 33d2de837aa06208ff713bdcb95f10aa862ae1cb
Change-Id: I06121d7ec695b7a93459f7ac10c93401fa3863cf