[roll] Roll fuchsia [ffx][virtualization] Fix build
https://fxrev.dev/1153934 made ffx_tool() require an `sdk_target_name`
when `sdk_category` is specified. `ffx_guest_tool` is not in an SDK, so
it does not need either.

https://fxrev.dev/1153935 increased the scope of marker checks, so now
`ffx_guest` needs to specify "internal" to avoid defaulting to
"partner" because it uses `fuchsia.virtualization`, which is not in any
SDK category.

Original-Bug: 314822328
Test: Add "//src/virtualization/ffx:ffx_guest_tool" to `host_labels` and `fx gen`.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156574
Original-Revision: 1a4b848a93971e54951ea39414933aad3c762051
GitOrigin-RevId: e2b532aefd466d73ed16b03cb7972f8c92bf3305
Change-Id: Ie099939cadce6567759b9e996072750ef603966d