[roll] Roll fuchsia [kernel][arm64] hard set SCTLR_EL1 when initializing cpus instead of orring in bits
More and more bits are being added in newer versions of the ARM spec,
but they're all defaulting to a safe state if 0, so instead of orring in
the initial state of the cpu, hard set it to just the bits we need.

Also update the register definition for newer bits in the ARMv9 spec.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154860
Original-Revision: f7bd57538e32d222b1eedc456bedcb5031502d75
GitOrigin-RevId: a6e5fbbc7baa6808b4cf7c5d73acf39f516c57be
Change-Id: Iabbde51ec3423429d6cc6091fe0b3173c05c9241