[roll] Roll fuchsia [fxfs] handle unaligned allocate calls
We need to be able to handle if the offset and/or length passed to
allocate is not block aligned, since there is no requirement that it is
and regardless, crashing like we did before is not the right response.

Posix fallocate stipulates that the filesystem can allocate more than
was asked for to aligned with block sizes, so we round the range
appropriately, taking care to make sure the recorded size is still
correct and testing that the tail block is properly zeroed.

Original-Bug: 293943124
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156212
Original-Revision: b5051ec74cadf8e93ebe198afe5be5b54513698e
GitOrigin-RevId: 6c8a6b7c847b02dfd8c5df8f1b2ae6da7ccd7612
Change-Id: Ib531b2792d7c67df636d90096d2c9fed6d7e22f9