[roll] Roll fuchsia [fdf] Major refactors of driver test realm
This changes the following stuff in the driver test realm:
 - split out the various servers provided in the driver_test_realm.cc
 - added more methods to the internal protocol used by the resolver
 - the resolver is now responsible for compiling the list of drivers
 - the list of drivers is delivered dynamically through fidl
 - removes some unused routes

This work is done in order for us to be able to collapse the topology of
the driver test realm to no longer have two layers of realm builder.
That is done in the next cl on the stack.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144152
Original-Revision: ad44d4441a3e6874b54b21988a80dce6fc9a800f
GitOrigin-RevId: 3f7d57a3739ab6a0f48cd3bcb3be91fc56f7f69c
Change-Id: Ib6e82312f1778bb2ea48ce034a72315fc07cc3e0