[roll] Roll fuchsia [superproject] Roll third_party/pigweed pw_bluetooth: Fix RFCOMM information length off by one
A misreading of the spec led to confusion on whether credits was counted
as part of information_length. It is _not_, but it is included in the
calculation for max allowed payload.

Fix this and standardize on `information` as the name used in the spec
vs payload.

Also add helper flag for checking whether a frame is UIH or not

Add standalone unit tests for RFCOMM.

Original-Bug: 378692444
Original-Bug: 378691959
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/249092
Original-Revision: 40d4c1261275065d30e77ea68fbea69061b70e78
GitOrigin-RevId: 73a2c5b4937c00041af241a374ebd14e601ccba8
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156496
Original-Revision: 37777e006f07ca9712138336dbfb4820f4c76288
Change-Id: I1d00f9f8a80e91f03a8ffa28f5af212073738682