[roll] Roll fuchsia [fbl][canary] Print the observed value during a panic.
Print the observed value of the magic number when a fbl::Canary fails
an ASSERT predicate, instead of allowing the code to re-load the magic
number post-predicate failure for panic printing purposes.

In a perfect world, this should never (ever) matter.  That said, when
your Canary assert predicates are firing, the world is a long way from
being perfect, and it can actually be important that we report the
exact value used in the predicate evaluation in order to help
understand what might have went wrong.  See the attached bug for a
practical example.

Original-Bug: 313083326
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1158041
Original-Revision: b4a64b05fe58310aad29e5b493fb6c33aaf7ad6d
GitOrigin-RevId: b8e10a89082f9e13c372b94f898ef7f6c2b75c08
Change-Id: I7b390ffa5c90825dc818bb6c555a1f32fc83a5d3