[roll] Roll fuchsia [storage] Move partition table detection to fshost
The system partition table is the one on which the primary system ought
to reside on (i.e. typically the GPT-formatted non-removable storage
device).  Currently the paver has heuristics to find this, but these
heuristics should live in fshost since that is a question of storage
policy and fshost is the storage policy engine.

Fshost will automatically resolve fuchsia.storagehost.PartitionService
and fuchsia.storagehost.PartitionsManager requests to the system GPT
component when it is bound, so calls to those interfaces will block
until the GPT has been found and its partitions enumerated.  The paver
will use PartitionService for enumerating partitions and connecting to
them.

fuchsia.fshost.Recovery/InitSystemPartitionTable is a special method
that is only supported when fuchsia.fshost.Netboot (the config value) is
set.  This method is essentially just a proxy for
fuchsia.storagehost.PartitionsAdmin/ResetPartitionTable, allowing fshost
to enforce its own policy restrictions on when partition tables are
reset.

The logic in the paver to deal with uninitialized or malformed GPT
devices was removed.  This is handled transparently by fshost; fshost
will present an empty GPT to the paver in this case, and the paver can
only do one thing in this case, which is call InitSystemPartitionTable
if the system is booted into a supported configuration.

Finally, this change corrects the directory structure of the services
exposed from storage-host to conform to RFC-0041.

Original-Bug: 371060853
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151454
Original-Revision: e530a51323aae5ad25845911b0691877e9b8a5a9
GitOrigin-RevId: 0297f8e78c3005df6e21745a4fddd4b74ff6c095
Change-Id: I02a60ea3b8ce01a9e7ef07370258a4491e9739b4