[roll] Roll fuchsia [dl] Introduce AdjustLoadBiasAdaptor
This is a wrapper around the ld::ModuleMemory object to be used by
the `DecodeModule` function in the DlImplTests fixture.

AdjustLoadBiasAdaptor will attempt to perform a read on the underlying
memory object using the initial `ptr` value passed in: if that read
fails, it will try to read again with `ptr - load_bias`, as is needed
for linux.

Original-Bug: 324136831
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147858
Original-Revision: a99ea2f8e82febef33ea086d1547e6c5b7869965
GitOrigin-RevId: 0595474fd95382dec85ab3ca61c21a8b576bb600
Change-Id: I6a87cb8ac3ebf42a475510c247a151580398e11d