[roll] Roll fuchsia [kernel][phys] Drop NVRAM from hand-off RAM description
An NVRAM range should not be treated as RAM, and longer-term we will
make that distinction more official. In the meantime, we begin to not
treat it as such by dropping it from the hand-off RAM description, in
turn dropping it from tracking within a PMM arena. One consequence of
this is the need to update the logic around protecting the physmap
regions corresponding to gaps outside of PMM arenas to explicitly avoid
any NVRAM region.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159514
Original-Revision: cb70f5d318dea51911426bfb243602c4cc127ec4
GitOrigin-RevId: c29c27f64e28260edfb698adde85f0bd20e4212d
Change-Id: I6484c109a84c971e57f520367cd7abbd5e21293c