[roll] Roll fuchsia [archivist] Set up separate loop for handling sockets
This sets up a separate thread for processing log sockets which reduces
the per-socket memory overhead and improves performance.

This also allows us to introduce a delay which prevents Archivist from
waking immediately after every message.

As a result of this change, we can reduce the number of threads used for
the executor down to 1 since the executor is no longer used when reading
sockets.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149015
Original-Revision: ec65ec02383869a631a530e9904e363e62d3cac0
GitOrigin-RevId: af927dcc3466cd06dca0c93bca9daa8f6cb837fc
Change-Id: Ic922bdfe44e7befb8de89bbc451dc1c531f29257