[roll] Roll fuchsia [starnix][cgroup] Build more of cgroup hierarchy
Created `CgroupRoot` for the root of a cgroup hierarchy, as it has
different properties than the rest of the tree. Both `CgroupRoot` and
`Cgroup` implements `CgroupOps`. Their main difference is that `Cgroup`
owns the `FsNode` of its `CgroupDirectoryHandle`, whereas the
`FileSystem` owns the `FsNode` of the `CgroupRoot`.

Each `CgroupDirectoryHandle` has a `Weak<dyn CgroupOps>` to get back the
associated `Cgroup`, or `CgroupRoot`, which it can use to implement its
interface files.

`mknod` is no longer used to create the initial `cgroup.procs` file.
Instead, `cgroup.procs` and `cgroup.controllers` files are created
alongside the directory, and can be seen with `ls` without first being
written to.

Cgroup filesystems are now Uncached, because cgroup interface files are
dynamically generated when the directory is created, and are removed
together with the directory. Additionally they do not emit any inotify
events on Linux. As such, we should implement the cgroup filesystem as
Uncached, and fill in the logic for `create_node_ops` and `lookup`.

Original-Bug: 377712920, 375677856, 373897750
Tested: syscalls test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146230
Original-Revision: 57a3977adcf018718c25fec2c0e0244fb778a8aa
GitOrigin-RevId: f1eb760f3a89c632f462130ad825d80e74628ff4
Change-Id: I68383fcb4a93e2ae0156c3ca8a119dcbf315dab3