[roll] Roll fuchsia [archivist] Introduce a new LogStream protocol
First implementation of a new protocol in Archivist that allows to
fetch the original FXT records directly without going through the
overhead of ArchiveAccessor. This protocol sends the records over a
socket and extends them with the attribution: moniker and component
url.

At the moment I plan to use this protocol from Starnix. But we could
also make use of it from feedback and ffx.

Original-Bug: 315520045
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159075
Original-Revision: 485cd2ad26244100b72666643bea332feb13470c
GitOrigin-RevId: efa34edff0259f722a2b1569c802a2d627a45f92
Change-Id: I3a85589e520ffc4eb9698e2fcdf4315d8c63bb1b