[roll] Roll fuchsia [ffx] ffx --strict target show
Add --strict support to `ffx target show`, which basically means
replacing the `TargetProxy` with a query to the DirectConnector. I chose
to expose the connector by adding a new FfxTool field that provides
the direct connection if one is available, as well as a
`Deferred<TargetProxy>`. If we have a direct connection, use it to
gather connection information; otherwise use the `TargetProxy`.
(Some of the information was also available directly from RCS, so
the code also is changes to use RCS where possible.)

Testing this required exposing a `MockDirectConnector`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160472
Original-Revision: a4c1ac66cf9345dbfa8bb2531f009ae521f7e5c0
GitOrigin-RevId: d24d242ff219cc551d845db1e7cb2f900f5dd88c
Change-Id: Ie437187ddef4177001bc2127f337880254601ef4