[roll] Roll fuchsia [thread_sampler] Disable Ints While Writing Samples
Interrupts are enabled while we are taking samples because we need to
allow interrupts to usercopy. However, because they aren't disabled,
this allows the possibility that we begin writing, set the pending
write, then are context switched out. Another thread would then attempt
to write a sample on the same cpu, violating the assertion that there
isn't a pending write.

Since the trace buffer is mapped and pinned, it is safe to write to it
while interrupts are off, so we now do so.

Testing:

Without the interrupts change, ThreadSampler.HighFrequency consistently
hits the assertion fail. With the added disabled interrupts it now
passes.

Original-Fixed: 379166110
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159872
Original-Revision: 49cb658aab947d662cebbfee465d83a554d07808
GitOrigin-RevId: 4fa65910cd32aff10802aa7551e42559934f5bc2
Change-Id: I117b2a5c42d54c4ece1c847a468fac341aef1c6f