[roll] Roll fuchsia [dl] Libdl Supports Startup Modules
This CL will pass a container of ld::abi::Abi<>::Module data structures
to the RuntimeDynamicLinker ctor: a permanent RuntimeModule data
structure is created for each ABI module and added to the modules_ list.

These modules cannot be unloaded for the lifetime of the program.
Appropriate symbol visibility is assumed to already be set on the abi
module when it is passed to the RuntimeDynamicLinker.

Original-Fixed: 324136831
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148633
Original-Revision: 89bbe2af30e522d76c43c6b65ee9674854800546
GitOrigin-RevId: a2134c2df0e2dd32f77af36ed9c515712c654dff
Change-Id: I6ec3597f73651190a07f63b8474f8e9ef26854e1