[roll] Roll fuchsia [usb][adb] Fix race in usb-adb-test
Currently, the test teardown logic calls the UnbindOp() on a background
thread, and then proceeds to tear down the fake root device - which
races with the background thread. If the dtor wins the race, it will
cause a failure because the UnbindOp() was called but never (not
yet) replied to. All we need to do is synchronize the UnbindOp() running
on a background thread to the testing thread via
WaitUntilUnbindReplyCalled(). This blocks TearDown() until the driver's
unbind logic actually replies to the unbind request.

Original-Bug: b/380480663
Test: fx test usb-adb-unittest
Multiply: usb-adb-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162295
Original-Revision: 32478ccf0ba429e93b4212ab5bf8a0c5992e76c4
GitOrigin-RevId: dfa17c2514e936a4336773b60552e80745bc4ca9
Change-Id: I92ac3349ebfb808ffb01f4423a829ad00fa85664