[roll] Roll fuchsia [dl] Use AllocChecker in PopulateStartupModules
Establish an API contract with the caller of PopulateStartupModules, to
pass in and check a fbl::AllocChecker to verify the success/failure of
this function.

PopulateStartupModules arms the function-level alloc-checker with the
overall failure or success of the function (using local alloc-checker
instances for actual allocations).

RuntimeDynamicLinker::Create will use its function-level alloc-checker
to perform the allocation of the RuntimeDynamicLinker and pass into
PopulateStartupModules.

Original-Bug: 324136831
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160474
Original-Revision: 3c52e1dbdc71dbd034d1e63baf43205284370976
GitOrigin-RevId: ad70a70bf2286305aa5dec90cb60425622d63db8
Change-Id: Id113fc41238371dfe117a361b8271cc104201507