[roll] Roll fuchsia [fxfs] Take a read lock when creating prefetch keys future.
Takes a read lock on the object while we know it still exists and pass
that read lock into the prefetch future so that the object cannot be
deleted in the interim.

Also deleting the `pre_fetch()` wrapper method for keys. As it is not a
safe fire-and-forget operation given the desire for looking at
non-existent keys to be considered a failure of the system. Moves the
logic to make it safe up to `pre_fetch_keys` so that has the only notion
of prefetching left in the system.

Original-Bug: 378508610
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160854
Original-Revision: 9452bca4691cb40f7c9ff35da54ada6a9bb6e0d4
GitOrigin-RevId: 41c5d19f7e9c8c39ac1c2975d8187878dc9fb253
Change-Id: I005ab4fe29bf35991f42e6474b2c918a80d8748f