[roll] Roll fuchsia [devicetree] Fix error check when visit fails
The devicetree::Walk does not end when we return false. It only stops
walking the subtree from the node we returned false. This essentially
overrides the status and therefore the error is not propagated up to the
board driver. This CL caches any error code that occurs during the walk
and propagates it to the board driver.

Test: Introduced an error and ran manager-test to confirm failure.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1161917
Original-Revision: 7c0a9aec4dba8ff5e09178ed2be1aacd46a7e2e9
GitOrigin-RevId: 62845e3f3ede75051005a0697a6b8537c03bec10
Change-Id: I9bbc1681431c3452bc5470ae4a0ecc20d4854642