[roll] Roll fuchsia [fbl] Allow AllocChecker::arm calls with smart pointer types
Smart pointer types are explicitly convertible to bool, but not
explicitly convertible.  Add a templated signature that allows
use of any type explicitly convertible to bool.  This makes it
more ergonomic to use with smart pointer types, which are only
explicitly convertible to bool and not implicitly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163334
Original-Revision: 4ec5ccb8b4adef5a73ecd657488837dabdd6cfdd
GitOrigin-RevId: 2ead076c4d5a565481ddf08b017a342b1d3a7a50
Change-Id: Iafe6fe22ed8204c440ab0af7ef23c6799441b33f