[roll] Roll fuchsia [debugger] Deprecate NotifyComponentDiscovered
The underlying ComponentManager event was deprecated some months ago
(see the bug).

This ended up breaking recursive filters such that tests that have a
programless root component, because we never get the notification for
the matching component URL that is installed by fx test. This manifests
in tests running with `fx test --break-on-failure ...` not breaking at
all on any kind of failure or crash.

The recursive filter installation now happens in response to
ComponentStarting events, but this isn't enough to completely solve the
problem, because these events are not emitted for programless components
that are started lazily. TestManager is changed in
https://fxrev.dev/c/1163032/ to start the test realm's root component
eagerly so that we are guaranteed to get this notification for tests.

Original-Fixed: https://fxbug.dev/375663708
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153733
Original-Revision: dec1089311a79b48da885c0b9645debcf8f10df2
GitOrigin-RevId: 9472e3f7a0b0975199e959cd04e1bbed5e39b149
Change-Id: I06c09894ac7ef86b30083da3630ae84811ce9c31