[roll] Roll fuchsia [netstack3] Decrement count when connection removed
Currently, the num_connections field tracking the total number of
connections present in the conntrack table is not decremented when a
connection is removed due to Connection::update returning
ConnectionUpdateAction::RemoveEntry, which happens (for example) when a
TCP connection closes.

This means that even when the table is not full, the num_connections
field can grow over time until netstack *thinks* it is full and starts
incurring thrash by removing existing connections in order to insert new
ones. Fix this by decrementing the counter if a connection is removed.

Original-Fixed: 380887806
Test: netstack3-filter-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163474
Original-Revision: 9c1b0d9e6718fc6cf3c06c4a337a6e810ab6ed1d
GitOrigin-RevId: fadcc645dd507151bd2f9f7b9d5f932b66328117
Change-Id: I8768722873c95b04d274b22e90cc52074148306e