[roll] Roll fuchsia [fdio] store node name in parent map entry
Identify the node in its parent using its ID rather than its name. This
allows a node's parent and its identifier to have the same lifetime
without breaking the parent's maps by storing the name in the parent;
copying the id is cheap, so it is stored on the node as well as its
entry.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154941
Original-Revision: 440ef74322b1a405006506733184c78b252a2c27
GitOrigin-RevId: 15a5b1b8ad89aaa0e5da9518737b2b5938b09325
Change-Id: I4d0e9d46e21f5641089cfebaf4e266d4ad45a0b3