[roll] Roll fuchsia [sestarnix] Revise the permission check & todo-deny audit logs
Introduce an audit callback to the permission-checking helper, that
will be called with an AuditContext structure, that the caller can
choose to add details to before emitting the log.

The todo_check_permission!() helper is updated to use the audit
log hook to emit a track_stub!() report for each denial, while
limiting the number of todo-deny audit logs emitted for a each
location.

Original-Bug: 380399690, 362707360
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165232
Original-Revision: 05a687fbac6756f99776e7cca4bc68c24f02f71f
GitOrigin-RevId: ab80dd084835f32767b986f828cc5e66205ee212
Change-Id: Ibcb574a4748f1112d50b9b4ebdee2d927a9af23b