[roll] Roll fuchsia [bazel] Avoid hashing the same content twice during workspace setup.
The content hash files for `fuchsia_sdk` and `fuchsia_in_tree_idk`
are the same, which resulted in hashing the content of the same IDK
directory twice when this script was running.

As a minor performance improvement, ensure the hashing of the content
only happens once, and create a symlink from one content file to the
other.

NOTE: The `bazel_action.py` script depends on the name of the content
files to match those of the repository that depend on them.

NO_IFTTT=no real need to change toplevel.WORKSPACE.bazel here!

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165953
Original-Revision: dccd2da5868a8caf2c58c91c185089c01aae8516
GitOrigin-RevId: 4a74402605586f6857e62dd8a9f4ea88a03abb3d
Change-Id: I47446867f91ad72e0e16b9a1bae92ead27f08a75