[roll] Roll fuchsia [bazel_sdk] Rename cc_features_template.bzl into cc_features.bzl
Since the file no longer uses substitution expressions, it can be
loaded as-is. This renames the file accordingly and moves it to
the @rules_fuchsia//common:toolchains/clang/ directory.

This will allow reusing the same file for the host C++ toolchain
definition in a future CL.

Original-Bug: 333059239
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1166532
Original-Revision: 9d6265684455d366d3175e0e590666b75fbc364b
GitOrigin-RevId: 23088c25b53b652dfdb6d58dcf516a30da28cba9
Change-Id: I188a112817f3b6932b07d78e967a202e152cf9ae