[roll] Roll fuchsia [aml-uart] Use a single dispatcher
Rather than have separate dispatchers for the wake lease time, fidl, and
interrupts, they were all changed to use the default dispatcher instead.
This allows us to simplify code and remove locks.

Additionally, the manual wake lease acquisition code was changes to use
the fdf_power::WakeLease class. As a result the ability to inject and
modulate a timer was lost. That said, it is still possible to simply
wait out the timer in order to correctly test the functionality as the
timeout value is fairly small it's not a big deal. Some coverage was lost, and
while it could be recovered by injecting inspect and observing the inspect
values, that would really just be testing the WakeLease class and isn't
worthwhile.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164953
Original-Revision: 54dbe9772a14bb87b8afd9440fc8e9b52536d921
GitOrigin-RevId: efc51ba8ac67c9d4491bb971b1f495a3945e8949
Change-Id: Iaceb404616946eedfbf13fbfc353131b13c01473