[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/pylibs/typing_extensions/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162780
Original-Revision: 05b11dcbfefc4d1970a06f53898a49c1c6e7cf5d
GitOrigin-RevId: 5049c99590583052b5d105369dbc3cf448d0c747
Change-Id: Ief053d16990fb02cdf52340e204017b23c916776