[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/pylibs/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162620
Original-Revision: ef580521c45bb52a6f921b5675306df54f0a078b
GitOrigin-RevId: 425b32a1ec805934d4871d5835e8926fc292448b
Change-Id: Idb250edb78314d21295a5b4b6191161a9c15e937