[roll] Roll fuchsia [owners] Remove chok@google.com from tools/devshell/python/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162617
Original-Revision: cc6d211f9d961ffa275a8d2bb8dbcae74a6d1db8
GitOrigin-RevId: caed5d566bb4bf9b831c4f41d63165d246d0467a
Change-Id: I47cb4c442b1f6914178347ab55185d08de249049