[roll] Roll fuchsia [owners] Remove chok@google.com from build/python/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162776
Original-Revision: c702cdf0d70cd5594fd6bb0a0f2fdf6611bb8d89
GitOrigin-RevId: 94a91f66abbfdf51e8cc81446adebbc739f7c3f1
Change-Id: I1d32abb8ac96f101e0bfac1cb3de114dfd7f2ce2