[roll] Roll fuchsia [network] Change the default netstack to Netstack3
This is to catch anything that isn't using the pre-defined products.

Since we don't want to transition terminal yet, we need to
explicitly pin it to Netstack2.

The bad_signature_update test change is required because the image size increases slightly when Netstack3 is used.

Original-Bug: 370546232
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128734
Original-Revision: 9dc522acc08c7eac059331bfa4a05479acc2bac2
GitOrigin-RevId: acfa3ec077bf7cd1e995e31ee0e9088b7ad490b7
Change-Id: I68d3a200840fd32d7a8ce938cdf710f9da5cba7b