[roll] Roll fuchsia [bazel_sdk][versioning] Use actual levels in api_level_test
Previously, `fuchsia_api_level_test_suite` used a fake set of API levels
in tests. These were added as the `valid_api_levels_for_test` attribute
of each level, and that was handled by the production code.

As a result, most test cases were not testing the actual values as
extracted from `version_history.json`. For example, "HEAD" was not
supported because it was not in the fake set of API levels.

We need to add tests against the actual levels, so just use the actual
levels for all tests. The main benefit of fake levels was knowing which
numbers are valid to cover that case, but we can obtain fuch a value
programmatically.

Original-Bug: 354047162
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088693
Original-Revision: cecb18f18c80c63441013d5a14726f2689e1799b
GitOrigin-RevId: 74f81dc492701ce9681884d4cec6b19729f54166
Change-Id: I5b3cdf1d78c7b1266c9c2e575ec20f373c306748