[roll] Roll fuchsia [display][banjo][fidl] Revise layer composition op representation.
This CL's main goal is to fix the naming issues of the Banjo type
ClientCompositionOpcode.
1. Each instance represents multiple operations, but the name suggest
   that each instance is one operation.
2. The name over-indexes on the way that the type is currently used.

This CL also aligns the FIDL interface to the Banjo interface, and
cleans up the comments for DisplayEngine.CheckConfiguration(), which
is where the composition opcodes type is used.

Original-Bug: 316631158
Test: fx test //src/graphics/display (no functional change)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169272
Original-Revision: 512662b5e40a9378762482dd98b4b134ecf96d6e
GitOrigin-RevId: b8dde9238f3cd30061b8e78a6261267d38f61684
Change-Id: Ic38e09f061fe7bf6e0e7a633711fec99468abdf0