[roll] Roll fuchsia [kernel] Fix overflow when updating the platform timer
The timer code did not use the overflow aware zx_ticks_sub_ticks
function when translating from a boot timer deadline to a raw ticks
deadline. This meant that it was possible for the platform timer to
inadvertently be set to an incorrect value, leading to stalls in the
system.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168218
Original-Revision: f7dfff172270ddc14501b5fd5c7342f16a55a118
GitOrigin-RevId: ccce256c81b840750c366bd1e7a88b2a3337b25a
Change-Id: I8de2f18fbeccb56856b27f7d4d6d1f7f934e4671