[roll] Roll fuchsia [wlan][policy] Pass in roam sender to roam monitor init
Adds the roam mpsc::Sender, which the roam monitor service loop uses to
send roam requests to the client state machine, as an argument in the
`initialize_roam_monitor` method, rather than building it internally.

This allows us to intercept that the Sender for testing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155215
Original-Revision: e0242af48b321282661ea6d66296745624519547
GitOrigin-RevId: f5246b1ed4147ae454a0004efe4052c6b0e55e51
Change-Id: I309b5cfd091ab4dac10c7f2865e171f973672329