[roll] Roll fuchsia [fxfs] check stream size before updating on allocate
Allocate potentially needs to update the stream size - if the allocation
range is past the current end of the file it's expected to extend it.
When checking that, writes that extended the size of the file could be
sitting in the pager vmo, if it hasn't flushed yet, so instead of
checking the internal handle for the size, we should check the stream
for the most up-to-date size.

Original-Bug: 293943124
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169419
Original-Revision: b497f507b7ab7c92768df0b1b98d1044afa13602
GitOrigin-RevId: 6d88285c4acd584481077398c38f2022aef39dc4
Change-Id: Ib14d7f4afa0365721773ac5dff1eff27a4b1c165