[roll] Roll infra/recipes [owners] Remove tmandry@google.com from recipes/contrib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1169785
Original-Revision: 48b4f56c2daf9e7188db3e6e413d3ca7ead9d2b3
GitOrigin-RevId: 181975fd6974f4f108a339cf77777aa8bba8e528
Change-Id: I19c68b2c504dbce631a146b48be641b30c3d89f4