[roll] Roll fuchsia [ffx][repo] Filter daemon repos that aren't running
A previous CL improved repo server tracking for daemon based servers,
especially ones that were in isolate dirs. However, this had the side
effect of listing configured repositories that were not actually running
as part of the daemon process.

This CL filters those repositories based on the status of the daemon
repository server task.

This makes it better, but there is still the possibility of edge cases
if there is a daemon running in another isolate dir.

Original-Bug: 382422026
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169017
Original-Revision: 5cea041b5f011797b7f477654a7f801b1e628cc1
GitOrigin-RevId: e977a08ed365701a9f32cd65fd68585770baab30
Change-Id: I53007e41a9085ff224f9d134b0357dc9f2837f39