[roll] Roll fuchsia [netstack3] Cache direction in per-packet filter metadata
Currently, both conntrack and NAT logic independently attempt to
determine the direction of a packet with respect to its conntrack
connection. Because this check is stateless (i.e. is performed again at
every NAT hook), if a packet is rewritten due to NAT, it is likely to no
longer match either the original or reply tuple of its connection. This
results in the netstack being unable to correctly perform NAT on it.

Instead of recalculating it each time, have conntrack return the
direction along with the connection itself when a packet's connection is
looked up, and then stash it in the per-packet metadata that persists
throughout the IP layer. This allows NAT to correctly be performed in
all hooks that a packet traverses, and also allows slightly simplifying
some conntrack code.

Original-Fixed: 371017876
Test: netstack3-filter-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168458
Original-Revision: f1fbd4f9027eb267d785d9c1cc13d4afb91410ae
GitOrigin-RevId: 0ecf97942b63c7dda0a3c553687d1072117775eb
Change-Id: I51f24af65d71809c35aeecaf2818a7fd5cd93a64