[roll] Roll fuchsia [display][fidl] Store config stamps as in-line values.
Previously, the config stamp in OnDisplayVsync() event was passed
in a FIDL `box`, which is a reference to an out-of-line object.
This added unnecessary complexity to the display coordinator and
engine drivers.

This change removes the `box` from the config stamp field. It's now
stored as an in-line value.

Test: no-op.
Original-Bug: 341156315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170274
Original-Revision: cca458d53ed0a4183ab50807ea534fcf79d5b907
GitOrigin-RevId: ca5378cd95b7850a6070a9fca7ebb286f0fd2c1a
Change-Id: I2eede80bf27e07e0126e96aeeb27df2af4ad7138