[roll] Roll infra/recipes [rollers] Don't use checkout.fuchsia_with_options()
That function is specific to Fuchsia (as of https://fxrev.dev/1115852
it assumes there is an integration repository in the checkout).

But these recipes are used for non-Fuchsia repos in the roller test
builders in recipes presubmit, so they should use the more generic
`api.checkout.with_options()` function to support that use case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1170352
Original-Revision: 50b6b8c487e60060fbfc9a7c3527814006f13d54
GitOrigin-RevId: 360b2f6b293dd71e68b996ef0f07523998eae58f
Change-Id: Ia2388eca07c75bcbe6d58e3663504152aff42aa5