[roll] Roll fuchsia [bazel][minimal_workspace] Use vendored skylib library.
$FUCHSIA_DIR/third_party/bazel_skylib already contains
skylib 1.3.0 so there is no need to download it with an
http_archive() directive. Instead create a symlink to it
in the workspace and use local_repository().

This should make running `fx rbe auth` faster, as it avoids
downloading / extracting the archive on every run.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169476
Original-Revision: 8e300f4c011a58e479def7cab59cbebb890da340
GitOrigin-RevId: 0da783ef9c8255056ff0c18854840c0fa49ff272
Change-Id: Ie97f37e538e9646d500cff87b8c12c8bda22b5db