[roll] Roll fuchsia [starnix] Support futex_wait on CLOCK_REALTIME
Add support for passing CLOCK_REALTIME deadlines to futex_wait in
starnix. Previously, this relied on translating the UTC deadline to a
montonic deadline and waiting, which is fragile and will be buggy once
we start pausing the monotonic clock.

This CL supports waiting on CLOCK_REALTIME deadlines by translating them
to ZX_CLOCK_BOOT deadlines. This is still a bit fragile, but we cannot
directly wait on a UTC deadline so this is the best we can do. Switching
to waiting on ZX_CLOCK_BOOT meant that we could no longer use an
InterruptibleEvent to block, as that struct uses a zx::FutexWait to
perform the wait. Instead, we switch to using a Waiter (which defaults
to using a PortWaiter) to block until either:
1. A boot timer is signalled, or
2. A futex_wake notifies the port.

Original-Bug: 361583830
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164957
Original-Revision: e679e919095440e150f61cf17b507b7c540300b7
GitOrigin-RevId: 0d9054d35fe688c2a1e114ca7b84fd1d1b5c6469
Change-Id: I6767388147e44cb8f66036e88ced5a14f00042eb