[roll] Roll fuchsia [sestarnix] Handle Tasks being gone, for Task-labeled FsNodes
Some FsNodes are labeled with an associated Task, whose current
label is the FsNode's effective label. An edge-case exists when an
FsNode is access-checked, that is labeled with a Task that has
already terminated.

We now return the policy's "unlabeled" SID in this case, allowing
more code to rely on all FsNodes having been labeled.

Original-Bug: 381210513
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170712
Original-Revision: 550595368397a5a738f472574e3ebe51ff0dfe9d
GitOrigin-RevId: eaff67fd3ccc6ddf3f1d282d1eb8908cc6ab1b83
Change-Id: I3139aaea22b339dfd8f45c148423ec38fdc85eca