[roll] Roll fuchsia [starnix] Implement MREMAP_DONTUNMAP
For private anonymous memory, MREMAP_DONTUNMAP remaps the mapping while
leaving a special mapping behind. This mapping will generate a pagefault
on each access that will be resolved by either userfaultfd or, if the
range is not registered with one, by zero-filling it. This CL only
handles the latter case.

For memory that is not private anonymous, the behavior is not documented
but seems to be similar to the case where remap is called with source
length of zero.

Test: Mremap.MremapDontUnmap*
Original-Fixed: 297372077

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163952
Original-Revision: 717579320fbbef0a954800d88f3d1abcb93892cc
GitOrigin-RevId: 175776f364465a4fde0533b7fa01a62fe4314097
Change-Id: If89376b4942026f1fadab008cd6710995728dcac