[roll] Roll fuchsia [netstack3/conntrack] Remove num_connections
Instead of capping the number of connections, we cap the number of
entries in the table, which is what we actually care about. This mostly
means replacing one constant with one that's twice as big.

Since this change updated the name of the inspect value from
"num_connections" to "num_entries", also update the sampler
configuration to export that instead.

Original-Fixed: 381946038
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170952
Original-Revision: e4a1a5e624201500c5417658ab93cd54b3e795cb
GitOrigin-RevId: ed6c2fe403724f720ef9922fe92fe29057aef9b9
Change-Id: I8522896bc516edb2709f7896306b18c37f09f690