[roll] Roll fuchsia [owners] Remove chok@google.com from tools/build/ninjago/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162772
Original-Revision: 4dc7969b1eca45aec5a5b79037db8db6b8a1d999
GitOrigin-RevId: 7919d131246f2916bcd34fc80218ed63abf106f5
Change-Id: I13ac345c4fb5f1e7ef9258f14d2725c1529fbcd2