[roll] Roll infra/recipes [owners] Remove tmandry@google.com from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1169762
Original-Revision: 3055232d3e421eef7f40100605c1a33402ca8e88
GitOrigin-RevId: aa7dcd74db1127b68d98e8de97547d66be5ad637
Change-Id: I132c0fa6805ad731e86f118ef9e06404d6aa14a7