[roll] Roll fuchsia [bazel_sdk][versioning] Handle unspecified API level after transition
Previously, fuchsia_transition failed if an API level was not specified
via any of the options. Failing in the transition prevented some
negative tests. In addition, the empty string was also handled the
FIDL rule.

As previously done for invalid API levels, allow the transition to
complete then fail when when analyzing the fuchsia_api_level rule.
The name of the relevant target is not known at this point, but the
Bazel error will indicate the target for which analysis failed.

Original-Bug: 354047162, 382773036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170002
Original-Revision: 0ae08c07324f467ac81fec0d84adf46e1f052e9b
GitOrigin-RevId: 4d28b846f92bf4556737616f0878ea4994e9a295
Change-Id: I97405c67df3a86171b0bc5bcc0aed6bcbf1d591b