[roll] Roll fuchsia [kernel][vm] Correct reference value align bits
The extra two bits used to be used for split bit tracking directly in
the page list node reference value. This was moved to be tracked as
metadata associated with a compressed reference in the compression
storage system and so this alignment can be reduced.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171573
Original-Revision: 9030ccd456e1b05f0f7a23b632e366cf4c9a24c0
GitOrigin-RevId: 49bb3cd5d4b7bcb537d1a40f29431f23c9b22aeb
Change-Id: I4f56660e2dafa6b8f4c2ed5d8dc8d5a4ffdaf36e