[roll] Roll fuchsia [ffx] Use longer knock timeout in `target wait`
Currently `ffx target wait` uses a 3-second-long knock timeout while
polling for the device to be up in `ffx target wait`. If the None
being waited on needs a security-key tap in order for SSH to succeed,
or if the target is sufficiently busy during boot, this can be too
aggressive a timeout. (This is observed in practice by noticing that
running `ffx target echo` in a loop will succeed much faster in these
scenarios than `ffx target wait` -- by around a minute.)

Address this by allowing the daemonless knock method to use its own
default timeout of 6 seconds to account for the additional setup cost.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169238
Original-Revision: ae8c81894277072432fa79a23af1c4c257c9bcb7
GitOrigin-RevId: a0f83e8e394b361e0e5c590bb7294f486fc42a05
Change-Id: I8ae8254d582c3b5293d05e157c17e2b565304ac1