[roll] Roll fuchsia [firmware] conditionally compile bootfs APIs
Passing files from the bootloader using the bootfs ZBI type was not
intended to be stable, but has become load-bearing on some legacy
devices. New devices should not use this, but the helper functions in
libzircon_boot required manually pulling in the bootfs.h definitions
(which are not part of the Firmware SDK) whether a device wanted them or
not.

This CL conditionally compiles the bootfs helpers, so that legacy boards
can define it and bring in bootfs support, but boards that don't need it
(e.g. vim3) can ignore it.

This also updates the Rust code for testing the bootfs APIs, which had
bit-rotted and needed a small update to compile.

Original-Bug: b/378736089
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173135
Original-Revision: 4718b84cba93fc92ebd4a1be49cf528c48468465
GitOrigin-RevId: 70547e87809fd76318af5c507d41852c6b158195
Change-Id: I458e53ce2d0717969981de922bf649026f86fb2b