[roll] Roll fuchsia [starnix][tests] Stop relying on bionic uapi in syscall tests
This stops adding the bionic uapi include dirs to the include path for
syscall tests. Since we do not have a full Android toolchain trying to
use these header files as if they were system header files leads to
conflicts and confusion as the headers evolve. Instead, tests that need
particular features that are not available in the sysroot we currently
target for Linux (glibc 2.31) simply define the things that they need
locally with two exceptions:

1. ashmem_test.cc is testing an Android specific feature (ashmem) so it
   uses the ashmem header directly from the bionic uapi folder. It uses
   a full source-absolute include path to avoid polluting the include
   path for other headers.

2. bpf_test.cc uses a large number of features that are difficult to
   define locally as the header file uses enums instead of preprocessor
   symbols to define new functionality. This test also uses the bionic
   uapi header directly.

There is still some risk that the bionic uapi headers for ashmem and bpf
may cause us headaches when combined with a Linux glibc based toolchain.
The risk is lower when using only these two headers directly. If we
continue to have trouble, we can move these tests to a location where
they can use a full Android toolchain.

Original-Fixed: 350479500

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1172434
Original-Revision: 986862ad9dce9717890a67d0ef1828a20ec8572a
GitOrigin-RevId: c0a22f0792879aea5dc9ca03b977a575960233f8
Change-Id: I56cebad5c990aacdc738e59e0ec03593c0415e43