[roll] Roll fuchsia [starnix] Send a Freeze signal with a Waiter to block the target task
When a cgroup is frozen, create a Waiter for each task in the cgroup.
Send the freeze kernel signal to each task with the Waiter to be
blocked at, waiting for an interrupt or a notification.

Handling of SIGKILL when frozen will be addressed in future CLs.

Original-Bug: b/333766695, b/377755814, b/377755814
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173136
Original-Revision: 624c20695c59811522973af2d998895449994798
GitOrigin-RevId: b71cda1c9da9afd46f25d3717b83e0f706cecf96
Change-Id: I97dc5a0c945ca49b8ea538c5e5b8e7842898546e