[roll] Roll fuchsia [diagnostics] Move Starnix's stub tracking to standalone library
Starnix has a diagnostics facility for tracking "stubs" that are not
fully implemented. When these stubs are reached at runtime, the record
is recorded for inspect and potentially logged. This moves the logic to
a standalone inspect_stubs crate under the diagnostics contrib directory
so that it can be used in other contexts.

Starnix exposes this via the starnix_logging crate to the rest of the
kernel. One difference that this change introduces is that the
`starnix_logging` GN variable no longer controls logging from stubs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171241
Original-Revision: a64edd391300f7f78f125deb881a0bdb59be1694
GitOrigin-RevId: 6fab3e5f6100794c1690651da05699642b8e6291
Change-Id: Iad55f618958976a2d7e18983fb84e3c8a375dd24