[roll] Roll fuchsia [fuchsia-controller] Enable logging upon construction a Context
This changes is a very basic setup for logging to a `fuchsia-controller.log` file
when initializing a Context that talks to a fuchsia device.

Note: Not every Context construction will enable logging since, in
general, initialization for a Context is skipped unless a target is
specified.

Original-Bug: 376290078
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146612
Original-Revision: 12284517878ff6a6684f5e0065dbe4c9142ce0ee
GitOrigin-RevId: 4032aa9a8ed7a0f257bce0b9ea84c01307fbb276
Change-Id: I542aefd793068ac90400bb7771580e1b88e167e2