[roll] Roll fuchsia [assembly] Add dtbo support
dtbo = device tree binary overlay

Boards can supply a devicetree_overlay file, which is plumbed through
assembly and spit into the product bundle.

Partitions configs declare which partitions / slots are dedicated to the
devicetree overlays.

When `ffx target flash` is run, we flash the dtbos to the correct
partition.

When we construct an update package, we map the dtbo to `type=dtbo`, and
the paver must know how to map that to the right partition for the right
slot.

Original-Bug: 383751674
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173555
Original-Revision: 8b4baa57c00b906fda103a7accd4aa9da7edae2d
GitOrigin-RevId: 3f65c526e68b2e63094cd774567863a0c2cf127c
Change-Id: Ib6950c9fd32ca1622b8676f7c441f24a8394b01b