[roll] Roll fuchsia [fint] Ensure that `fint set` calls `build/regenerator`.
This is necessary before adding any more logic to the regenerator
script. For example, [1] tries to move the generation of the Bazel
workspace to build/regenerator, but fails with an error when invoking
Bazel that says that the workspace does not exist.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1169075/3

Original-Bug: 382019521
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174452
Original-Revision: 47e70a485cb1556f74ea0ef92861af9dfa9211cc
GitOrigin-RevId: 7bd3f648be5fad8ecee1928abac52654a18c9f90
Change-Id: Ic80ecd770f9da1321524cc54ba93893f7fa1bb00