[roll] Roll fuchsia [python] Simplify self extracting binary build
The original way of doing self extraction was to embed the ZIP file to
be extracted as a data field in a Rust binary, where there is one binary
built per test. This required a full rebuild of the Rust binary for each
test change.

Two features of ELF and ZIP formats allow us to simplify the
construction of these binaries:
1. ELF files can have arbitrary data appended, which does not affect
   their execution.
2. ZIP files can have arbitrary data prepended, which does not affect
   extraction.

Instead of creating a separate binary for each test, we instead create a
base binary, and for each test we create a new copy of that binary with
the ZIP file for the test concatenated.

This speeds up local builds of the RTC test by over 65%, representing a
reduction in development cycles of over 60 seconds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174973
Original-Revision: 13951f3a8d3da798614a69b86306200dbed61caa
GitOrigin-RevId: d7111b7e0c4fa0692abdb5dc5c455eef385058b5
Change-Id: I1b5d2bfbb1467315e638af3f475ac46e7d2a3744