[roll] Roll fuchsia [fint] Add `compilation_mode` to replace `optimize`
"optimize" was originally chosen as the field name for configuring debug
vs. release builds based on the assumption that the `optimize` GN arg
would eventually be used for specifying debug vs. release but that
hasn't happened and `optimize` is used for a different purpose. So
rename the `optimize` fint parameter to `compilation_mode` to avoid
confusion with the `optimize` GN arg.

A soft transition is required to avoid breaking the infra recipes and
configs, so keep accepting `optimize` for now.

Original-Bug: 382773442
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170792
Original-Revision: 673d3d0712f0bf83249bcd2b3e2ae6be0beff5b1
GitOrigin-RevId: e60247a3628fdbf5ae7fdf68ce1dea4906c87fc8
Change-Id: If85d81c29d3f07931b4b84e55151273928cf6940