[roll] Roll fuchsia [netstack3/conntrack] Improve connection eviction
We found that connection eviction was too greedy as originally written
and would lead to connectivity problems when the table filled up. This
is always going to be a tough case, but there are a few things we can do
to improve matters.

In short, we attempt to evict a GC-eligible connection first, and then
pick the connection that's "least established". See the comment in
Table::finalize_connection for more details.

This means we'll always do a full scan for every new connection when the
table is full, but given the current table size is relatively small,
this is unlikely to be a problem in practice.

Original-Fixed: 380888426
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1173793
Original-Revision: 42833b6546792b61ae024d5f93f443180f3325a0
GitOrigin-RevId: e56e64ff8c7d8b1f92f9912ccc97b59003ac82f6
Change-Id: I4b9f42832352975fd89a98148c6f4cb31aaa147b