[roll] Roll fuchsia [assembly] Allowing developer overrides to override kernel args as key=value pairs
Many kernel args are in the form of 'key=value'.  This tracks
the kernel args internally in that form, so that developer
overrides can provide a new value that properly replaces the
existing one in the resultant zbi.

This also allows for the detection of conflicting values for the
same key to now be caught during assembly time.  Before, the
following two would have not been flagged as problematic:
- 'foo.bar=true'
- 'foo.bar=false'

Because these were different strings, they wouldn't be flagged
as a conflict.  Now, they'll trigger duplicate-value-for-key
errors during assembly, unless one comes from the developer
overrides.  In that case, it will always win by replacing the
value in the resultant zbi.

Original-Bug: b/382037107
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168227
Original-Revision: 25830794e5f8ff373f9581dfb6701b4cf1bc076c
GitOrigin-RevId: 7d02c8df01acc40c20eecf9cd94e6161ac296a65
Change-Id: Id2fc688881b13f7b621620bfca0f0f9084207975