[roll] Roll fuchsia [starnix] Check on the timer deadline before raising suspend signal
When the heap has only one timer and Android keeps updating the
timer with exact the same deadline, kernel will raise the signal to
prevent the suspend. We need to check on the deadline to ensure
no change is required to the driver.

Original-Bug: b/384556971
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177573
Original-Revision: 8084215c3261d8a19c3e0e172056cecd092646b1
GitOrigin-RevId: 43f0a417bee4448f21b8fe851b57dd1f78ceee09
Change-Id: I948792c468fa67e014ff84e11ebe9d1269392780