[roll] Roll fuchsia [kernel][phys] Drop to EL1 early in physload
We drop to EL1 - disabling any EL2 monitor - early in phyload, before
setting up the address space. This simplifies the hand-off to EL1, as
then we'll set up the address space in EL1 as opposed to EL2 and don't
have to futz with arm-specific reinstallation. This is prudent also in
that the kernel proper no longer executes in EL2 and we have no need to
do so in physload/physboot.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1176696
Original-Revision: 5abb290a7c4c48c2755f128391decc8669b82a6e
GitOrigin-RevId: 4cd20f3e3369c6c4ef0d68a0b72a6568078f257d
Change-Id: I540727f0f6516b4c578993f2ce5cf42022816da2