[roll] Roll fuchsia [assembly] Stop adding build-info/minimum-utc-stamp
We have two backstop files:

  /config/build-info/minimum-utc-stamp is supplied by the product,
  but is never used.

  /boot/config/build_info/minimum_utc_stamp is supplied by the platform,
  and is used as the backstop.

We are removing support for the former, because it is unused. The first
step is to make sure the file supplied by the product is not added to
the final images. Then we can remove the usages in product, and remove
the config field altogether.

Original-Bug: 356621614
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179433
Original-Revision: 3446459f9849d900ea2f2e2611b099d82d82333a
GitOrigin-RevId: 9a7bafd342a83543f47b8b1177960095b6220376
Change-Id: If820bfee06f3dd655984d3237ee313e4df0a0883