[roll] Roll fuchsia [virtio-gpu-display] Fix incorrect cleanup during driver shutdown.
Due to a typo, DisplayEngineEventsBanjo::SetListener() does not
correctly reset the listener when given a nullptr. The bug results in
event listeners still being called. This CL fixes the bug, to address
any false rejections in CI/CQ that may be caused by this problem.

In the future, the Banjo adapters will be lifted into a library with
full test coverage. This bug was discovered by the work for that
project.

Original-Bug: 42079190
Test: none (test coverage will be provided in future CLs)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178539
Original-Revision: 534aae311de21e9a8fc6076fd035b085b5058d4e
GitOrigin-RevId: 7e5aad4508b55a792591ed0f3ad35392191145f8
Change-Id: Ib74acb7d47bfddaa0840650300e4aab80091a958