[roll] Roll fuchsia [owners] Remove luluwang@google.com from src/connectivity/bluetooth/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/luluwang@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179719
Original-Revision: b1de10a7acb0028a9d738bfd1162e7c060db98c1
GitOrigin-RevId: c42ad13758a2260bf503c2152c1d29fe1bf43e61
Change-Id: I0910c3b748de4e961c9a5f458e01b6172093443d