[roll] Roll fuchsia [owners] Remove anmittal@google.com from examples/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179726
Original-Revision: 444439820618d0c54e7564353743da08d11376f6
GitOrigin-RevId: 9f4cfd8a9929ea5684110d42d5756ee76bb83e2a
Change-Id: Ifebe64894f00a75cf813c57a60488df105cf2424