[roll] Roll fuchsia [owners] Remove tmandry@google.com from src/lib/fuchsia-async-macro/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179728
Original-Revision: 346de84a373dca26731e64098f9d923f487a67d4
GitOrigin-RevId: d6d606a6b8dd11d907815b32c3586eacd5c558c2
Change-Id: I1521646660ff0833e346df155d1a2d371676c29f