[roll] Roll fuchsia [owners] Remove tmandry@google.com from tools/devshell/contrib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180369
Original-Revision: 48302f0d13caf293b54a7843714c5baa6cf759e5
GitOrigin-RevId: 17c9e6a70ea7880fd63b8468b10638d669ba337c
Change-Id: I409618dec5200b4375ebb1e917dab35c8eddea17