[roll] Roll fuchsia [owners] Remove anmittal@google.com from sdk/fidl/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180370
Original-Revision: a9fd59c52bbe23771e9b2b2506dcf45f4460b3c6
GitOrigin-RevId: 4e75152290d652882a5f0b110a685586efbf9716
Change-Id: Ie94d5b63acea0499111f5d92d31d0545c9c3a713