[roll] Roll fuchsia [owners] Remove anmittal@google.com from sdk/fidl/fuchsia.debugdata/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179722
Original-Revision: 065088cc7c51b70dcc06c0e39c1beaa4d6a07d0b
GitOrigin-RevId: 2cc43aacac23a2875a8391ab754c15d2f5ce9135
Change-Id: I6f70a633747d3f6386acc368bbd61745410a623e