[roll] Roll fuchsia [display] Create standalone protocol adapters library.
This CL extracts the adapters in the virtio-gpu-display driver. The code
is similar to the driver, with one exception -- the
DisplayEngineBanjoAdapter class became a template, so it can accommodate
different DisplayEngineInterface implementations.

This CL has fairly empty tests for the adapter implementations. Unit
tests will be fleshed out in separate CLs, so we can iterate on the
approach to mocking.

Original-Bug: 42079190
Test: fx test //src/graphics/display/lib/api-protocols
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178735
Original-Revision: eac9caaa3581818b49e3f79b9807709e130568be
GitOrigin-RevId: 161aadc679bbd6e1cc8e03d961e5d53d5035a7fe
Change-Id: I8dc3991aa951fa314d4e0018a90226d04284df18