[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/pylibs/mypy/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162775
Original-Revision: 59df3394f205455cfef4550aa941c5dbdb404e32
GitOrigin-RevId: 660f69a5cd05f2e822fc2a09128016f14ba7c0f0
Change-Id: I3db4708913842daadcdaf212f8b8c78b5f7bb2df