[roll] Roll fuchsia [owners] Remove chok@google.com from scripts/lib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162777
Original-Revision: a98872564700cf30dd097bac0ddab4f389d93fed
GitOrigin-RevId: 109826a9e86dddf78143cc300e60c5a7ba72af0b
Change-Id: Ie6fc0ad4a54d3f2b5bdb42370fb7645e977b1f35