[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/pylibs/autoflake/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162778
Original-Revision: 001a65aa84a7ef08dcefe8b4fc64e70cad20fd58
GitOrigin-RevId: 65892d330c21c24e33a38ffe57f6594cbfd2a674
Change-Id: I61f61dffc65edc632dd380e830712311dab1df0f