[roll] Roll fuchsia [owners] Remove tmandry@google.com from tools/multifuchsia/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179730
Original-Revision: b28169b887f8d8d28769d4eda923be45e5753a19
GitOrigin-RevId: 47cbbb7b08132b46767bb7c050e24e328f4093cc
Change-Id: Id75583134306329e00292f92ad30a26962ae6a34