[roll] Roll fuchsia [owners] Remove tmandry@google.com from tools/rust_extract/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180354
Original-Revision: f5b321487f9075df846011774f7dc9d149745115
GitOrigin-RevId: 16f7ba4098f9377b83218c5428666cd372f7ac5c
Change-Id: I7cc4de9197ee3eac8510aacf5b79cec50155aef7