[roll] Roll fuchsia [owners] Remove tmandry@google.com from build/rust/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179723
Original-Revision: cac893c53d0a90225224dc4aabe4e38c62859477
GitOrigin-RevId: ac1832e6a4eea6ae6cf1a89289353672b0040120
Change-Id: I297a94b91e10bcf816af21381b037d9e7e8980cb