[roll] Roll fuchsia [fidl][rust] Place server and client in dispatcher
Dispatchers need to pass a reference to the server or client to the
handler so it can properly close the connection if an error occurs. This
threads the server or client through so that handlers don't need to
store their own copies at the high level.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168655
Original-Revision: caba3f204941ffa827101ae64547ab2ce6d471bd
GitOrigin-RevId: 8f0b0afb63a3e6da762a4f634031032d290f38e6
Change-Id: I568bbe2ed6431c36127b0ab3aa9858d83200973e