[roll] Roll fuchsia [owners] Remove maheshsr@google.com from src/testing/end_to_end/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182593
Original-Revision: c59cb7453c91d931705e5eab933881ab7fc7c5bb
GitOrigin-RevId: 50bfa4bfdada1c3e175aa4c070e6ceacf6610f6b
Change-Id: If569310de4512212eefa6da632f7490d561b77be