[roll] Roll fuchsia [owners] Remove tmandry@google.com from src/lib/scoped_task/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179729
Original-Revision: 288e6ac5e73c293534f406a0452d7948ac553874
GitOrigin-RevId: 84384831b084cd3e1faa64db665d035f84c76249
Change-Id: I41d95ce91666274a1dec3eac3b915238c4701f0d