[roll] Roll fuchsia [owners] Remove anmittal@google.com from src/sys/test_manager/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179720
Original-Revision: 6bddb68a4ce8d8a1400fe0bd764d262ff3ca1c57
GitOrigin-RevId: 1ca39dd1fc7620cc1a152199b83dc995d49e86b7
Change-Id: I578143596e7a93c4959091161e2c7df006591ccd