[roll] Roll fuchsia [owners] Remove maheshsr@google.com from third_party/pylibs/typing_extensions/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182599
Original-Revision: 23e33dca93b0be95ce9f3000286a9f00a4d789e4
GitOrigin-RevId: 7564838cf34cf1bd76eb8e157127667f75fa56fc
Change-Id: Ic82bc2cfd9604b971efba7bb6c39427a02e9ebe6