[roll] Roll fuchsia [socket-proxy] Wait for response on registry call
After each call to the registry, wait for a response from the server
watcher to ensure that we receive the number of updates we are
expecting.

Move the call that doesn't result in a DNS update to the end so that we
can use `now_or_never`.

Multiply: socket-proxy-integration-test
Original-Fixed: 369741015
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126723
Original-Revision: dda8f77c3cdf59a1f9f2eb39081da9d2944ab31e
GitOrigin-RevId: 0d90a345816065738f49daac9e8b52991bf5df93
Change-Id: If0cde9258527cd25926ca750909d980b8b0e3c64