[roll] Roll fuchsia [owners] Remove maheshsr@google.com from third_party/pylibs/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182595
Original-Revision: 04910b8ce60f00fe73b7527ae0a864661ed4e01c
GitOrigin-RevId: 02b330ca3f5fecec5d2d4cc3be9799bcdbf98c4d
Change-Id: Ia87b6a8a377c0e9f54266224070d56420cb7562a