[roll] Roll fuchsia [owners] Remove maheshsr@google.com from third_party/pylibs/mypy/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182612
Original-Revision: 51e5d8154f5cead0919769ba7d957bb52ebacaf7
GitOrigin-RevId: 5d7d1f07558746f060d67554fbdfe41d7334d58b
Change-Id: I03b812de8ebc9b9bae85429ecabf5f2364448bb5