[roll] Roll fuchsia [owners] Remove maheshsr@google.com from third_party/pylibs/autoflake/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182596
Original-Revision: 24a65d8d7d0d97739df0d6d76623625599e4dde0
GitOrigin-RevId: 95d1597890b48b66b9bf5971ef42b160f9eadeea
Change-Id: I8679d7e8d9c34e255309db1f702acba0d82bd219