[roll] Roll fuchsia [owners] Remove maheshsr@google.com from src/testing/end_to_end/honeydew/honeydew/interfaces/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182594
Original-Revision: 2b41d3399106c185f0efca296783ae94e91c1cee
GitOrigin-RevId: d8246ed23ac4cb9ac0d4504582bf6880d311315c
Change-Id: I2f6632ff3c82b7636b44c3aaa9abde477f511127