[roll] Roll fuchsia [owners] Remove tmandry@google.com from docs/development/api/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/tmandry@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180356
Original-Revision: 651edc97f2b1fe1fc5ae26733f3e04efca4332bb
GitOrigin-RevId: 5c306888ceb084b1c6da151b784e055e2e5a1ce6
Change-Id: I8cd7e2bd053568bf00c3237862198221269c174e