[roll] Roll fuchsia [kernel][x86] Unify PAGE_SIZE definition
Changes the definition of PAGE_SIZE on x86 to match the other two
architectures. The visible change here is the addition of the `L`
suffix that makes the PAGE_SIZE constant a long instead of int, which
prevents writing code that is typed correctly on one arch and not the
other.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177472
Original-Revision: 6a350869832ef81f52f3ad139cadac10c95a642c
GitOrigin-RevId: c3c28cbdd11c6f4b1a9886a55c42f7a75265e79d
Change-Id: Id345f94a7cde6599d141807c9c8782d122b1d1b8