[roll] Roll fuchsia [fdomain] Don't poll the transport in transaction futures
There was a mistaken idea at one point in FDomain's design process that
you could poll the same future from two different tasks and the result
would be both tasks waking up when the future was ready.

This wasn't the case and we already fixed a few bugs related to this,
but there's one more architectural spot where we were relying on this
assumption and it was causing some subtle bugs.

The upshot of this from a user perspective is that when you create a
client, you now *must* poll the future you get along with the client
object to completion in its own task. Previously we insinuated that this
was optional so long as you always awaited futures you got from
individual method calls on the client. That's a weird "feature" to offer
anyway, and we never rely on it elsewhere, so good riddance.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1181773
Original-Revision: 8ec9dbc353cc299496c4576e23085cecb2d3d7ba
GitOrigin-RevId: 0b76c8158bfc110870d9dc4dbc7c90b6427f50c1
Change-Id: I2f341825b061ab50cac9f03ecef5b0134dd6f81a