[roll] Roll fuchsia [kernel][phys] Drop to EL1 early in physload
We drop to EL1 - disabling any EL2 monitor - early in phyload, before
setting up the address space. This simplifies the hand-off to EL1, as
then we'll set up the address space in EL1 as opposed to EL2 and don't
have to futz with arm-specific reinstallation. This is prudent also in
that the kernel proper no longer executes in EL2 and we have no need to
do so in physload/physboot.

This is a reland of fxrev.dev/1176696, which was reverted due to
instability around a related change. Nothing has changed with this one.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182457
Original-Revision: b076fcd46ef4c5bb25b6449f5323f6e42bf08030
GitOrigin-RevId: f6fcc77d6f019e541e4a84cb4a67f9d111d014da
Change-Id: I5d4d1aa3274a2f81d701e8ba3e7892eb4bf7d2e5