[roll] Roll fuchsia [bazel] Generate remote_services.bazelrc file.
This is a reland of [1] which was reverted. Compare to patchset 1
on Gerrit to see what was changed.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1177873/7

------ original commit message -------------------------------

Instead of appending its content to bazel-workspace/.bazelrc,
write the content to a separate file, and make the `bazel`
launcher and `bazel_test.py` scripts use it directly.

This also moves the generate_remote_services_bazelrc target
to //build/bazel/bazel_sdk:remote_services_bazelrc, as it is
only used to run the Bazel SDK test suite without generating
anything with the internal Bazel workspace, as it happens
when generating and validating the Code IDK.

Original-Bug: 384956017
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183712
Original-Revision: 455976f51a9067df3400df06033c40e2789268a6
GitOrigin-RevId: aea64fc9c3cdc6b6cc0b9724022516d21c6fa854
Change-Id: Ie2907fb7067ee7f1b83768dfbc9288ec8fa84b58