[roll] Roll fuchsia [sestarnix] Fix create_test_executable() to set the context
Fix the create_test_executable() test helper to set the caller-
supplied security context using the "fscreate" mechanism, rather
than by manually setting the xattr, to ensure that the label is set
correctly on the in-memory representation of the node.

Add an explicit test for create_test_executable(), as well as for
the new ScopedFsCreate mechanism, which will be reused in future by
the overlay filesystem labeling hooks.

Original-Bug: 388241218, 364568931
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183892
Original-Revision: e18142bbbd7198fa390e352a94f6946a9b7c8bbe
GitOrigin-RevId: 1638f1491cbf53e4a2e26cb68fa67084774e0545
Change-Id: If1b9ffb471cecb98d6ac5987921e2959010db3a0