[roll] Roll fuchsia [kernel][vm] Remove extra IPI in ForceFlush on x86
The ForceFlush method in x86 previously sent two IPIs when
invalidating a TLB entry in a restricted page table - the first was
to clear entries in the restricted page table itself, and the second
was to clear entries in the associated unified page table.

This change consolidates the invalidations for both page tables into
1 IPI that causes both PCIDs to be flushed on cores running either
the restricted or unified aspace.

Original-Bug: 319325273
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177986
Original-Revision: 3039d4ba135f0fcb68aa1fdacfec08eed38c34db
GitOrigin-RevId: 7565c849846ad13798905151d772e75cbe1dde06
Change-Id: Ibe3accfa8e0020f6ccbb8527878bc8b0381db9f6