[roll] Roll fuchsia [framebuffer-display] Switch to library-based Banjo adapter.
This CL deletes the custom Banjo display engine implementation in the
framebuffer-display library, and switches the library and drivers to the
API in the lib/api-protocols/cpp library.

This work gets the driver's code closer to virtio-gpu-display, in the
hope of making future API migrations easier.

Original-Bug: 42080631
Test: fx test //src/graphics/display/lib/framebuffer-display
Test: fx qemu -g -N --no-virtio
Test: ffx target ssh display-tool squares
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183413
Original-Revision: 0a1937ae657013365512454b70279a28870b861c
GitOrigin-RevId: ecbb8a03d40058fb35e087ba79bf76cfc9de3466
Change-Id: I3e07f83252240aa8e9573d256e75aa77916acf8f