[roll] Roll fuchsia [bazel_sdk] created generated_constants file for fuchsia_clang
The generated_constants file needs to be created for the fuchsia_clang
repository. This is done for all code paths except for when we fetch
directly from a cipd tag. This adds the logic to ensure that it is
added.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1184782
Original-Revision: d85e4baacb3a7db179f2c91189d81b9aee6d917f
GitOrigin-RevId: 2ba03d1294688fd717319cba9f996ad1c7845b83
Change-Id: I5538c40e5cb29608c65c69203bf6170b46b81bc6