[roll] Roll fuchsia [ld][testing] Use explicit ::testing::Sequence in MockLoaderService
The ::testing::InSequence object causes all `EXPECT_CALL(...)`
on *any* mock object to be sequenced with all others.  Instead,
use an explicit ::testing::Sequence object belonging to the
MockLoaderService and pass this in `.InSequence(...)` on every
`EXPECT_CALL(...)` on the mock.  That way, expectations are only
sequenced with respect to other expectations on the same mock.
Since the underlying mock is not exposed by MockLoaderService,
it's easy to ensure this is on all (both) its wrapper methods
that do an actual `EXPECT_CALL(...)`.

Original-Bug: 382340693
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186315
Original-Revision: 80829e60afb556b3195d36ad9419ea055b87811d
GitOrigin-RevId: f862c2c985eccbb5d63166f1d79c64e5102ce135
Change-Id: Ife596d8008a613b86584fe9b9a7f4a73c7cd1306