[roll] Roll fuchsia [role_manager] Move tests into integration directory
Move the integration test in the tests/ directory into the
tests/integration subdirectory. This is in preparation of moving
the tests in //src/zircon/tests/profile into the tests/ directory.
This, in turn, will put all of the config tests in one place and
make testing the new board profiles simpler.

Note that this CL is a functional no-op; it just moves stuff
around.

Original-Bug: 323266882
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183047
Original-Revision: 119d17e9bbd76931a8c39e441695a14bc2cbe557
GitOrigin-RevId: cf5216d3d65ac6dda9c20c77952b2a40e5b51640
Change-Id: I317940a50f47c34a15e6063a166f5b409dc5462e