[roll] Roll fuchsia [tee][runtime] Hang storage objects off of context
This adds a 'storage' entry to the api_impl::Context struct and hangs
the persistent object tracking logic off of the context. This
consolidates static state in the runtime and prepares for implementing
transient objects and operations that have to interact with persistent
and/or transient objects.

We may also want to replace some of the runtime mutability with static
mutability checking inside the persistent object management. This defers
on any changes in that direction to keep the changes simpler.

Original-Bug: 360942417

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185927
Original-Revision: c18a5bac5b7593274d7c572883cac0340cca5521
GitOrigin-RevId: 6ba1099ca582f3bd77a4e0565a08764ffbf8d7b7
Change-Id: I53749793b5b9fe30c8e1258257eb8a1b11becca1