[roll] Roll fuchsia [zircon][vm] Insert new VmCowPages nodes as hidden
When creating the first bidirectional clone of a node in the VmCowPages
tree, we need to split the current node into a hidden parent, which
holds all the pages, and a child that will hold new, dirty pages.
Previously, we created an empty node for the child and mutated the
current node to make it hidden. After this CL, we create the hidden
parent as a fresh VmCowPages object and move all the pages from the
existing node to this parent.

This CL is expected to regress a number of VMO clone microbenchmarks.
These regressions are necessary in order to restructure the VmCowPages
data structure to unlock larger performance wins in future CLs.

Original-Bug: b/338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186497
Original-Revision: ac93b88033033bcc18eddec4ce97a99653e4db53
GitOrigin-RevId: fc7b46101355f9d266dd0ebfa1109968868f534f
Change-Id: Id0338dfc4c1cd38bda8e9e011be93ef31db774ba