[roll] Roll fuchsia [Lacewing] Run commented flaky test in unstable group
As part of fxr/1174812, this test test was commented from not running in
CQ as it caused some build bot issues. But if a stable test starts flaky
then it should be moved under unstable group, so that it will still be
run in FYI builders. Commenting out will make this test not run
altogether including in FYI builders

Original-BUG: 384233293

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187913
Original-Revision: d506a8904dfb02767ba0bcd30e63aea91adca2c2
GitOrigin-RevId: a8de7ad3bbd82e17c8fd67da0679572ef45eecc6
Change-Id: Ib1a7065df8e3c400581e394b16179ab9a1053dd5