[roll] Roll fuchsia [bazel_sdk][tests] Fix bazel_test.py for the in-tree @fuchsia_sdk
This CL changes the implementation and interface of the
bazel_test.py script to ensure that it correctly runs against
the in-tree @fuchsia_sdk and @fuchsia_in_tree_idk repositories,
when invoked with the new `--fuchsia-in-tree-sdk` flag.

The `--fuchsia_sdk_dir` option is renamed to `--fuchsia_sdk_directory`
which is consistent with `--fuchsia_idk_directory`. For now it still
expects a self-contained SDK directory (i.e. one that does not load
rules from @rules_fuchsia).

+ Update README.md accordingly.

Original-Fixed: 383536158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1171837
Original-Revision: 0d648f621526e0d596a911db8042cd6e34111efd
GitOrigin-RevId: 099297a816b6fd6a34f1bdb4dda296912c06f8a6
Change-Id: I9b07b212180f95514985293f1a59feb5b5c1865d