[roll] Roll fuchsia [component] Simplify log messages
Too many monikers being logged! Not every level of logging needs to
log the moniker of the component with an error. It's sufficient to log
the moniker only when there are multiple paths into the error (and
when that error might not be wrapped by another error that also logs
the moniker).

Original-Bug: b/384927637
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185554
Original-Revision: b8d175f6aa2e60f81ca1c9fefeac8a674370fa99
GitOrigin-RevId: aa8b94c7f2d5000ad0344d08fe0cb5d0dd6f2edd
Change-Id: Ifdffff369400d3ad50944477af72e8f9bf80f644