[roll] Roll fuchsia [fuchsia-controller] Disallow undefined symbols in C extensions
This change adds an action that generates a linker script to
create a weak symbol for each dynamic symbol in the python interpreter
prebuilt that's part of the Python C API found. By doing so, a shared
library for a Python C extension can be built while disallowing
undefined symbols.

This likely all sounds a bit obtuse. After all, it's generally okay to
have undefined symbols dynamically linked to symbols in other shared
libraries. The problem is that the Fuchsia tree does not contain
libpython and so we cannot currently build the shared library in such a
way that the linker can dynamically link the Python C API symbols to
libpython.

Original-Fixed: 295680434
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178099
Original-Revision: d9e78d3283a95ed9ab4fc993fc0294dbd1ed5d59
GitOrigin-RevId: aba30051bedeee03c38307353882e6e202e9322a
Change-Id: Icd16b88e5381f80fed5f3fbb0a29f272bde5b473