[roll] Roll fuchsia [owners] Remove pineapple@google.com from third_party/rust_crates/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189386
Original-Revision: 1d60cf0fb698bc3c2fda1cb6b3471ef2a6f7873a
GitOrigin-RevId: 950ec172fb492df1a564b5333baa0035704707e8
Change-Id: I79a5a4c01f005abe798a6db86cf4789ec189ac0d