[roll] Roll infra/recipes [owners] Remove pineapple@google.com from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1190594
Original-Revision: 6963c07a85120b720bce34807ba5007e2b656876
GitOrigin-RevId: 1a2c3ac4e517cb113ce6cb1211b450e8444fa735
Change-Id: Ia113cafcd4b86f55a7b85cc6e30565e71d66de7e