[roll] Roll fuchsia [tee] Implement the transient storage interface
Well, most of it. This change introduces the key types associated with
transient objects, but with stubs to be properly implemented when we add
support for particular cryptographic schemes. The transient storage
interface is in turn fully implemented in terms of these types.

Original-Bug: 371213067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183052
Original-Revision: 09da47744ffbf67441943a5b81b0282ff3f9f251
GitOrigin-RevId: c58772a595ac8fdf1150b3c3e926183cfb6e5e36
Change-Id: I659cb17323b85cdf02812c958dd0485c3e16def0