[roll] Roll infra/recipes [owners] Remove pineapple@google.com from recipes/contrib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1190595
Original-Revision: f70507d5b0f3bd8390ddf170a3ce55eb23974cd2
GitOrigin-RevId: 5bc87d1c5fe70035e7422709f06ebe9b9cb47f51
Change-Id: I2df1b859881add6acf78815f3f59c00fcce17f40