[roll] Roll fuchsia [virtual_audio] Simplify device shutdown
Virtual audio devices will be given a callback to be invoked when they
have completely shutdown. A device is completely shutdown when its fidl
server and driver have completely shutdown. These drivers will also be
given a callback that they will invoke only when the driver is
completely shutdown. When a driver is completely shutdown is determined
by the different implementations of the driver.

The virtual_audio driver itself will keep track of which devices have
been shutdown and remove them. Originally, devices that shutdown were
not removed and instead marked as unbound.

This simplifies the logic for waiting for all devices to shutdown. This
also reduces the need for a cyclic dependency between
`VirtualAudioDevice` and `VirtualAudioDriver`.

Original-Bug: 388880875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189216
Original-Revision: 207815f700c463efa9a49f75ae356f60b0086cff
GitOrigin-RevId: 98f45a86645ecf397b9d6687e2bec0fd2b2138f5
Change-Id: I217a9959c6f54a3be6d2a236a2ef5e5cbd25d1cb