[roll] Roll fuchsia [sag] Don't block AcquireWakeLease on satisfaction.
This makes AcquireWakeLease safe to call from OnSuspendStarted, in
addition to guaranteeing that suspension is blocked earlier in the
call.

If a client needs to know when the lease is satisfied, we can add
an additional method to support that.

Multiply: fuchsia-pkg://fuchsia.com/system-activity-governor-test#meta/test-root.cm

Original-Fixed: 389163722

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190492
Original-Revision: be2210b5296925a329dda61016cc964522bf01c0
GitOrigin-RevId: 29e222e1b95b2c22ae1c45d105eacdcbf03094d9
Change-Id: I4a29a053cd5e7ec45cfbfc7e1b7890df39c083b0