[roll] Roll fuchsia [owners] Remove pineapple@google.com from docs/development/api/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189385
Original-Revision: 0c8c3b03e3754c3a1903b7ff8063524ff8ce8e41
GitOrigin-RevId: ffe5685d3733b2ad7c20e7cdf64e9e4dd7bbc9af
Change-Id: I5f6ac151e4dfc0eb5e8f28f3a4550d3d1df41114