[roll] Roll fuchsia [owners] Remove pineapple@google.com from tools/devshell/contrib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189387
Original-Revision: 9187729c65c52ea1bfeca7495f932ce7144aa9b7
GitOrigin-RevId: aaf3e54ce0129de1e7d11225984bd26cbb63a140
Change-Id: I931908a0ffa0a97e67445d3c2ccd2314027f1759