[roll] Roll fuchsia [zxtest] Invert conditions in streams.
This allows us to generate if/else branches that are always paired,
avoiding unexpected results when a single if branch is expanded.

if(true)      if (true)
   MACRO   ->     MACRO- if (false)
else       ->            else

To:
if (true)
    MACRO- if
    MACRO- else
else

This also uncovered an exisitng bug in a macro
LIB_ZXTEST_HEXDUMP_PRINTER and introduced a warning dangling-else.

The first one, is due to the macro parameter being named size, and the
variable used in the macro body being byte_count. Just happens that it
matches the context where it is used. This has been fixed. Some
indirection is needed, to generate storage for the variable when the
value comes from expression such as sizeof(T).

The second one, needs some specifc tricks to trick the compiler into not
identifying the dangling else.

Original-Fixed: 388300636
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191212
Original-Revision: d05114e0787e57eeeb88bb7eac72b7f2e403b563
GitOrigin-RevId: b46dcd28746dc4d32712cfcd9e20c5b0231041af
Change-Id: I5f032781fed79a3bbcc0c57d1e857df987247e39