[roll] Roll fuchsia [fuchsia.io] Add flags for unnamed temporary files
We want to add support to open an unnamed temporary files. This file can
be linked into the filesystem to make it persistent. However, it can
also be created to NOT be linkable.

This defines a new member of Open3's `Flags` to indicate to open the
object as a temporary unnamed object. A protocol is expected to be
specified with this flag to indicate the type of object to be created.
Currently we will only support `PROTOCOL_FLAGS`.

If we want to create an unnamed temporary file that is not linkable,
this is indicated by setting the flags `FLAG_CREATE_AS_UNNAMED_TEMPORARY
| FLAG_TEMPORARY_AS_NOT_LINKABLE`.

Original-Bug: 293950309
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185112
Original-Revision: 74c4b137c17b4e52fdeeb7855bb42fe05a8885ad
GitOrigin-RevId: 6a2b01f08317837bc6eed8c786953687a298babd
Change-Id: I70f15d51d87c89fa6b643d5414aaf776a4b1bcb2