[roll] Roll fuchsia [storage-benchmarks] Use config capabilities
Rather than sniffing for the system configuration, read config
capabilities (fuchsia.fshost.*) to detect how to run the storage
benchmarks.

To achieve this, I added a root shard which routes some capabilities to
the core realm, which is bundled into the "testing_support" AIB.  These
routes aren't particularly sensitive, but I thought it would be better
to start a precedent of keeping capabilities that are exclusively routed
for testing, in shards which are bundled with testing support.

This change also allows the benchmarks to run on storage-host builds
with Fxblob and fixes a few minor issues.

Original-Bug: 372555079
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190816
Original-Revision: 5186e5abe33fe46cc551b9ba11cdda121983725d
GitOrigin-RevId: cd4a117de5e2965967bc0a5ad3f6c1856126cf55
Change-Id: Ic15b97d046c8b164795f98f91407486a18765403