[roll] Roll llvm-project/libc [libc] Use the NVIDIA device allocator for GPU malloc (#124277)
Summary:
This is a blocker on another patch in the OpenMP runtime. The problem is
that NVIDIA truly doesn't handle RPC-based allocations very well. It
cannot reliably update the MMU while a kernel is running and it will
usually deadlock if called from a separate thread due to internal use of
TLS.

This patch just removes the definition of `malloc` and `free` for NVPTX.
The result here is that they will be undefined, which is the cue for the
`nvlink` linker to define them for us. So, as far as `libc` is concerned
it still implements malloc.

GitOrigin-RevId: b2986f256013199bcc6b7b96363fd06d9d28048a
Original-Revision: 15c0944acb56faa6b9409a9bafa997ff2caed7bd
Change-Id: Icfbdfba34b34af9b0b47fd898dbed888c0300d37