[roll] Roll fuchsia [multifuchsia] Touch args.gn upon checkout and rebase
This ensures that upon the first `fx build` in the new checkout or
rebased directory an `fx gen` is run to ensure the gn build graph is
correct. This works around an edge case that results in ninja ending up
in a loop.

Original-Fixed: b/391278102
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190733
Original-Revision: 93d5e64b8fb12232d0749b2dddcc2ea1a5365585
GitOrigin-RevId: 5c2cb5cf171116d6d0f0eb559440c97b8f7fa205
Change-Id: I04db16c66c6b113054b5aa342ed7935cbbcbdd04