[roll] Roll fuchsia [assembly] Move serial core driver to base from bootfs
The serial driver is now a boot driver which is not the ideal case. The
reason is that it should act as the fallback driver for uart child node,
which means it will bind to uart child node when no other driver binds
to the node, as a boot driver it'll seize the node from all the base
drivers because starts earlier during device boot. This CL moves serial
driver to base driver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177132
Size-Review: James Sullivan <jfsulliv@google.com>
Original-Revision: 20718b91c2526c520d7798d1fd38df85ee8d12dd
GitOrigin-RevId: 09f8a30f2f15116d3ac346a86b8961518baa5316
Change-Id: I1f89189ec9c865215f70658b2eaca38c1893e63d