[roll] Roll fuchsia [uart] Introduce Rx/TxInterrupt to properly annotate methods with TA
Instead of passing arbitrary lambdas we are passing an object wrapper
that hold the lambdas and other objects. It provides the TA annotations
to enforce holding certain capabilities when calling such objects.

The API is a compile-time contract that provides methods to access the
capability disable the IRQ (TX or RX).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119761
Original-Revision: 680354df007515c73190caff4404408c56285303
GitOrigin-RevId: afdf8d2180843b77326e54b6827330ffc13420a7
Change-Id: I152f2b6e6d5ceb4a9f1809667aa42bee7253bf79