[roll] Roll fuchsia [graphics][msd-vsi-vip] Add GtestPrinter
Add a GtestPrinter that will log the Gtest output to the driver's
log. This makes it easier to see test output with

 ffx log --filter msd_vsi_test

The main downside to this is maintaining the printing code, and
the extra padding that is logged to the left, but this can be
somewhat mitigated with the --hide-tags option.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193116
Original-Revision: 35d734f2e5642700b2c64e378b7698ba0ce56a37
GitOrigin-RevId: f776cc3f1d194db6b6c26a616d445c3ea6d54de7
Change-Id: I9a330dbd14be00880af8235c44f1adbb824235dd