[roll] Roll fuchsia [ffx][build] Remove default sdk_category from ffx_plugin()
Require that `ffx_plugin()` instances in an SDK specify the
`sdk_category` rather than defaulting to "partner". Instances that do
not specify a category are for internal use only.

Use "not-yet-specified" for the 200+ existing plugins that do not
specify a category to avoid needing to determine the correct category
for all of them in this change. Remove the category from the few
instances that previously specified "internal".

Specifying the appropriate category will allow more accurate application
of the SDK markers and related category enforcements.

Original-Bug: 314822328
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191537
Original-Revision: 4dfbf99edb2e9bd5cfc6bfd346fc241117890cd6
GitOrigin-RevId: fe9d331ce0f000cbb58ddd8c55f0ffe5e47f89bc
Change-Id: I35252be0d9cabbd18185ab526bda4a7c9faf5d5b