[roll] Roll fuchsia [fshost] reenable unformatted_small_disk on minfs
This test was disabled "temporarily" a while ago because of a flake bug,
but it was never enabled again. That flake bug is long closed, and some
sleuthing indicates whatever underlying issue that this was disabled for
was fixed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192774
Original-Revision: d42ac4ed0aa137cd89eff67eb3e9634a1e9c20ba
GitOrigin-RevId: 43332da5de61667c8b13c305016c76144f9c05fa
Change-Id: If18ce88099b459d7a73d5849075d93e20d95be5e