[roll] Roll fuchsia [ffx][fho] Refactoring DeviceLookupDefaultImpl
This refactors DeviceLookupDefaultImpl out of fho and into
target_holders, where it is primarily used when implementing
TryFromEnv for target related objects.

This is part of the project to eliminate the dependencies of fho
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Original-Bug: 392318888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193580
Original-Revision: e51cf2e44d22d1318df38dca345ec9ba179f1da2
GitOrigin-RevId: dc5fa9473758f20de2f3fdd19e0514b96aa74074
Change-Id: I79ca247fb5b437a625aedf99d97bbefc25425645