[roll] Roll fuchsia [fshost] launch fxblob in write_data_file dynamically
write_data_file exists to write files to the on-disk filesystem while
booted in recovery mode, which runs Fuchsia out of a ramdisk filesystem.
When in recovery mode, the ramdisk filesystem is the "main" filesystem,
so the static fxfs child of fshost is serving blob and data from that.

If fxfs gets a start call when it's already running, it shuts down the
current filesystem and starts a new one with the new block device. That
is what was happening with write_data_file - it was running its
temporary filesystem instance in the static child to write a file and
then shutting it down, but that replaced the fxblob instance that was
hosting the system. Instead, it should launch a dynamic child in
fs-collection to host this temporary filesystem instance.

The write_data_file tests were printing errors about failing to shut
down the blob and data volumes because of this, which are going to soon
become failures, preventing this from regressing.

Original-Bug: 391891190
Test: fx test fshost_integration_tests_fxfs_write_data_file
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193376
Original-Revision: e18baca8dcd04d5ccfc3d818dc5cdf1912529444
GitOrigin-RevId: 4c8af968bb26fc869255161b4dd1ffce654b6a00
Change-Id: I2f325dc9060f6e1e6d3cb65315f2e536dd56f9a9