[roll] Roll fuchsia [kernel][phys][arm64] Hand off address space(s)
As already done with riscv64 and x86, this is a major step toward
physboot properly creating the kernel's address space: it hands off the
address space as it is currently constructed, upper and lower - along
with the kernel's mapping with its final page protections in place -
enabling the kernel to bootstrap from the already-constructed identity
map.

This is the second reland of fxrev.dev/1173133; the reverts were due to
instability around secondary boots. It is now believed that this was
due to stores of boot info on the boot CPU - as well as instruction
memory itself - not being flushed in time for loads/execution on
secondaries. This new version addresses that.

Original-Bug: 42164859, 389702612
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190177
Original-Revision: c097d362cafd86317a15132219d038d493db1baf
GitOrigin-RevId: 35f870861691f5c68da76d3d451f3558f54c38c2
Change-Id: Id58bec85d562a14d40b9b61472eb9a773806563d