[roll] Roll fuchsia [owners] Remove pineapple@google.com from tools/devshell/contrib/lib/rust/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/pineapple@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189382
Original-Revision: f32448fe89d38228aa067544d64b3f172e75f37a
GitOrigin-RevId: 45aa1b9c307f08910d041fe7638b1472beaccbcb
Change-Id: I81a060427caacc389ea18ab52064543f11753f79