[roll] Roll fuchsia [fuchsia.io] Replace GetFlags/SetFlags at NEXT
Rename existing GetFlags and SetFlags methods to DeprecatedGetFlags and
DeprecatedSetFlags. Replace with GetFlags2/SetFlags2 by renaming them
back to GetFlags/SetFlags.

This CL does not remove or change any existing usages of either method.
Transitioning callers to the new method will be handled as a follow up.

zxio and fdio both gracefully fallback to the old method if the server
does not have support for the new ones, however this may incur a latency
penalty. Once we complete out-of-tree server support, we will transition
all callers off of the deprecated method and mark it as removed.

Unfortunately this means we'll incur a temporary size increase while we
need to support both methods, but this should be resolved as soon as we
complete out-of-tree support.

Original-Bug: 376509077
Test: No functional change, rename only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189354
Original-Revision: c6384627453dfbff9abf0f4f8eb6df124d7813eb
GitOrigin-RevId: 0f4cb004e0c31c6df0a2e9dbbd7a00a6e7aca0a3
Change-Id: Ia558be674337fb2655f271afb4e4f59dda31a935