[roll] Roll fuchsia [dl] Assert module_ in SetAbiName
SetAbiName is a no-op if `module_` isn't set on the `DecodedModule`
instance. A caller may set the name for the module before setting the
abi module and will not know that the name setting does not stick.

This CL adds an assert that `module_` is set in `SetAbiName` so callers
are enforced to set the module before its name.

Original-Bug: 338238626
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192132
Original-Revision: af1b65508f01dedcfeceaa7013b31ecc78fe6f6e
GitOrigin-RevId: ddcb2f63378bbe2c7f54c0199422738074475ca1
Change-Id: Ib267a81670fe7698594e443e8a4a6bc6566e30f4