[roll] Roll fuchsia [display][coordinator] Sync after ApplyConfig() in integration tests.
Currently, the ApplyConfig() FIDL API and its higher-level ApplyLayers()
are prone to racy usage, leading to flaky tests. For performance
reasons, ApplyConfig() does not wait for the FIDL call to reach the
Display Coordinator and be executed. SendVsyncFromDisplayEngine() does
not go across the same FIDL pipe, so the resulting VSync signal may
reach the Coordinator before the config is applied.

Integration tests currently use a LayerCount() helper (which reaches
into the Coordinator implementation) as a way to wait until a
configuration is applied by the Coordinator. However, the helper is
only reliable when waiting for a configuration with a different number
of layers from the previous configuration.

This CL aims to remove this source of flakiness, and uses the following
approach.

1. ApplyLayers() is modified to issue a GetLastConfigStamp() and return
   its value. The call goes across the same FIDL pipe as the
   ApplyConfig(), so it is guaranteed to complete afterwards.
   A future CL will switch the ApplyLayers() implementation to use
   ApplyConfig3(), at which point ConfigStamp values will be generated
   in tests.

2. Direct usage of ApplyConfig() is followed by a GetLastConfigStamp().
   A future CL will remove the direct usage, and consolidate the
   integration tests around ApplyLayers().

3. DisplayLayerCount() is removed. Using GetLastConfigStamp() is more
   reliable, and the tests don't need to reach inside the Coordinator's
   presentation state.

Original-Bug: 388885807
Test: fx test display-coordinator-unittests --count 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193939
Original-Revision: 321ab900ccb277565143d2edde2028232f13dfb5
GitOrigin-RevId: de928e1786576ef69051bc64a672cc8948738429
Change-Id: I2f09b555b0bb34d70645b6ffcd1799e0560c8c3f