[roll] Roll fuchsia [windowed-stats] Introduce filter mapping to reactors.
This change introduces the `filter_map_data_record` reactor combinator.
This combinator is the first that affects control flow on the basis of
inputs (rather than outputs like `then`, `and`, etc.). Filter mapping
can be used to react only to particular data that may not always be
available. For example, an input event may have variants, some of which
are to be sampled and some of which are to be ignored.

Note that discarding data events does not affect system events, which
are still observed.

Test: unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189412
Original-Revision: cf0fc2740791bfe0d2143cab2e1740b5d8c2aa7b
GitOrigin-RevId: bd9437f69b05d99000d7cbf352e67f0ecde288eb
Change-Id: Ic92c6de3c665b83bf58c1dbd3c1889a054f33bc2