[roll] Roll fuchsia [fdf][compat] Perform all dlopen and dlsym calls in the framework.
This patch adds a modules section to the program stanza of the
component manifest for drivers. This section provides an optional
list of additional modules to load in addition to the driver as well
as a list of symbols the primary driver would like to obtain from
those modules. The symbols are then populated into the existing
symbols field in the start args.

Additonally, in order to support a feature in the compat driver, we
also provide support for overriding dynamic libraries provided during
the dynamic linking phase of the extra modules. This allows us to
redirect libdriver.so to compat.so.

Lastly, the compat driver was updated to take advantage of this
feature and avoid calling dlopen and dlsym itself.

Original-Bug: 360403906
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186157
Original-Revision: f8fdcbf1930cf3cf8c0e63470cbdc6a1a4fd6db1
GitOrigin-RevId: ca3e27989dd72236c5a9f6ec703d0f76a346e389
Change-Id: Ie3ac3b60476ea9b03515cd955a322f51fd951795