[roll] Roll fuchsia [ffx][adb] Use target proxy to get ssh address
This switches the adb connect command to the the TargetProxy
interface to get the device ssh address and nodename.

This is more reliable since it queries the device directly vs.
TargetInfo which was a cached version of information from the daemon.

Original-Bug: 393142801
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195509
Original-Revision: be6ad4f2ef56da0a2823ab1a2f29eecb711ddfc9
GitOrigin-RevId: 6982ba54af2c797bc5140e45ebf09cff1a21921b
Change-Id: I6d34033e3401f3791ad1b0e777a53f14d049ca7d