[roll] Roll fuchsia [kernel][lockup] Refactor mp_all_cpu_startup_sync_hook
Refactor mp_all_cpu_startup_sync_hook to use the new
CanDumpRegistersAndBacktrace function and expand on rationale for why
it's OK to panic there on production builds.

On production builds, if one or more secondaries fail to start up,
we'll attempt to dump diagnostics from the CPUs that failed to
check-in and then panic.  On development builds, we'll emit an oops
instead.  See the comments in the code for the rationale.

Change x86's DumpRegistersAndBacktrace implementation to not require
that interrupts are disabled.  This fixes a potential bug where we
might otherwise recursively panic.

Manually tested by hacking up the secondary startup to force them to
fail to check in, then verifying that we oops or panic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194248
Original-Revision: 57951b353b3569540217b1e6efad83c6aa6545b0
GitOrigin-RevId: 9a1f0d4383e2f182624c1b073888351b04dda85e
Change-Id: Ic265c28a29e91162e2db40aef63ca12481b6a3d4