[roll] Roll fuchsia [starnix] Remove starnix.runner from SDK partner_internal
The unstable FIDL library "fuchsia.starnix.runner" was in the
"partner_internal" category because `ffx_starnix` tool is in the
"partner" category. Use of this library by that tool was removed in
https://fxrev.dev/1170856.

This addresses one instance of an exception to "partner_internal"
stability rules.

Original-Bug: 378470841, 364727245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196796
Original-Revision: 98b0ad819193e5abecb5c4cc153ad4eb332ef8d3
GitOrigin-RevId: 236421127e52b11f5a170a08fd8c9a8ce41a95c2
Change-Id: Ia85dcb1266f3b7fd8f54623957583533fcc9700a