[roll] Roll fuchsia [docsgen] Remove use of //sdk:host_tools.internal
`//sdk:host_tools.internal` is only intended to be used by ffx tools as
part of the build. `//sdk:host_tools_for_single_toolchain`, which is
already a dependency, is a superset of that target. Switch the "--sdk-manifest"
argument to that.

Original-Bug: 317385989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195097
Original-Revision: 2ec4367008b8f716b6dd5d8567f6504488d091c0
GitOrigin-RevId: e6254b1c5c41ba6a4957c279eb73dbb820cb6963
Change-Id: I0e3d266aa66568278a1ea560886a182b84153d44