[roll] Roll fuchsia [docs] make tracing duration examples more concrete
I was confused by the `var`-iables in the example,
because the variable name didn't match the screenshot.

To eliminate that confusion for myself (and hopefully
others), this patch replaces the `var`-iables with
literal text that matches the example.

Note that while examining the code for the `var`-iables,
I realized that some of the C and C++ examples had an
unnecessary C-string delimited. So I removed those along
the way.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196287
Original-Revision: cbf97081a617c834791b75e3a2af3a2e39da2fd2
GitOrigin-RevId: c54332bb5d841bf358dd3a515aeee022293b16c6
Change-Id: I05f760d9748321a5f9ea921ce730c7c8545e9119