[roll] Roll fuchsia [ffx][fho] Removing reference to TargetInfoQuery
This removes the reference to TargetInfoQuery in the DeviceLookup trait.

All the usages of this trait used Option<String> to create a
TargetInfoQuery, so the trait was updated to use Option<String> and
the underlying implementation now accepts `impl Into<TargetInfoQuery>`
to localize the usage of this data type.

This is part of the refactoring of fho to remove all target and FIDL
dependencies so non-device related tools will build faster.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198425
Original-Revision: e3465c6b71a7da95809c8c4f108e9e73752e575c
GitOrigin-RevId: b75357fcfafba160f0779c1b2287744b2ecf4293
Change-Id: I07b46d270921bfb5ba74991fcea468bd9742b5c1