[roll] Roll fuchsia [devicetree][pci] Parse arm gicv3 interrupt map from devicetree
This updates the PCI visitor to parse the interrupt-map property which
specifies the interrupts for this device. Only ARM Gicv3 interrupts are
supported in this change. We currently use this interrupt controller in
qemu and crosvm with PCI devices.

Parsing this property requires looking up information from both the pci
node as well as the interrupt controller that the interrupt-map elements
refer to. This adds the parsing logic in the PCI visit which means that
it must find the interrupt controller within the devicetree and parse
some of the fields. This is partially redundant with the interrupt
controller visitor which also parses fields from the interrupt
controller field. This could also be handled by having the PCI and/or
interrupt-controller partially process their respective nodes in the
initial visit and then finalizing parsing in the FinalizeNode second
pass through the tree. Either way knowledge of these fields has to span
the PCI and interrupt controller visitor.

Original-Bug: 392960178

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196706
Original-Revision: d8d1e9f52f79265a6cc2fefba779a9614fb629f1
GitOrigin-RevId: dfc0db99950cb76df85dc3c0fa939ca327a45d42
Change-Id: I6ff08c1f1d2e84047c06b6ffc4e08f5f3e7f6404