[roll] Roll fuchsia [owners] Remove mkearney@google.com from docs/go/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198719
Original-Revision: 2e36fd71de62fddf28593d6cb94074761266e070
GitOrigin-RevId: b28cc2a2883af422507b97e1df6650a12276d003
Change-Id: Ib58b4652e59e434d207391b537e7168c263280ea