[roll] Roll fuchsia [owners] Remove mkearney@google.com from tools/doc_checker/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198721
Original-Revision: 14d73c585d953c77aad4987ae1e7b26388f66f1c
GitOrigin-RevId: 886afc9fe75a3e9afbc60de7191d0276880a6285
Change-Id: I3e8437a72890f6a69bc4f46083584f83551f8943