[roll] Roll fuchsia [owners] Remove mkearney@google.com from docs/contribute/governance/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198334
Original-Revision: 0116498096244e0d3fb4377fb3f08d196469c468
GitOrigin-RevId: 0e70a61672f0e9e4009b322143c5b33b50dbabe5
Change-Id: Id4a517a542eb4e62f9775ee185962186dbfba123