[roll] Roll fuchsia [owners] Remove mkearney@google.com from tools/docsgen/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198335
Original-Revision: c684f9bad4b3f9c138597dd5d8cde70ee556f199
GitOrigin-RevId: c31353b0ee737e8a48124c172341dd9d59586e20
Change-Id: Ia0c5b134a97c054250373eb976f1cf3266f4ef06