[roll] Roll fuchsia [owners] Remove mkearney@google.com from tools/mdlint/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198723
Original-Revision: ab851851738b32a4d6f3721cffa61dc57d6f9088
GitOrigin-RevId: 789587db4f2de8ab9541904ec428343233854859
Change-Id: Ifa8d87614b71719be50091a68cfe45102beb4608