[roll] Roll fuchsia [owners] Remove mkearney@google.com from tools/fidl/fidldoc/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198722
Original-Revision: a4ddbb28a0cea1941bbf642af5da8b81a06c59d7
GitOrigin-RevId: 72b01072f3d9a0860a1ccf8842e24699ea56f024
Change-Id: I28f41721162478025368245847682f698a97e5dc