[roll] Roll fuchsia [rbe] Swallow remote command output if only being used for cache hits.
With the remote_cache_only RBE mode we only use the remote
command result if it was for a cache hit, so there's no need to print
the output when the remote command fails but the local command succeeds.

This provides the expected DX when using deny_warnings=false along with
rbe_mode=remote_cache_only when there are warnings present in the current
checkout. The remote build will fail but that's expected and we shouldn't
print those errors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199250
Original-Revision: 11d34a6732cc0c78b6d6deee1bb3494c00f59f45
GitOrigin-RevId: 0ce1b6cc1cc62329e40cedccdeb5a043baade211
Change-Id: I4cc0f369b577eab318341a9126ae7fc62252e9c7