[roll] Roll fuchsia [starnix][range_map] Remove iter_mut() method
This method breaks the invariants of the range map because if the client
modifies adjacent values in the range map, those ranges ranges should be
merged but won't be. Also, this method is expensive to implement when
backing the range map with a copy-on-write data structure because
mutable iteration implies a write to the entire data structure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198942
Original-Revision: 0954b56aee6b2d161d3b83fa7b20671ceb734444
GitOrigin-RevId: f98df8189384f8835417683e565e7fb3b76138c5
Change-Id: Ifa49d37e1417e121b945f865e78bca0afc47330a