[roll] Roll fuchsia [superproject] Roll scudo [scudo] Refactor the secondary test (#125595)
Remove all redundant code and create a couple of structs to handle
automatic init and destruction. This replaces the test fixtures in
prepartion for passing in multiple configs for some of these tests. This
is necessary because not all of the gtest features are supported here,
and there is no easy way to create a test fixture with a template.

GitOrigin-RevId: 97f87a63ee403c701ccac6efe5232a7e64633265
Original-Revision: 8e022782d84ccd02f2032012f33ee47002406adb
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199524
Original-Revision: e898de8d50d1fc68b037cb1c2f40ad2ee3ad105a
Change-Id: I0a1ace78eb66442e8e60885757605a50859fc138