[roll] Roll fuchsia [devfs][cf] Add ServiceMemberWatcher
Adding a way to watch a service directory for
instances of that service, and then return a
client end to a member protocol interface
that is part of that service.

This is specialized for the case where there is
only one interface in a service that a client would
be interested in connecting to.

ServiceMemberWatcher uses ServiceWatcher for the parsing
of the directory->Watch command.
ServiceWatcher is modified to increase ergonomics.
ServiceWatcher could now be used as a fallback for
services where multiple interfaces are provided.

Original-bug: 361633233

Testing: added test to incoming_integration_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125747
Original-Revision: 8acb33ac0b5c84637c5d84d26f43c99df62f320a
GitOrigin-RevId: 0d4672d1f88351ff2e8270b2a6073fb27aadf7e9
Change-Id: I4271edca990a7db79038781d4a973362fbe00b4e