[roll] Roll fuchsia [zxtest] For streams, conditionally allocate the assertion
The approach here is much more in line with what vanilla zxtest does for
evaluating a predicate.

(1) Delegate to a template function.
(2) On failure, generate information and return.

For streams, we follow a similar approach here, where we delegate to a
template function, which will evaluate the predicate. If the predicate
is true (e.g. success) we return a nullptr, otherwise we return a
std::unique_ptr<StreamableBase> that will contain the right information
and allow for streaming data into it. The lifetime of this unique_ptr is
tied to the if/else branch introduced by the macro.

Before we were unconditionally instantiating ostringstream, and 2
fbl::String (actual_symbol and expected_symbol) on every call.

This approach seems much simpler.

Original-Fixed: 394101741
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198573
Original-Revision: 56db0f57d9eba4b83d5eb358442969217bac9b7a
GitOrigin-RevId: 18cbb10b7037144cd92949f8c97a76e9c5851d32
Change-Id: I940ec67cbce2b8b696b05944f533a0068a555183