[roll] Roll fuchsia [fidl][cpp] Stop reformatting golden files
We currently reformat golden files so that changes to the clang-format
implementation and rules don't result in golden failures. This is
taking 7-10 minutes of CPU time, though it's parallelizable so in
practice it only adds a minute and a half to local builds.

After this change lands any clang-format changes will need to be done
atomically with regenerating goldens (with `fx check-goldens`). This has
the advantage of allowing us to review the impact of updating
clang-format as part of landing it.

Original-Bug: 393476961
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199485
Original-Revision: 9342ed34915fddd41eff22c3284dedfef37b875b
GitOrigin-RevId: 00cd699922fe1bf6222db7db7748c47b1f283447
Change-Id: I1027c298a26bb175a3deaa77a032b4cd11f4e1f6