[roll] Roll fuchsia [vfs] Reduce the size of spawned futures
The size of the wrapping future created in ExecutionScope::spawn was
twice the size of the task passed into it. This was because the wrapping
future was capturing the task in an upvar and then moving it to a local
when polling it requiring space for 2 copies of the task.
https://github.com/rust-lang/rust/issues/108906

Writing our own future prevents duplicating the space requirements. The
task spawned in VolumesDirectory::add_directory_entry is now 15312
bytes, down from 30640 bytes.

Original-Bug: b/393365596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198686
Original-Revision: fe8f3638c56955f5fbbc4f03428e7caf869e8cbb
GitOrigin-RevId: c89c6fd011718cde4f0f61941f314b03ead66fe4
Change-Id: Ie505025a80fd331b2872101adf74687d677a86ae