[roll] Roll fuchsia [tee] Partially implement TEE_CheckMemoryAccessRights
This provides a partial implementation of the
TEE_CheckMemoryAccessRights function. This can be used to check if an
access to a particular address range will fault as well as to query for
exclusive access. This implements the former check. To check for
exclusive access, we'll need to perform more tracking of mappings
created within the runtime - in particular we'll want to track memory
reference mappings.

Original-Bug: 384584494
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1175104
Original-Revision: fd044a21ba568279fb1ae145af5e7f9cbab9633a
GitOrigin-RevId: 5e8378b28c0efed2c908f528fd945e5116351aaf
Change-Id: I133f91c606fad33cbe5db23b76f19c36deca9c3d