[roll] Roll fuchsia [fxfs] Add support to create unnamed temporary files
An unnamed temporary file is first created in the graveyard and marked
to be purged. It needs to be marked to be purged so that we can reclaim
back any storage used once the open counts drops to 0.

This change also changed `open_count` to change one of the top bits to
indicate if the file is an unnamed temporary file. We need to be able to
identify the type of file it is on `link_into` to determine if the
object needs to be removed from the graveyard (if it is an unnamed
temporary file). As we now have two bits in `open_count` that are used
to represent information about the file object, refactor `open_count`
such that it will be wrapped in a struct and we use helper functions to
get open counts and other information.

Original-Bug: 293950309
Test: fx test fxfs-platform-test storage_conformance_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177512
Original-Revision: 523ea446c771925c960c255f137ba88cff6e437b
GitOrigin-RevId: 2c98cfdbbbebf061ccd66e65626edbc4b8834727
Change-Id: Ibe0ca2016f9427de48063bf42c678e259037864e