[roll] Roll fuchsia [sdk][component] Temporarily disable lsan for fdf_rust_component_tests
LSan was detecting a leak in this test with fxr/1197316 applied. That
particular change doesn't directly affect this component. It makes
improvements to logging so it's likely this leak was always around, but
we only just surfaced the diagnostic. We should fix the leak, but for
now we can temporarily disable the check for this test.

Original-Bug: 394632612
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200444
Original-Revision: 428e913aa75c486981228434a65e4e67dc165d0a
GitOrigin-RevId: 18d5b57151065d0e5b356718d8c81c649b8a0b19
Change-Id: I513d095aa62bf68227143413cddd02e0922e8fa7