[roll] Roll fuchsia [netstack3] Straighten ICMP echo NATing
Most of the support was already in place, but we didn't have integration
tests running, and the forced directionality on request/reply with
panics on the rewrite was subject to real runtime panics if replies are
observed before requests in conntrack.

This CL enables all the tests that we can to put more load on ICMP Echo
NATing, and changes the conntrack strategy to use both the src and dst
port ids as an ICMP id. This gets ICMP echo NATing to a more reliable
place, without having to teach conntrack about directionality just yet.

Reply-tuple rewriting is not mod-walled into conntrack so we can enforce
the mapping per protocol.

Original-Fixed: 380506614

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199346
Original-Revision: 6628e7ebe0fd02d8c87c4cb40be58af23cd16c59
GitOrigin-RevId: 728554f7b872f44dd17edbf079ecd24a5c9b5473
Change-Id: Ia47a4f1613efd287c64abad75e12d7be02010b9b