[roll] Roll fuchsia [test_pilot] error restructuring
The principal goal of this CL is to replace some strings that were used
to identify error sources with the source errors themselves. This posed
a challenge, because the widely-used error enum would no longer be Eq
or PartialEq when some source errors were referenced.

To make this tractable, the single error enum UsageError was factored
in a logical way into BuildError and UsageError. A BuildError was
defined to wrap UsageErrors. This approach organizes errors in a
reasonable way while still allowing tests to use Eq/PartialEq when
asserting error results.

Test: tests updated

Original-Bug: 327640651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200364
Original-Revision: f504b178adf5973459db0a3ebe5d61e343163bdd
GitOrigin-RevId: 4e69075870ae16f676032e5626c028cf90abedc3
Change-Id: I9999e4cb28c82537d62a2b277064468ea959006f