[roll] Roll fuchsia [kernel][vm] Re-design clone flags
Changes the flag to CreateCloneLocked to be the kind of snapshot
semantics that must be provided, instead of directly requesting a clone
type. This allows for better logic inside CreateCloneLocked where
instead of 'upgrading' the clone type, it can just pick a clone type
that matches the request.

The logic of picking the clone type is refactored to be in terms of
finding the intersection of what kinds of clones are possible given
the properties of the current VMO, and what semantics need to be
provided.

A consequence of this change is that it is now difficult to enforce a
previous check that prevented SNAPSHOT_MODIFIED on a root with pinned
pages, even though a SNAPSHOT_AT_LEAST_ON_WRITE would be permitted. The
previous check was, however, a byproduct of the previous code structure
and does not need to be enforced.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196435
Original-Revision: 1dab957bb1775a7e1904bb9012c59c71d19b49d7
GitOrigin-RevId: 917f9bf402d4d13bfe18de57edce34417af69f91
Change-Id: Iaac563e6af1a2cb930d254850eed1b7c8eee2fa2