[roll] Roll fuchsia [ffx] Fixes no-daemon regression for ffx-strict.
This fixes a regression in ffx-strict that causes the daemon to start
even when using ffx-strict.

In order to handle the dependency cycles, this moves `NetworkConnector`
into its own crate. It's not the best solution, but it's at least a
short path to fixing the overall regression.

Original-Bug: 392190786
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194249
Original-Revision: c2fceab8b65a9d6218f170b5b2ebc08a5849fa23
GitOrigin-RevId: 87db529c6c0075c921fd121f4942d274546cdaf1
Change-Id: I208abc27923e5a1f87bbfb36aba1b85b090b97a7