[roll] Roll fuchsia [ffxutil] refactor ffx invocation
There are a variety of functions in ffxutil/ffx.go for running ffx
(Run(), RunWithTimeout(), RunWithTargetAndTimeout(), etc.). The
ffx-strict work will add another set of parameters (required config
options) which is hard to implement with the current model, since the
options are not composable. This CL introduces an `ffxInvoker` struct
which holds the options, and methods that can run the command (run()) or
build the exec.Cmd (cmd()).

The existing functions are retained for compatibility and convenience,
for now, but some of the more complex and seldom used ones will
eventually be removed, with callers switching to the new interface.

Original-Bug: 391964053
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192682
Original-Revision: 55f5b45b4cc459b8713d119ba8183ab10008a24f
GitOrigin-RevId: 993f668e42c20908738b22f03eac09cf0b830bd2
Change-Id: I4fcddb195c1f5c8ac9034de9eb975838ed47d56b