[roll] Roll fuchsia [network] Fix dns client shard typo
Fixes a typo introduced in I4318b6ace3e456929cb756c2a0f649f6bfc452ae in
the client shard of the DNS resolver. This was only used in a couple
tests, so the blast radius is minimal, but it means we were likely
missing logs for the DNS resolver for tests that were using it since
that CL landed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201226
Original-Revision: 88868401f807d1d22af7a67d212b94297bddb803
GitOrigin-RevId: 927e34287ea0eddcd686c128ef1c99a9e9ae0b00
Change-Id: I6c7512a7641e0226f073f24226e5663a5dc54d0f