[roll] Roll fuchsia [fake-display] Disallow releasing in-use resources.
The protocol between the Display Coordinator and engine drivers
specifies that the Coordinator will not release resources (images,
capture images) that are currently in-use (by the currently applied
configuration, or by an in-progress capture).

This CL adds a missing check enforcing this constraint in the
fake-display driver. Thanks to the check, incorrect releases will result
in an immediate crash, as opposed to having ServiceAnyCaptureRequest()
crash later. This change will hopefully make it easier to debug any
incorrect release.

This CL also replaces some comments with error logging. This also
makes it easier to debug errors, and aligns fake-display with other
drivers.

Last, this CL fixes an incorrect release in an unit test.

Original-Bug: 394413629
Test: fx test //src/graphics/display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200746
Original-Revision: 6919fd7cf4fac46b0ed5ed6eb35bd6969ffa0e36
GitOrigin-RevId: 336967833f34add0a164152be27d7a0aee721be4
Change-Id: I8cb112cad163d04c1a14899a1301887c9bce85b8