[roll] Roll fuchsia [ffx][writer] Refactor Format enum to command
This moves the Format enum to ffx/command:lib so the dependency
direction can be reversed and have writer depend on command.

This is not ideal, and perhaps there is a better crate for the Format
enum. It is however, good to switch the direction of the dependency
so the TryFromEnv implementations for the writers can be moved
to the writer crate and out of fho.

Original-Bug: 392174596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201487
Original-Revision: cc12e8d1334a9f7c72c16f88781cabe5b27b9906
GitOrigin-RevId: 92f90cb5510c9bade1d057091cd0907c9dfea90a
Change-Id: I773511b2b4c44347c56c5c2a8bd111b80673e7d1