[roll] Roll fuchsia [bazel_sdk] Prioritize HEAD variant over arch binaries
Use prebuilts for the "HEAD" API level from "variants" if they are
present. This makes it possible to use the appropriate prebuilts when
they are built, which could include via developer override in-tree or
when building the full SDK once support for "HEAD" when it is included
in the IDK.

Previously, the "binaries" metadata (for prebuilts built at "PLATFORM")
was processed second, meaning it would replace any metadata that existed
for "HEAD" in "variants". If both were present, this would cause the SDK
to use the "PLATFORM" prebuilts when targeting "HEAD" instead of the
prebuilts built at "HEAD".

There is no change in behavior since the IDK does not currently
generate prebuilts for "HEAD" and related metadata in "variants".

Original-Bug: 385406226
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200490
Original-Revision: e7bc75b39e8e05e7073be1f5088093247d4e081c
GitOrigin-RevId: b655a1a0aec6062a7268b8d76a665f9a522be5c8
Change-Id: Ic52d249a4a7e1d0cc3e32b34451d7ed614b0ae40