[roll] Roll fuchsia [bindgen] Control --explicit-padding with a knob
With the linked bug, we are in a case where we are using bindgen to
generate bindings that are indirectly (via bssl-sys) used by third-party
library (openssl). The latter code expected the construction of structs
not to have padding fields to initialize, and so we have a need to
control the generation of explicit padding fields. We default to 'on'
though, as this is generally useful for self-documenting ABI.

Original-Bug: 394116035
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201306
Original-Revision: 2322b00531394fe716f973e8b412075ea3ebc281
GitOrigin-RevId: 07fc55557235aa6ba6df687951ed9beff7fa3c4e
Change-Id: Ieceb75cf5ef1815de035949b8e0f5b72b1b35f79