[roll] Roll fuchsia [overnet][usb] Retry sync packet on garbage
If garbage data is received on the usb channel when attempting to
connect, attempt to re-send the sync packet to restart the connection
dialog. For some reason the first read packet is getting eaten in the
usb_rs layer, so without this it doesn't synchronize.

The missing first packet should be fixed as well, but re-asserting the
state makes sense in this situation anyways. This fixes reconnection on
daemon restart, but there are still issues when the device goes away and
comes back in the higher overnet layers.

Original-Bug: 393442313
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200492
Original-Revision: 1033c303f428b4e6ee77faec34dffb73dc962dea
GitOrigin-RevId: a0ed26f251c192afc269b119e00ce63471a0efb5
Change-Id: I07948b0f4c9aad97399a2604182d4ebecbd5963c