[roll] Roll fuchsia [netstack3] Remove unnecessary unlocked access markers
There are several instances of dedicated lock levels being declared only
to be used to implement UnlockedAccess for a particular resource that is
nested somewhere inside StackState. Instead, provide a handful of
UnlockedAccess impls to the top-level fields of StackState and replace
usages of the existing granular ones with those.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202265
Original-Revision: bfa526298562c32e93630d804499a27059d35d4a
GitOrigin-RevId: 4aa7f7b23741787b240744d9e67c3652b4d77dfb
Change-Id: Ifcc603dc89274414a4f0ae44d94eb36930c57686