[roll] Roll fuchsia [bus-pci] Split PCI bus driver for user and kernel implementations
This splits the two definitions of the PCI bus driver (kernel and user
space PCI) into separately named drivers that are included in the
configurations which need them. The driver using user space PCI support
is called 'bus-pci' and is included on all x86_64 and RISC-V
configurations as well as the vim3 and pkvm boards. The driver using
kernel PCI support is called 'bus-kpci' and is used on qemu-arm64.

Previously, the bus was always called 'bus-pci' and its definition was
determined by the platform_enable_user_pci GN argument.

This change modifies the 'arm64' and 'qemu-arm64' boards, both of which
are used in qemu to emulate an arm64 device.

Original-Bug: 42182308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199464
Original-Revision: 4c448a5077aa530da1567eb88dadd00a75f1a4e6
GitOrigin-RevId: 4a9adbc8abb82f6247f07c366ff199847920a334
Change-Id: Iba1b58965e7946fbf5e1548fa323e477af24ed0b