[roll] Roll fuchsia [kernel][thread] Construct free_dpc just before it's used
Move the construction of the free_dpc in Thread::Current::Exit to the
point at which its first used.  By moving this down, we eliminate a
blocker for making Dpc's fields const.

No behavior change.  This CL is the first in a series of small
improvements to the DPC service in preparation for resolving
b/395669867.

Original-Bug: 395669867
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203291
Original-Revision: 93b5b7c6d3e931ecee73f02166f8eb2f58c3b61d
GitOrigin-RevId: 9b069fc9999a8c9e027bc63ca23a54d6d9291f5d
Change-Id: I9d6a66c73945561c9a3c6b36c3bacd98dfd2865f