[roll] Roll fuchsia [fxblob] Reduce memory overhead of merkle trees
Open blobs keep their merkle trees in memory. These merkle trees account
for roughly half of fxfs' memory usage. The inner nodes of the trees can
be dropped since only the leaves are necessary to verify the data
blocks. The Vec holding the leaves is constructed with Vec::push which
has a growth factor of 2 which means that it could be twice the size
that it needs to be. Dropping the inner nodes and shrinking the Vec
holding the leaves reduces the overall merkle tree memory usage by 25%.

Original-Bug: b/393365596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200346
Original-Revision: 1db176812b548af8248a97c27dc30cd422b19176
GitOrigin-RevId: 154a11e2f6972c5fe589c94cc7c42f8cf90bba64
Change-Id: I0bb18d606317847efc6e0ff94f3d262f26bda4e7