[roll] Roll fuchsia [superproject] Roll llvm-project/libc [NVPTX] Make ctor/dtor lowering always enabled in NVPTX (#126544)
Summary:
Currently we conditionally enable NVPTX lowering depending on the
language (C/C++/OpenMP). Unfortunately this causes problems because this
option is only present if the backend was enabled, which causes this to
error if you try to make LLVM-IR.

This patch instead makes it the only accepted lowering. The reason we
had it as opt-in before is because it is not handled by CUDA. So, this
pach also introduces diagnostics to prevent *all* creation of
device-side global constructors and destructors. We already did this for
variables, now we do it for attributes as well.

This inverts the responsibility of blocking this from the backend to the
langauage like it should be given that support for this is language
dependent.

GitOrigin-RevId: 3f76b4f1b4d75b5a7633d2b995ff049cc299f2bd
Original-Revision: 4dc3dc2ce2cb7ba01842d10fe479d9ffbfd275c7
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203786
Original-Revision: 215b1eda7c1c85702a4509c8910a70c8ecc93881
Change-Id: Id3cba0bcff9f72d81fb57e5232529eaded719ab8