[roll] Roll fuchsia [display][coordinator] Switch empty config tests to color layer tests.
This CL removes the usage of empty (zero-layer) display configurations
from the Coordinator integration tests. This is a step towards removing
support for empty display configurations, for the reasons described in
the attached bug.

This CL introduces the use of color layers in the Coordinator
integration tests. The model is a bit confusing, due to an unfinished
migration from primary/color layers to unified layers in the Coordinator
<-> display client FIDL API. The model will be cleaned up by future
work, which will be easier to do once we remove empty configuration
support from the Coordinator.

Original-Bug: 394413629
Test: fx test //src/graphics/display/drivers/coordinator
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203366
Original-Revision: 8dbdb4a780aaa38c4354ec20f6e8ef743fc0570c
GitOrigin-RevId: d36ad47d51426a14eab5c02ea0899e7d278cd996
Change-Id: Idc1c6bb114e61a18a14b765a4169829513d05e36