[roll] Roll fuchsia [botanist] Start checking for SSH key after sending serial commands.
Sometimes we see the serial commands to provision and cat the SSH key
get stalled for some time. However since we start reading for the
response before sending the commands, this may cause us to miss the
response because we time out before we can send it. Instead we should
start the timeout after sending the commands. The side effect is that in
the case that we hang completely, this will cause the timeout to write
to the serial socket to be 2 min instead of 10 seconds, multiplied by 5
tries.

Original-Bug: 395706575
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203870
Original-Revision: 55d7a41e9359886381b606657c96c48e65975ac2
GitOrigin-RevId: 8438d280f17a5f6b405b9be0318235b2be5be0a7
Change-Id: Ice31625b48cd1b8120438992b63dca24463c842e