[roll] Roll fuchsia [starnix] Implement PTRACE_O_EXITKILL.
This CL implements the PTRACE_O_EXITKILL option in Starnix. When this
option is set, ptracees receive a SIGKILL when the ptracer terminates.

Note that this implementation is wrong, because PTRACE_O_EXITKILL
triggers when the ptracer exits, but starnix uses the thread group
leader as the ptracer. Some tests had to be skipped because of that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201230
Original-Revision: 0d50f425a88096b17f671662a2afa6147d50d258
GitOrigin-RevId: c544d55eef695230ac29746d445bb72b266b783c
Change-Id: Iec7286d52ff33d93e3d5ac18cfd28f78886f49cc