[roll] Roll fuchsia [fxfs] Link count for unlinked files with no links should be 0
Previously, this would report 1, but change to have it report 0.

In Fxfs, reference count refers to all internal references to an object.
So, even for objects living in the graveyard, the reference count is 0.
However, we've been using this reference count as the link count of a
file. This is not quite consistent with what link count refers to. For
example, with files that have been unlinked at there are no longer any
links to it, the link count should be 0 (Fxfs will previously report
link count as 1 as the object lives in the graveyard).

Add test in starnix syscall test suite to ensure consistency with Linux.

Original-Bug: 391458506
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202865
Original-Revision: bc3d78963ce3cd06b0189e6d45f8fb29bc81d4b0
GitOrigin-RevId: 763588039a287f5a7f8709c0a6464845eb101fa0
Change-Id: I8a8aca4234d4e0b63130fb3a6296b7dd0132dd23