[roll] Roll fuchsia [assembly] Unify sampler config in the diagnostics subsystem
Working towards using config capabilities instead of config_data for
Sampler configuration.

As part of that, we can move the work that sampler does to expand some
templates given the component_id_index values to assembly. To accomplish
this, we'll need access to the ComponentIdIndex in the Storage config
from the Diagnostics subsystem configuration.

This change moves this configuration of Sampler from Storage to the
diagnostics subsystem preparing for that change.

Original-Bug: 395159893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203964
Original-Revision: 1d326d8520fbba4e56377976b7d2f22ac359df35
GitOrigin-RevId: 583045fa7c92c5184cfe1a0a593636fa6653d6b3
Change-Id: I65dac0b344ce54cd185faf3a70c7a233bc862ce0