[roll] Roll fuchsia [display][coordinator] Improve integration test comments.
Thanks to a recent change in the Controller testing API, one of the
tests is no longer racy. This CL adds documentation to the testing API,
and removes the raciness warning.

Another test was very likely to report false positives. This CL greatly
reduces the odds of a false positive by adding a hard-coded timeout to
work around the lack of a testing API for event states.

Last, this CL renames a few variables and rewords a few comments, to
make the test logic easier to follow.

Original-Bug: 388885807
Test: fx test //src/graphics/display/drivers/coordinator
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205566
Original-Revision: 07adf8ab930beb52b51113a486b5753d71ee73ad
GitOrigin-RevId: 42ce3fb6958c62758360a976093a3a08f7003546
Change-Id: Ie2825493572a43d03b0c2f5af534ec7d7f1a41b7