[roll] Roll fuchsia [kernel][vm] Virtualize VmObject lock
Changes VmObject to not assume that the lock it should use is in the
shared state, but rather requires the implementation provide the lock.
In the case of VmObjectPaged it further defers the lock to the linked
VmCowPages. The lock returned is ultimately the same lock from the
shared hierarchy state, but this enables it to be changeable.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204906
Original-Revision: 7e5355c99b7576346b88e2d808b5c0220b6ae28d
GitOrigin-RevId: 3426609c81724e2616999bafa4b5d316474a9ada
Change-Id: I27247410b66cd2fa2036f9dd5b5159ffc804172b