[roll] Roll fuchsia [boringssl] Access test data as files
Following upstream's lead in the past, we previously had a scheme in
place to embed the contents of a large number of test data files and
access their contents at runtime through a generated function that knows
how to access these blobs. This file is unwieldy at 27MiB and upstream
has done away with it (and the way to regenerate it) at HEAD. We don't
need to do that to access test data though: we can define them instead
as `resource()`s that can be accessed through normal file I/O and make
the definitions of those table-driven and based of the list of sources
generated in BUILD.generated_tests.gni.

Original-Bug: 339028859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202464
Original-Revision: bcd3afd54f30213551d02df20d147234cb95351b
GitOrigin-RevId: cc5b44d0e39dfa19d794614b594bef720ae2b2d4
Change-Id: I5900bd88f1856cb61b14d072c592d21a53ed6822