[roll] Roll fuchsia [ffx][fho] Removing ffx_writer dependency in fho
This removes the dependency on ffx_writer in fho, allowing the
refactoring of TryFromEnv implementations for the Writers to move
to ffx_writer.

This also moves ToolIO trait to /src/developer/lib/writer since it is
used by on-device commands as well as ffx.

Finally, this adds the assert to fho and ffx_writer that they are
host toolchain only.

Original-Bug: 394964930
Original-Bug: 392174596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204351
Original-Revision: 124993707f9b585790fc680c682fe7ae2165433a
GitOrigin-RevId: 85640477e6c239276641df14dc643c7f7aa331ca
Change-Id: I560b9fcbf60a532b85e3e0b2d40b2a5efddb043e