[roll] Roll fuchsia [sampler-config] Remove Optional selectors
This is another instance in which implementation details of sampler had
made it to the Configuration definitions. Instead of keeping vectors of
optional selectors, we can just remove selectors we don't need to track
anymore without impacting the configuration data types.

I realized this in the follow-up, but took the changes to this CL to
keep the follow-up change smaller.

Original-Bug: 395172409
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205846
Original-Revision: a47c7cb7f0b384b8db456d62e6677865aed920d6
GitOrigin-RevId: 57bf27ffcfe91664946e980dbcb5c308a8fc628b
Change-Id: I3fa4cc814daf09e4691a6f3c2149ef46e2c0a51e