[roll] Roll fuchsia [power_metrics] Be less strict about wakeup events.
If there is a non-duration event with the same name as an expect
sequence event, ignore it instead of failing. This allows us to
use the duration events that are guaranteed to be coupled with
flow events.

Original-Bug: b/390223971
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204871
Original-Revision: 3c31b96f8cce6c28a764e0d46bd868b501d43ead
GitOrigin-RevId: 6c769228a051748573ef93380238fb6e1a1436f4
Change-Id: I53af2d4b6c736df2706c558a343227f12d70d75a