[roll] Roll fuchsia [build] Clean up errors with bazel directories
Previously is a directory was specified, but missing we'd get a log
message that did not look like an error, and gave the wrong message.

"""
INFO:

Non-directories are not allowed in --directory-outputs Bazel path, got:
"""

Now we prefix with 'Error:', and add a specific error for missing
directories.

"""
Error: Directory provided to --directory-outputs is missing, got:
"""

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207045
Original-Revision: 66890add98bc1442f33a4291f4264cf8422a086d
GitOrigin-RevId: aba1e30b97e2d9e4dfd2313c95c43b5c6323f65f
Change-Id: I255744439bfa7cdd755abfc07dbfec1c6e4991f9