[roll] Roll fuchsia [owners] Remove maheshsr@google.com from tools/testing/resultdb/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/maheshsr@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1182592
Original-Revision: 7d86bbdebae86f5d9408791673d50013072f32a2
GitOrigin-RevId: 999d9cd977cb7bbef7ef8a21227afbf0d905d4d1
Change-Id: I15e186a41ad1d8cee13cd29e0de08efa7b1a9c42