[roll] Roll fuchsia [kernel][vm] Re-order reference creation
Change the order of code in VMO reference creation so that fallible
checks are performed before constructing the VmObjectPaged. This is
necessary as destructing a partially constructed VmObjectPaged is
unsafe.

Comments and checks in ~VmObjectPaged are updated to reflect this, as
the existing check no longer makes sense as cow_pages_ is const and
assigned at construction.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206864
Original-Revision: 22a2ab8b794b7d53d7c08829d546c43996026c0d
GitOrigin-RevId: 4c8f662dc279845d354df5a0f8f8e9ce398e6cbf
Change-Id: I0aa33f9cdbd6a343bb4de388639cb1a3c18ec2e3