[roll] Roll fuchsia [netstack3] Avoid dropping NDP watcher responder
In the NDP watcher implementation, flush_batch promises to check if a)
there's a batch of watch entries queued to send, and b) if there's
currently a pending watch request. However, it does not correctly check
a) before it decides to take() the pending watch request, which can
cause us to drop the responder (which clients of the ext crate end up
observing as a peer-closed).

Test: netstack3-bin-instrumented_test
Original-Bug: 341759423
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207485
Original-Revision: af7c635647b9a830105eae90f3415d63b67fc9f6
GitOrigin-RevId: 0ae10676a9b78bd19d0111aecda02f6b566e454d
Change-Id: Ifa108c7068c5b9b4ab36c36735223ef882662c75