[roll] Roll fuchsia [sampler][assembly] Prepare for soft migration of fire config schema
At the moment the sampler config looks like:

```
sampler: {
    metrics_configs: [paths],
    fire_configs: [paths],
}
```

In the follow-up we'll be using:

```
sampler: {
    project_configs: [paths],
    fire: {
        component_configs: [paths],
        project_templates: [paths]
    }
}
```

`metrics_configs` isn't used at the moment in any product.

`fire_configs` is used in products at the moment, but only passing a
`component_config`.

This change introduces the new schema for FIRE configs and handles the
component_configs part of it that is the one that we want to
soft-migrate.

The follow-up will delete the old schema entirely.

Original-Bug: 395159893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207544
Original-Revision: b7e840274d1c17bc4c25b898c902ce31562e7cfe
GitOrigin-RevId: 05e5e328a363e474dc471c1ac9f5a104bd17de33
Change-Id: Ie5f6cf8a44154f3280a6b0aacc91e1092cadfb17