[roll] Roll llvm-project/libc [libc] Produce standard-compliant header guard macros in hdrgen (#127356)
Macros starting with alphabetic characters such as "LLVM" are in
the application name space and cannot be defined or used by a
conforming implementation's headers.  This fixes the headers that
are entirely generated, and the __llvm-libc-common.h header to
use a conforming macro name for the header guard.  That is, it
starts with "_LLVM_LIBC_" instead of "LLVM_LIBC_", as identifiers
starting with an underscore followed by a capital letter are in
the name space reserved for the implementation.

The remaining headers either will be fixed implicitly by removal
of their custom template files, or will need to be fixed by hand.

GitOrigin-RevId: bbad61b3a698e7d8ad44160f9464847f197a543b
Original-Revision: c137c112faf4b47cc446b54a16d7c532819b874f
Change-Id: Ie9a3a0dbe183480eaf245348794a1fb7d408b80a