[roll] Roll fuchsia [kernel][vm] Simplify cow pages deferred deletion
Instead of using the generic VmDeferredDeleter for VmCowPages, relies
on the fact that VmCowPages have an explicit dead transition that is
invoked by the VmObjectPaged, and are not actually dealing with
recursion in their destructor.

Performing the dead transition loop explicitly in the ~VmObjectPaged
simplifies the VmCowPages::DeadTransition, as it does not need to
perform lock dropping, and saves some memory in the VmCowPages as it
does not need to have the potential to be in the deferred deletion
list.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206885
Original-Revision: ac1b66dffd527a801a052117aedb063889b08db5
GitOrigin-RevId: 0c14d80e142fb3fd03518144841de5a36f7b38d4
Change-Id: Ia1190b48272ff20eae2b2f7ee5f4994901744541