[roll] Roll fuchsia [starnix] setpriority permissions
This fixes a Starnix crash that can be observed by starting a process,
setting the started process' RLIMIT_NICE to zero, and then calling
setpriority on the started process - the clamp fails to find an integer
that is below-or-at the rlimit of zero while also being at-or-above the
minimum raw priority of one.

From the reading and interactive experimenting I've been doing recently
it seems to be the case that the rlimit is only appropriate to use in
making the allow/disallow decision; the rlimit isn't appropriate to use
in the arithmetic calculation of the new priority value being set.

Original-Bug: 322894197
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1197195
Original-Revision: 162a89127acd103ba69b42e42f63ed14618b635c
GitOrigin-RevId: e6eb8b3456b5290dcf1ebff30173da7fd2d0dbb4
Change-Id: Id4a32d2273020d909519f33846da53e0518c90de