[roll] Roll fuchsia [owners] Remove zhiyichen@google.com from src/devices/lib/broadcom/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/zhiyichen@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207865
Original-Revision: 702189a1d251aaed0c9c30025ae4b738db73db6d
GitOrigin-RevId: 024d57945458fcba20b567ac6210b73f0a4457b4
Change-Id: I682760828bdf5d8da338bebfdf6b51ee35579b64