[roll] Roll fuchsia [kernel][arm64] Separate boot CPU from secondary entrypoints (1/2)
At one point it made sense to have a unified CPU start-up path for the
boot and secondary CPUs: there was a fair amount of commonality, and we
could easily switch on MPIDR_EL1 for context-specific codepaths.
However, we are now in the process of both cutting down on that
commonality and the logic in the boot CPU path altogether, so it makes
sense now to separate them.

This is the first in a series of two changes that accomplishes this,
separated out for easier review. The first patchset creates secondary.S
with a plain `cd zircon/kernel/arch/arm64; cp start.S secondary.S`. The
second patchset does a _minimal_ removal of content in secondary.S to
ensure that things work. Proper clean-up will happen in the second
change of the series and reviewers of this one should look at
https://fuchsia-review.googlesource.com/c/fuchsia/+/1204364/1..2

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204364
Original-Revision: b55c590bc3e56f75fd494d9682ad3886e976a15d
GitOrigin-RevId: 15e8bf126087aeaa355f763bba0692a3723858f9
Change-Id: I02bc1db77ad133f4008f3f8aa21a5964f12b05db