[roll] Roll fuchsia [kernel][vm] Augmend mapping WAVL tree
Used an augmented WAVL tree to store mappings in a VmObject instead of
list. This allows for efficiently finding the subset of mappings that
might overlap with an invalidation range, instead of having to always
process the entire list.

When there are many non-overlapping mappings this dramatically reduces
the number of mappings visited, and greatly improves performance.

Original-Bug: 310255065
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144972
Original-Revision: ca99733f331d720dc91945e9fd9dce24090b791b
GitOrigin-RevId: 5bb59bc9679564d83ef802d2aa8d09ebf5e25459
Change-Id: Ibc64b8e8b21b82d944129875f4f4c826b51fa5fa