[roll] Roll fuchsia [msd-arm-mali][power] Migrate system test phase 2
Migrates the test to the new power non-hermetic testing framework. This
is done by componentizing the test and connecting it to the capabilities
it needs. Also the custom product and board is removed as that is not
needed anymore.

Original-Bug: 335526423

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188995
Original-Revision: 790179b0b25889a7e3ec575a1e8a3f9d10a06b67
GitOrigin-RevId: 6a46999d90461dec0d19ece1ee9392f8dae0a8b6
Change-Id: I972afa05578119cc28c62909c643c62f96788225