[roll] Roll fuchsia [fxfs] Reduce memory usage of the DirentCache
Don't initialize the HashMap backing the cache to the DirentCache's
limit. Each volume has a DirentCache but only the blob volume gets close
to the limit. Reserving space for 2000 entries for the unencrypted
volume which only has 2 files is a waste. Growing the HashMap as
necessary doesn't significantly affect performance.

Shrink the HashMap backing the cache when shrinking the cache size. The
number of items in the cache is reduced when under memory pressure but
the HashMap backing the cache still had capacity for the larger number
of items. The HashMap is now shrunk to further reduce memory usage when
under memory pressure.

The cache size when under normal memory pressure has been reduced to
2000. The blob volume stays just under this limit at boot.

Original-Bug: 393365596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203845
Original-Revision: 7c54e826e6c40aea3bc65b932cc9aeabae12e44c
GitOrigin-RevId: 76861d81a067469b6a0689e2a77be31e9a4c7bea
Change-Id: I7d490b05e453ee038e7fbf344a727f6df1bd7a1a