[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/mobly/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162615
Original-Revision: 97b4a1f8e4aeb9ae98b0e1323343f3d653d90198
GitOrigin-RevId: aa071de392b71343d5b4559ecfacf995dc77bea0
Change-Id: I746a6635aeb33a1cafa8709d6649c63fec956489