[roll] Roll fuchsia [tee] Implement TEE_IsAlgorithmSupported
We implement TEE_IsAlgorithmSupported() to indicate support for the
current (limited, and growing) set of supported algorithms, replacing
a previous panic. We will naturally update its implementation as we
support more algorithms.

This makes a number of WTPI tests that were previously passing start to
fail, as they had been indirectly relying on the call to panic, which is
something that it should never do. This also makes a handful of xtest
cases begin to pass by virtue of the test case gracefully skipping when
encountering an unsupported algorithm.

Original-Bug: 360942581
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199824
Original-Revision: 0a1285e0528b08e6ea10f9d67809a9eac12ac0d6
GitOrigin-RevId: 942e58a044d6bf8898ad5bcffa767978a163c7f7
Change-Id: I77643aa4b5b68b57adbf292786f34d10d60da2c3