[roll] Roll fuchsia [fshost] potentially set a type guid for test disk
The gpt matcher will match anything that doesn't have a type guid, or
anything where the type guid is "empty" (all zeros), which is what the
v2 ramdisk reports when it's not configured with a guid. It's enabled by
default so it ends up grabbing any devices that fall through the cracks.

This adds a default type guid that is non-zero, and the ability to
configure it. If the device is uninitialized or gpt is set, the guid is
cleared so it'll trigger the matcher.

At some point we will make the gpt matcher configured via args instead
of heuristically picking a device, which will remove the need for this.

Original-Bug: 391891190
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209967
Original-Revision: ad04bae9abc3a896118d26c35540570bcbfec98a
GitOrigin-RevId: 4f3e081be898b71e3f3c136ac5cf82c8cc8d4b7f
Change-Id: I6ebbe0390771e40b3f840fda32c93a667ff304a8