[roll] Roll fuchsia [owners] Remove zhiyichen@google.com from src/connectivity/wlan/drivers/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/zhiyichen@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207864
Original-Revision: 5eb9ef44a1d51af44835deeb4b7445e9482e5244
GitOrigin-RevId: 2962fe73a94b8fbeb2159d780f26d5e2d4ebfbc0
Change-Id: Ica1d6446f21f5da8fca3ef1aadb7c47230b6d964