[roll] Roll fuchsia [testing] Sanitize child component names in enumeration_test_adapter
It is natural to give the name of the case in the enumeration config the
same name as the canonical test case name, so we do that. Further. gtest
will generate a parameterized test case name with slashes (like
HmacTests/HmacTest.HMAC_SHA256Basic/0). Currently however - also out of
want of a canonical name - the enumeration adapter forwards the name of
the test case as the name of the child component in which it's run - and
slashes are invalid in component names.

Accordingly, we sanitize the constructed component names, replacing
slashes with, say, underscores.

This changes starts making an number of tests start to actually run
instead of short-circuiting and erroring out. In the case of the
oemcrypto-integration tests the runtime is now no longer reasonable in
its current form, so we disable most of its cases for now (pending a
closer look out-of-band from this change). In the case of the HMAC
tests, these generate UBSan failures and more investigation is needed.

Original-Bug: 397716819
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210568
Original-Revision: f59d0264def6b83533a2a7913cf95c50df8cbca6
GitOrigin-RevId: fcae3e82adc320334a5d1cde0b6b3a62080be660
Change-Id: I7d66b09fc4714a6a1b66aeccf9c2b70e2da3463f