[roll] Roll fuchsia [system-update-committer][cleanup] Remove hardcoded Config
Currently, this config is hardcoded to either require
all health checks to pass or to ignore all health checks.

Health checks are already ignored if in the recovery configuration
so it is redundant to add a config to ignore health checks for
recovery-fdr.

This CL removes the config while still following the same logic
paths. Many tests are removed because they tested configuration
values that can never occur in production.

This is part of the work to improve health verification.

Run-All-Tests: true

Original-Fixed: 397222040

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208704
Original-Revision: e8cedfafc17a1def2760d42bbd7fbbbf54a88290
GitOrigin-RevId: 502874ff9079d6cda86cf0828b43605c47bce721
Change-Id: If5534daf196358d6a192aaf650ff84e359f160af