[roll] Roll fuchsia [fuchsia-controller] Don't create types until they're declared
The dynamic FIDL bindings no longer build types when encountered outside
their declaration. This prevents the issue of building the same type
twice resulting in failing comparisons of seemingly equal values. And
this simplifies reading the bindings as there is no longer a special
case for encountering types that haven't been created.

This change overhauls the unwrap_type() method and simplifies the
resolution of each ForwardRef when encountered. First, unwrap_type() is
renamed to unwrap_innermost_type(). I felt the name unwrap_type() made
the function sound too simple. unwrap_innermost_type() now just returns
a type, and thus reduces the scope of cases callers must handle. The
primary trick this relies on is specifying the `module` keyword argument
of the ForwardRef constructor so that resolutions knows where to import
the ForwardRef from. Many tests of unwrap_innermost_type() were added
too ensure it only succeeds with there is exactly one "innermost type".

In spirit of the overhaul of unwrap_type(), this change also refactors
how Zircon types so that any Zircon type is now just unwrapped to an int
and treated as such for the purpose of decoding.

Original-Bug: 396778959
Test: Added tests for unwrap_type
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204426
Original-Revision: 0835dbb1f9193afcff5ae68dac84bd0c49837c6f
GitOrigin-RevId: a39f7d1e4c895ad80fd0b759eede0fba53952f12
Change-Id: If8b6e8afa5d59f8ffbab56a0876897d2046c94f2