[roll] Roll fuchsia [assembly] Add amber-files back to multi-product builds
As is turns out... we had fyi builders that run tests that need
universe packages. I am adding this back with if-conditionals
to avoid including packages from the product bundles. This is because
in a multi-product build, we may not have a product bundle set. And
if we do... that "main" product bundle is only one of the many
possible product bundles. Best to not include any product bundle
until we understand more.

Original-Bug: 398295610
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212564
Original-Revision: db8812b2f383a400764b749ffbda6a80b9e52c32
GitOrigin-RevId: c14dbe6f4ed0b3684bf52c8e3af72f2b49768b8d
Change-Id: Ic78fe54d3f5f48012fe79b0695e878206f3b2504