[roll] Roll fuchsia [logs] Flush logs before exiting on CTRL+C
Gracefully shutdown the logger when CTRL+C happens
instead of abruptly stopping which is the default behavior
of the runtime. This notifies the async thread when CTRL+C is pressed,
allowing the writes to stdout to finish before exiting.

Original-Fixed: 360372610
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208984
Original-Revision: d93cacfe18a191357b5372366da718c16c8f0eb7
GitOrigin-RevId: 0e39ec2e433c0c45eaaf5df48514e2b1d865479a
Change-Id: I8f0b93bfe6aecc49dc9888bb9a1a8276cd09fcc7