[roll] Roll fuchsia [audio] Remove ADR support for Dai drivers
This is the first of four sequential CLs that reduce the
compatibility burden for audio.

Fuchsia has supported four audio driver "models": Codec,
Composite, Dai and StreamConfig. Of these, three (Codec, Dai
and StreamConfig) have been deprecated since API level 20 --
all current and future audio driver development should use
the Composite model.

AudioDeviceRegistry will be used in future products. In
addition to Composite drivers, it has long supported
StreamConfig (and Codec), to maintain the possibility that
we retrofit the legacy AudioCore (which serves StreamConfig
drivers) to eliminate a significant portion of older
duplicate code. However, it is now clear that the existing
AudioCore is certainly adequate for legacy products, and
for any future products that require mixing we would create
a new implementation. (See fxr/1209988, 1209989, 1209990
which remove the remnants of this 'v2' refactor and
migration, which has been in-progress, then halted/waiting
for more than two years.)

With the clarity that AudioDeviceRegistry need only support
Composite devices, we can simplify its internals and reduce
its API surface area. This is a perfect time to do this
simplification, because ADR is currently only used in new
products (and these only contain Composite drivers).

This first CL removes support for the Dai audio driver
model. All traces of this driver type are removed, as
shown by the removal of 'Dai' from the DeviceType enum
and DriverClient union.

Original-Bug: b/390573488

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212026
Original-Revision: 616e105db1c83444098a229a9ddc284457b61e39
GitOrigin-RevId: 26a18c8b206bc0567c33b860df36384742269347
Change-Id: I1a460a76808ad25445b340f9417dd79e4a26340e