[roll] Roll fuchsia [block_server] Propagate block flags from device
As part of this, I made the `name` field in PartitionInfo non-optional,
as the only device which behaves like a partition but doesn't set is the
ramdisk.  I had to slightly modify FxblobMatcher to accommodate this, as
now the label will come back as an empty string rather than an error.

We might need to eventually allow the C block_server bindings to create
DeviceInfo::Block, since some future block devices that we will port to
using block_server might only be blocks and not also partitions.

Original-Fixed: 376745136
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213148
Original-Revision: 0e79d08288b6051612b08a55b9f08e2d75d73046
GitOrigin-RevId: 41d751fdcfb5a2a593f75ae8ed2b725b43a048a8
Change-Id: Id93b06e6350f5df8b08d9964299857bc08a67b11