[roll] Roll fuchsia [fidl][rust] Include generation in transaction ids
This adds a 7 bit "generation" in upper bits of the transaction id. It
increments each time we allocate a new txid for a channel. We mask it
off when we convert it to a slab index. This means that subsequent FIDL
calls won't end up with the same txid, even if they're in the same slab
spot.

This means that in the common pattern of a channel being used for
repeated hanging-gets each of the gets will have a distinct txid, which
allows for easier tracing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200091
Original-Revision: 9ec9e7a8e8644620e1a9802dca3838eec401f219
GitOrigin-RevId: 1e0ce3877ee168f54e672c999bf3d449dc638876
Change-Id: I5ea4839f8da32329ed316925a54fff5a3d333efa