[roll] Roll fuchsia [devices] Ensure /svc and class paths are opened with PERM_READABLE
This is required so that service instances can be enumerated correctly
when we transition fdio_service_connect to use the new Open method,
which enforces the new set of fuchsia.io rights. It currently uses the
DeprecatedOpen method which does not enforce these rights. The same
principles apply when enumerating devices in devfs.

Original-Bug: 324111518
Original-Bug: 376575307
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211866
Original-Revision: e40524217396be20af99531debeaadf2f553794c
GitOrigin-RevId: 9c6acdba186f87fa539a9b179cc5034e34331718
Change-Id: Ia4a5a39f8460773496d9912cc603e2053d5c1255