[roll] Roll fuchsia [kernel][phys] Simplify HandoffPrep::DoHandoff signature
There are no non-ELF kernels any longer, so we can do away with the
indirected boot lambda and just pass in the ElfImage object itself.

This is a reland of https://fxrev.dev/1178154 - with no diffs - which
was reverted only to revert a parent change.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204504
Original-Revision: 522f7155f263171d332304803dc7a572d251e8e5
GitOrigin-RevId: b1ce9eebac85dc725dc9f1efac3cfbad4e2c6c94
Change-Id: I1597172d04446edecf4e262cfed44aea49c1bbaf