[roll] Roll fuchsia [ffx][e2e] Starting emulator synchronously
This changes the e2e_emu lib to run `ffx emu start` in-line vs. as a
background child process. The advantage to running it like this is that
the command does not return until the emulator instance is up and can be
connected to via RCS. When run as a Child process, it returns quickly,
potentially before the emulator is ready.

Original-Bug: 398962542
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214185
Original-Revision: db0edca10205515c85a82a60f2ce2829c5a3259a
GitOrigin-RevId: e345334449f42e577999b3d7a21774bab55431b6
Change-Id: I3a8a732edf7c247d963869b01b2293af7d5b9587