[roll] Roll fuchsia [kernel][vm] Dump pager and port handles on timeout
Dump all processes that have handles to the pager object and the
associated port object on a pager timeout.

This walks the root job tree recursively, but the system is in a bad
state anyway, so it should be okay to be inefficient here for better
debuggability.

Original-Bug: 393202368
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214095
Original-Revision: ce007e9e8e793fc395053bace29d76a4db5fedbd
GitOrigin-RevId: 07cd4d517b7b969ef54352d4c3f9a6620bb6e449
Change-Id: I56c77c7ee36bb2096dd4bd3181b4448305cfa2a0