[roll] Roll fuchsia [starnix][functionfs] Separate reads and writes
Reads must not be reordered relative to other reads, and writes must not
be reordered relative to other writes.

Because reads and writes were handled on the same thread, but using a
concurrent dispatcher, they may execute out-of-order relative to
other commands of the same type.

This change updates the logic to allow reads and writes to execute
independently of each other, but keeps the ordering consistent within
each type.

Original-Bug: 395635052

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214444
Original-Revision: a952a3a0699cf348f1c5021766f1fd91dbe157ae
GitOrigin-RevId: 2c663d68af6403ab363eb16af62d45af0b30a939
Change-Id: I7f88a913a6d0247cb6c80095fe171c28d35806a1