[roll] Roll fuchsia [virtio][pmem] Add basic virtio-pmem driver
This adds a driver for the Virtio PMEM device which is a mechanism for
sharing a region of "persisted" memory with a guest VM. This driver
reads the configuration of the device and creates a VMO representing the
persisted memory region. The VMO is not exposed outside the driver and
the "flush" operation is not implemented.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213504
Original-Revision: a898e14622583ff3e37252847ffb80722601b7cd
GitOrigin-RevId: f80e8f835911b8dbb108ce6ea1a5c923230b2e73
Change-Id: If5e98cec7088341cde58711c32f8a8a04fd6ed74