[roll] Roll infra/recipes [rust-toolchain] Filter out rustc-perf from vendoring
Building rust broke it now requires all submodules to be present when
vendoring, but we don't check it out because it has a complicated
licensing story. This patch excludes it from the check to allow rust to
work. We should be able to remove once:

1. https://github.com/rust-lang/rust/pull/137583 has landed
2. our stage0 has advanced passed this checkout version.

Original-Bug: b/399417231
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1215144
Original-Revision: a3421adf94a9221a2038108da4566da6089aae02
GitOrigin-RevId: bb529ac503d02b941202e3be6bfcd542c8a90d60
Change-Id: I79ac0b563b26c15ff6c62dc8fe9e4c987f9fae34