[roll] Roll fuchsia [kernel] Remove idle time assert from scheduler
This idle time assertion has been causing a lot of flakes in CI/CQ,
so remove it. To handle the case where the actual_runtime is smaller
than the idle time (which is the case the assertion tested for),
bound the active_procesor_time_ns to 0. In other words, if the delta
would result in a negative number, clamp it to zero.

This workaround will be removed when the attached bug is resolved.

Original-Bug: 379576294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215427
Original-Revision: 39fbe71d9ec49e88dd84bf8481fe3c36fb4ea28e
GitOrigin-RevId: bfb2b1e62ac90998e5b352b1913eced188e60141
Change-Id: Ie8a351767131d40f0d366a7fb80b9cf846082c42