[roll] Roll fuchsia [fx][rbe] Check RUNT network on corp machines.
In order to detect early a very annoying issue that leads
to failed remote builds, make `fx rbe auth` also verify that
the machine, if on a corp domain, is not connected to the
RUNT network.

This is done by calling the check-runt-script introduced
by tqr/948730, if avaiable in the current checkout.

In case of error, this willl print a message pointing to a
bug comment providing a temporary solution.

+ Remove a few shellcheck warnings.

Original-Bug: 399096925
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213367
Original-Revision: 9ac6d0a9cd92628873b5068c81a3afbb8b973094
GitOrigin-RevId: 53a776839ed92d743dcb43650e49992eb6015e76
Change-Id: I1cf53e53b01907b746964ff51176d6af37d14f2c