[roll] Roll fuchsia [test_pilot] add TestConfigBuilder
This CL adds TestConfigBuilder to test_pilot. The builder implements
most of the functionality required to collect tests parameters and
assemble them into a test configuration. It leaves out the specific
set of parameters that test_pilot must understand, leaving that to
another struct.

This CL also add the Logger trait and a few implementations of it.
Loggers are used to debug the collection of test parameters.

Test: new unit tests

Original-Bug: 327640651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200685
Original-Revision: a78080016185c7339bbf8b91529bd6ec641393e6
GitOrigin-RevId: 4f59dbff6e76645652a8dbbc81952f0766d9f1ab
Change-Id: I4455d346cd5230b0fea7e49adef78416c89edb82