[roll] Roll fuchsia [tee] Introduce ErrorWithSize
This is a simple abstraction packaging an error with size, useful for
operations that return a required minimum output size in the case of
Error::ShortBuffer.

While we're at it we revisit the related decision to have structured the
inner get_object_buffer_attribute() to return the written subslice of
the provided output buffer parameter in the case of success. While it's
the opinion of the author that returning the size of what was written
and leaving the synthesis of the written subslice as a boilerplate-y
exercise for the user is a particularly C-flavored API, that is the
flavor of API that we're ultimately implementing. (Plus the Rust
standard library seems to have capitalized a bit on this with the
definition of the io::Write trait.) So we update the success parameter
to just be a usize of how much was written, which will also ease the
implementation of several API bits to come.

Original-Bug: 360942581
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215425
Original-Revision: d78d3ad43b7e28e3b1d15b97337d6cb04239678c
GitOrigin-RevId: 8133974d3966df2e6600a502bd0221bf0e0eb660
Change-Id: Icea4e9786180f608628e57340ff1063d9c6394c7