[roll] Roll fuchsia [fdomain] Rename .client() -> .domain() and make infallible
There were some concerns about how overloaded `client` can be in
contexts where we needed to get the client for a handle or FIDL proxy.

In addition, this is no longer fallible. We now return a placeholder
client that has a permanently-failed transport if the client has gone
missing. This should result in `ClientLost` errors being deferred to
handle use time as we have tended to prefer.

Original-Fixed: b/388594413
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188754
Original-Revision: 429a7a6e774820076924a6d038762445727523c8
GitOrigin-RevId: 8818501cb917f9f00376cb467db7e4937437b582
Change-Id: I21aa18fcae593b743fc24dc571916c094d4fa783