[roll] Roll fuchsia [libvulkan] Use service instead of files for testing DeviceFs
The callback Vulkan-Loader currently provides for Fuchsia uses the
`fdio_service_connect` family of functions, which should not be used for
opening files. This callback is used to connect to the loader's
DeviceFs, which is intended to connect to a specific GPU device.

Implementations are only supposed to connect to devices (services)
through the loader's DeviceFs, so this isn't a problem in practice.
However, we do happen to rely on this for our fake implementation which
tests the Vulkan loader.

Currently this works because we do not correctly enforce filesystem
rights with the fdio_service_connect functions due to many callers
relying on implicit rights. This test is one of the last remaining
callers of fdio_service_connect that use it for "connecting" to
files/directories. Once we migrate this test, we can actually begin to
enforce the io2 rights model correctly.

Original-Bug: 376575307
Test: fx test libvulkan-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213344
Original-Revision: 7f568c51c2ccbdd6ada25cc3c9536de60b907151
GitOrigin-RevId: accfb9bfa42ab52b6a48194c861bc6e11ee3689e
Change-Id: I6bd15c0ffb150e1a61d71b77702ddd5f19da1604