[roll] Roll fuchsia [assembly] Add support for bootfs files/packages in bootstrap and lower configurations
The bootstrap, embeddable, and empty configurations have a need to add
files or packages to bootfs.  This is rare, but needed for tests or
these configurations where the package system isn't fully available yet.

This allows product assembly to add to those sets based on files and
packages from the product (as opposed to the platform) but only in
these restricted modes.

Original-Bug: b/399480273
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215544
Original-Revision: 4918645d7f15a2dca2d74f7710951bd708b090df
GitOrigin-RevId: 87da83158fa815a333e15ff386204010dea328ef
Change-Id: I5ee6a47c56524a66ec435c23b2923c809ef6fbc5