[roll] Roll fuchsia [power][broker] Remove "unset" values when adding elements
The way the code was structured was leading to emitting unnecessary
events when adding a new element to the topology. When we add a new
element we always know its initial and required level, so we can omit
initializing the element with "unset" and exexcuting two updates to its
metadata fields in inspect.

I think there's likely additional opportunities for improvement and
simplification of these codepaths.

With this CL we now get:

```
{
  event: "add_vertex",
  vertex_id: ${element_id},
  @time: ${time_seconds},
  meta: {
    current_level: "${initial_level}",
    required_level: "${min_level}",
  }
}
```

Before we got:

```
{
  event: "add_vertex",
  vertex_id: ${element_id},
  @time: ${time_seconds},
  meta: {
    current_level: "unset",
    required_level: "unset",
  }
}
{
  event: "update_key",
  vertex_id: ${element_id},
  key: "current_level",
  update: ${initial_level},
  @time: ${time_seconds}
}
{
  event: "update_key",
  vertex_id: ${element_id},
  key: "required_level",
  update: ${initial_level},
  @time: ${time_seconds}
}
```

Original-Fixed: 398008841
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215304
Original-Revision: 230cfa012b58a28d766fe801047982d5ad255521
GitOrigin-RevId: 49b34510dd737aeaf030c5c132084b62ea940302
Change-Id: Ieb9a5a077d38053e1e33335c1d5933088c8825c3