[roll] Roll fuchsia [kernel][vm] Rewrite PhysicalPageProvider
Changes the design of the PhysicalPageProvider to directly return all
the loaned pages to the PMM, instead of relying on the PMM to clean
them up via pmm_delete_lender.

As a consequence of this change the coordination between the
PhysicalPageProvider and the page requests can be improved to remove
some existing race conditions that meant it was necessary to avoid
canceling active page requests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174214
Original-Revision: dd8bc1ca68eff2413e12bc31753ea79742a77024
GitOrigin-RevId: e8c610f28d88a18e9f6d6f2982b8ddb43b549d11
Change-Id: I580c86bcf24b164bbbb04fb99d4f6dacac44262b