[roll] Roll fuchsia [kernel][vm] Update loaned state accessors
With the removal of pmm_delete_lender and other changes to the loaned
pages interface the access rules of loaned and loaned cancelled flags
can be more strictly defined.

Due to the need to pack into a single byte an atomic still needs to be
used, but logically manipulations of a particular flag are externally
coordinated and hence all reads and writes can at least be of a
relaxed memory ordering.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1174215
Original-Revision: 60ec223dedbd8961c487b8135f36787161132a5a
GitOrigin-RevId: e5efd548a0c3fc49201d523d218b15480fc77491
Change-Id: Ie9ac14152f00d9114a179bd022c3d1a947bc6253