[roll] Roll fuchsia [build] Add variant-libprefix conf to profile-use
This patch adds variant-libprefix config to profile-use
variant to ensure that the shared libraries are installed
into the appropriate path when built with that variant.

It also adds in nstrumented tag to the profile-use variant
because the special handling of shared library paths is
done for the variants based on the instrumented tag.

Original-Bug: 335919004
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217126
Original-Revision: b6304742267404db447bc3a02a994fa451c2ded2
GitOrigin-RevId: 31dc482a7d1dd3e27ee69f560de5820dcf579bd5
Change-Id: I903402888385d9461bb6314968c58bf8321594f5