[roll] Roll fuchsia [ffx][build] Remove sdk usage from cli-goldens
This reworks the cli-goldens tests to use ffx and the
subtool manifest instead of the SDK or the idk_atom_collection.

Also corrected metadata about ffx auth.

This improves the build performance since it no longer has to wait
for the IDK or SDK to be built.

Original-Bug: 392136182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216846
Original-Revision: 404159383db4b34d04d635314f3cd3b631dba306
GitOrigin-RevId: 11c71f536b33d98489708d7e76017f5d2f072a5a
Change-Id: Id60285b758e25d059fcc83f296d52406e9a709e7