[roll] Roll fuchsia [ffx][idk] Treat unstable FIDL libraries as not in IDK
Use the "not-in-idk" marker for FIDL libraries that are in an SDK
category - and would otherwise have a different marker - but are
unstable. This helps ensure compatibility requirements are met, which is
the point of the markers.

Add existing violations to the allowlists.

Original-Bug: 314822328
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211284
Original-Revision: 802091e0b29d232dd377e7581051fd55a46eac8a
GitOrigin-RevId: 26df4924aaf28956675dedbedd53addaf124f26f
Change-Id: I4ea4acf52eb56985a0e01e5df75c0ed89f6b5e3d