[roll] Roll fuchsia [kernel][vm] Return locked clone
When creating a clone from a VmCowPages returns the currently held lock
for the clone, allowing the caller to complete any setup under this
single lock acquisition.

This causes no change in behavior as with the hierarchy lock the parent
lock, which is already held, locks the child as well.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1218447
Original-Revision: 57d9e7a526e1198dd627f517183c4190221d7816
GitOrigin-RevId: 0f8cd2382882c041dbd154c2ba7ec944a7adbeaa
Change-Id: I58aa2a5cf376d62e5e23d688f8eedab002aa5756