[roll] Roll fuchsia [assembly][timekeeper] Firm up use_always_on_counter
We use the always on counter to emulate an RTC chip on platforms that do
not use the classic PC-like RTC circuitry.  But, for this to work we
need two ingredients: the counter itself (a board feature) and a
decision to use it (a platform config option).

This change makes the final selection be a logical AND of the two
settings.  This allows us to introduce the feature tag
`fuchsia::always_on_counter` without changing anything until
the platform bit is turned on.

This is not strictly part of b/388562148, but is part of the changes
needed for some platforms, so I'm attaching it to that bug.

Original-Bug: 388562148
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220268
Original-Revision: 76fb4af0e8b859d8a3757415d76c557658f1f52f
GitOrigin-RevId: ae0c0ea530e659616baca46705e24ff464fd3c3c
Change-Id: I96ff5160a69cc9ec4d12e35da25d77fdc99a0ae9