[roll] Roll fuchsia [starnix] make memfd a tmpfile on a private tmpfs
This ensures that memfd is properly labeled for SEStarnix: we observed
that memfd is labeled as if it was a file create on tmpfs.

As supporting evidence for this change, consider the freshly-added
memfd.cc test and the policy change in
https://android-review.git.corp.google.com/c/platform/system/sepolicy/+/889855.

To get the right name in /proc/<pid>/fd, we fix the path-returning code
to add a `(deleted)` tag at the end of unreachable and deleted paths.

To allow testing, I change the base composite policy to give a specific label to
tmpfs.

Original-Bug: 398816545
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214904
Original-Revision: 846b86c96d8e1ed801d4634b7d4a10882af16ee3
GitOrigin-RevId: 9c2cc759231783cfe9a67069d803c4cd2927c59a
Change-Id: Iac0c2ffb8b553aec228c131c98a5247ace2ce5e6