[roll] Roll fuchsia [component_manager] increase liveness check frequency
We're still seeing crashes due to component manager writing to a full
channel, but we're missing the log message that's supposed to give us a
breadcrumb on where to look.

Increase the frequency of this log message to every 500th write. This
will both make the log more noticeable, and ensure that the log message
is written shortly before the crash (the write limit of 3500 and a log
interval of 1000 meant it got written 500 writes before the crash).

Original-Bug: 393176930
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220965
Original-Revision: c16229844763f8eb5779a0235297f6666fb5c257
GitOrigin-RevId: a4e15641e2190eb505f1ffcb34d1949ea29f6f75
Change-Id: I4d4221e2d2d1e3c41bc9a2e94e8813758534a899