[roll] Roll fuchsia [assembly] Assembly containers do not write paths with colons
Bazel does not like paths with colons, therefore before we write a
file to a container we convert any colons to a directory separator.

For example, if previously we would have written to:
  container/path/inside/with::colon

We now write to:
  container/path/inside/with/colon

As seen from the example, we also properly handle multiple colons in
a row.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220845
Original-Revision: 3090b1915184cbd5ef8a02272835ce72695ac097
GitOrigin-RevId: c35f03f96a9a13c6f4358fe55be18cdafcd67f90
Change-Id: I5c046df805c537db363c97df2a9aebab3fe56bcc