[roll] Roll fuchsia [console] Use service capabilities
The topological path suffix was removed as the only user of this
API specifies a virtio-console and it will only ever spawn a single
virtio-console. At that point, we just know the console is from a driver
rather than the kernel, and that's sufficient for us to just connect to
the first service instance to enumerate, so we switch to providing a bool
instead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216996
Original-Revision: 2fa13363a6d674ca4a9b276c13eee2106624721b
GitOrigin-RevId: e230cce5cb9adf7f093f1a2ab06ecd862ce10c1f
Change-Id: I127c0905fb5590b2f6c02703b6cfbbaa521f087a