[roll] Roll fuchsia [bazel_sdk] remove example of dep selecting on api level
This example adds an unnecessary constraint on an older API level. The
use case is not needed at this time so we are removing it to avoid
adding extra technical debt. We can always add it back when this use
case is needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221250
Original-Revision: 7a4ba8d71c6a633d36ec9fe43d422a0166d7b8a3
GitOrigin-RevId: 5517a2be061ee48d242b046920f299d240eb8423
Change-Id: Ibcb9b330ed1bcde5cdb372b882cc44178384a7cb