[roll] Roll fuchsia [dl] TlsDescResolver
This implements a new dl::TlsDescResolver class that uses the
entry points in tldesc-runtime-dynamic.S (recently added).  It
has its own unit tests but is not yet used by dl::LinkingSession.
It's a cheap ephemeral object meant to be constructed directly in
the argument to elfldltl::MakeSymbolResolver.  Its constructor
takes a reference to a per-module ownership object (linked list
of std::unique_ptr) that must be kept alive as long as the GOT
where relocations were performed with this resolver may be used.

Original-Bug: 342480690
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217964
Original-Revision: b4555558caa7fc8079263db1b27af14c39956198
GitOrigin-RevId: ace04527f48c11dfd96feac7afb5c142c78b4a8a
Change-Id: I911510b3040c44a825dc1a5b02646cd4e6548213