[roll] Roll fuchsia [ld][dl] Factor out ld::MakeDlPhdrInfo
This adds a common routine for constructing the dl_phdr_info for
dl_iterate_phdr, and uses it in libdl.  It also extends the
passive ABI with a `loaded_modules_count` field so that the adds
count for dl_phdr_info can be produced directly from the passive
ABI when there is no libdl.

Original-Bug: 342469121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222045
Original-Revision: ae385db33cdd338b5c0b04325e539034a65c1369
GitOrigin-RevId: 905275e79e87fbf79c984f0f91326cc957a42cfc
Change-Id: I53df7ce0288e280bcb94bf9857f6818b14e90e37