[roll] Roll fuchsia [power] Balance system-activity-governor tasks
The optimization of the SAG has greatly reduced the execution time at a
client side for calling TakeWakeLease. But this resulted in many tasks
scheduled but not completed at the end of the benchmark test for
TakeWakeLease, and happen to affect the result of ToggleLease because
Power Broker was still busy handling the requests from the tasks on the
SAG side.

To avoid this, a semaphore is added to control the flow so that tasks
can get executed. The test is also changed to hold to a lease so that
the optimization of the lease reuse can be tested. A different test for
the original behavior may be added later.

Original-Bug: b/395214643
Test: refer to the bug.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217325
Original-Revision: bfe019fa469ea07dadeaccdf17b2080b0a6b27e9
GitOrigin-RevId: 500a2e497309b7176e403e34ea57bf385b67b3eb
Change-Id: I2cb06234db52c8644dcc03b8069351c1b80a4ea3