[roll] Roll fuchsia [fdf][rust] Use an `Arc` in with_raw_dispatcher
A lot of contortions in tests happen because of the non-'static
test dispatcher. The reason for making it non-'static was to
prevent it from escaping the test, but it's not really worth it for what
we're doing in most of the tests.

Instead, give the test an `Arc` of the dispatcher and then make sure
that there's only one outstanding reference to it after the test body
has run.

Multiply: fdf_rust
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222022
Original-Revision: c23f8c08c15ab0f7ba4f439bdd56858ce8069f1f
GitOrigin-RevId: 7d476dbf187821f0b2b71cdd1e34f20a4bf053cd
Change-Id: I4f6248f4a26290b1d46ed15b2ca36d43d0d9f526