[roll] Roll fuchsia [sysmem] change pixel_formats log from error to warning
This is more than info but less than error, so plumb warning level and
use that.

The client can log at error level if the client both (a) sees this
result in an overall allocation failure and (b) considers that an
error.

Original-Bug: 401057708
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223344
Original-Revision: 251739a9773fa83be585cb39566e4e330896c292
GitOrigin-RevId: 7c25d6303c7eea74d555798438255684dcd31b1f
Change-Id: I1c6c2443b8ed4c088a7e8b8962267af55be3dbd8