[roll] Roll fuchsia [fdf][rust] Tasks have to be static, unfortunately
It was incorrect to allow clients to post tasks to the driver runtime
dispatcher that are non-'static, because it's fairly trivial to create
situations where the task borrows something that lives less time. The
only reason it appeared like this should work was because we end up
erasing lifetimes in the implementation so the compiler couldn't tell us
it was wrong.

Multiply: fdf_rust
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222023
Original-Revision: 6ea20dc22c49f6eb4b398a09091d9e4a1bff718e
GitOrigin-RevId: 6eb284f20f61d7e9e51f301bba8e4fd4a239f399
Change-Id: I4128032818642edd8ccd1c7a9a808f64dab7c6c5