[roll] Roll fuchsia [netstack3] ResourceCounterContext non-mut ctx
Update the bounds on the ResourceCounterContext so that it operates on
an immutable reference, rather than a mutable reference.

In doing so, `increment` is renamed to `increment_both`. Now that it
takes an immutable self, the compiler had a hard time determining which
`increment` was being called (either the one provided by
`CounterContext` or `ResourceCounterContext`). Giving it a separate name
disambiguates things.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224384
Original-Revision: 5899cbdd44fed6444b9142c44b2dd5b2abdb2fc1
GitOrigin-RevId: 35cefa89675cce5d95cc344532baae6d15cf19b5
Change-Id: I0af8e558071b07fee5ef6218a1b96e15bb0b0f78