[roll] Roll infra/recipes [mirror_version] Only consider last Copybara footer
Cherry-picks to sd26 have been failing:
https://ci.chromium.org/b/8720695068026969361

This is because the most recent commit on sd26 currently lacks a tag.
That happened because the release-mirror-version builder that was
supposed to copy the tag from GI's f26 branch failed:
https://ci.chromium.org/b/8720750037535589777

It failed because the triggering commit had *two* GitOrigin-RevId
footers:
https://turquoise-internal.googlesource.com/smart-integration/+/4246d9256708eff8c093fad2a158938d28de49ed
The recipe used the first one it found, which was not the correct one as
it corresponded to a commit that had already been mirrored. The correct
behavior in this situation is to only use the last Copybara footer if
there are multiple.

I had to manually add the tag via the Gerrit admin UI in this instance,
but this CL will prevent it from happening again going forward.

For posterity, that commit has two Copybara footers due to the following
chain of events:
1. http://tqr/953096 rolled the antlion repository in GI on f26.
2. The resulting commit was copied to the f26 branch of public
   integration by Copybara:
   https://fuchsia.googlesource.com/integration/+/84feb778b724eb565cc6b022c1547b81539b7e96
   This is the source of the first Copybara footer in the commit
   message.
3. The submodule-update-f26 builder updated the antlion submodule in
   fuchsia.git accordingly in https://fxrev.dev/1224764.
4. The submodule update CL got cherry-picked onto f26:
   https://turquoise-internal.googlesource.com/integration/+/e85f2b6a785a233a7e9da59e6990bd80b69e2e42
5. That CL got copied into SI's sd26 branch by Copybara, producing a
   commit with two Copybara footers:
   https://turquoise-internal.googlesource.com/smart-integration/+/4246d9256708eff8c093fad2a158938d28de49ed

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1225385
Original-Revision: d37010352b1f434cb7ea8c8c385433192372b077
GitOrigin-RevId: 64cea9ff201647a4049c2cfd04fa3d3c1b0c8e13
Change-Id: I99f7ea9fe4c83897276398d85f20e5ec490356b1