[roll] Roll fuchsia [kernel][vm] Always dead transition parent_
Always returns, and clears, any parent_ during a dead transition. The
previous logic around avoiding this is an optimization that was
applicable when the deferred deletion was a more heavyweight list, but
the logical complexity is no longer worth it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223984
Original-Revision: b5af05f2303c45f04f47556f082b9b52a06fac42
GitOrigin-RevId: ad03636ed232062df51c42ecde1076ca1ae16971
Change-Id: Ieb90e1f083b99df69f983254ff43d3e086ad5e56