[roll] Roll fuchsia [kernel][vm] Owned page iteration takes parent lock
Allows the different owned page iteration methods to take an optional
reference to an existing locked parent. This prevents a double lock in
situations where the caller already holds the parent lock, but also
needs to walk the hierarchy.

A helper object is added to simplify the locking management of the
ForEveryOwnedHierarchyPageInRange method.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222705
Original-Revision: 2ffae9f209118156844152dad1885f594ef7d99f
GitOrigin-RevId: 2e2101cec00cfc38daa5a656e80c3ff1703bfc1f
Change-Id: I5558704d4cde4e4878485290ec6b9bae3957fcb4