[roll] Roll fuchsia [bazel] canonicalize label in refresh-bazel-compdb
The refresh-bazel-compdb script allows a user to pass in a label. This
label must point to a fuchsia_package but the fuchsia_package target
adds a suffix to the name to make it work with the workflow tasks. This
can confuse the user so we removed the need to pass in this suffix.

Fixes: 402509201
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225871
Original-Revision: 80f1b86d3f41332953caa72b8579c9df4a01f84c
GitOrigin-RevId: 88b2b0d5d4f7c37b0a12038c07281fa9effd2e40
Change-Id: Iad0e608cb861f3174c3a9c9ccdba7c387c8ae346