[roll] Roll fuchsia [firmware] add `cipd_utils.py manifest` subcommand
cipd_utils.py has some utility functions to help analyze and compare
manifest versions of CIPD packages, but no central way to actually
generate the manifests in the first place.

This adds a `manifest` subcommand to generate a manifest on the given
repo that's in the format expected for the rest of the tool.

Original-Bug: b/393584187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224484
Original-Revision: 10263b5387a66c09589281bf793322ce4544f35a
GitOrigin-RevId: bbc7a22f7e14b3afd547e75e8bf361ad5eb4871f
Change-Id: I1a38be5ca177837a0962bffa3ee42967cee3ad76