[roll] Roll fuchsia [firmware] cipy_utils.py out-of-tree support.
cipd_utils.py contains some functionality that we will commonly want in
our firmware repos; the easiest way to do this will be to support
running out-of-tree and copying it around as-needed.

In this CL, we allow the `cipd` utility to be on $PATH instead of
requiring a Fuchsia checkout, and support an `--allow-dirty` flag to
support generating manifests from a dirty repo, which is useful when
developing and testing release scripts out-of-tree.

Unrelated, this CL also logs stdout/stderr when a subprocess fails,
which makes diagnosing errors much easier.

Original-Bug: b/393584187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1226664
Original-Revision: f3bafbaaba8acad7ba15107aaefdffa349d52cc0
GitOrigin-RevId: dd22dcf39458db0185635d26056839ca2c34b38f
Change-Id: Ibdb96655e46065fde40952bbf62bae1379a7d276