[roll] Roll fuchsia [assembly] Switch scrutiny structured config test to use assembly's example aib instead of core shard
This is the last user of the core shards and core realm definition in
the legacy bundle.  So switch it to a component that's part of the
example / testing AIB, so that the core_shard() and core_realm_definition()
templates can be removed.

Original-Bug: b/385412024
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178716
Original-Revision: 8eab27d8d7cb9b7773d256d8a6db31918eed4db1
GitOrigin-RevId: ff81052bda6c02a09278ad4e0105bd3743eb397d
Change-Id: I539c73dfab551ee272c056bfc86afbb5cf369f6f