[roll] Roll fuchsia [sestarnix] Implement and call bpf_map_alloc hook.
`bpf_map_alloc()` now assigns a security context to the BPF map and it
is called by the `sys_bpf()` syscall.
In this CL we also introduce a wrapper for the ebpf_api::PinnedMap, so
that we can store the BPF Map's security state.

Note that the security state itself is for now unused. It will be used
by the bnf_* hooks in follow-up CLs.

Original-Bug: 364570322

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224304
Original-Revision: 584580eafb06d7bb0d13df36494bd3ca10a1e61c
GitOrigin-RevId: 5ccc35c9e39879f3005ec1670303b50783baf61e
Change-Id: Ie6c597f266c12f48381d639cae5fe71e27c9f7b2