[roll] Roll fuchsia [tee][binder-proxy] Access shared memory region using virtio-pmem
This routes the fuchsia.hardware.virtio.pmem service to binder-proxy in
the tee realm and adds a configuration 'use_virtio_pmem' to
binder-proxy. If this configuration value is set to true, binder-proxy
waits for a device to be published in the pmem service and collects a
VMO from the service. The VMO is not yet used for anything.

Original-Bug: 397758089

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216548
Original-Revision: 306704d993cf76b186f2c874b8f141c886e4e40d
GitOrigin-RevId: c54e0c42db83dd7efe85704434f359ffba53b9b8
Change-Id: I0602d9c3a62d1a22f1c7f5fb8327dd588368c132