[roll] Roll fuchsia [sestarnix] Implement and call bpf_prog_alloc hook.
Similarly to `bpf_map_alloc()`, the `bpf_prog_alloc()` hook assigns
the security to the BPF prog. The hook is called from the sys_bpf()
syscall.

Note that the security state of the Program, is not accessed right
now, but it will in a follow-up CL.

Original-Bug: 364569869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225207
Original-Revision: 2cc1a226ef7f520bc3b0fbcaabab9d8de8a531cc
GitOrigin-RevId: 8c4e9218dbe45a418595dfd7bd326954e038dd43
Change-Id: Iaf482f7c2194026a917d8464f4c7ffd69ad707a1