[roll] Roll fuchsia [debugger] Add ElfMemoryRegion
This provides a way to directly access raw ELF files, decompressing
sections as needed. This is particularly useful for the unwinder which
expects to be able to read ELF data directly from a Memory
implementation (which is just a simple opaque reading API).

This CL doesn't hook up any of the users of this class, that will be
done in followups.

Original-Bug: https://fxbug.dev/292107156
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195606
Original-Revision: 5b05e0c30bb7d7579efb8b0c65475a988d6d8713
GitOrigin-RevId: 198b77b334c0d5f3420e823a21d487f4086c676e
Change-Id: I3dd8e5bc134360ae9471eb373cc095f1a8b5d11d