[roll] Roll fuchsia [tools][fuchsia-sdk-run] Removing fuchsia-sdk-run
This tool was intended for use OOT when switching between multiple
Bazel SDK projects. However, the project has shifted priorities and
OOT multiple projects are not going to happen as hoped.

This tool imposes some constraints on the design and structure of the
IDK and SDK which are causing friction to change.

When the need arises for a tool like this again, it may be restored,
time will tell....

Original-Bug: 386416011
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227064
Original-Revision: 4236a7c5a6d82bd5346fb9d72c87311fa0f8fb25
GitOrigin-RevId: 17ad2dbc25f494974ad926c72120fe3e32e8e18a
Change-Id: I5bb9e8c313a0f6c709b62e715a94b6ed2f5dd3f3