[roll] Roll fuchsia [ffx][pkg] Removing unused code due to daemon server removal
This cleans up all the code that was used to manage the configuration
values for the daemon based repo server.

It also removes the data structures that were used to expose the repo
server functionality to the daemon protocol model.

Original-Bug: 391921340
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1226985
Original-Revision: 8a4848db07faab430e97c5903394182f32d1fe33
GitOrigin-RevId: 3bf37023586db5fbba8263f4b2b3697e9912be43
Change-Id: Ie5d31365ce5db772e2eef7535670d337dcaa8edb