[roll] Roll fuchsia [bazel] workspace_utils.py: Fixed missing tracked files
The generate_fuchsia_workspace() function returned a set of
file paths that was missing a number of important input files
that are read by the function. As a consequence modifying any
of them would not properly trigger a regenerator invocation
on the next `fx bazel` or `fx build` call:

- build/bazel/templates/template.bazelrc
- build/bazel/templates/template.bazel.sh
- build/bazel/templates/template.platform.mapping

This fixes the issue by:

- Adding a `read_text_file()` function and `input_files`
  property to the GeneratedWorkspaceFiles class.

- Using them to read input files in record_fuchsia_workspace()

- Add the content of generated.input_files in the result
  of generate_fuchsia_workspace().

Manual comparison of the `build.ninja.d` file generated before
and after the CL confirms that the missing input files are
now properly listed in that file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227466
Original-Revision: ed6a77ef2eddf25ef685a21b3614654f342bce14
GitOrigin-RevId: e3797addf948e3c6ea70514a8d3036210b3bb108
Change-Id: Ida24dccdf51161f2e6723722a3359925f26a08e6