[roll] Roll fuchsia [honeydew][python] Fix coverage.sh script.
This allows for fuchsia-controller to be imported using the system
search paths in addition to the current shared lib directory.

This also updates the paths used in the install.sh and coverage.sh
scripts in use with Honeydew so that they are able to locate the updated
`fuchsia_controller` library.

Original-Fixed: 394153187

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228985
Original-Revision: b685c45c1d20acf895dc2c43a6bda938b066bfc9
GitOrigin-RevId: c0d686141e5da3eea6fa24534cc262c88f078ff4
Change-Id: I04608f41e33f1c9f2912f866fea5f1f409dc2811