[roll] Roll fuchsia [starnix][cgroup] Freeze newly created Task in frozen cgroup
As a freeze operation is propagating, `clone_task()` can happen
simultaneously, and the newly cloned task can "escape" from the freezing
cgroup, allowing it to be scheduled. This change fixes this by:

1. Locking the Cgroup hierarchy's PidTable, so that processes cannot
   move between cgroups during Task creation.
2. Initializing the new Task with an existing KernelSignal::Freeze.

Verifying this behavior in a test is hard, because we cannot guarantee
the ordering between a `fork()` and an ongoing `freeze()`, and depending
on which happens first, the newly spawned `Task` can both be scheduled
or never scheduled.

Added stub log for CLONE_INTO_CGROUP, it is not implemented in this
change.

Original-Bug: 402530518
Multiply: cgroup_freezer_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230023
Original-Revision: 7702af677a5c4076eed242e30b7dd1da8473aff0
GitOrigin-RevId: 7efcff115ee1ab03a2f1d32375b1537e5eeb2b8e
Change-Id: Ic72be5da170da3c9dad8b545a04458bfeef89238