[roll] Roll fuchsia [fdf] Fix issue with disabling a nested composite
This fixes an issue where a composite that is colocated inside another
composite could not be disabled fully as the composite rematch was not
being performed. On the driver index, the disable handler does this, so
here I've added the same into the driver manager after it receives
confirmation from the driver index.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221227
Original-Revision: df23f39466fd5937c3472f31350605aa2463305b
GitOrigin-RevId: 066cdc66b6a0de76ca2b43c1a1195e399e508898
Change-Id: Ib34eb0a8b140b7eb4c5c62992aff568298d09889