[roll] Roll fuchsia [kernel] Create singleton KTrace class
This CL modernizes the ktrace library API by introducing a new
singleton class called KTrace that encapsulates the KTraceState.
The Emit* functions are now static methods of the KTrace class.

Because this is a refactor, it should be a functional no-op.

Original-Bug: 42069955
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231580
Original-Revision: 9420e57ab27e3612fb146d347e581f39b4ef489d
GitOrigin-RevId: a56d6bb705663c6adf52278cb85da8694129f263
Change-Id: Ibd799c135091c691df8941bd6fb154f303916fd6