[roll] Roll fuchsia [kernel] Ensure .dynamic ends up as read-only
The section may statically be writable, but in any case after relocation
it should be read-only. This change ensures that by updating the layout
to place .dynamic in rodata if it was originally read-only or in RELRO
if writable.

This fixes a regression with BFD on arm64 and x86 where .dynamic was
writable, leading to a writable rodata segment.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232064
Original-Revision: 23089061850f387755d0726cc4cc7f8b25ec3211
GitOrigin-RevId: 205fe56fe2d1ecd81c83cbadffbe6a87bc73b5b4
Change-Id: Iec2857e797df68811b7fc808e18f98c333a1a76b