[roll] Roll fuchsia [kernel] Move ktrace_init into KTrace singleton
This moves ktrace_init, which is the LK_INIT_HOOK that initializes
the tracing subsystem in the kernel, to a static method in the
KTrace singleton. Some of the operations in the init function have
been reordered: the early exit when ktrace is disabled now runs
before the RegisterCallback and the CategoryBits are set up, since
they will not be needed if ktrace is unused.

This is the final callsite that made use of the public accessor for
the internal KTraceState. With its removal, we can finally make that
accessor private and prevent any files that include ktrace.h from
being able to access that internal state.

Original-Bug: 42069955
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232105
Original-Revision: a35f49ae59dcacbe85b113cd0b75caf1b6dce3da
GitOrigin-RevId: dd029f2c6c48ae4e8a72d9f900484a71d5a01b4d
Change-Id: I30cc05b58079de806e18e32c96814f1573f59942