[roll] Roll fuchsia [fidlgen_go] Treat flexible methods/events as strict
To assist with ongoing FIDL protocol migrations, we allow fidlgen_go to
generate code for handling flexible methods/events by treating them as
strict. All remaining Go code is used to communicate with in-tree
servers, so we always know the exact set of methods which
servers/clients on both sides need to support.

This is required to allow us to transition certain io1 methods to be
flexible to allow a graceful phase-out. This will allow us to complete
the ongoing fuchsia.io migration without being blocked on the Netstack2
removal.

Original-Bug: 400784122
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225704
Original-Revision: 46fcc60fdab5f89a52fe0c785c5fa26314571089
GitOrigin-RevId: fa14e8d13e1f7ae298f514bc00f21349d5689572
Change-Id: Ief64aab575a869ef1a7c65e2926d43c61d98519d