[roll] Roll fuchsia [assembly][bt] Add config for Hands Free
Add a configuration to enable the Hands Free profile
(bt-hfp-hands-free). This config shares the same codecs and controller
features as the Audio Gateway role.

Define an AIB for the component and include it in the build if specified
via assembly. This AIB is experimental and has not gone through security
review yet. Add it to userdebug and eng builds only.

NOTE: This CL does not enable it on any products.

Original-Bug: 401064356
Test: fx build
Test: fx test src/lib/assembly

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221429
Original-Revision: 831ba4a1d37572aad93c6e8cf529c16479cab131
GitOrigin-RevId: 403d78fc09456a960a532c1d6021334f391d3cc6
Change-Id: I2f4b016dd379705180d8417880418f19d6e4dca7