[roll] Roll fuchsia [build] Move zircon sysroot/sdk to sysroot_sdk so its files aren't included by RBE for linking
The gen/zircon/public/sysroot folder is included as a whole directory
structure, no files should be in it that aren't part of the sysroot
itself, including the sdk atom that describes it.

The sometimes-present nature of the sdk atom files causes cache misses
as depending on when build actions run vs. the sdk atom creation ends
the files may or may not be present, changing the overall closure of
files that are part of the remote action.

This shifts the sdk atom from a subdir to a sibling dir, so that its
files are excluded from the operation.

Original-Bug: b/405227511
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233610
Original-Revision: 301385ab5b78a009a7b97553ece834e286829f01
GitOrigin-RevId: c5d3da9e410aaa5962a882134986aedc0d32aea6
Change-Id: Ifc97bcc56f9e6044b4af9d0b7945c51d9a59dbeb