[roll] Roll fuchsia [kernel] Add a tracing resource
Introduce a system resource that will gate access to ktrace
operations. Note that this CL does not make use of this
resource, it just creates it and routes it through component
manager and its tests. It also modifies the security eng policy
to correctly restrict this resource's usage to shells, tests,
and the ktrace provider.

Future CLs will modify the ktrace syscalls to use this resource.

Original-Bug: 364894040
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228966
Original-Revision: 10f5cc59eff77aa181458322acef7125315fef10
GitOrigin-RevId: 571ccb5fceee0ede97ca1f714c3bf7c913f22ee7
Change-Id: I83b5cc376f96a0a97e811821e7f1df203cc07854