[roll] Roll fuchsia [network/test] Create event stream before starting component
The iPerf tests rely on being able to observe the client exiting, but
they open the component-stopped event stream after starting the client.
This means there's a race where it's possible for the client to exit and
the event to be sent out before we actually open the stream. Fix this by
opening the event stream before starting the component.

Original-Bug: 404377171
Multiply: iperf-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233773
Original-Revision: 04cce37d454f5e561e95ee1243486bff958a9623
GitOrigin-RevId: 28badbbac330d801f9e906ce5b47748352cbdfeb
Change-Id: I7e43158333fb771e172b653114dd92471f9aef26