[roll] Roll fuchsia [test_pilot] add run_test and helpers
This CL adds run_test, the function that actually runs the tests, and
some related helpful types. run_test functions in the sense that it
will actually run a test and its post-processors, but it does not yet
do the output processing work required of test-pilot. This will happen
in a future CL.

Helpers include:
- OutputDirectory, which owns the structure of the output directory and
generates paths on behalf of run_test.
- StdWriter, which is used to route stdout and stderr to files in the
output directory and to another writer (in this case, writers for
test-pilot's stdout and stderr streams, respectively). StdWriter also
deletes the file it creates if it is never written to.
- InvocationLog, which records the invocation of the test and post-
processors and which is serialized into a file in the output directory.

test-pilot's name has been changed from test_pilot.

Test: new unit tests

Original-Bug: 327640651

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233585
Original-Revision: f005996074d06519ed7b04b75905b45556251eb4
GitOrigin-RevId: bebbb9b951d3a28c418f63b59eb9b0b0f934c8d6
Change-Id: I7eaaa12644f4ddfb6dc186399f4b040d948df93f