[roll] Roll fuchsia [symbolizer] Update build rules and paths
The primary change of this CL is to update the incorrect assumption
that in-tree, the SDK root is always the $root_build_dir. Since the
SDK may be built in a sub directory (as we are currently making
improvements to IDK construction and usage in-tree), the location
of `.symbol-index.json` should be calculated from the build directory
from the configuration context when in-tree, and fall back to the IDK structure otherwise.

This also adds a build rule to add `symbolizer` to
$root_build_dir/host-tools to be aligned with how other host tools are
located by `ffx` before the complete IDK is built.

Finally, there is an update to OWNERS to replace people no longer on the
team.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232765
Original-Revision: 394d82eddf600893c121207284a681ef13c49c51
GitOrigin-RevId: 62d00adb8adc775169b51965f11066ba0cec0e73
Change-Id: I16ff32b3e98ed3a8795148777a40af18b62f49d8