[roll] Roll infra/recipes [submodule_update] Reword commit message
Now that submodules are rolled back and the pinned versions no longer
impact checkout state, it no longer makes sense to have the commit
message for submodule updates mirror the corresponding integration
commit - in fact, it just creates confusion because it's unclear whether
the fuchsia.git or integration commit actually updates the version.

To make this more clear, have the submodule_update commit message
explain that the update is only for the purposes of Code Search and does
not impact any checkouts or builds.

Original-Bug: 406021557
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1235246
Original-Revision: c5cfdcdc667883543a03833f99f1b6e959642518
GitOrigin-RevId: 1d0e824cda345b394c9c15c809c5f2fde0365493
Change-Id: Idb8217f7648697b5f87c1955092f132fd8c6b191