[roll] Roll fuchsia [net][ffx] Cleaning up ffx in conformance test
This cleans up the staging for using ffx in the network conformance
tests.

This is in preparation for streamlining the IDK dependencies in-tree
when using ffx in a test and remove the hard coded manifest used by
ffx in-tree.

This updated test depends on the "HostTool" SdkRoot in ffx. This means
there is no manifest used to configure the SDK, but relies on host tools
being located in the same directory as ffx. In-tree this  should be
`$root_build_dir/host-tools` e.g. out/default/host-tools.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231284
Original-Revision: bf850049b0aee8ddeec527093f5f4ac5b11d3b9f
GitOrigin-RevId: d97ec5670dc10f5747549551493ad5a5ce59a92d
Change-Id: I52cbf93f38bb9a1c3246347a7f0f9fff2ef13be1