[roll] Roll fuchsia [packet-formats] Extract NDP option type consts
Sometimes we want to be able to refer to a known NDP option type as a u8
without needing to include it in the union of NdpOptions that are
necessarily parsed when parsing an NDP message using packet-formats.

To facilitate this, define an option_types mod with those consts, and
then use those consts in the definition of the protocol enum (which
requires some small changes to the `create_protocol_enum!` macro).

Original-Bug: 397475289
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235464
Original-Revision: dce5f4a35d5a957ab7ca26606da65c27949dcec7
GitOrigin-RevId: 14341d03cab1367debd7dc2dee62f94872f0e14e
Change-Id: I1b979474aa5a62e3eb9461ddecbce13e1604e2ac