[roll] Roll fuchsia [wlan][sme] Don't use MacSublayerSupport in SME
Refactor SME to no longer depend on MacSublayerSupport. Specifically,
this removes the functionality to intersect rates and capabilities for
the SoftMAC AP implementation, which is currently only used in tests.

This also removes the QueryMacSublayerSupport calls on startup in
both the FullMAC and SoftMAC MLMEs.

As a byproduct of this refactor, fuchsia.wlan.common/DataPlaneExtension is
no longer used by either MLME or SME and will be removed in a future
change.

This is part of the effort to cleanup driver_features in
fuchsia.wlan.common.

Original-Bug: 305303708
Original-Bug: 406220225
Test: Unit tests on CI/CQ
Test: Antlion quick tests on CI/CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235564
Original-Revision: 4af8a562c78fea3f6f5a27e714eed00536046257
GitOrigin-RevId: 35a86d9ef3f98e6c779c55f0534ae96be43a5c64
Change-Id: Ib57ee3b6649d468e6af3e737516c916de1a1d12a