[roll] Roll fuchsia [kernel] Introduce static KTrace instance
Modifies the KTrace singleton to expose a static KTrace instance.
This will allow the object to eventually store ktrace metadata
like categories, whether or not writes are enabled, process and
thread names, etc.

For now, this is a functional no-op. It just starts the process of
moving functionality out of the internal KTraceState in preparation
for the usage of per-CPU buffers.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236640
Original-Revision: 1fca48d35dd2aaf37475ecc8f61601303880cc3c
GitOrigin-RevId: 1cd4fb1f150263cd723fc1ea3d55eef3c8967e37
Change-Id: Iac54179d13cbcf882702f764f91831687b65e587