[roll] Roll scudo [scudo] Use a tryLock in secondary release to OS (#132827)
In the caching part of the secondary path, when about to try to release
memory to the OS, we always wait while acquiring the lock. However, if
multiple threads are attempting this at the same time, all other threads
will likely do nothing when the release call is made. Change the
algorithm to skip the release if there is another release in process.

Also, pull the lock outside the releaseOlderThan function. This is so
that in the store path, we use the tryLock and skip if another thread is
releasing. But in the path where a forced release call is being made,
that call will wait for release to complete which guarantees that all
entries are released when requested.

GitOrigin-RevId: 96be331a6c19607267f44ed16d2a93ceaa00bef9
Original-Revision: 2dd0bdf9c99c51d6ad2ecac1dbfbe8c32b82bb01
Change-Id: If0d78083265eb3381e8d54212d25188faad79d55