[roll] Roll fuchsia [ffx][sdk] Update finding host tools not in the sdk
This changes the get_host_tool to only return paths of tools
in the SDK manifest. If there is no manifest, which happens in-tree
or in an isolated test environment, in this case the tool is looked for
in the same directory as the running binary (i.e. ffx).

This fixes a problem of blindly checking the sdk root for the host command
since there are binaries which are compiled for Fuchsia, not the host environment.

Original-Bug: 406159261
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237364
Original-Revision: dbf618e3fa497ab479add18fb6ac44f34db6b54f
GitOrigin-RevId: 7e689ce9b9fe52d2f309b2840a9377d9764e6db0
Change-Id: I6d4f0f79db6f14c228fd148f4cdc79204f179ce0