[roll] Roll fuchsia [goldfish] Remove CreateBuffer2() support.
CreateBuffer2() render control command is used to create external
data buffers shared between Vulkan instances and non-Vulkan clients.

Currently this render control command is not supported on
gfxstream-based AEMU (FEMU). Given that this is not used by any
in-tree or out-of-tree clients, we decide to remove support for this
method for now to unblock the AEMU migration to release-gfxstream
builds.

Test: goldfish-unittests
Test: goldfish-control-device-test
Original-Bug: 405344910
Original-Bug: 406545904
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237584
Original-Revision: bbdefae70905110ba94ffd92609a6430c2fcfc15
GitOrigin-RevId: 860cd8b610bd9fd0248a94f16bf1554eaa3d0a21
Change-Id: Icd9456c4beb11c34bdd1ff75c10b3fd02a21ddeb