[roll] Roll fuchsia [heapdump] Transfer the name and vaddr of ExecutableRegions too
These fields will be needed when we start using the
//src/developer/ffx/lib/symbolize library for symbolization
instead of delegating to pprof.

For now, the only effect of this CL is that the Mapping
entries in the generated protobuf will have the "filename"
populated. Note that pprof replaces it with the detected
path when it resolves a mapping through its build ID.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235585
Original-Revision: 48384d5dc6797b693e71c6209479be763b6cf184
GitOrigin-RevId: 4412d2da74f72dd989bc7fbe2552ac66319b6461
Change-Id: I911e6a854e2fde42be3685606e1706f33afd5a7e