[roll] Roll fuchsia [fidl][python] Implement protocol (ignore methods for now)
fidlgen_python is in bringup and will only contain an empty class for
each protocol until a later change lands to implement the methods of
each protocol. I decided to land these in two separate changes because
this change allows the boilerplate for protocols to be put in place
before making the jump to implement all possible kinds of methods.

Original-Bug: 355717613
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215765
Original-Revision: 02e21079ec9edf849236b2b7dd90537fdfdfec81
GitOrigin-RevId: 058e07b6e4335bfcb34354c05bccfa5c7361b055
Change-Id: I8e22614152087b27740b0af217afe2c14279643b