[roll] Roll fuchsia [ffx] Allow VSock address types
This is a pretty bloody change. A lot of FFX code assumes that targets
are identifiable by addresses, and those addresses are IP addresses.

That said not much of substance should happen in this CL. It's mostly a
no-op until we start using vsock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210315
Original-Revision: cbf616ef157561d5d99d76eaafbc66c59215999e
GitOrigin-RevId: bac3091febaee3b9f97b67124ec67633fda54fe1
Change-Id: Ia8dadfcfb1e6d817fe264a94b8c940965915c702