[roll] Roll fuchsia [kernel][phys][vm] Formalize VM hand-off
This change introduces phys handoff abstractions describing the mappings
it created along with the VMARs - subVMARs of the root - that these
should end up in, as well as updating the kernel's VM initialization to
simply iterate over these and follow through with VMAR creation and
mapping reservation in a table-driven fashion. The table-driven aspect
simplifies things, is more flexible (being in turn driven by the source
of truth for those mappings), and readily extends support to the other
mappings we intend to make in physmap (e.g., periphmap and stacks).

This replaces the previous VMAR creation + reservation logic in VM init
that explicitly did this for the kernel and physmap alone.

Example VM init logging as of this change:
```
[00000.000] 00000:01025> initializing vm
[00000.000] 00000:01025> VM: handing off VMAR from phys: physmap @ [0xffffff8000000000, 0xffffff9000000000)
[00000.000] 00000:01025> VM: * mapping: physmap (rw): [0, 0x1000000000) -> [0xffffff8000000000, 0xffffff9000000000)
[00000.000] 00000:01025> VM: handing off VMAR from phys: kernel @ [0xffffffff00000000, 0xffffffff004af000)
[00000.000] 00000:01025> VM: * mapping: segment (p_vaddr = 0x0) (r): [0x3229000, 0x32a2000) -> [0xffffffff00000000, 0xffffffff00079000)
[00000.000] 00000:01025> VM: * mapping: segment (p_vaddr = 0x79000) (rx): [0x32a2000, 0x361b000) -> [0xffffffff00079000, 0xffffffff003f2000)
[00000.000] 00000:01025> VM: * mapping: segment (p_vaddr = 0x3f2000) (r): [0x361b000, 0x3626000) -> [0xffffffff003f2000, 0xffffffff003fd000)
[00000.000] 00000:01025> VM: * mapping: segment (p_vaddr = 0x3fd000) (rw): [0x3626000, 0x36d8000) -> [0xffffffff003fd000, 0xffffffff004af000)
```

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178312
Original-Revision: b470f93e01a5b16499ebd7be6f9371ee8bddf2c0
GitOrigin-RevId: 170fe6ffabe1949f941fd281af89344b50fa9ca8
Change-Id: Icdde1e69c4c206925fd653d594a19d105da09ca6