[roll] Roll fuchsia [bazel] add hlcpp visibility support to generated sdk
This CL adds a mechanism which lets us inject visibility into the hlcpp
bindings to restrict their usage. The mechanism at play here lets us
decide visibility on a per SDK basis which is important to make it so we
don't break our OOT users while still restricting the in-tree users.

The CL restricts usage to the fuchsia_sdk and internal_sdk since there
are libraries that are still using hlcpp but it prevents usage in other
locations in-tree.

Original-Bug: 406789405
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238428
Original-Revision: 9a3a215ee90e1d305d35b455823489171d67885d
GitOrigin-RevId: 3e6fc56b12667d533b6b27905d96bee50b2a194d
Change-Id: I6aab185765cf77635d5c66335492a05ddc4570a4