[roll] Roll fuchsia [sestarnix] Check sys_rawio for FIBMAP ioctl
This is the last substantial permission-related part of getting the
capable_sys test of the SELinux Test Suite passing. Since it's pretty
obscure (we're not yet aware of any FIBMAP callers other than the
SELinux Test Suite) and we'll need to fake a successful return of the
ioctl we introduce the selinux_test_suite kernel feature flag to allow
Starnix to take action based on whether or not it is running the
SELinux Test Suite.

Original-Bug: 364569559
Original-Bug: 392615438
Original-Bug: 398663315
Original-Bug: 404795644
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222964
Original-Revision: 47dd09f1bc5dd1f239d7588d761ee8ef25e9d20b
GitOrigin-RevId: ae0fdb9833293cd8b7556a6724d3d904ffc81470
Change-Id: Ie87b499eb3ae972ff323ffafeec9f5b45dc7b8ad