[roll] Roll fuchsia [assembly] Set release version during config generation.
This cleans up the logic that sets the "release_version" field in
board_configs and board_input_bundles. Build rules can set the values
explicitly in GN or Bazel targets, but the config generator tool will
set the default value "unversioned" if no value is provided.

This also allows targets to set these versions as a string directly in
the target, or as a path to a file somewhere in their workspace. This is
necessary to support some board definitions in other repositories.

Original-Bug: b/399808455
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238429
Original-Revision: 4077eaa2afb94144c002d6e1d8187a05a780c317
GitOrigin-RevId: 68542562eb57c8aaea65e3f730b3b7d8efc867a3
Change-Id: I92af35fc54e60e895f23353736da44f2c69ee911