[roll] Roll fuchsia Reland "[aml-thermal-s905d2g-legacy] Use SDK pdev library"
This is a reland of commit 3e0f29b0ce98bb02d172ad553dff557a57cffbaf

The original CL caused the smart display buildbots to fail because the
aml-thermal driver would crash if it failed to get the smc resource.
Before the original CL, the driver would ignore the fact that it could
not get an smc resource.

The original CL has been updated to not crash if it fails to get an smc
resource.

Original change's description:
> [aml-thermal-s905d2g-legacy] Use SDK pdev library
>
> Make the aml-thermal-s905d2g-legacy driver use the
> //sdk/lib/driver/platform-device/cpp SDK library in order to interact
> with the platform device.
>
> Original-Bug: 366517646
> Change-Id: I349d30acff517fe03a1471cf7c2da30c6baefe11
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239304
> Reviewed-by: Andres Oportus <andresoportus@google.com>
> Commit-Queue: Nick Eaton <nickeaton@google.com>

Original-Bug: 366517646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241564
Original-Revision: f4f4ebf865bce4eee8b395431e112e51f2cba018
GitOrigin-RevId: 3f75fa998e96d8ee7543df81a1d27d8c89e048a4
Change-Id: I36c26c2f100f3d59611f36af86640aac3e0567a0