[roll] Roll fuchsia [flash-kola] slight reboot message rewording
"Rebooting to latest bootloader" was causing some confusion as it made
it seem like the bootloader was always out-of-date, even if the dev
hasn't changed anything in their tree.

Removing the "latest" part to make it more clear this is just part of
the normal flashing process and doesn't have anything to do with
out-of-date images.

Original-Bug: b/407614253
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240886
Original-Revision: a6324899f6f2a30cff2688caf0786493362ce139
GitOrigin-RevId: 8b36b65063e3788fe92724b902f0ba74659fcd02
Change-Id: Iececc6ca94cbb968f5674dd51b5300d8e38bd841