[roll] Roll fuchsia [fdf][rust] Fix lock contention on task drop
When dropping the future of a task because the dispatcher is shutting
down, we need to make sure we aren't holding the lock on the currently
active future in case something tries to re-wake the task. Otherwise we
get a lock poisoning crash.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241765
Original-Revision: 65fef7c5785d2e2666dafd54dfa690002360756e
GitOrigin-RevId: c417731a88d6ef21050d0e2781fe3bbafdf0a315
Change-Id: Iae1f34e3ca5297b22e4291ce4548a059372fe425