[roll] Roll fuchsia [netstack3] Avoid debug-printing IP addresses
We're currently debug-printing any fuchsia.net.interfaces.admin/Control
request that we process, which, in the case of the AddAddress and
RemoveAddress methods, includes an IP address that will be formatted via
the Debug impl for [u8; 4] or [u8; 16] rather than pretty-printed. This
means it won't be redacted along with other IP addresses. Remove this
debug print since we already log in the processing for most of these
requests anyway.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242086
Original-Revision: a47e7af2283c0bbc977e3f1a2851e2cbb0a9258d
GitOrigin-RevId: 633243764aa8026b978e96eaaa6c105a0fbb2c6a
Change-Id: I9e517cce831fa1736755b308799d0d4485fab1b2