[roll] Roll antlion [owners] Remove dhobsd@google.com from packages/antlion/controllers/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/dhobsd@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1241747
Original-Revision: b0fb1e0bb7b1dda6f8d6b374f1f9599e4db4bb64
GitOrigin-RevId: 8b91c22dd7f74c15ec7fa7e095dae12bfd011528
Change-Id: I5a568fbe041390291513898aa26766fa5d745755