[roll] Roll fuchsia [sestarnix] Create all Fuchsia-space Anon nodes as private
Migrate all Anon nodes wrapping Fuchsia primitives (those with
internal names of the form "[fuchsia:...]", which do not align
directly with Linux equivalents) to be private, so that they are
un-labeled and not access-checked.

Labeling & checks will be re-introduced for this class of file node
as part of a broader review of Fuchsia-space resource treatment.

Original-Bug: 404739824, 404773987
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242104
Original-Revision: a1ff29af2300dfd951ee09e224805baec957bd9b
GitOrigin-RevId: 2afd4c9cb6d866950f8c0a453d53c65497f76909
Change-Id: I8ec142dba75a40b1be39c76afdf8237ca48e7263