[roll] Roll fuchsia [kernel] GenericWatchdot::Pet should check if dog is enabled
Lockup detector calls hw_watchdog_pet that boils down to
GenericWatchdog::Pet in this case.

This lead to some confusing experiments when bringing up the watchdog,
in which the watchdog would be left enabled, but tell the kernel it was
disabled to observe the kernel rebooting after the bite time.

Original-Bug: 407841192
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237564
Original-Revision: 65f5537e42411c16aa9f2d25786ad09f7c6739d0
GitOrigin-RevId: 275db3b94c8f2e200f3c92e99a7e2ae32306858a
Change-Id: Iabe43a250f9138d0fdd8c1694bbcfdac185eaadd