[roll] Roll fuchsia [kernel] Rename ktrace-tests to ktrace-internal-tests
The ktrace tests today actually test the internal KTraceState
implementation. This rename reflects that behavior, and also prepares
us to introduce new tests that test the KTrace singleton's
functionality directly. This will be useful when we start to
implement per-CPU buffers, which will not utilize the internal
KTraceState object.

This is a pure rename and is therefore a functional no-op.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243126
Original-Revision: 50048c3b13850e90634af56bbe3e1b04aad3803d
GitOrigin-RevId: 106692dc0818047f3170af6e184cb42b285c910e
Change-Id: I55955aa0731fe6845d699170ec1b950c545f7d65