[roll] Roll fuchsia [starnix][iptables] Implement the MARK target
This is gated behind a starnix feature, which is not enabled in this
change. This change adds a unit test to test the iptables parsing
correctness. A follow-up change will enable the feature for test
containers and add an integration test.

Original-Bug: 398328518
Test: starnix-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238847
Original-Revision: 5d35795a5647b7ae96f6fd9f1c9505eea0f74ae6
GitOrigin-RevId: bd89475f90bbf1755340f39dcb59e2246d36616b
Change-Id: I4713d402cebac2977d82c7851e3fd26ea90d8843