[roll] Roll fuchsia [kernel] Add Reserve method to KTrace
This CL adds a Reserve method to KTrace that wraps the pre-existing
internal::KTraceState::Reserve method. This means KTrace now
implements the fxt::Writer protocol, and the static instance can
now be passed in to the Emit* functions directly.

This is a functional no-op, but will make it easier to add support
for per-CPU buffers in the future.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243284
Original-Revision: 94ae7e14929e8382b757e5969079aa76a937b9ef
GitOrigin-RevId: dfff6e76a23d1ae20b104bafa0daefef64f85030
Change-Id: Ide8c65e450d980af865ce5a597f3f6c478b4fe31