[roll] Roll fuchsia [kernel][vm] Unlocked FreePages
Removes the lock requirement on calling FreePages as it is not actually
needed and existed as a 'just in case' for future changes, but none
have yet needed to leverage this being locked. As the lock might be
dropped in between a page being removed from the page_list_ and
actually freed, requiring that the lock be re-acquired just to perform
the free does not provide any value.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239006
Original-Revision: 5168b06ac6e9faaada2ec832005f8efde8ed7d74
GitOrigin-RevId: 7fdc7b822c2e3296d02ddcf458f1842c45be52ca
Change-Id: I3d957fea20dcb46e7dbd547a8d4b5375ade4d8da