[roll] Roll fuchsia [ffx][repo] Fixing handling no target device
This fixes the early validation that there is a target device
when starting the package server without the `--no-device` option.

Some changes to the structure of the ffx_command_error::Error data
made it impossible to go from a fho::Error to a FfxTargetError in
order to determine the cause.

This change updates the FfxError type to use the underlying FfxTargetError
as the "source" of the error, which allows accessing it and downcast it
correctly.

Original-Bug: 407770719
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243884
Original-Revision: 197b9542179e7a7f26aed78976127f058e9ac0c1
GitOrigin-RevId: 1bc7f47f7ac5c595aef67a5e63dcd26fd7d148d0
Change-Id: I7dfbd5462bba5efe58b96e4992961ad578b3d627