[roll] Roll fuchsia [assembly] Add release version numbers to Product configs.
Product configs already include version information in the form
of a file.

For consistency, and to remove the need to read that file during
bisection, we copy the contents of that file to the top level of the
"product" section of the config json.

Original-Bug: b/399808455
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238666
Original-Revision: 5cd854452c3256c45913a4fd1960a9d3c92e8b22
GitOrigin-RevId: 4ce782edcd4b4fd2033cdbc519f2ac50e9595837
Change-Id: I47493e65d98668b568dff758a7ce78388fb9708e