[roll] Roll fuchsia [starnix] Have WRITE permission for set_content_size
Modifying the content size is a modification to the VMO and is supposed
to require the WRITE permission, as per RFC-0238. This changes the
SLICE, which is not writable, to a writable clone, which will allow the
kernel to begin enforcing the permission.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241186
Original-Revision: 14e02a9408a8894d5c53646e9b6cba4ea600e40f
GitOrigin-RevId: 9bcb0651c56c29c3ebd0e9964d924ec01e51d922
Change-Id: I8a429583acf08742931c5c9b3b11a156fdf9136a