[roll] Roll fuchsia [realm_builder] fix race in realm builder event stream test
The realm builder test for event streams started two eager components,
with one connecting to an event stream and expecting to see a start
event for the other. As these components start up at the same point,
sometimes the former would start before the latter could connect to its
event stream. In this case, we fully expect the second component to miss
the event, as it happened before the event stream was opened.

This commit fixes this race by making the first component lazy and
having it be started by the second component only once the event stream
is connected to.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243844
Original-Revision: 98c6a02ff1633d1e1f36f2713de45225097ca07e
GitOrigin-RevId: 5e777b9ad8af62e117017b6263dcec5841412a1b
Change-Id: Ia0dd0a63ebcb981096a0e55afedddbc1824cae53