[roll] Roll fuchsia [netstack3] Prepare congestion module for SACK
Rework the congestion module to operate based on the SACK scoreboard to
estimate how many bytes are in flight, which falls back to the same
estimation when there are no SACK blocks.

We delegate the decision of which segment to send to the congestion
module, along with calculating the congestion limit and mss.

While I'm here, rework a bit the counter semantics so they better
reflect the retransmission cases.

Introduce a test that verifies that congestion control is in fact
limiting the amount of data sent per RTT by evaluating the average
flight size when not in slow start.

Original-Bug: 42078221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239886
Original-Revision: a3c11088affb540dab6e726a08b9b83d864fb84f
GitOrigin-RevId: 72c9de7dc7d8db2c5e2b915ad2e1b475f9b96bf5
Change-Id: Ibbd3281920f144d322ff8e84c3dde857c884546b