[roll] Roll fuchsia [sestarnix] Rename the object class and permission types
Kernel-specific class enum is now KernelClass, and the kernel-
specific permission enum KernelPermission.

The kernel-or-other enum is renamed to simple ObjectClass, to
reflect that it refers to a policy class that may or may not be
one of which the kernel is aware.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245284
Original-Revision: 5319f924f7a10c6bbf5699c20e7ae1d0130486db
GitOrigin-RevId: 1758e74a87838cbd599a5edcfa348f62301f9f09
Change-Id: I774fc08fc7863d0ee68463352900747e9cf058fc