[roll] Roll fuchsia [ffx] Remove hardcoded sdk/manifest/core
This removes the reliance on the assumed existence of sdk/manifes/core in-tree.

The IDK/SDK building is being improved, and the tools used within Fuchsia need to be
flexible enough to allow the IDK to be moved, and most tool executions not require the
IDK manifest to find other ffx- subtools, and the related host-tools.

Original-Bug: 397989792

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221424
Original-Revision: b37256c8a7ae1c0242ad55e562669069afaec29b
GitOrigin-RevId: e772c02a0e7a1bd1d65fd1012f824aced0bb4fe6
Change-Id: I7d7913781eba4bd6d9d7c4e615c0b55e19b6f039