[roll] Roll fuchsia [starnix] Always decrement counter when timer is cancelled
Previously, when we were using an event, the timer that was cancelling
the outstanding timer would clear the event signal for both the new and
the old timer.

Since we are now using a counter, each message has to be acknowledged
with an explicit decrement.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245345
Original-Revision: dbcd95a37e0bd446b6fb58f943882c58154e96c2
GitOrigin-RevId: d326aa95bcec0154edf9e63a4d2da3f175a994f4
Change-Id: I050110572ba47ae556b8738f891e8423022395ee