[roll] Roll fuchsia [build][python] Fix typing issues in assembly modules.
This removes typing issues in the module files, but also
in some of the scripts that use them.

Note that the @serialize_json decorator is *not* compatible with
mypy (as it breaks static analysis), so is no longer used in
the definitions of ImageAssemblyConfig and AssemblyInputBundle.
A comment has been added to warn about it usage in serialization.py.

+ Remove broken commands in assembly_input_bundle_tool.py
  which relied on non-existent methods in AssemblyInputBundle
  and thus would fail at runtime ("diff", "intersect" and
  "find_blob").

Original-Bug: 366268742
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241526
Original-Revision: 98c7dcda4dbe4871e3ec1ac931795bf748b7d2b2
GitOrigin-RevId: 126ee4b807cd863dd1d179db52c022615247352c
Change-Id: I9a9701274c773d241a128f3e281b4bd0b3401684