[roll] Roll fuchsia [kernel] Change ktrace categories to use acq/rel semantics
It turns out that my reasoning for making the load and store of the
categories bitmask relaxed in https://fxrev.dev/1243384 was wrong.
Specifically, the order in which we set the categories bitmask is
important, because we want arbitrary trace records to enter the
ktrace buffer only after we've finished emitting the thread and
process metadata.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1248026
Original-Revision: aa5a64ea8b2d50242cf74f47ba9d955d8019007a
GitOrigin-RevId: dcd3061b2481368889eb77389b0d72e2fa916c14
Change-Id: I9367220526b9c0c4d5de870f08d0098e23f7f1e4