[roll] Roll fuchsia [vfs] Remove dependencies on the rust vfs
The rust vfs was converted to a dylib which caused performance
regressions in fxfs. Ideally, fxfs should only depend on the vfs
statically. Fxfs depends on severals crates which depend on
fuchsia_component which depends on the vfs dylib. fuchsia_component
only depends on the vfs for ServiceFs which none of the crates that fxfs
depends on make use of. This change switches fxfs' dependencies to only
depend on the client portion of fuchsia_component
(fuchsia_component_client).

Original-Bug: 396172331
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235845
Original-Revision: ba7171e3aa8e512443793dc988742d3c808991ec
GitOrigin-RevId: eaf656a19ead9d9625f747e59dc4aeb398f7b9ab
Change-Id: I19a01728381d9f08b6cfcb5915c5eb64fcb01d0a