[roll] Roll fuchsia [display][fidl] Remove VirtconMode.INACTIVE.
INACTIVE is the sharpest edge in the list of members. This CL removes
the member, and switches the default VirtconMode from INACTIVE to
FALLBACK, which is the most intuitive mental model.

The attached bug describes the long-term plan for removing VirtconMode
altogether.

Original-Bug: 408475481
Test: fx test //src/graphics/display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/689987
Original-Revision: 77e39397a334ba76e398fe50288ede74b61e85fc
GitOrigin-RevId: e5c2e00d4c49cff5ba9d02e8fa443e8f16f8e0e7
Change-Id: I54bee175da567dbc9bd41a45b2b756b3aae518e2