[roll] Roll fuchsia [starnix] Send attribution update on process name changes
With this change, Starnix will send a memory attribution update when the
name of one of its Linux processes changes.

This change also refactors some of the memory attribution code. In
particular, the memory attribution module is event-driven only, and no
longer wakes up every 5 seconds to scan the list of processes.

This is a reland of fxr/1217147. fxr/1217147 held the PidTable lock for
the whole duration of the event processing loop, including the sleep,
which was incorrect and degraded performance.

BUG=393078902

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239525
Original-Revision: c0f99d72002a7224864968d743e7a654e59397fe
GitOrigin-RevId: d2f2f3be02bc55b9fc311d57b1051b997f621445
Change-Id: Iaf75fe8f26c444b05d10690ea200e87d63cad0ba