[roll] Roll fuchsia [dma-buffer] enable uncached contiguous buffers
Zircon has supported multi-page contiguous uncached VMOs for several
years, but the dma-buffers BufferFactory did not provide access to the
functionality.

This change adds a bool, much like on BufferFactory::CreatePaged, to
enable or disable caching on new ContiguousBuffers. As this is a
signature change, all callers are updated to call with caching enabled
-- the prior default.

Original-Bug: 409561709
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249004
Original-Revision: 534355df1168bfa8bd2bc937fef4a6210ec04010
GitOrigin-RevId: f392551dbe33146d332850dd76c4232041b32daf
Change-Id: Ia5dbc5dc5d09aa8465a1448704d6c326efba19e2