[roll] Roll fuchsia [bazel] Fix bazel_action.py query cache key computation.
Use the content of input starlark files used for cqueries to compute
the key used for the query cache. Without this, modifying these files
would end up with stale cache entries being used in the next incremental
build.

This is normally not a problem for most incremental builds, because
any change to build files would re-run `fx gen` which wipes the
cache entirely, but a change that only changes the starlark files
would trigger this bug, so fix it there, as it was discovered
during local development.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249944
Original-Revision: 86377dc79f48ccf1ca479f35a6767218ce549e62
GitOrigin-RevId: 4b44817b7d616366900a21049385a4b7b30b6be9
Change-Id: Iae629e3d540d5fd53e7ee434501442f61283a85b