[roll] Roll fuchsia [ffx][flash] fix flashing sparse images
Now detect if an image in the product bundle
is already a sparse image and is greater than the
max download size of the Target. If that is the case,
expand the image to a temp file and then re-sparse
the image to a new set of sparse images, then flash
the re-sparsed images.

Original-Bug: b/390670098
Original-Bug: b/394699849
Original-Fixed: b/390670098
Original-Fixed: b/394699849

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243987
Original-Revision: 0471baa8d7f7da518fcf5e44a8e1538028992b7b
GitOrigin-RevId: 87c22b1b6f9444a9371fce4657b0c67d39b5acdd
Change-Id: If2ecc01e1381e97e6c58b4bac38c25584ca71927