[roll] Roll fuchsia [ffx] Populate ssh host address even without an ssh host pipe
We need to know the address of the host from the target's perspective to
make package management work. Usually we get this as a side effect of
establishing an SSH host pipe, but with USB or VSOCK we don't actually
do that, leading to a surprising breakage.

Now we have a separate path where we can SSH in *just* to get the host
address without starting a pipe. We should only hit this path when we
don't have an SSH host pipe so existing workflows shouldn't change.

Original-Fixed: 1250227
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250227
Original-Revision: 1dabfe9a3c97fd35c217b78cbe247b7dd1e582a7
GitOrigin-RevId: 3d2638e85c1d5446ca25373b4a388ab1a2e283f0
Change-Id: I026ddbfcedc8362f63cb27f5bf2d16472bf32b34