[roll] Roll fuchsia [kernel][vm] PageLookup uses LockedPtr
PageLookup uses a LockedPtr for the owner instead of a raw VmCowPages*.
FindPageContentLocked takes the required ancestor lock and populates
LockedPtr if an ancestor owns the page. Otherwise the LockedPtr is left
unset and the caller can assume that |this| is the page owner, which it
already holds locked.

LookupCursor functions change accordingly to use this new LockedPtr. One
consequence is that the owner() helper is no longer valid and is
removed.

Original-Bug: 338300943
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249114
Original-Revision: a2b74540dc59a2f472811f4c902a8b20064d8931
GitOrigin-RevId: 940a1e56ded5de451ff63015c0d19efbebd51d7a
Change-Id: I3202c2eed041cf8d933eb50996c4343e3db7353a