[roll] Roll fuchsia [audio][test] Make drivers tests 25% faster
In AudioDriverTests, at the end of each FIDL exchange with a
driver (i.e. when the test case is complete), we always wait
for additional FIDL activity such as an unexpected callback,
or a disconnect. Rather than just "waiting for something bad
to NOT happen", we can be proactive and use a simple "call &
response" to ensure that the driver is still responsive over
FIDL - we use GetHealthState for this purpose.

By moving to this, eliminating our 100ms wait per test case,
this shortens the execution time for AudioDriverTests suites
to 75% of the previous run-time.  \o/  yay!

Original-Bug: 370741565

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251224
Original-Revision: 13ffa86dade8049b1fc3a80a856e84ab701f1ec9
GitOrigin-RevId: 94c20b15fad1d85ee3501b744f0ee6d6f86d3cd9
Change-Id: I43c8a87e3d724d9534f1a9cc5053e6484e1563e7