[roll] Roll fuchsia [audio] ADR eager-start config (1/3)
Add a platform configuration flag to specify whether the
AudioDeviceRegistry service should be eagerly started.

A flag is being added to the platform configuration schema
for audio. Because out-of-tree product definitions include
settings for audio, a soft transition is required. This CL
is the first of three that are ultimately needed.

Original-Bug: 390573041
Original-Bug: 410822855

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228525
Original-Revision: 31590d1a9a80ada1f32fbf4d9a02527a13cc89c3
GitOrigin-RevId: 34f868cb6fe4d4af89c164b010a8cdde8afcb6e0
Change-Id: Ieacda01434a474046970b54073491ddfdb995034