[roll] Roll fuchsia [bazel] Fix `fx bazel-tool set_gn_targets <label>`
This tool command was broken by a previous CL [1] that added new
fields in the bazel_build_actions metadata of bazel_action() GN
targets, which broke gn_target_utils.py, invoked at runtime.

This was undetected because this is a development tool that is
only used by the build team, and never runs at build or test
time, ether locally or in infra.

This fixes the situation by:

- Adding dataclass field definitions for the extra metadata keys
  in gn_target_utils.py

- Documenting the new keys in //BUILD.gn

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1220864

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253186
Original-Revision: 749d5113313e844b976737ace51a42867d1992a1
GitOrigin-RevId: 72bc0812fdcae8f622c514e2e54c99275ff919b6
Change-Id: If64ae3c9ba5221905c65802a1c6535a1a889a0c0