[roll] Roll fuchsia [kernel][arm64] Change arm64 spinlocks to be 32bit
This matches the size of the other two architectures, which use
cpu_num_t (uint32_t) as the basic spinlock type.

Was previously using unsigned long for historical reasons, to make it
match the size of the architecture when building with arm32 and arm64
at the same time.

No functional change except for the size of the spinlock and any
alignment or size changes this may entail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251013
Original-Revision: e05da63dc6e15ab21a8d6ebef9717258ea244fac
GitOrigin-RevId: 0dd7f3b2e201e7db3258242a1a0b1c64d07ffa97
Change-Id: I0d76a89319151eac11368a8669f143f9b5230651