[roll] Roll fuchsia [bt] Add one-to-one matcher for hanging get.
This change adds a library struct to match pairs of things that can
become available in any order.  Clients can enqueue to the left or the
right queue, and when there is an item available in each queue, a stream
will yield a result.

Tests: fx test one-to-one-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926955
Original-Revision: 2ca0c56d5510a0c95d4ed48d6b39957907144e1e
GitOrigin-RevId: efaa413d70d7b7d6f81e7a3a260cb1b8f0e9a79f
Change-Id: I5bfa6997619fcb7ca4b0598a13c70618b20ca924