[roll] Roll fuchsia [ffx][flash] Improve flashing speed
We used to use a buffer which is 4KB, this is proven to be a bottleneck
of flashing performance. Now we change it to read file into a 50MB
buffer, and let the TCPStream to handle the chunking. From local
testing, the chunk sizes is about 1.6MB on average. From local testing
with pontis, this decrease the flashing time of large image from 279
seconds to 75 seconds.

Also we noticed the set_active call can fail on sorrel, we change it to
be best effort and log a warning when fail. Also added a 3 seconds
timeout on set_active call, as this call is not computationally
expensive and should response pretty fast when succeed.

Original-Bug: 409544982
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1252985
Original-Revision: 84b8c6436fb0c9b5e16ddcedcbc38af413fc59f8
GitOrigin-RevId: 77a6b7444266b11266f3b75887f809041efe6a71
Change-Id: I0e3983bb2a7e5290d5974fda559781c9b2f574a1