[roll] Roll fuchsia [starnix] Mark sockets when feature enabled
Before this change, starnix has a local state for SO_MARK which does
not affect routing. In this change we want SO_MARK to be used by the
netstack, but put it behind the feature flag so that there are no
user visible behavior changes in routing.

The feature is enabled for the debian container so that syscall tests
can test against the change.

Original-Bug: 410576727
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238848
Original-Revision: 34763bc37fe31f64b5e7df27d6daac70f8de466b
GitOrigin-RevId: 0e952da9e94a4119be2e42eac72f6c365054481b
Change-Id: I9500c2e4a198f36807c2249c457b73d503b5671c