[roll] Roll fuchsia [update-rustc-third-party] Download cargo-bazel
... instead of building it with crate_universe bootstrap. The bootstrap
process creates a new bootstrap repository and does a full rebuild of
cargo-bazel every time, which can take around a minute of even longer.
This change switches to download this binary from upstream release with
a matching version. As of version 0.59.1 today, the binary is < 20MB,
so should be pretty fast to download.

Note Mac support is included in this change but not tested.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253005
Original-Revision: 907638c9a5fff09b4b3bc56237b9503b3d62cc3f
GitOrigin-RevId: 10659acba3f370a64d9d7b596f21950bc1709c4d
Change-Id: I06d3b83ae15bf8033c7d855159154e5a31f7bf8e