[roll] Roll infra/recipes [coverage] Never skip uploading in CQ
Uploading multiple times should only be an issue for absolute coverage,
which is only uploaded in CI. There's no need to query for past builds
with the same fuchsia revision in CQ.

Also reduce the limit for querying past builds. It's very likely that if
there is a build with the same commit it will be the most recent one,
but we query a few more just in case.

Original-Bug: 411464876
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1255585
Original-Revision: b9125ca1c1986ca5bdc06f907387134359a022cb
GitOrigin-RevId: 886f30d3bbe6f45c789d20ffb7c9146e1dc03f35
Change-Id: I294ddfbc5615a166d4e1fb64c45bae747d4e6e0c