[roll] Roll fuchsia [zx] Expose signals from zx_object_wait.
Currently, wait and handle_wait in the zx crate only return Signals when
the syscall result is ZX_OK. However, the docs make it clear that the
result is defined when the status is ZX_OK, ZX_ERR_TIMED_OUT, or
ZX_ERR_CANCELED.

ZX_ERR_TIMED_OUT is useful in practice: it's common to wait with an
INFINITE_PAST deadline in order to see the signals currently asserted on
an object.

ZX_ERR_CANCELED is less likely to be useful, but since it's defined
according to the docs, we should probably expose it at this layer.

Original-Bug: 410869980

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253204
Owners-Override: James Robinson <jamesr@google.com>
Original-Revision: 2e2af48a7c86ec8067e58efdfdd1664e8671c3ad
GitOrigin-RevId: 27908b90b606996151e2462dc651e7da42f72073
Change-Id: Iff9e0997a3e0ccafe3412fc60aa73177fee39628