[roll] Roll fuchsia [kernel][heap] Trace how long growing the heap takes
This CL adds a trace duration that captures how long it took to grow
the heap during an allocation and how many retries it took to do so.

It also slightly refactors the calculation of `growby` to make it
easier to understand.

Original-Bug: 399427107
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253439
Original-Revision: a2bddada8e0cdfc4103f4f09f96865c2d90b1430
GitOrigin-RevId: 1c0e6654eb960cc616791548d31ba40e1edc5d65
Change-Id: I4f63001881c4a7fac6e30463660ded3e4fdfd29c