[roll] Roll fuchsia [audio][test] negative tests for driver SetGain
Restructure the StreamConfig gain test cases, and add a set
of negative tests for the SetGain method.

After these changes, audio_driver_tests validates that a
`SetGain` call is entirely ignored (no gain notification) if
it contains an invalid gain_db (infinity, nan, out-of-range)
or mute/AGC (setting these when the device does not have the
capability). It also verifies that if a field is absent from
the `SetGain` call, this is treated as no-change.

Additionally, we enhance existing `SetGain` cases to ensure
that values received in subsequent gain-change notification
matches those which went into `SetGain`.

This incrementally unblocks enabling AudioDriverTests on
additional CQ builders, to increase driver test coverage.

This should be our last enhancement of AudioDriverTests test
cases for StreamConfig (or Dai or Codec) drivers, as these
driver types are deprecated -- all investments should be
for the Composite driver type. In fact this CL is only being
landed for completeness (along with a large set of other
changes to AudioDriverTests), as it was previously-developed
and forgotten.

Original-Bug: 370741565
Original-Bug: 309860869
Original-Bug: 313655758
Original-Bug: 313655706

Test: fx test -do audio_driver_basic_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957396
Original-Revision: a40ce1691fe7e015030f7f9d52d115ded17a1a25
GitOrigin-RevId: 333e17abfaa9584be55475dc9d663c27fd27066c
Change-Id: I58cbde49be56103587f31cefb5b83b224f45b2be