[roll] Roll fuchsia [kernel] Implement KTrace::ReportMetadata for per-CPU mode
Implements the ReportMetadata method for KTrace when running in per-CPU
mode. This is a combination of the ReportStaticNames and
ReportThreadProcessNames methods in internal::KTraceState.

This CL also updates the tests to verify that the metadata is emitted
exactly once when tracing is started. This required overriding the
ReportMetadata method in a derived class because the thread and process
names are written into the global singleton's trace buffer, which we do
not want to manipulate in tests.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250804
Original-Revision: 77bf3682902e8f99d2f7bb049354209b31bf4801
GitOrigin-RevId: d2331a27b7f937898ba98949948b009cf57e9293
Change-Id: I53cfca6ea5ca540cdb446253dfcdaa78cee57aea