[roll] Roll fuchsia [memory] Use a value struct for zircon resource names.
A Rust string instance is made of a struct and a heap allocation.
It uses respectively 3 ptrs for the struct, and 2 ptrs for the heap
block header, for a total of 40 bytes.

Hence it is always more efficient to use ZXName of 32 bytes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1254306
Original-Revision: c7ab779e3cd1db7d61184a49026edc1212c97a5c
GitOrigin-RevId: 0d05085eef10070946c56da0eb1960edbbf42164
Change-Id: I2c73af9bdfd7f267b22f1fcdc7c5c16019f86fab