[roll] Roll fuchsia [fidl][rust] Moveable decoded wire values
This change introduces the `Decoded` type as a variant of `Owned` which
does own the backing memory, and provides methods on `DecoderExt` to
decode into a `Decoded`. To that end, this change also introduces the
`commit` method on `Decoder` which allows resources to be dropped even
when decoding fails part of the way through.

The API usage examples and benchmark have been updated to reflect the
new decoding API.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253165
Original-Revision: 707a88f98e1e431a0a6a5362a7ae7787e1334be1
GitOrigin-RevId: 5ee9db6e5976e7179c207a3dcaead17af742a18d
Change-Id: If5b0f80b8639ea8dd00c57f06b489de7bdbfcda9