[roll] Roll infra/recipes [release] Handle fuchsia.git as an <import>
http://tqr/979232 switched fuchsia.git's Jiri manifest entry in
integration.git from a <project> to an <import>. Now the <project> entry
for fuchsia.git is actually in fuchsia.git, but the pinned revision is
still attached to the <import> in integration.git.

The release operation recipes previously assumed that every repository's
pinned revision was attached to a <project>. Now that that is no longer
the case, the release operations need to be able to handle fuchsia.git
in its new state.

This is very much a quick-and-dirty hack, long-term we should figure out
a way to make this more generic and avoid hardcoding fuchsia.git's
project name and manifest path.

There's also going to be more work required to make these recipes work
with projects that are pinned in fuchsia.git.

Original-Bug: 385378627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1258664
Original-Revision: efb8df58b7efc21182cbfe94991f4bae36c3e2ee
GitOrigin-RevId: 8bf502f56f19e791586e96d1b549d878498e8c09
Change-Id: I40ba436b027929f8ea2ca53ebd2bd991ea42e3fe