[roll] Roll fuchsia [starnix] Promote the Socket's FsNode to a strong reference
Maintaining a weak reference to the Socket's FsNode is unnecessary
since the only reference from FsNode to Socket is from the file-
system node to which a Unix-domain Socket is bound, to the Socket,
and in that case the FsNode differs from the one associated with
the Socket itself.

Holding a strong reference avoids raciness between referencing the
FsNode of a target Socket, and it being torn-down by its owner.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261264
Original-Revision: ade556b77559b725156f8d35c13c37ee12b85b88
GitOrigin-RevId: 60f686da0ea90d2f01f244a7c16c771d224b5279
Change-Id: I45776ca44d5993ece43f2aea3082f34f322d6167