[roll] Roll fuchsia [idk][build] api_level is always a string in IDK metadata
https://fxrev.dev/1082028 added support for strings while keeping
numbered API levels as integers. Based on experience with this in an
SDK, it will be simpler if the level is always a string. The value is
mostly used to build paths, so this shouldn't be a problem. Anything
that needs the integer value of a level should get them from
version_history.json.

Original-Bug: 326277078
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087174
Original-Revision: c168823f9d8cbffe4fdcc8ef15863c9be3310896
GitOrigin-RevId: 654fcd937348b005c0266493f5f48327d72cadae
Change-Id: Ie5ab4c06e83696c1bb20a9b088bb68628e17c22b