[roll] Roll fuchsia [fxfs] Fixed duplicate tombstone issue
Prior to this CL, if you were to create an unnamed temporary file, and
then close it without linking it anywhere, some time later when the
timed flushing occurs, the file would get tombstoned for a second time.

The root cause if this issue is that prior to this CL, we would allow an
open count to be taken on a file that had already been purged if it was
an unnamed temporary file.

The fix is to disallow this, but in order to do so, we need to change
how the file is created.

This also fixes an issue where we errnoneously took the truncate guard
when queueing a tombstone, instead of when we actually executed the
tombstone.

Original-Bug: 412830759
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259844
Original-Revision: ea33213a53100617a356a6820860580f5e42ce01
GitOrigin-RevId: c5b4d70063ce3fac499c48d21b8daf4a37820bed
Change-Id: Ib4a74f8b655d4f8fabbb5a340fba835d4b23b419