[roll] Roll infra/recipes [android_artifact_downloader] Add support for "source" artifacts
We currently lack a means to specify LICENSE files in the CIPD packages
we make with this recipe. Not a build artifact per se, but important
still. This change introduces the basic generalization of a "source"
artifact that can be specified like the other specifications: it
indicates a project and it is in the revision of that project associated
with the build in which the source will be located.

We keep this flexible as there might be multiple different license files
with differing names that we may want to install in different locations.

Original-Bug: 432538327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1338856
Original-Revision: 442fecbfd1eb67dba34cef1ec0ba246b4be4a8d9
GitOrigin-RevId: 5f93f5ce52b2d8b4c5cc5b9a691ec7bb71d5e420
Change-Id: I49a6b235e758b3dab0978ff0f694242ad6d0878e