[roll] Roll infra/recipes [android_artifact_downloader][reland] Add support for "source" artifacts
We currently lack a means to specify LICENSE files in the CIPD packages
we make with this recipe. Not a build artifact per se, but important
still. This change introduces the basic generalization of a "source"
artifact that can be specified like the other specifications: it
indicates a project and it is in the revision of that project associated
with the build in which the source will be located.

We keep this flexible as there might be multiple different license files
with differing names that we may want to install in different locations.

This is a reland of fxrev.dev/1338856. The original version featured a
typo that led to non-executable files having the wrong upload paths.

Original-Bug: 432538327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1340827
Original-Revision: 0bede47a9dbd64b1a8e356a1a8e934322086e7c7
GitOrigin-RevId: c009bdf36dfa79db88fead5269cc7f642182bbb9
Change-Id: I3c5935b239c5b53836871990e283bbf4204b12a1