[roll] Roll infra/recipes [trigger] Fixed a bug in toolchain triggering logic
Before this patch, when a single revision was built on a builder
for multiple times, only the oldest builder status was preserve.
The revision will not be considered as green if there was a failed
build in the past. After this patch, the revision is considered as
green as long as there was a green build in the past. The original
behavior is the culprit of current rust stage0 builder outage.

Original-Bug: 437392924
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1341865
Original-Revision: e12ae1e3c9b90f9e16273dc8b012f594f301d7d4
GitOrigin-RevId: a48d5c8b2712a525e7aef4c368d54971cfa14ef3
Change-Id: I178a6f0a127d4bafea380bab92fd6c319cb99d93