[roll] Roll infra/recipes [crosvm] Just build ToT
In this change we abandon the building of a fixed revision. That seemed
prudent initially, but now we indeed want ToT, as the only compatibility
guarantee we really have is ToT crosvm aims to be compatible with ToT
pvmfw on Android's git_main branch. (Also crosvm aims to be stable on
ToT main.)

Original-Bug: 432538327
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1342836
Original-Revision: 006c3c331987cc3603d4b1bd6e257dc4326d263f
GitOrigin-RevId: b118520c1c6bfde4b38b2b27bb60790985166f43
Change-Id: I7c57e73ad61219d3a94acd481b5130c4f7af0aaf