[roll] Roll antlion [wlan] Fix errors on unreferenced build targets
Some of the build targets are not referenced and thus don't have type
checking and gn applied to them. This changes fixes those errors.

The build targets are still not in CQ though, so this is a temporary
patch, primarily targeted at ensuring the correctness of the migration
of antlion. The migration includes locally building all targets in
antlion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1353724
Original-Revision: 6e5af05f770788cdfb4a30716cbb43d913e12023
GitOrigin-RevId: c68367e9394d1f8d5c042c8a440b2d3b5d93b0d9
Change-Id: I045a553829eb00f19668af13681323f474f6cb3c