[roll] Roll infra/recipes [coverage] Provide clearer error message.
The message about coverage builders not rebasing on HEAD can be
misleading to look like the real cause of the build failures. Instead,
having the actual fail reason come first will probably make it clearer
that that is the real failure and the rebase message is just a
suggestion.

Original-Fixed: 433961609
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1353872
Original-Revision: 46a60ea24796887e6b6162a400e02ee010a0d425
GitOrigin-RevId: 3b9e4660b28b3004f2027a7af6767e126ec0299e
Change-Id: I4ae89a161672f0802d5c27a020542c444dc734b3