[roll] Roll infra/recipes [crosvm] Ensure use of sysroot
We were not providing our build sysroot explicitly to cargo, resulting
in it linking with the system sysroot, and possibly also compiling with
the latter in build scripts as well. We fix that in this change.

This ultimately resulted in sysroot version mismatches at runtime.

Original-Bug: 432535854
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1357764
Original-Revision: 5f0231289f9f22056df3b3458d9589a1322a058d
GitOrigin-RevId: 54357f5d59e1b0b6cfce87b7c3f65c6ae99fda1c
Change-Id: I62698f81d81f11056a33398677ebb5479096447d