[roll] Roll infra/recipes [checkout] Skip clean check on Mac
See http://b/442754664#comment10 for full context on the failure mode
that this works around. We're ok with disabling this check on Mac
because support for checking out and building Fuchsia on Mac is being
dropped shortly, and in the meantime there are very few Fuchsia
engineers using Mac and so a low risk of bad things happening if one
repo is considered dirty on Mac even after a fresh checkout.

Original-Fixed: 442754664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1367417
Original-Revision: 59a2a908be0eb21c4436b837c8262b2feb4ea4d6
GitOrigin-RevId: cbbd024ab6524309824bb3c43f34a12a7c7e925c
Change-Id: Ie315f639781bc2b3e7c346a1f6e7b2e6499bee61