[roll] Roll infra/recipes [checkout] Use current_revision instead of current_revision_number
`current_revision_number` sometimes points to a revision that's not
actually the latest, if change details get retrieved shortly after a new
patchset gets uploaded. For example, here is the output of a change
details request for http://tqr/1038010, made exactly 1 second after
patchset 2 was uploaded: https://paste.googleplex.com/5742242425470976

Relevant parts:

    "current_revision": "aeb3c7415f2eef334cd5904e0f96e94826e1838c",
    "current_revision_number": 1,
    "revisions": {
      "7badc9b6881900ddf73c7ff1f219d26283150c3b": {
        "_number": 1,
        "commit": {
          "author": {},
          "committer": {}
        }
      },
      "aeb3c7415f2eef334cd5904e0f96e94826e1838c": {
        "_number": 2,
        "commit": {
          "author": {...},
          "committer": {...},
          "subject": "...",
          "message": "..."
        }
      }

Weirdly, the `current_revision` and `current_revision_number` fields
actually point to different revisions. I imagine this is due to some
sort of race condition in the Gerrit backend.

This is an issue because we use the `CURRENT_COMMIT` query option in the
request to get the full commit message for only the current revision and
not for any other revisions. Since the `current_revision_number` field
points to a non-latest revision, if we use that field to determine which
revision to use then we won't be able to get the message, which
currently causes an exception:
https://screenshot.googleplex.com/6K3qMVG7swZpQEV

As a workaround, use `current_revision` which seems to be more correct
and up-to-date.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1367416
Original-Revision: 883c67204b03caf7771c4ae3e7f7910c7adce2b1
GitOrigin-RevId: b85a80eef72964fd1e77c3ede081903670242313
Change-Id: I9f0c61e2837a760b0c706b08d0054d608c4854e8