[roll] Roll infra/recipes [fuchsia/fuchsia] Run tests in parallel
If this build is a CQ retry and is able to skip building fuchsia, run
tests on the previous child build and run tests on a new build of
fuchsia in parallel. Should all the tests ran on the previous child
build pass, the tests running on the new build of fuchsia can be
canceled, and the passing test results can be used. Should any test ran on the previous child build fail, continue
running the tests on the new build of fuchsia and use the resulting test
results.

Original-Bug: b/394863496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1367296
Original-Revision: 3b88aa70c651224306b6ce141e7785d58a00b365
GitOrigin-RevId: 572f483b058dca612600e8726f42e2406939c335
Change-Id: I80a60ec436b6675ab022b3ef3e538729ede437a7