[roll] Roll infra/recipes [rust_toolchain] Fix rust repo fetch problem.
Rust build rely on git commit history information. We have a
'git fetch' step to fetch every refs from remote after checking out
the rust repo. However, the refs for the submodules are not necessary
and they are causing problems when some submodule refs are not uploaded
to upstream. This patch disables the submodules in this step to mitigate
this problem.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1370177
Original-Revision: 8188e9e9e1797b3656e36ddccd3d12db16e40fb5
GitOrigin-RevId: 0671ccb68a83d9e4044174b861fbecfd11bb547d
Change-Id: I0ccecf4407e9d89a15f838bd64788d90964be2ab