[roll] Roll infra/recipes [fuchsia/fuchsia] Do not upload to catapult for bisector builds.
Builds launched by the bisector builder will be out of order, as in
they'll be run on older commits. This causes out-of-order uploads to
catapult which aren't handled properly so we shouldn't upload to
catapult in this case.

Original-Bug: 445994940
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1372718
Original-Revision: 38c8f0eb21cfd9e86efebe4cd8a670acae94eb84
GitOrigin-RevId: c4c8d8544cbe05def06cd3dfce1447c05082b58a
Change-Id: I58c7ac34f25211f4563b6fe79de6f88e599f1a88