[roll] Roll infra/recipes [checkout] Attempt to patch skip_patch_projects.
Since we run fuchsia builders on recipes changes, we don't want to try
to patch in the recipes change since the recipes repo is not included in
the fuchsia checkout. However, there are some builders that check out
the recipes repo that we would like to patch the recipes change in in
order to test it out. Instead of skipping processing the recipe change
altogether, we can just ignore it when it fails to patch because the
project doesn't exist in the checkout. This change should also have the
added benefit of patching in Depends-on footers for non-integration
changes that are part of the fuchsia checkout.

Original-Bug: 436936960
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1381840
Original-Revision: 10de973a4b4e70557561606148f900aff9f4baf6
GitOrigin-RevId: 55f2c4053331872dd92c00e560f554ed30c0089a
Change-Id: I6fd1a9c240101dba325a4240eb97da06af7489eb