[roll] Roll infra/recipes [infra] Recipe to create compat branches
This recipe should be triggered when a CL lands that may contain an
update to sdk/version_history.json. It checks whether there is a
"supported" API level that's not mirrored by a "compat/" branch,
and creates the branch if necessary.

Original-Bug: b/447826691

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1352707
Original-Revision: b669e1793e5c71cb0562d415455eb216ddfc1e53
GitOrigin-RevId: 899c189d5a1a5b1bfde9fa4f877eabc0bb4ea75b
Change-Id: I9fe8d1640ebc1c3b37a4f6ab28d4681067232e7a