[roll] Roll fuchsia [kernel][vm] MutatePages for VmPageSpliceList
Add a function to the VmPageSpliceList, MutatePages, that allows a pass
through a finalized VmPageSpliceList to modify pages in the splice list
without transferring ownership or considering them 'processed'.

This is useful in cases where the caller might have to deal with
ZX_ERR_SHOULD_WAIT, as the current VmPageSpliceList iterators transfer
ownership to the caller, and pages can not be re-appended in the
Finalized state.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398093
Original-Revision: b2baf189c9da319d9df36782ac4ded7e7a04a028
GitOrigin-RevId: 4280fd74b44dd5e1c38a363b3c316dc997b534fe
Change-Id: I68f35eb2520d4db182770057c0eb1e4882b00c9c