[roll] Roll fuchsia [sestarnix] Add a test for type transitions on FUSE
FUSE filesystems are typically genfscon labeled, and do not support
seclabel. In this case, type transitions are applied to compute a
label and perform may_create checks, but the label is subsequently
discarded in favor of the genfscon label. This test implements a very
simple FUSE filesystem to test this behavior. The test is currently
expected to fail on Starnix.

Original-Bug: 414336570

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398814
Original-Revision: ac8ec035577dc710d5a8dfb7df8f7e89596b2f2b
GitOrigin-RevId: df39c323332a6d0390eb03236d2ee3f48a2d0f8f
Change-Id: Ic6502d68c0e36e5f5d6dc27f6bde3ba6951259b4