[roll] Roll fuchsia [bootfs] Replace `bin/startup` with eager
To make this possible, change eager startup failure behavior to just log the
error instead of propagating the error to the parent's start. This is
necessary because some `core` components in certain builds have a child
definition but their package is not part of the build. This should
probably be fixed, but the new behavior is arguably more conceptually
coherent -- the lifecycles of parent and child instances are normally
decoupled and if the child is started by some other means it doesn't
affect the parent.

The benefit is we can reclaim a bit of memory from bootfs, as well as
simplify the topology slightly.

Original-Bug: 438195589
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396566
Original-Revision: a89c273fef3e686c3739f03733015a0533eef894
GitOrigin-RevId: 1a2f2143eae30dd4c01f2a7a1ed3316a29f9392d
Change-Id: I75fdddceb042eec750c4e57ccacfd10d31f65e95