[roll] Roll fuchsia [kernel] Introduce <lib/page/size.h>
This change introduces a simple 'library' meant to consolidate reference
to and dependency on system page size *within* the kernel (with
visibility constraints as well). This preserves the current expectation
that page size is available as a constant, but is intended to break its
relationship to the zircon public headers. Follow-on changes will
replace usage of (ZX_)?PAGE_SIZE (and friends) in the kernel with the
definitions in this header.

Original-Bug: 42146863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405318
Original-Revision: 9fb8afc6c349aeb6b3c206f0c32a2dae8ec4c935
GitOrigin-RevId: c140277f1c5cf6e8c1bc08965abcfcc41be6bcf8
Change-Id: I07f5a71fe7a3db21df314e6c75d99d8cc468b511