[roll] Roll fuchsia [kernel] Replace usage of ZX_PAGE_SIZE
Page size in userspace should be a runtime determination and the export
of page size in <zircon/limits.h> is deprecated. We replace that
deprecated usage with the kernel-internal libpage definition of page
size... with the exception of the last use of ZX_PAGE_SIZE as a default
boot option, to be handled in a follow-on change.

Original-Bug: 42146863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405319
Original-Revision: 645e833a9a39e27631dbff28d319a7e4e6d03706
GitOrigin-RevId: 9035cd280769da1e5d6331e7e11c3b691c4eb003
Change-Id: I14b781316fd11b9bd47db966c5444869731fb86c