[roll] Roll fuchsia [inspect] Make the error message more clear when property match fails
This updates the output to inform the user when they've attempted
to match a property, but the name is actually a node.

We can't/shouldn't use AnyProperty to match empty nodes, because
that breaks the distinction between nodes and properties. But,
we can still suggest the node-equivalent of AnyProperty since
the presence of the node is detectable.

This also updates the test that brought this issue to my attention.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1408719
Original-Revision: 7ee0a53df0fbeb3ba1448289533485ed0cf4c04c
GitOrigin-RevId: 9be3d79de19d8bc12cff2d422f41b822ad22babe
Change-Id: I571b0a25bec4026c430f814f74e2e743654566de