[roll] Roll fuchsia [wlan] Now get_band_cap_for_channel also filters by operating_channels
This change doesn't affect current behavior in practice because
scanning for a particular network is already filtered by the operating
channels reported. Additionally, a check for a valid US channel was
removed from `validate_radio_cfg` which validates the channel
configuration of an AP starting. Since `get_band_cap_for_channel`
already filters by operating_channels, such a check is redundant.

The only corner case is when SME receives a scan request with only invalid channels
and sends an empty channel list to the driver. In that case, the driver might scan
on all channels and present the client of SME with scan results on channels that
potentially don't align with the reported operating channels.

Original-Fixed: 42174927
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1407280
Original-Revision: 86dfffaddcd40551d7a2561c6ba3159828f43c92
GitOrigin-RevId: 0fac17c0b582b731555d5e582a3c96aa1c5c14bc
Change-Id: I2621bb1882d1e2427b1c059a5cff6f38664b7611