[roll] Roll fuchsia [ffx] Change `ffx target list` to use is_default
Instead of doing a name comparison, use the TargetInfo::is_default
field when indicating that a target is the default.

Note that _multiple_ devices can be listed as "default". This is
not a regression, since before this CL, multiple devices could be
configured with the same node name, and they would then all be
shown as the "default". In practice, this simply means that
specifying the default device may not actually disambiguate the target,
and ffx target list will reveal that by putting the "*" next to
multiple devices.

(This was surprisingly weird code, because of all the macros in
target_formatter. At some point we'll need to clean all that up.)

Original-Bug: 448164233

Test: updated existing unit tests, and a couple to explicitly test
the new mechanism.

[g-review]

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409458
Original-Revision: 8795dff07204f134b1cfd3ee3cf409ecdbbc6bc6
GitOrigin-RevId: a5f56abc8a17b2f9548996521d9b3bb35bba0a44
Change-Id: I278295641fd5a0a33c01c1a903ae31a1ffc59721