[roll] Roll fuchsia [vfs][rust] Remove uses of FutureExt::boxed
Rust doesn't seem to inline trait methods that often. Using
FutureExt::boxed causes the future to be constructed on the stack and
then copied to the heap allocation in the FutureExt::boxed method. When
Box::pin is used directly, rust will make the allocation first and then
construct the future directly on the heap. This avoids the memcpy of
the future.

This optimization is probably only saving nanoseconds and is unlikely to
be detectable in most storage benchmarks.

Original-Bug: b/456517172
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409734
Original-Revision: 3aaa05af0c9ecfb5dad5a96a24a46e4df9944344
GitOrigin-RevId: bc66f83dbe096bf52104147e08af16dd84ffe434
Change-Id: I56256858a9c26a25aa33189388ca3a81dfc42c68