[roll] Roll fuchsia [netstack3] Test-only Eq impl for SocketCookie
The motivation for this is for use in diagnostics tests, where we don't
want to have to take ownership over the value just to test that the
cookie is correct by exporting it.

It ended up being a somewhat viral change because it required adding a
new testutils variant of the netstack3-sync and, transitively,
netstack3-trace crates.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409554
Original-Revision: fd4cec3faa5b5378a5e9b2a62f47c60980ed00ce
GitOrigin-RevId: 09d0d36293c9c8535e2e05852766e7149c3da35a
Change-Id: I23b437d7750f513d447e0da7c63d9d8c893384d6