[roll] Roll fuchsia [blobfs] align default cache policy with fshost
fshost defaults to (and so almost universally overrides) the cache
eviction policy to never_evict, so there is no reason to have a
different default in blobfs.

This was causing a bug with the new fvm2 component, because it wasn't
forwarding the override of this option, so it was evicting by default.

This option should just ultimately be removed, but this fixes the
immediate issue.

Original-Bug: 456807639
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409897
Original-Revision: 2ba0173f75ec30ee0f934d34fc3cc39e351a7353
GitOrigin-RevId: 9a302a12c67a5b96dded9345b965ce98975c7307
Change-Id: I60fa74c0f189a7eb815c57c76311c9dcb9ec9259