[roll] Roll fuchsia [host-target-testing] Make uncompressed blobs available if needed
After https://fxrev.dev/1351551, uncompressed blobs are not included by
default in local builds. This causes local OTA tests to fail in test
cases where a new update package is created on-the-fly.
This change ensures that uncompressed blobs for ZBI and vbmeta are
available for the test to modify as needed and create the new update
package.

Original-Fixed: 446897798
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409914
Original-Revision: 87de24d25b4f9f709d7236ed3667b950e64a01da
GitOrigin-RevId: e4dd9a3014b35e73129528d0578446f370228bdc
Change-Id: I9455dfd749e5e4beef3138ad6d4627f308773061