[roll] Roll fuchsia [lockdep] Check if reporting is disabled for AcquireAfterLeaf
If a mutex is tagged as LockFlagsLeaf and is contended, the lock
acquisition path will try to update the thread's runtime stats when
blocking. This acquires the ThreadRuntimeStats lock which is tagged with
LockFlagsReportingDisabled, but a violation is reported nevertheless
without this fix.

Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409703
Original-Revision: d4313a1741fa239472026ca0f6496c8b4d3f79a5
GitOrigin-RevId: b3d812edf55d851d9e186431f10af33822b3fbca
Change-Id: Ie125db72aeac7cdd93580347975e2e9d47f94938