[roll] Roll fuchsia [assembly] Explicitly set product_bundle version strings.
Currently, PBs are assigned a version number which matches the version
of the SDK that was used to build it. This works, but it's a bit
disconnected from the version of the contained assembly artifacts.

The new `ffx product-bundle create` tooling changed this behavior to use
the version string from the product config. This is better, but in some
situations I think it would be best to explicitly set the PB version at
the target level. This also allows the versioning console to transition
from "build_api/build_info.json" -> "ffx product-bundle get-version",
since we can explicitly set the version to match the previous format.

This CL explicitly sets the version string on the build target for all
product bundles in the fuchsia tree. In the future we can make this
field required.

Original-Bug: b/456261023
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1407674
Original-Revision: 4fe3346c49e4209fd2b8d5f2656c712a3d9661c2
GitOrigin-RevId: 04b086fc3b76450de421c142b174541cd60fcf37
Change-Id: Ie97cfc74e08217f23ec2f0eef3ba0a740df7ea89