[roll] Roll fuchsia [kernel] Use PAGE_SIZE from <lib/page/size-asm.h> in assembly
We extend libpage with a generated header with #define-versions of the
constants defined in <lib/page/size.h>, intended for assembly usage. In
migrating the kernel's assembly over to using that we remove the last
usage of the <arch/defines.h> PAGE_SIZE #defines and so remove those
definitions too. Deferring removal to a follow-up change was also not an
option as this would result in conflicting definitions of PAGE_SIZE.

Original-Bug: 42146863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1410037
Original-Revision: ea95d5e278cfa6af2312de4c5e41e01e3ecf5e9b
GitOrigin-RevId: 68b11cfb3980bc16839573a7f6f1e15330937039
Change-Id: I5434465ab8409bc5349d67d8202fc6af208c8b94