[roll] Roll fuchsia [build][prebuilts] Fix incremental issue when generating Breakpad files.
The //build/prebuilt target invokes the generate_breakpad tool to
generate Breakpad symbol files from the prebuilt debug symbol files
that are found under //prebuilt/.build-id.

It generated a depfile to ensure that the action is re-run by Ninja
if any of those files changed. Unfortunately, this is not incrementally
correct, as the content of the directory depends on the content of
the Jiri manifest and whether `jiri init -fetch-optional=debug-symbols`
was used (the directory is empty by default otherwise). The depfile
would not be able to detect that new files were added by a `jiri update`.

To fix this issue:

- //build/regenerator.py adds a task to glob() the content of
  //prebuilt/.build-id, if it exists, and writes the list of file
  paths to a file, under //prebuilt/prebuilt-build-ids-list.txt.

  The file's timestamp is only updated if the list of files changed.
  No content hash is performed, because it is assumed that all
  files there are read-only with unique content-based names
  (e.g. //prebuilt/.build-id/xx/yyyyyyyyyy.debug).

- prebuilt_binaries.gni is modified to use prebuilt-build-ids-list.txt
  as an input, even if the tool never reads the file. This ensures
  that Ninja will re-run the action if //prebuilt/.build-id was updated
  before a regenerator run.

  The action is also switched to use `hermetic_deps = false` as it
  cannot be hermetic. The previous use of a depfile hid that fact
  but was not incrementally correct.

- The `-depfile` option is removed from generate_breakpad as it
  is no longer useful, and cannot be incrementally correct.

Original-Bug: 454618550
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409574
Original-Revision: ea7d3e0bc5ba0b0e44b7dd847a68b5384dcaa8a9
GitOrigin-RevId: 11641ba78d2e7e23422725e1a0fda99fd0b8b307
Change-Id: I3048ef26ed563bbdb5e8b8f84486b36a8cf7e7e5