[roll] Roll fuchsia [windowed-stats] No longer accept timestamp on InspectedTimeMatrix::fold
`InspectedTimeMatrix::fold` acquires a lock on `TimeMatrix` before
folding the timed sample. To avoid the possibility of MonotonicityError,
change the API to accept the sample only so that it can generate the
timestamp after acquiring a lock.

The existing `InspectedTimeMatrix::fold` that accepts a timestamp is
renamed to `InspectedTimeMatrix::fold_at` because the event builders API
(which is currently unused) still uses it.

Original-Bug: 456481189
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1408463
Original-Revision: 95d67576def3edc36426c4dc12f7301e38660e92
GitOrigin-RevId: 77dbd626ae382acce0b537def4e9fceb81c62b05
Change-Id: I46e954d54b78c7437ebeaed8dab8669a9f4334ba