[roll] Roll fuchsia [rust-analyzer] Quick fix for check-on-save
With buildfile set to /dev/null, rust-analyzer will try to cd to /dev,
which causes check-on-save to fail because it's trying to run fx from
outside the Fuchsia tree.

It's not totally clear if this is the right long-term fix, but it at
least unblocks things for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413494
Original-Revision: 82e1bd5a8326327432afb11978efe83ebc6046e9
GitOrigin-RevId: 00feeb4141e589df7ce451636da6bd92bee99445
Change-Id: I88a0565d8cc22b1c70ffc4fd14eb4c1e85b96cf4