[roll] Roll fuchsia [starnix][perfetto] Improve trace state handling
This handles when the TraceState of the trace engine updates
faster than the TraceObserver events are processed. This should
be extremely rare because the perfetto_consumer holds a ProlongedContext
which prevents the trace session from ending.

The panic which was happening in this situation is replaced by
cleaning up the perfetto_consumer and re-initializing it.

Original-Bug: 457768643
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413516
Original-Revision: dad9efc3a5282ee5206c1cdee46d3c6d9532cd43
GitOrigin-RevId: 55c432d12164db685c4358bcabe82dfbbcf8bece
Change-Id: I5e87e69216c36fac3188e67bf9984ea2495655c4