[roll] Roll fuchsia [fxfs] Fix last_object_id computation after unlocking
When unlocking a store, we were setting the last object ID by using
`update_last_object_id`, but that assumed that the object ID was *after*
it had been encrypted, but that's not the case when a store is being
unlocked.

I think this could potentially have led to filesystem corruption because
I think it could reset the last object ID to an earlier value than it
should which could potentially lead to objects being overwritten when
they shouldn't.

Original-Bug: 453445060
Original-Bug: 450789112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413694
Original-Revision: fa984ad3ec313d54255789db394f343cb8ecc9fb
GitOrigin-RevId: ad5755b2b5c4fffe6120a78354fcdfc2297a3365
Change-Id: I6bc17c29cf0ad95870633e3c068c4aa02f39cf3c