[roll] Roll fuchsia [bazel][fidl][zither] Add remaining Zither backends
Add Zither backends for Rust, etc. `legacy_syscall_cdecl` is the
only backend still not implemented. The Go and Go runtime backends are
not implemented because we do not believe we will need them.

The following targets fail to build. This will be addressed separately.
* //sdk/fidl/zbi:zbi_zither.rust
* //sdk/fidl/zbi:zbi_zither.rust_syscall

NO_IFTTT=This change brings the Bazel file closer to parity with its the GN file.

Original-Bug: 454449781
Test: fx bazel build --config=fuchsia //sdk/fidl/zbi:zbi_zither.asm //sdk/fidl/zbi:zbi_zither.asm.gen.check //sdk/fidl/zbi:zbi_zither.c //sdk/fidl/zbi:zbi_zither.c.gen.check //sdk/fidl/zbi:zbi_zither.kernel //sdk/fidl/zbi:zbi_zither.kernel.gen.check //sdk/fidl/zbi:zbi_zither.rust.gen.check //sdk/fidl/zbi:zbi_zither.rust_syscall.gen.check //sdk/fidl/zbi:zbi_zither.syscall_docs //sdk/fidl/zbi:zbi_zither.zircon_ifs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1410095
Original-Revision: bfb432567249c7515ac844aaca389e899ea9ec78
GitOrigin-RevId: 76e30608e657ae82147dc97aa0f5da57047a9c2e
Change-Id: I7449e28a31285aa4f1bb3eaaac5cd34370e0eb3e