[roll] Roll fuchsia [netstack3] Combine socket mark matchers
Keeping the knowledge about the number of marks out of the API will make
it easier for us to add more marks in the future.

While we're here, also add a TODO on the routes rules FIDL to combine
those two mark matchers for the same reason. The core implementation
already combines the matchers.

Original-Bug: 458029146
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413114
Original-Revision: ca895301220dcabae66018975c773ddbcc8c70c7
GitOrigin-RevId: 8f3104b632ffe2f97723e3728916c937633faa1e
Change-Id: Ia9c9c28346e53e1c9c2f65113f4d822d52782138