[roll] Roll fuchsia [Toolchain] Restore mac toolchain components for Rust, llvm, and LLDB
The toolchain components for mac still need to be updated together, and
tagged in CIPD. The manifest can have these removed again once we have
alternate infrastructure to manage CIPD tagging and builder triggers
independent of the Fuchsia manifests.

For some components, like LLDB and LLVM,  may need other changes to
the toolchain infrastructure and recipes to be in place before we
remove them.

For now, we simply want to restore something as close to the previous
status quo as we can, to ensure we can roll all of the required parts of
the toolchain to our customers, like Pigweed. Once we've unblocked our
ability to roll, we can determine the correct steps to take to decouple
the toolchain builds and infrastructure from the Fuchsia manifests.

Original-Bug: 458061496,458075605
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1414198
Original-Revision: 6cb10390334dd613a8645109feeb57a7b164b5bd
GitOrigin-RevId: 327e215494fb9b0c558374e145cd85cc7d52533a
Change-Id: I4b0b3d82fe5248e4beeda7f8eeffd7877f716ff3