[roll] Roll fuchsia [paver][moonflower] use hardcoded inactive partition type GUIDs
Previously, the paver was using the boot_a/boot_b partition flags to
determine the inactive type GUID, but we've run into some cases where
the boot partitions themselves have mismatched GUID and flag states, in
which case we would grab the wrong inactive type and totally mess up the
GPT.

This CL uses a hardcoded list of known inactive type GUIDs instead so we
know whether to swap GUIDs or not even in the case when the boot
partition is corrupted. It's still unclear how the boot partition gets
into a bad state in the first place, but this should hopefully prevent
us from totally bricking the device when it happens.

Original-Bug: b/444508079
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413481
Original-Revision: a40c84f67114953ff820c3e18b8604e26f428715
GitOrigin-RevId: 33a6def8ed117ea27cff3123b32f6f92f628bb7f
Change-Id: Ieddd83bc5d3c197f3d16691e5618e604ecd82207