[roll] Roll fuchsia [bazel][fidl] Add fidlgentest_go_test() macro
Add a Bazel equivalent of the GN template.

More work is necessary to be able to run the tests successfully,
especially on infra.

It is not a symbolic macro because go_test() is not a macro and thus we
cannot inherit its attributes.

Original-Bug: 455677059
Test: build --config=host //tools/fidl/lib/fidlgentest:all
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1414580
Original-Revision: 2bb932f3e8f0eda47eba4460882f006b09eeae48
GitOrigin-RevId: ddb5d65159400193cae7190ce8a57300c5adf978
Change-Id: If1775b14ac020add2cf46403e60574fa8bdf607d