[roll] Roll fuchsia [kernel][vm] ZeroPagesLocked: Use markers more often
In ZeroPagesLocked, we sometimes have to choose between representing
zero with a freshly allocated zero page and with a marker. This CL
represents zero with markers whenever possible.

The old policy was to try to "justify" markers by pointing to committed
pages. This was believed to bound the total number of markers allocated.
We measured the number of new un-"justified" markers, and found that
this CL only increases the number of markers in rare scenarios.

Original-Fixed: 458059877
Original-Bug: 423017076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411519
Original-Revision: ef7080f1dd438525528004b7ca8fe0959cd4b83b
GitOrigin-RevId: 9e79c17f45d6b445f86e01f5766624ea031853db
Change-Id: I7b13fba9c0f4bee514705f34aa317f1ca77cb1b2