[roll] Roll fuchsia [Netstack3] Keep sharing domain tokens for the lifetime of the socket
Previously SetReusePort2 was taking token koid and dropping the token
after that. Now it will keep the token for the lifetime of the socket
(or until the token is replaced). This will allow Starnix to observe
if the token is still being used.

Original-Bug: 451615802
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1413439
Original-Revision: c82f72abf92e722b0f61e797bb5a6f8f8e49a35d
GitOrigin-RevId: 85a9ea7527e09a288b0627a31c31258d3c992759
Change-Id: Icae284d502917267f0a82bb6afa0dcdff730fa23