[roll] Roll fuchsia [thermal] Record max temperatures.
Records max temperatures at the same frequency that we record
historical temperatures.

Historical readings are changed to be averages, so there are natural
ways to reason about these in relation to the maxima. (Conversely, if
we were continuing to record a sample, we'd have to choose when to
record each sample in relation to the intervals on which maxima are
computed, and there isn't a natural way to do that.)

Original-Fixed: 451015386
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391615
Original-Revision: 22613eeafaedc800be263983edbe48f0f4da6538
GitOrigin-RevId: 0c2d735aefd2ba77789bc32b5658fa0a015ab2a9
Change-Id: Id94f16178578ccdde622116da84a77021aa87e95