[roll] Roll fuchsia [libc] Update build rules for llvm-libc printf
The implementation headers are in fact used outside the
implementation, so treat them as interface headers.

The upstream snprintf test has been changed to use test harness
features that aren't supported in the zxtest version of the
tests, so skip it for now.

Original-Bug: 42141211
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416674
Original-Revision: 264be46457f4a299ab38645cf89e42137bd6168b
GitOrigin-RevId: d0b58077c26105fb1f3f71420b7e9f16ce3310a8
Change-Id: I21f2ea82554f1677e6205ca49d1ade0375e67325