[roll] Roll fuchsia [ffx][forward] Recover from disconnects
Introduce a sentinel listening socket on the target so we can always be
sure to detect disconnects.

The main loop watches that sentinel and triggers a delayed reconnect.
The reconnect can take a bit, we're relying entirely on the internal
timeouts from the Connector tooling here.

Adds a new argument `--once` to the CLI to opt out of the auto-reconnect
behavior, causing it to bail on the first error.

Original-Fixed: 456723550

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415775
Original-Revision: f1373fadd48eb613d0b2750ea68cf2dc1ac83643
GitOrigin-RevId: 0a2ad69a8dd329953845906625da16c1e8dea47c
Change-Id: I703f9adae5fe4135b656acce8a751b36d3aff2e9