[roll] Roll fuchsia [debugger] Improve zero line LineTable entries
We now do something closer to what GDB does:

Instead of returning one singular line table row entry for an address,
we find two entries: one for the "symbolizable" address (which doesn't
necessarily have to be is_stmt) and a "statement" address, which is the
closest following is_stmt entry after the address we're looking up.

This lets the code still be shared for breakpoint and symbolization code
with a new option to ResolveOptions to specify which one of the above
entries should be used when resolving a location.

Original-Fixed: 436290252
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1410834
Original-Revision: aacaa508f1a8730ae93a31e5e0cf7e64c727c06c
GitOrigin-RevId: d129c7c2ed6795a93f9da5ced18e43755b477041
Change-Id: I8894ed582eb4e74adda697d18b9c1b7d9f23bf16