[roll] Roll fuchsia [ffx] SSH key verifier command.
For eng and userdebug builds, downloads the authorized_keys from the
device via HTTP, and uses the result to verify whether or not matching
public keys for the device are stored locally. Searches the ssh agent,
the home directory, and the fuchsia directory (if set) for ssh keys.

Things to add:

-- Checking pub/private keys match.
-- Auto-fix issues with keys missing from home dir.

Original-Bug: 436649104
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364587
Original-Revision: 7afcad33e8f353db4456fe2426be383be1f76069
GitOrigin-RevId: 99766330b26e9470d607f4d819d5c4705d83cba4
Change-Id: I2bad1588dfa10c1dac382dfd0de85113573038dd