[roll] Roll fuchsia [fuchsia-merkle] Add MerkleRootBuilder and MerkleVerifier
MerkleRootBuilder is similar to MerkleTreeBuilder but it doesn't store
all of the interior hashes. All users of MerkleTreeBuilder except for
fxblob only care about the root so this is slightly more efficient for
that purpose. Fxblob also needs the leaf hashes which MerkleRootBuilder
is also able to generate.

BufferedMerkleRootBuilder is similar to MerkleRootBuilder but is able to
buffer data. This is useful when the client is receiving data that may
not always be block aligned.

MerkleVerifier uses the leaf hashes generated by MerkleRootBuilder to
verify the data later.

ReadSizedMerkleVerifier builds on MerkleVerifier and can save a
significant amount of memory when data is always verified multiple
blocks at a time. This is intended to be used in fxblob.

Mundane was replaced with bssl because bssl's Sha256 implements Send.

Original-Bug: b/437178186
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371261
Original-Revision: cda74b9429b6d4d7ec8f340df8afcd6cc4685324
GitOrigin-RevId: 7854806efef7758a8f573629bae2b34ad6e16456
Change-Id: I173183920bf9f12a23d0d77f20f471200aac2741