[roll] Roll fuchsia [logs] Report error string to Feedback
Report an error string for the decoding
error to Feedback when a decoding error happens.
This is best-effort, as constructing a CString in Rust
is fallible. On the case of failure, no error string
will be returned.

Original-Fixed: 459462291
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1418677
Original-Revision: 5eed34ed1aa5f67a9cea287e2bdfd144f96a99b9
GitOrigin-RevId: 0bcfb2092192f9fcf44ea3029853b45b87aef473
Change-Id: Ibed1fdad8c62f11c3e0ff0ebca37333c69d43763