[roll] Roll fuchsia [ramdevice_client] remove unneeded isolated devmgr dep
The ramdevice-client-test library had a group that included the driver
shim library, but it is not clear why it was there. The comment
explaining it didn't seem to be right. A couple of deps were missing
from sysconfig-client that it got transitively, which were just fixed up
directly.

The allowance for these to use the shim was also removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387104
Original-Revision: ce6688cb5ed03ec988237527c698fc7b087bae90
GitOrigin-RevId: b3191370797bf17c63a398d3c5bf3576b2428a04
Change-Id: I10a54f4bf4630d771165f289bdcbebd24d40b2b8