[roll] Roll fuchsia [usb] Fix the USB device qualifier return data
When responding to a GET_DESCRIPTOR query for a DEVICE_QUALIFIER,
the usb-peripheral driver code had taken a shortcut which led to
returning incorrect data.

Rather than maintain an instance of a device qualifier data structure,
the implementation copied a prefix of the device descriptor - which
starts with the same fields - and overwrote several fields that
needed to be different in the device qualifier response.

However, two fields were not corrected: b_length, which still held
the size of the larger device descriptor structure, and
b_num_configurations, which was forced to zero.

It was this last action that caused the LeCroy GUOT-05 test to fail.
The test requires b_num_configurations to be at least 1.

A review of the correct setting of this field suggests additional
work is required to properly support the spec and any necessary
product-specific overrides. For now, a workaround is in place
along with a bug comment to provide a better solution later.

Original-Bug: 434990314
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1407778
Original-Revision: 1837892efa2dc2b01ccc1524fd125c621005ef8c
GitOrigin-RevId: c3b8541d4e8ec38d3a74fc652124645adee72822
Change-Id: I43d088912a4983401ddbdd1a37012ef5265f53df