[roll] Roll fuchsia [fxtest][gemini] Highlight key stack trace lines in Gemini Analysis
This change makes it so that the key lines of the stack trace are
highlighted in color in the stack trace, instead of being listed
underneath. The lines outputted in red is deemed the most key line to
solving the issue, and the lines outputted in yellow are lines that
could potentially be used to debug the issue.

Test: Added unit tests to gemini_analysis_test to verify the new output functionality.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416955
Original-Revision: b99a121d85fc0ab8ed7d4d6b6dad035f93d209ba
GitOrigin-RevId: d9c3d77c2744ea8ff95e3659397bdf1712db6dfd
Change-Id: I499c4f5828c4147e19203565b44fae46088ae616