[roll] Roll fuchsia [phys] BootZbi::Load boot tests
Add boot tests that exercise the BootZbi::Load and Init path.
The cases under consideration involved whether the kernel can or not be
loaded in place, and whether the Data ZBI fits in the allocation or not,
and whether it can be extended in place.

The kernel dimension is handled by generating a ZBI that is kernel
aligned (it can be loaded in place) and conditionally inserting a 1 byte
long discard item before it(cannot be loaded in place).

The DataZbi checks are performed based on the initial allocation
dimension, and the extendendability can be guaranteed by overallocating
and then shrinking the allocation such that there are N bytes available
at the tail. And it can be guaranteed that it wont be extendable in
place(and force relocation) by allocating the tail when shrinking the
allocation, hence no bytes are available after it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333918
Original-Revision: 2d7040f283a53c622f2fd27514e3d6f3bf8485cd
GitOrigin-RevId: 36533df743f9b654adb0921a346fd7e16ead927a
Change-Id: Idb6ec525404816862b61fd3d9ec16241a7d953b4