[roll] Roll fuchsia [starnix] "Sniff" a longer process name from argv[0].
Linux's uapi is limited to a comm value of 16 null-terminated
bytes, which is not long enough to differentiate between many
processes when the comm is set naively from the binary's
basename.

We use the comm value for Zircon process names because it's the
only logical task name in the uapi that always involves a
syscall. argv[0]'s basename is often the most useful name for a
process, but the most useful names in systems like ART get
written to argv[0] well after the task has started and don't
involve a syscall.

Using the comm value for Zircon process names makes it harder for
Fuchsia-side tools to understand which process is which if e.g.
there are a number of processes whose binaries share a common
prefix. This is especially challenging when Starnix's own config
relies on process names for things like rt_role_overrides.

This change attempts to improve the situation for all of the
above challenges by sniffing a longer comm for internal usage
when argv[0] is a superstring of the user-provided comm value.
In practice, this seems to get us better Zircon names for many
tasks in our current system without risking clobbering names
that are intentionally provided by userspace.

Original-Bug: 455850151
Original-Bug: 321273720

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1406011
Original-Revision: 43a842cdbab0d869cbb2794a3253e5888baecb13
GitOrigin-RevId: 78dbffa6a1b327eca4a8f307c3badc37bfd7fa7c
Change-Id: If31959edd8e635935c3ba525e2ef234660f1d559