[roll] Roll fuchsia [starnix][rcu] Add RcuCache
RcuCache is a cache that uses RCU to provide thread-safe access to a
hash map. If the number of entries in the cache exceeds its capacity,
entries are evicted in FIFO order.

In the future, we might elaborate this data structure to retain recently
accessed items, for example using the CLOCK algorithm.

Original-Bug: b/445230228
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420035
Original-Revision: 5ae7bc3112a102d061ec44ccf5ad1c7f224e425f
GitOrigin-RevId: 6899857d9daf7cf516e247636939f3b430e97d36
Change-Id: Ife2b2009280ceff105a4313787427513de338c90