[roll] Roll fuchsia [kernel][hypervisor] Disable preemption in AutoVmcs/AutoGich
AutoVmcs and AutoGich are guard-like objects that ensure the "owning"
thread is not preempted.  Additionally, they are designed to catch
programming errors where the owning thread blocks while they are
active.

Prior to this change, these objects relied on the scheduler and
arch_blocking_disallowed to defer preemption if a reschedule event
occurs while active.

As a preemption deferral tool, arch_blocking_disallowed is flawed in
that there is no mechanism to ensure that the deferred preemption is
handled in a timely fashion.  Explicit disabling of preemption via
AutoPreemptDisabler is the preferred tool.  A future change will alter
scheduler behavior such that arch_blocking_disallowed is only used to
assert that a thread does not block when it shouldn't.

To prepare for that change, we update AutoVmcs and AutoGich to use
AutoPreemptDisabler.  Note, arch_blocking_disallowed is still used to
detect programming errors where the owning thread erroneously blocks.

Original-Bug: 42143537
Original-Bug: 457759524
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416637
Original-Revision: d14d9a6122c087c74927907703fb7f2e1d5235ba
GitOrigin-RevId: 496f9310e3650c35273deed20cd6951f34edd247
Change-Id: Ie67dbb3dd6fc703c4699e4122f55c3f8b87609a6