[roll] Roll fuchsia [owners] Remove seanolson@google.com from src/connectivity/wlan/tools/wlantool/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/seanolson@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421044
Original-Revision: 74c1fd4d2886d9f80291838e66b22a05e42e0909
GitOrigin-RevId: 18463951e96a41b1cd000aaacd5b03f6245c23af
Change-Id: I041aaa20edbe1a513e1bb77d57a14892930c0c29