[roll] Roll fuchsia [owners] Remove seanolson@google.com from src/connectivity/wlan/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/seanolson@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421146
Original-Revision: 0d53ad7330fa436eaecd4fcb5723aa60f0fa29dc
GitOrigin-RevId: 41ae8958a2c5e850a90e54acb8b9d582e2e70c20
Change-Id: Id096c7be8456bcf0cee0fcd6797ba717a4f61a62