[roll] Roll fuchsia [devicetree-boot-shim] Extract boot config from initrd
Extracts the BOOTCONFIG file if present, and makes it available to the
shim. The file is immediately copied out of the way.

Boot tests for boot-config were added, that validate that for different
scenarios, the BOOTCONFIG file is handed off properly.

Original-BUG: 408051286
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1404202
Original-Revision: 47693c92a5e436a401d0dde7940bae728690d80d
GitOrigin-RevId: 65af6eede323bcc785e47ed231e05ead9a179e82
Change-Id: I1e986d4873a249c8cf3e7eb420a6aee5b8943a60