[roll] Roll fuchsia [zbi] Allow for providing an embedded ZBI
Allow for running the ZBI tool against a potentially embedded ZBI(in
some other file). The check is less strict about the size of the file
being exactly the length described in the header, and just check what
the header states to the contents length. This allows in the case of a
ZBI with an appended BOOTCONFIG to be checked, so that their contents
can be verified.

Original-BUG: 408051286
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1417074
Original-Revision: 040dce56c8cc2103b71608d52102052b7239bd9f
GitOrigin-RevId: 54a457db458b69ef234ba2a15f29bd5002b2fef1
Change-Id: I0a5147c195fdfde832c007eea34bda9eaced1e2d