[roll] Roll fuchsia [kernel][scheduler] Make Rescheduling with arch_blocking_disallowed an error
Prior to this change, Scheduler::Reschedule and
Scheduler::RescheduleMask and would pend a preemption when called in
an arch_blocking_disallowed context.  That is, they pend a preemption
and return rather than continue down into RescheduleCommon.  This
change removes the special case logic for arch_blocking_disallowed.

Now, calling Reschedule or RescheduleMask with
arch_blocking_disallowed is an error and will result in the existing
RescheduleCommon assert firing.

The motivation for this change is to remove the only remaining
non-assert use of arch_blocking_disallowed and to simplify the
reschedule logic in preparation for use of mp_reschedule_self.

Original-Bug: 42143537
Original-Bug: 457759524
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415586
Original-Revision: 49e0b74b3a6322e861d2e0df8b75a531ab9ef1d3
GitOrigin-RevId: eb9a35602d8c3f4707437eeceb0bdc2af99a633c
Change-Id: Ibfc56e0cd609b325ec43c17d0fb8454c5dc0969a