[roll] Roll fuchsia [honeydew][wlan] Remove wlan_core set_region affordance
The `wlan_core` `set_region` affordance was a duplication of the location
set_region affordance. As setting a region is not part of the Fuchsia
WLAN SDK, it doesn't make sense to have an affordance for it.

This change also refactors the corresponding test in `test_wlan_core.py`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416936
Original-Revision: cca8f2f19b6b67ac1c1b1ec06cdb4325906350a1
GitOrigin-RevId: 86c31f1bd8dd2f54f23042062db19f67d6d3dc08
Change-Id: I4786ca160669839f3e8951f23738f7b4f0773c05