[roll] Roll fuchsia [feedback] Split RebootReason
With the addition of GracefulShutdownAction, it was becoming difficult
to keep track of when the action was/wasn't applicable to RebootReason
and the associated functions in reboot_reason.h/cc. This CL defines an
interface containing most of the functions from the now-deleted
reboot_reason.h.

The interface has two implementations: one for zircon reboot reasons,
and one for graceful reboot reasons. The action is thus only
available/used for the graceful implementation.

Original-Bug: 414413282
Test: `fx test //src/developer/forensics`
Test: `ffx component stop archivist.cm` @ 908973498ebc8df8

[g-review]

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415521
Original-Revision: a1257ff3dbad615bdff7a36492959a94bda7be0e
GitOrigin-RevId: eb6feeace5c7b961f1abeebd98f80f914a3782a8
Change-Id: I108fe9642a318be9f0aaee76079bece67f25637c