[roll] Roll fuchsia [wlanix] Process nl80211 requests concurrently.
It is possible for an NL80211 request to arrive while wlanix is
already processing one. In particular this can happen while we are
in the middle of a scan. After this change we will respond to those
messages before the caller times us out.

Test: existing + new
Original-Fixes: 458433475
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1415519
Original-Revision: a55fb7296cef9469d2df5e7976985146e9a8d308
GitOrigin-RevId: fe48b957283171de880ed24104df7f991e01b994
Change-Id: If2e9adcff03cff148aa99c0aa458b8c8272cc2f1