[roll] Roll infra/recipes [init_project_branch] Get project name from remote.
When creating the branch for a repo, the project name should be the repo
name from the remote url. The jiri project name is an arbitrary name and
though it usually matches the repo name, it's not guaranteed to, so we
should parse the name from the remote url instead.

This should fix errors like
https://ci.chromium.org/ui/b/8698720432921493841.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1421503
Original-Revision: 571bd0fdd75f1bb243d7622f9104f3eeb11418a2
GitOrigin-RevId: 3bf04e598842d5c84dc2f49f1e8acc6d1999b4ba
Change-Id: Ic88339c089a8b8ec8bc99d79390589ddc2ab270d