[roll] Roll infra/recipes [cipd_util] Don't set refs on non-scheduler builds
If a build wasn't launched by LUCI Scheduler, it was probably launched
manually using `bb add`. The main reason to do that (versus triggering
manually in the LUCI Scheduler UI) is to trigger on a specific older
commit instead of on the tip of the trigger's target branch.

In that case, updating the CIPD ref would be incorrect because it would
effectively move the ref back in time to a prebuilt from an older
commit, which could in turn cause any rollers that track that ref to
roll backwards in time.

The safest thing in this case is to just not update any refs for
non-scheduler builds.

Original-Fixes: b/460180461
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1422494
Original-Revision: 7d372b95f3abf84ae164ae03240923d2917ece3d
GitOrigin-RevId: e959eb4bbcf9f6365b8b3cf8be852eb02af4a601
Change-Id: I63c2d6b7143c53d0b719212e01253e9c2a93ec1a