[roll] Roll fuchsia [kernel][mp] Disable preemption for local task in mp_sync_exec
There are callers of mp_sync_exec that rely on the calling CPU to not
reschedule while running a task (see https://fxbug.dev/460442928 for
an example).  In the past, a scheduler exemption deferred a local
reschedule when arch_blocking_disallowed was true.  That exemption
is/will-be removed so we need another mechanism to ensure the local
task does not reschedule.  AutoPreemptDisabler is the preferred
mechanism.

mp_sync_exec will now disable preemption when running a task on the
calling CPU and when processing its mailbox.

Add an in-kernel unit test that verifies the new behavior.

Original-Bug: 42143537
Original-Bug: 457759524
Original-Bug: 460442928
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1422178
Original-Revision: e90f7ee1030e7afcf5ba6afa2eb9d6bf4ec1e7b2
GitOrigin-RevId: 922ec74c40a1fce5b8ca731bfd531fecca157aec
Change-Id: Ib75fb65892e78124a83d0d1089d4d8efbae2390c