[roll] Roll fuchsia [storage][f2fs] Replace usage of PAGE_SIZE
Per RFC 0016[1], page size in userspace should be a runtime
determination and this constant should go away.

The f2fs code does not clearly delineate page size from block size, so
as a temporary measure all mention of PAGE_SIZE/kPageSize are replaced
with kBlockSize - and a runtime assert around the relationship is added
to F2fs::Create(), to be revisited if and when we want to make this code
supported in non-4KiB-paged contexts.

[1] https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0016_boot_time_page_sizes

Original-Bug: 42146863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421525
Original-Revision: 5b1b89958ed3d41a0e2c46defcda9fdeeee7e75c
GitOrigin-RevId: 42edeed2cfc01537a3413370000c5f0d1e9ea58a
Change-Id: I033f74232a765b9f95966fd55080932f4fc01c4e