[roll] Roll fuchsia [blobfs][fxfs] Add NeedsOverwrite call to BlobCreator.
This is meant to be a way to get the ondisk status of a blob. It could
be This will default to `false` for now to signal that no overwriting is
required until the more complex decision-making can be instrumented in
blobfs.

This protocol is entirely in-tree and can be changed around or removed
if we decide on a different path later.

Original-Bug: 457411364
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411814
Original-Revision: 8ef334e0e839e8871aed46a9a8b88b5296be1800
GitOrigin-RevId: a37a8419a86532fe879a646d9a40c1af364639db
Change-Id: I4b42f3a7bd2b5154d25fb1e4116d3e3257ffe11e