[roll] Roll fuchsia [wlan][test] Close tracing Session proxy so tracing socket will close
Stopping a trace will not trigger the socket producing the trace to
close. Dropping the Session proxy is the only way to terminate a trace
to trigger the socket to close.

WLAN's tracing facility was only half-working before this change. By
happenstance, for hw-sim tests, the wlancfg component always panic's
after it exits, and this was causing the Session proxy to close and
write the trace. That of course is not the sort of behavior we should
rely on in our testing.

For FullMAC tests, there was no component to crash and cause the Session
proxy to close. Those tests would hang waiting for tracing socket to
close. This change prevents that hang.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1423095
Original-Revision: 1584e61616214c2830bdcf1ef743506868427f0a
GitOrigin-RevId: 727400892a72e3df65bd6ffe09031d2648ae1bc4
Change-Id: I9874b6b51e63d03963fe026a72692a1608a8391f