[roll] Roll fuchsia [ffx] Resolution only makes sense for products
Some of the Resolution code was written as though it was possible to
"resolve" a query into a target not in product mode. That was mainly
dead code (and/or just led to error-handling later), so this CL
removes such code, clarifying the data model.

On revisiting the code, it turns out that it makes sense to simply
return an Option<ResolutionTarget> when dealing trying to turn
a target into a Resolution -- it is not an error. This CL also cleans
up a bit of that code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421522
Original-Revision: 240f9a4704d896f13a1ebf70e2eeee5a6011e814
GitOrigin-RevId: 92e19e5bdc13963f727f0d92594b8fe995f27eb9
Change-Id: I6dd0550c3d3e2a5a73f10ceb372c53dcb867419f