[roll] Roll fuchsia [graphics][magma] Replace Fuchsia usage of PAGE_SIZE
Per RFC 0016[1], page size in Fuchsia userspace should be a runtime
determination and this constant should go away. magma is intended to be
cross-platform, but catering to the lowest common denominator means that
PAGE_SIZE shouldn't be unconditionally relied up on common code. To that
end, we replace usage of PAGE_SIZE with `magma::page_size()`, which
abstracts over the cases where PAGE_SIZE is not defined.

[1] https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0016_boot_time_page_sizes

Original-Bug: 42146863
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421777
Original-Revision: 80402328c742466b21da4d25797a9ed896fe9479
GitOrigin-RevId: 4c7aa8a1c6933369e02d2b278e0b79fc5f91f5a6
Change-Id: I277a43e7a06ee349d58b5d0153ef844b885f5674