[roll] Roll fuchsia [netlink] Support SIOCGIFNAME
Support the SIOCGIFNAME which supplies an interface index and requests
an interface name to be written to the provided buffer.

Add ioctl tests for successful and failure cases. Combined the
SIOCGIFINDEX and SIOCGIFNAME calls into a single test, as testing
SIOCGIFNAME standalone would require assuming an interface's id, which
is not stable. This follows the pattern for the equivalent gVisor test.

Original-Fixed: 325639438
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421527
Original-Revision: b91fd8d89f8587070a7210c4187755b8d50a6642
GitOrigin-RevId: 21f4a9330e2aa4de7db07a737a6560696c69d95a
Change-Id: I51f336bbca0243c649fcd50fdb4da33aa7320a92