[roll] Roll fuchsia [trace][lacewing] Switch to sync socket
When trace manager receives the stop request, it begins flushing trace
data to a socket, which in turn causes RCS to stall as it fills its
outgoing buffer. In practice, this causes the stop request to stall out,
which results in an eventual ZX_ERR_PEER_CLOSED.

We can mitigate the effects of this while we dig into the internals by
instead starting a background thread that begins reading out of the
socket as soon as the stop request is received, and yields occasionally
to the main thread to handle the stop request.

Original-Bug: 446873535
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421694
Original-Revision: e28422e4bdd4519a18179751edb654975f1093e9
GitOrigin-RevId: 92c26439e91f760c366b62ee3f6043cfdf103c76
Change-Id: I1e554bac566dc1c9768ae0f619a15e9cf5f9875f