[roll] Roll fuchsia [kernel][vm] Re-validate mapping in PageFault
Changes the locking structure in the page fault path to re-validate
whether the mapping is still valid or not after acquiring the object
lock. This allows for the aspace lock to be dropped prior to acquiring
the VMO lock, instead of after, preventing contention on the VMO lock
from propagating into the aspace lock.

Re-validation of the mapping leverages the property that the VMO in
the mapping is nearly immutable, with the exception of being set to
null when the mapping transitions into the dead state. As this
transition is not reversible and happens under the object lock, it
can be coordinated with an atomic, instead of requiring the aspace
lock.

This change completely removes the aspace lock as a source of lock
contention in page faults for relevant Starnix workloads.

Original-Bug: 452015888
Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420454
Original-Revision: d20c741d789f31a7c5ebf280a2306fda01e672a1
GitOrigin-RevId: 8788b46d424e8495ddb8e5afc7cc5e08ce29848e
Change-Id: I7f55fd95e681ce4842bc53dc781f48989f60cbe1