[roll] Roll fuchsia [kernel][debug][x86] Validate DR6
DR6 is set by the hardware when a #DB occurs to indicate to the
debugger the reason for the exception. As a consequence there should
never be a #DB with no reason in DR6.

Adding this check as a KERNEL_OOPS is part of root causing the linked
bug, where userspace is being told it took a debug exception, but the
DR6 it receives contains no cause.

Original-Bug: 453746441
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425017
Original-Revision: 5bdfb8914fca5cc266f43668d622bedeeb4997eb
GitOrigin-RevId: f8e02a454b869f21040ee9c96b768c78097af33b
Change-Id: I1fd402d9e31bc17ea948f30a858d276929a687c6