[roll] Roll fuchsia [fastboot] Support update-super command
Add support for the `update-super` command and allow it to be used to
determine the partition target when flashing the `blob` volume. By
default, when flashing `blob`, userdata is unaffected. To support
`fastboot flashall`, we can use the `update-super` command to convey to
the fastboot component if we need to wipe userdata or not.

Original-Bug: 458075061
Test: Tested locally on device.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1414417
Original-Revision: 50370bb5d9be7e9ea71551f6c6bacff5831a08c4
GitOrigin-RevId: f4417932d70cf876fb9869ebe94d17ba31c9a2a6
Change-Id: I43160a92e35fe18d2ba87c11279af6f32d5b9f64