[roll] Roll fuchsia [zircon] Job policy test avoids creating zero-sized physical VMOs
Instead of attempting to create a zero-sized VMO, attempt to create a
1-page VMO.

Even though the creation of this VMO will fail due to the invalid
handle, we are expecting a policy error, not an error related to an
invalid size.

We may soon disallow creating zero-sized physical VMOs. This is related
to our ongoing project of tightening the syscall APIs related to VMO
size.

Original-Bug: 460506492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1422517
Original-Revision: 9a78239f91c47f4c99ad7a786cb17216caf51390
GitOrigin-RevId: cf06712d865d384c1a2b9e5722bad099bdec0db5
Change-Id: If22abd2f425c608fc03c8fa593f3393b00a93e63