[roll] Roll fuchsia [uart][dw8250] Standalone driver for the DW8250 uart
DW8250 has some extra features and bits that make it somewhat
incompatible (but better) with the plain 8250 driver. In order to keep
the maintenance of the original 8250 a bit simpler, fork a copy of the
driver and customize it and remove dw8250 references from the original.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1412275
Original-Revision: 0a4670b1fbfd1616b9f1411937f291c6638585c0
GitOrigin-RevId: 4fbe64211661b5fc9c868bcf69504e314bd1f959
Change-Id: Ib2ac6d04a859ffc71c196c0fce3cfa05b6518205