[roll] Roll fuchsia [netpol] Remove CheckPresence from SocketProxy protocol
This endpoint is unused. We decided to have a Starnix kernel flag and netcfg configuration flag to determine whether these protocols should be used. When these flags are enabled, it is expected that the backing protocol will be present on the product to support those calls. This revert cleans up the unused FIDL endpoint and associated files that handled the calls.

Original-Bug: 385368895
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272004
Original-Revision: 0189e6688659737950d222c733f8be2baf99ac76
GitOrigin-RevId: 03b83c548eea254a0aba76e3f6d6ff38b65eb418
Change-Id: Iec895eb5c2cbddf4d1ec522d08c92a9fa2cfb26b