[roll] Roll fuchsia [internet-checksum] Migrate to criterion
The supported benchmarking framework in fuchsia is `criterion`. The
previous benchmarks in `cargo bench` can never be built in fuchsia
and run against a target. This change also allows the benchmarks to
be run in `perfcompare` builders to monitor the performance changes
in CQ/CI.

The benchmarks themselves are unchanged.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426034
Original-Revision: 03d6f9304340fdedeef69fce747d4b48c7071aac
GitOrigin-RevId: b772d80367fe7b839851a55512b4c48ec0d6bae7
Change-Id: I593ff3cced39adc34a1b63666a89b6765743592f