[roll] Roll fuchsia [wlan] Write beginning of duration in wlan_trace::duration
Instead of only writing the duration event at the end of a duration in
the wlan_trace::duration macro, the macro now always writes a duration
begin event at the beginning of a duration. This prevents ambiguity when
a component unexpectedly exits in the middle of a duration and there is
not duration present in the trace at all.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426355
Original-Revision: 5742a5b3af78a762e0af3b9b9ee09f7b5a95552c
GitOrigin-RevId: 42287f329195880e6ed01e35480e898076d2e453
Change-Id: I8b695ccea76c3c15d43c3e679bedb1f318c3b26e