[roll] Roll fuchsia [bt] Define a platform configuration for bt-rfcomm
Define a new AIB for bt-rfcomm, which can be optionally added to the
build.

Add a basic platform configuration for RFCOMM. By default, this is
enabled. This matches existing behavior in which it is packaged with
bt-init.

Add the RFCOMM AIB, if it is enabled, in the platform config rules for
Bluetooth.

Add a configuration capability to bt-init which is used to dynamically
route the bredr.Profile API. If enabled, Profile requests will be routed
to bt-rfcomm first. This behavior is unchanged.

Original-Bug: 457743695
Test: fx build, fx ota, verify bt-rfcomm starts
Test: fx test bt-init-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1412045
Original-Revision: 1a21f2d86fb633495d46f6d568ae0ce2f6604dd7
GitOrigin-RevId: 1a1e057cfbe2e65f7f7decd0ab1fc4fa6bfaadaf
Change-Id: Idbc135371560cf70596a28728dbdf71c4170dafb