[roll] Roll fuchsia [starnix][userfault] ENOTSUP on using userfault without SIGBUS
The only usage of userfault currently supported involves using
UFFD_FEATURE_SIGBUS. Return NOT_SUPPORTED as soon as possible when we
detect that we try to run without it.

Note that we can't error out in UFFDIO_API ioctl directly as running it
with an empty set of features is a supported way to get a list of all
supported features.

Test: UffdTest.*
Original-Bug: 391599171
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426295
Original-Revision: 19f52c69d21ca2bfc5af3576245f37150bf3e044
GitOrigin-RevId: 316e030ad095d10d4c677a47d7ea8c031a465555
Change-Id: If8eb9b56e1678e0da1eb07bdeac270322d59647b