[roll] Roll fuchsia [checksum] Remove `add_bytes_small`
The original implementation had a separate algorithm for calculating the
checksum when the buffer was sufficiently small (< 64 bytes). This was
originally added as an optimization, however current benchmarks reveal
that the main algorithm in `add_bytes` is more efficient, even when the
buffer is small.

Original-Bug: 462085360
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426474
Original-Revision: 7496eade2e6996c13157b308a98d65632d34dab3
GitOrigin-RevId: 339bc320748161ed78bbb440496ab561ae2bebfb
Change-Id: Iaa50861eaf0b2130afafebe56c1603576a14aab1