[roll] Roll fuchsia [kernel][x86] Downgrade spurious #DB OOPS to printf
When a spurious #DB exception occurs, x86_debug_handler will emit an
OOPS with some extra details.  This was added to help troublshoot
b/453746441.  These are now happening a lot on CI/CQ so we're
downgrading the printf to an OOPS in order to eliminate this as a
source of CI/CQ failures (think flake rate reduction).

Original-Bug: 453746441
Original-Bug: 461514946
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427774
Original-Revision: 71fad86440fdb6c6a1dcb8ff98605604f8920781
GitOrigin-RevId: 0e21d15258e5e2ec1b51f3e430d5ad6d5dddfea3
Change-Id: Ia440e6316163831a6628260e624ea3405683e189