[roll] Roll fuchsia [time][test] Re-enable `correction_by_slew` test
This change re-examines the exact sequence of events in the test, and
adjusts the manual clock advance and "run until stalled" to ensure that
two clock sample events are consumed from the test time source before we
proceed with other checks.

Ensuring that both events are consumed ensures that the clock slew code
is activated, which is what this test verifies. The time samples
delivered by the fake time source are deliberately chosen to elicit
clock slew, not a clock jump.

Original-Fixed: 397762299
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427042
Original-Revision: e0d624b11208a3ae326762bc51066efc3bf203e8
GitOrigin-RevId: 0638713fc48d89cf99fcbbc35ef52a58b2456846
Change-Id: I4788357598c5331154a73f2dfbed070979683ae4