[roll] Roll fuchsia [nmfs] Move filesystem to a module
Common style in Starnix is to place filesystems inside their own module
so that it can be self-contained. Update nmfs to be within its own
module and initialize with the kernel via the "Expando" pattern.

Add test target as well and ensure tests continue to pass. The concept
of a NetworkManagerHandle is now satisfied by the Expando itself so it
is no longer needed.

Original-Bug: 461796926
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416959
Original-Revision: 27396bd41cec5cbfef31b4230047f1b38b2b86ef
GitOrigin-RevId: 74ecacdebfca41d4176467376a881eaaba96d420
Change-Id: Ie0ac59722d658fd1ea2794ea44ed30aa159f6c4d