[roll] Roll fuchsia [libc] Generate <zircon/process.h> from YAML
This generalizes the build machinery for generating header files
using the llvm-libc hdrgen tool.  It exploits recent new features
in the tool to use it for headers that are not part of the LLVM
project and whose source `.yaml` files are in the Fuchsia source
tree instead.  This is used first for <zircon/process.h>, where
the hand-written header is replaced with a generated one.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1397053
Original-Revision: 5743a95ab9e1c3fd5e36aac4dcf94aa895ab9ec5
GitOrigin-RevId: 1836ac8df2372117998baa8b944fc30bfec21fc8
Change-Id: I14cccacfe45a5de4e27908a12dd3c7e1866a70c6