[roll] Roll fuchsia [memory_monitor2] Disable configurable collections
Before this change, we had two recurring sources of expensive memory
collections (excluding on-demand via ffx): one regular collection
every five minutes for cobalt, and another on a configurable rhythm
depending of the current memory pressure.

While this flexibility can be useful to adjust the scrutiny depending
on the current memory pressure, there is currently no known downstream
that uses this ability, and on some products, collection is so
expensive that we have to carefully time and count occurrences.

With this change, we now perform the one collection once every five
minutes, and we produce all reports from that one collection, reducing
the amount of CPU we use on collections without reducing (in all known
configurations) the frequency of the reports.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416614
Original-Revision: ecc00d8284fc6ce50f6b21c0491364995426a8c7
GitOrigin-RevId: 1dd4e0f428ca58054b08ffe52d75aa61f51355b1
Change-Id: Iec5aeb465c9c124e32137e94ee876e72c4967b79