[roll] Roll fuchsia [starnix] Simplify memory attribution
The restricted address space is at a fixed location, with a fixed size.
We thus don't need to dynamically query MemoryManager for this
information when attributing memory to Starnix processes.

Not requiring MemoryManager allows us to simplify the notification code,
as we can start sending the attribution as soon as the pid is known,
instead of waiting for MemoryManager to be instantiated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1416554
Original-Revision: 85d262760432c628176f3ef9d01804f25bab874e
GitOrigin-RevId: 020217b9c094b591f0f7022faf152209b9051f1f
Change-Id: I7bd1cf75876c7793f4c9071bb101f7a125b8cefb