[roll] Roll fuchsia [trace][Lacewing] Catch and ignore FIDL errors when stopping tracing
In the past, investigations into getting ZX_PEER_CLOSED while attempting
to stop tracing indicated that FuchsiaController was no longer able to
communicate with the device. Current evidence suggests that is no longer
the case, and so it's worth trying to catch and warn on this failure
instead of allowing it to escape and abort the test.

In all cases that we've examined, the trace_manager on the host side
receives the Stop FIDL message, flushes the trace back to the host, and
then responds. FIDL indicates to the trace_manager that the response was
sent successfully.

As long as Lacewing is actually able to read the trace, and is able to
send future FIDL messages without incident, we don't actually care too
much about this failure.

Original-Bug: 446873535
Test: run Lacewing tests that use tracing via fx test-remote
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427994
Original-Revision: 5dd424173051c686fa983fa404767ff587331d06
GitOrigin-RevId: 95a5bfb24f57d05c73e43fb73330021b7d46ae88
Change-Id: I3d02b631b572e97b225b763cefbe0677ddd72120