[roll] Roll fuchsia [fxtest] Allow Gemini API key to be passed to analysis script
This change updates the fxtest tool to correctly pass the GEMINI_API_KEY
environment variable to the Gemini analysis script when the
--gemini-analysis flag is used. It involves propagating the key through
the hermetic environment, storing it in the execution environment, and
passing it as a command-line argument to the analysis script. The help
text for --gemini-analysis was also updated to reflect the correct way
to set the API key, making it so users no longer have to pass it in
using --env.

Test: Updated unit tests to reflect change

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427028
Original-Revision: a4254b9962547a0055e72fcf75d259e20468e1be
GitOrigin-RevId: d22c82eb2253a5d0fc75ec50714588146ae49efd
Change-Id: I0096eb1ec6824b7300f650362a557b76f87422d2