[roll] Roll fuchsia [nmfs] Remove reconnection logic and mutex
The NetworkManager had logic to reconnect to the StarnixNetworks
protocol if the state between the socketproxy and the NetworkManager
ever got out of sync or the socketproxy Fuchsia component
crashed/restarted. This behavior has not been witnessed in builds and is
therefore likely not needed, as removing it simplifies the
NetworkManager data structure greatly.

Simplify the constructor by passing in the proxy and an inspect node for
builds where the network_manager is enabled and allow for the default
constructor when the network_manager is initialized from the filesystem.
In the next CL, the functionality will be updated to only instantiate
the NetworkManager when the network_manager flag is enabled rather than
instantiating it from the filesystem calls.

Original-Bug: 461796926

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1423436
Original-Revision: 62dbcfe12e8575ab9dac6369cca584a993a2310c
GitOrigin-RevId: 39ed805aa20029dd0a322f08dd80d96de19b92b0
Change-Id: Ifb805d2b224cc6e484509342d2cd4f09a6d7b8c2