[roll] Roll fuchsia [fxtest] Disable timeouts when debugging
This overrides any timeout set on the command line, and the default case
(where it is not specified) by setting the timeout value to zero.
Setting the timeout value ourselves also overrides tests that set it
explicitly in their test_specs.

This is the naive approach, but in https://fxbug.dev/462419969 we should
think of a better way to handle this so that the timeout case (when
reproducible at a developer's desk) is catchable in the debugger like
other failures.

Original-Bug: 462419969
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1428934
Original-Revision: aa04e389528d308bf902f8e4708a6d689e4381ee
GitOrigin-RevId: e164471618cdcc9a59478fc6f16fea75a6e16d40
Change-Id: Ibe9b86acde56e6976b0100cec16458bdfa802b91