[roll] Roll fuchsia [usb_vsock_host] Clean up randomly-allocated ports
When we initiate a connection from the VSOCK host, that connection
needs a source port and a destination port. The destination port will
be provided by the caller since it identifies what service we're
connecting to, but the source port is randomly generated and usually of
no interest to the user.

When we pick that random port, we mark it as reserved. When the
connection dies, we *should* mark it as free again. This is a bit hard
to do in our code and not a huge problem given the size of the port
space we have set aside for connections, but it's better to do for
cleanliness's sake.

Original-Fixed: 407622199
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1428135
Original-Revision: 3534440606ddf03c4fd784a93a941c9271a9ae23
GitOrigin-RevId: e8f4bc3a683e9fcea0493f693c4ddca7cfae6536
Change-Id: Ib849d554d9c53d66e358a1b4ea8a282c665b19cc