[roll] Roll fuchsia [test][revive] Add an "opt-out" annotation
For tests that have test-case-revive configured, we want to switch
from requiring an "opt-in" annotation on test cases that should be
revived, to having that the default and using opt-out on cases that
should not be revived.

This needs a soft transition, so for now, add the `opt_out` decorator
but still require the `tag_test` decorator.

Also, nerf the `_tag_name` parameter - simply using `tag_test` is
enough of an opt-in.

Test apps that aren't built with `params.test_case_revive = true`
will continue to run normally without revive, regardless of annotation.

Original-Bug: 460154311
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420880
Original-Revision: bdeed24229e9296f23e4ef3ad48295452f135dd8
GitOrigin-RevId: 30d5c4edb236cc71869c2f9aadd17e61c579fae4
Change-Id: If8dd013daf56c37fa50f478f56a00cca2e9a851c