[roll] Roll fuchsia [nmfs] Only instantiate NetworkManager with flag
By using "peek" in the Expando, we will only initialize the
NetworkManager struct when the "network_manager" container flag is
enabled in features.rs. Return EPERM on each of the filesystem calls
when the NetworkManager is not present.

The existing tests are split into the proxy enabled and non-proxy
enabled variant. Integrate any checks that are done internal to the
NetworkManager into the proxy variant of the tests and remove the
non-proxy variants.

Original-Bug: 461796926
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426775
Original-Revision: f08c62f90dfb7139d04b213658496b95e3b147ea
GitOrigin-RevId: 1c0a2fbbe021ec6669b8dc78ec12d881bfd3d329
Change-Id: I9e28afd369396c508de526454734bbc0bca04b36