[roll] Roll fuchsia [kernel] Print thread pointers in arch_dump_exception_context()
This makes the arch_dump_exception_context() for arm64 and x86
print the thread pointer registers that are not stored in the
iframe.  These are directly in the CPU state for the current
thread (only saved in arch_thread at context switch, not kernel
entry), so make it clear that arch_dump_exception_context() is
always called for an iframe from the current thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1429894
Original-Revision: 3ec21ba56f5050b8624aca781e6bdc9453aacdd5
GitOrigin-RevId: 65b975fa3294c0c9cb7dea53660ad3ba1e05d0dc
Change-Id: Ic11db044bec7ef3f4ebddbf39e5acfbb2b279223