[roll] Roll fuchsia [starnix] Remove /sys/kernel/mm/ksm
This should only be present if we support ksm, or if we need to
pretend that we do, but ksm is not enabled in our targets and the
directory doesn't exist.

This removes some spurious SELinux denials.

Original-Bug: 462356430
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427556
Original-Revision: ba0af1ce833862fd7c4bc199f73f58ff5a108fc5
GitOrigin-RevId: 8abec185e7947531fd5b83c10376b97f93c9c821
Change-Id: I4b542b7a50648eb84e62c192182082eda73400d8