[roll] Roll fuchsia [debugger] Tolerate non-existent threads in OnThreadExiting
This could be the case when we're keeping track of DebuggedProcess
objects, but are not attached to their debugger exception channel to
receive thread starting events, and then at some later point do decide
to attach, but haven't synchronized the threads yet.

Clients are already equipped to handle receiving the notification in
this case.

Original-Fixed: 463317618
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1431994
Original-Revision: 64a7bd282ee9e1eb01a5615bf4b689bf1d055215
GitOrigin-RevId: 9f9c868213d31fad725441dcd162e2b001568972
Change-Id: I5c27590e34fa26bef3fc64a1ffb1d24ffe09d8fc