[roll] Roll fuchsia [ffx][repository] Return user error in deregister if repo name prefix
This fixes the situation that ffx reports an internal tool bug when the
repo name for deregister is specified on the command line like 'ffx
target repository deregister fuchsia-pkg://repo-name'. This may happen
in manual use cases or when a script parses the output of pkgctl without
removing the prefix. Instead a user error is returned, explaining the
problem to the user.

Original-Fixed: 462716616
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430695
Original-Revision: 636c537870094fc1445cd832a50d6fde14d80e8a
GitOrigin-RevId: 62dc80dfeb9f6907646a8cf2a10a3d5a5b1a6d97
Change-Id: I2a72034f070ca9ab14ac0ca9ac96edb1026202ac