[roll] Roll fuchsia [ffx][trace] Fix symbolization crash
This corrects a misunderstanding I had about the
SessionParser::parsed_bytes field.

I thought it held the raw
bytes for the TraceRecord that was most recently returned by
SessionParser::next(). In fact it holds the raw bytes for all of the
RawTraceRecords between the previously returned TraceRecord and the
current one (including the bytes for the new one. This is because not
every RawTraceRecord maps to a TraceRecord. Some RawTraceRecords like String and Thread
interning records are handled by the SessionParser, so that trace
consumers don't have to worry about these implementation details.

The code has been updated so that in addition to setting the
parsed_bytes, the parse_next() function also sets the record_range field
to the range within parsed_bytes for the actual record that was
returned. This allows the ffx trace process_event_records function to
pass just the bytes related to the TraceRecord into the processing
function and preserve the other bytes.

Original-Fixes: 462217132
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1432774
Original-Revision: 9d86874e78b513ab5f442fddae37c92e9223e64b
GitOrigin-RevId: 947617f1987680087447c9e0e647128757a75f66
Change-Id: Ie7e9c957d630cd78258ef43157013cc5b20dc034