[roll] Roll fuchsia [testsharder] Add summary_if_skipped field in output
Create a summary_if_skipped field to replace the summary field. This is
done because determining if a shard is skippable is now done by looking
at the skippable field instead of checking if the summary field was
populated.

The summary_if_skipped field will be used to upload results to resultdb
of additional shards that the fuchsia.py recipe decides to skip
(additional to the list of shards that testsharder decides are
skippable) and will replace the summary field to be more explicit about
under what conditions this field is used.

Original-Bug: b/450062147
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1432754
Original-Revision: 22c4ca52ab30097797c53961c6b33b09c4e940b8
GitOrigin-RevId: f1668bfd0aae7211112aef8e5ce0b9a8daee10cb
Change-Id: Icc706cb22a2e94be1698ddd018bb158012008149