[roll] Roll fuchsia [audio][inspect] Make BufferTracker thread-safe
The BufferTracker class can be accessed from multiple threads,
but it was not thread-safe. This could lead to data races when
buffers are submitted and completed concurrently.

This change introduces a mutex to protect the internal state of
`BufferTracker`, ensuring that all access to its member variables is
properly synchronized. Additionally, `ZX_ASSERT_MSG` calls were
replaced with `FDF_LOG` to prevent crashing the driver in case of an
unexpected state.

Test: audio-inspect-recorder-test
Original-Bug: 462554546
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1429571
Original-Revision: 61029c596526adbd42915dcf438b209f9e3f66de
GitOrigin-RevId: 64e9fc4dfe997f0b1d8970f8ee8892bd6585b958
Change-Id: I95436840a8029987fddec79066441aa113c2f630