[roll] Roll fuchsia [lowpan] Fix input type for Verhoeff checksum calculation.
The otVerhoeffChecksumCalculate API takes aDecimalString as the first
parameter, which should be a string containing decimal digits converted
from the ephemeral key byte array.

Instead of passing a pointer to the first byte of the raw array, we need
pass a pointer to the string representation. This CL converts the key
array to a decimal string and passes the string pointer.

Original-Bug: b/373664917
Test: manual test with Thread Utilities app to verify the overall ePSKc
      work flow, works as expected.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420636
Original-Revision: 0900a2f1e2ebf58fa015ddb9b91f42ece63e113d
GitOrigin-RevId: 045a9328e41fae4ed25f3e81e09da4e4b86baca1
Change-Id: I8356d1e68bade2de57b0f7fe114062baa84df577