[roll] Roll fuchsia [starnix] expose real block devices
When starnix is routed a /block directory, it takes anything in it and
puts it in /dev/block, using the semantic label it was routed as in the
block dir. Starnix handles converting file reads and writes to the
device into fuchsia block client reads and writes.

Original-Bug: 450325631
Original-Bug: 407091711
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425841
Original-Revision: f832e51612ea8f24fdeb8f0cf3d6535f2c98810f
GitOrigin-RevId: 82b1f67d4c53011c78e8c02af2a0b5bf6e1251a3
Change-Id: Ia91c90473c563be205dd02606cf30832f4f43fc9