[roll] Roll fuchsia [wlan][test] Don't write a trace if a test exits normally
Writing a traces for passing tests is a waste of resources. We can use
other means to generate traces for passing tests when we need them for
comparison with failing tests.

A trace will be written for a test if the test did not complete within
2 minutes, the streamed bytes to the trace file exceeds 50MB in size,
or the test panics.

Original-Bug: 447408672
Test: Manually verified a trace is recorded in each scenario.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426754
Original-Revision: cd3e8a952cdf87df36085ebc581bf0ba0381832c
GitOrigin-RevId: 07358091a614f8b95fac8ed59a7013a27ffc7e32
Change-Id: I8bfb1d119645f426a19bc78df88955337b069bf2