[roll] Roll fuchsia [owners] Remove nikitajindal@google.com from src/storage/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/nikitajindal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433954
Original-Revision: b4a5c40c2df3a093014868c3d655c04f0efa2c7e
GitOrigin-RevId: 6f890293f3241bd025a9ba47a13a35a4e64e4a4d
Change-Id: Ia7e7441fd6f973139b65e0a6817f9625e7ccc440