[roll] Roll fuchsia [wlan][hw-sim] Create stress test to tease out flakes in hw-sim startup sequence
We've seen flakes in the startup sequence of a hw-sim test, specifically
while waiting for the WlanSoftmacStart event. This test runs the startup
sequence 50 times in an attempt to trigger a flake more often.

Inclusion of this test is temporary until we have more data on the flake.

Original-Bug: 447408672
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426757
Original-Revision: f89c85f723f05a5f33c9c590cca55db9d8a18c84
GitOrigin-RevId: fe8dead96ad03ae63989f1228b5bc12d68a021d7
Change-Id: I29edab7da1e796978cd9008b77a658f6c5387278