[roll] Roll fuchsia [network/tests] Only parse real TCP segments
Update the socket integration test's TCP parsing logic so that it
only tries to parse TCP segments out of data that *is* a TCP segment.

The original logic would try to parse a TCP segment out of every packet,
and expect a failure if the packet didn't carry a TCP segment. That's
not guaranteed behavior and shouldn't be relied on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433854
Original-Revision: 66528d9f9317816ab878d261865455cf78eb86e9
GitOrigin-RevId: 0f0915550992e1a3368f8b4ce049f8821341fa52
Change-Id: I4821231ef4e76f2d00f21da05bd1bd421585aa62