[roll] Roll fuchsia [fxfs] Ensure superblocks aren't confused with earlier filesystems.
If we find a valid-looking B superblock with a generation greater than
the A superblock, we will assume it's newer *and from the same
filesystem*. This can lead to problems if we overwrite one filesystem
with a new one using image_builder_mode, where we leave the second
superblock intact.

This CL tries to read the highest superblock generation. If they don't
exist, that's fine, we use generation 1. If they DO exist, we use
generation + 1 to ensure that the newly written superblock supersedes
all previous ones. Superblock generations are wrapping so this works
for all generation values.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433239
Original-Revision: 94065473c0f13273bc142e922862ee3a09582fab
GitOrigin-RevId: 0ef54f1e838b60d4332c72d33b20b0e992ace630
Change-Id: If6f09d922f9b7ece0130f8cd3f8ad06a90398c69