[roll] Roll fuchsia [memory_monitor2] Fix test broken by missing config
In change 1416614 we removed the dependency of memory_monitor2 on its
assembly-provided configuration (because it is no longer
configurable); as a result we removed the inspect node where we would
log the current configuration before.

We neglected to update an assertion in one of the end to end test that
verifies that the node in question is always populated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433894
Original-Revision: d6e3368ea3f74bed3cc8b37ca96e25773dd48b74
GitOrigin-RevId: b3451ef8d40ceec6bd1bc7bbc167b0bcc7c59bcd
Change-Id: I285951e29bb0f15f349658e8bc223baa11bd8ae0