[roll] Roll fuchsia [wlan] Create core e2e test for SAR settings
This CL introduces a SAR settings test, which validates that:
1. all SAR settings can be set via the core APIs
2. the device connects to an AP
3. the SAR setting is persisted after the connection

This test is written in sync python. To facilitate that, a new sync copy
of the Core Testing base class is added. This class (and the test) make
extensive use of Honeydew to avoid code duplication.

Test: fx test sar_setting_test
Original-Bug: 463663537
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433095
Original-Revision: 260cef01274df86355621952d9240d0d41243021
GitOrigin-RevId: 69aa0dcca74be6ea299bb265517f9609fde9d306
Change-Id: I1616ecacdebcc9594a72cd118f2d5c9c72a7d0ab