[roll] Roll fuchsia [components] the target of a route is required
This commit moves the target WeakInstanceToken from the Request struct
of a route to the argument set, and makes it required. This change means
that whenever a capability route is to be performed, the target (aka
initiator) of the route must always be supplied, and a default value
cannot be relied on. This is necessary because sometimes default values
for routes are set at a different point in the component graph than
where routes originate. For example if a dictionary is exposed and then
accessed through the exposed directory, when an entry in the dictionary
is opened the "default" target would be set where the capability was
offered into the dictionary, instead of where the exposed directory was
accessed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430014
Original-Revision: e41245e0c6aa81525efba8283e3a42ad9659c8ef
GitOrigin-RevId: 463d6cb8bc2677d63529e55be6faf79c371abe23
Change-Id: I9c12a5bcaf0a789c3989c8e28ebac9f06977e52d