[roll] Roll fuchsia [mkbootimg] import to third_party
This repo contains some tooling that is useful to create and analyze
Android boot images.

We already have a copy of mkbootimg.py in the tree but it would be
better if we could switch to a more formally imported/versioned copy
here. Additionally this repo contains unpack_bootimg.py which is useful
to analyze and dump boot images.

Initial version is android16-qpr1-release-fuchsia, which is the same as
the most recent AOSP push android16-qpr1-release but with a cherry-pick
to include a LICENSE file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430735
Original-Revision: 73402ff9eb93ea381df0cbadda37c613c3c85de9
GitOrigin-RevId: b585cf25f34bebf0fc9df2bec4f79bd9476d2c0a
Change-Id: I9b5db1a6215b4f32c58e240077244c5b8429d9c3