[roll] Roll fuchsia [ffx][doctor] Add check for ffx-usb-driver status
This change adds a new check to `ffx doctor` to report on the
status of the `ffx-usb-driver` process. It checks if the process
is running, reports its PID and the socket path it is listening on.

It also warns the user if the discovered socket path does not match the
path configured for ffx.

Test: Added unit tests. Manually invoked ffx doctor.
Original-Bug: b/460195632
Original-Fixed: b/460195632
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420157
Original-Revision: 65545174281ffb99ad10a34652fd1af61b6c8567
GitOrigin-RevId: beefd18546ac2cca3dcbadebd09d18687ad2a2d1
Change-Id: I4b10eaee1f0f9ce21e603b0bf7ea23363e5b1c14