[roll] Roll fuchsia [ktrace_provider] Quiet Logs About Exceeding Profile Allocation
Ktrace provider uses a thread profile to ensure it's able to timely copy
data. Under high memory pressure and/or cpu load, it can consistently
fail to finish its copy of the trace data within its allocation, causing
log spam.

This patch quiets the warning to debug. When viewing a trace, it's
obvious when we drop kernel data due to not being able to copy it fast
enough -- we get events emitted from the kernel showing the spans of
time we dropped records.

We could go a step further and keep a running average of how long it
takes to copy the data. However, because we have kernel trace data, we
can use that the compute in-post the time the ktrace_provider took to
copy the data, either by using the syscall events to measure
zx_ktrace_read calls, or scheduler events to see how long it was
scheduled for.

Original-Bug: 465141469
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438015
Original-Revision: 08faaf985a021c8543f258798d8a13e05d7e1a5c
GitOrigin-RevId: baaa07ed1cbe411f64a7e489e80ee2f997ab397f
Change-Id: I3c6d426e7678dddf39892e40443f21f39d52b3ea