[roll] Roll fuchsia Reland "[botanist] Restart the ssh controlmaster when detecting a reboot."
This is a reland of commit d081217f14dd484650e63794ff477852227c06c8
This differs from the original CL by having testrunner call to restart
the ssh controlmaster before each test if it detects that it's not
running as opposed to botanist continuously restarting it. This will
allow us to reconnect to the target as well if we detect that the SSH
connection was lost whereas before we only try to reconnect after a
failed test.

Original change's description:
> [botanist] Restart the ssh controlmaster when detecting a reboot.
>
> Original-Fixed: 463446410
> Change-Id: I9ef6ea1a8ba2d27061ba4752ae8c70c497403c85
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433314
> Commit-Queue: Ina Huh <ihuh@google.com>
> Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
> Reviewed-by: Carver Forbes <carverforbes@google.com>

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1434301
Original-Revision: 318a3f4415d5cbba73ecf51c60f9be8c6001c8a2
GitOrigin-RevId: 51c3566821399febdd8591dadd0ffbd020b469eb
Change-Id: Id3fd841609dadfc8600ff80b5a59c198a139643c