[roll] Roll fuchsia [cog] add git polyfill for use in workspaces
This CL adds a git wrapper which will be used inside of the cog
workspace. This script does not add any commands and is not wired up to
fx in this CL. That logic will be added in a later CL. To test, you can
just execute the script directly.

Original-Bug: 461899545
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438302
Original-Revision: 5e51889c237c257ad875a013722349a5e6bfab0d
GitOrigin-RevId: a732f0e46b6fba718d7f3ebf6f6007a8689f574a
Change-Id: I78f50595e864b18df7330b24ac939b3bc4cff1bd