[roll] Roll fuchsia [debugger] Ensure threads are synchronized on exception
Now that DebugAgent is correctly *not* attached to individual process
exception channels in certain modes, we are not receiving thread
starting events.

An implication of this is that if there is an exception in a process
underneath a job we're watching, we have to be very sure that we
actually have a DebuggedThread object for it, which means we might need
to synchronize the DebuggedProcess with the underlying Zircon source of
truth before collecting information that some FIDL clients might expect.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1435995
Original-Revision: 5e7e56fef2cde3eecbc68c434f7f985674ea8754
GitOrigin-RevId: f1591d2327be24841230dc4101e297be362fa693
Change-Id: I5dc4c0d00aacd2095bcadf2fa4231e4bc1dc093e