[roll] Roll fuchsia [debugger] Allow `continue` over exceptions
..instead of repeatedly hitting the same exception over and over again.

This helps exceptions in tests in particular which are effectively stuck
until a user does some form of a "detach" event.

In the CLI interface this is done by typing "quit" or "detach", but in
VSCode there's only one "detach" button, which is not obvious that that
is the one to press to get out of the debugging session. Now the "play"
button will cause the test to exit after two presses, due to
second-chance exceptions. Handling those is done in a followup, as well
as enabling the other buttons to not get stuck.

Original-Bug: 444070494
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421474
Original-Revision: ce95984b25788bb8ad590bcdb6a23fc3e0675c08
GitOrigin-RevId: b9658e015f4ed6e8fc76d7f80a8c1fa3aef56067
Change-Id: I0ed80b6cd1795bcec4b2f9d26ab5f9d7c48ebeb4