[roll] Roll fuchsia [driver_runtime] Include threadid and role in stall message
Includes the threadid that was stalled and the role of the thread
pool when warning about a stalled thread to make it easier to
attribute the message to a cause.

Also lowered this message to an INFO level message since it's not
really a critical problem until all threads are stalled, and that's
still a warning.

Original-Bug: 442604405
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1435398
Original-Revision: f0cdc918cad966286ea83ae3fcfde8cb143a39d4
GitOrigin-RevId: 463ea05d0a8f8a0287c7e948d7f3f61f29541f51
Change-Id: Ibffac30f291dd65fc26dfd212a1ee9f3d9fb9153