[roll] Roll fuchsia [debugger] Skip second-chance exceptions in tests
This makes use of the test_failure_stack_matcher to inform the process
object that it is a test, which lets its threads query the process for
additional help when deciding how to handle an exception. This creates
an illusion for tests that disregard second-chance handling and
correctly deal with debugger installed breakpoints, hardcoded
breakpoints, and other types of exceptions without needing information
about the specific test run time.

Importantly, this has no effect until we actually realize that we're a
test. This is important, because we don't want to skip second-chance
exceptions in all cases.

As long as we have a stack matcher for a particular runtime's test
assertion failure path, we can identify that we are a test.

Note that this does mean that the test process will crash "for real",
since the exception will be forwarded, rather than handled.

Original-Bug: 444070494
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409614
Original-Revision: 46ac9760a267d35a9251cfd48a6977faaa165704
GitOrigin-RevId: 2cfa43a3c89ea7f4f90ad4d9f610ea9bd2ef8666
Change-Id: I635473baca22be8ac1be6c329144c869317f769a