[roll] Roll fuchsia [toolbox] Expose tool protocols to toolbox
Gather all the protocols currently routed from into the SSH namespace
and make them available to the toolbox instead, so we can have a single
namespace that concentrates the tools available for shells.

As we're doing this, adopt the toolbox pattern of moving all the offers
to the toolbox dictionary closer to where the components are inserted
in the realm.

Original-Bug: 416063207
Original-Bug: 447668986
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427114
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: 2986f1ef942493c9270d521c947fbe83e765201c
GitOrigin-RevId: ead8c3e327d7093f71374df13c98cf593559f1bb
Change-Id: I09b9cbd28cf81c5037842fdf1f750b857347598e