[roll] Roll fuchsia [wlan] Add factory_address to fullmac FIDL query
This change adds a `factory_addr` field to the return value of the Query
method of WlanFullmacImpl protocol. This represents the factory MAC
address, as opposed to the mutable `sta_addr`. This is a pre-requisite
to enable MAC randomization.

This change also populates that field for the brcmfmac driver, returning
the `sta_addr` as the `factory_addr`, since MAC changes are not
currently supported for the brcmf devices.

Lastly, this change depends on the CL below, which stubs out that return
in the synadhd driver, in //v/g/.

Original-Bug: 455612638
Depends-on: turquoise-internal:I1e72cad890a1c9202e9880aa81f6e0cc51160892

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405994
Original-Revision: 1c6396804859bb2736317d9d7088effa2a5a5e9c
GitOrigin-RevId: 5d74c32f8a473266ba009e3b4a9f55707a26ec3c
Change-Id: Icf0e0cef87328d125df283ce9ae850e97e331e88