[roll] Roll fuchsia [fdomain] Convert Handle to arbitrary borrowed types.
With this change it's possible to run `as_unowned::<T>` for any
`T: HandleBased` for a borrowed usage of a handle as an abirary type.

This is primarily used in future fuchsia-controller changes where
FDomain handles are stored in a map for a given fuchsia-controller
`Context` object.

Original-Bug: 465842888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440484
Original-Revision: b09123b64bb4128579ebfa8346e5eb5d37a07029
GitOrigin-RevId: 636fb756bf2d3d67fc7b5af40bea7e925a5ae8cc
Change-Id: I6f92f57367f4a591083e1bbbaa5dc223722b3d10