[roll] Roll fuchsia [debugger] Step should not get stuck on exceptions
...which provides a much nicer UX flow when stuck on an exception, both
on the CLI and in VSCode.

Prior behavior continually tried to single step over an instruction that
raised an exception, which would get marked as handled and then step
again, raising the same exception, causing no progress to ever be made.

Now, we make use of second-chance exceptions to allow us to forward,
rather than handle, the exception, which still allows the user to take
action at the second chance, and for optimizations to be made in certain
contexts (like tests) to not get stuck on a test failure line.

Original-Fixed: 444070494
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411755
Original-Revision: 1234118473f4f5052bbf8c1ae2bd67dd6698b3f5
GitOrigin-RevId: 6dd38a578a445ff461213cf6d1d2af6a71475fcb
Change-Id: Ib030f8429422fcf9999d98b5e6e8f36fe48a0ece