[roll] Roll fuchsia [components] add OpenExposedDir to fuchsia.component.Controller
When a child component is created a fuchsia.component.Controller may be
attached to it. This allows the parent to interact with this child
without needing to constantly reference it by name, which somewhat
simplifies things and removes one possible error category.

Add OpenExposedDir to this controller protocol, to complement the
already existing GetExposedDictionary and mirror the functions on
fuchisa.component.Realm.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1432594
Original-Revision: 2d160257be343ed43310085bd95d85da12abf3b5
GitOrigin-RevId: 9392419faded7573a04e8179049673a145641e89
Change-Id: I35d75a8be64a55b4211040c6012099c987b040aa