[roll] Roll fuchsia [socket-proxy] Fully clean up state on connection
A reconnection to the StarnixNetworks or FuchsiaNetworks protocol
indicates a reset of networking state for the socketproxy. Previously,
we were clearing the networks list, but not resetting the default
network id for the respective registry. This led to sockets attempting
to get the mark of the default network, but not finding any network in
the network registry matching the id.

Original-Fixed: 465164164
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441413
Original-Revision: c7f3ae6f5c9f2a77359975097d3371f4149b3352
GitOrigin-RevId: d8410a4202a5543165c96d3c51b7cc00cd19482b
Change-Id: Ied90053900b51520f75648666f78dafd9372fa1c