[roll] Roll fuchsia [kernel][vm] Use prev_eviction_result for same page check
For IncorrectPage eviction failure, the page will either be moved out of
the Isolate list when the ownership change completes, or it will remain
in place when ownership changes due to clone creation/deletion. In the
latter case, eviction is going to be attempted on the page again, which
is not a bug. To correctly account for this, use the previous eviction
result for the same page check, and update the previous result even if
the failure reason is IncorrectPage.

Original-Bug: 434361683
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441521
Original-Revision: 1dc0006afa9be876c7ea4bb55407b1872a849e77
GitOrigin-RevId: 32a70f2176fc3c9505fc60528c8dfdc80192a5a7
Change-Id: Ibad93e82bebf59e90e388646c5c3dc8c5e6c5e25