[roll] Roll fuchsia [libc] Have MemorySnapshotStartArgOnSuspendedThread check all regions
This particular test checks that the start arg passed to a new thread is
accessible in the TLS region via __sanitizer_memory_snapshot. As libc
changes, the actual location of the allocation may change. Update this
test just to assert that the start arg is searchable via any of the
callbacks since it should be kept live.

Original-Bug: 342469121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441748
Original-Revision: 09413967c7c87a8f9be4d335aaa174515317533f
GitOrigin-RevId: 9f966de8b69b28b3d99ff1e0fa01e1dfff6dd3d3
Change-Id: I25143e0a457faf7fda757dbeff1fb1df9b5f8556