[roll] Roll fuchsia [dwmac] Increase reset timeout
Some devices seem to time out when querying the bus mode after reset.
Increase the timeout in an attempt to avoid this. For most devices this
should have no impact and for devices that are on the edge it should
should increase make them succeed every time.

For devices with faulty hardware or other issues this will slow down the
time to failure. Hopefully that will not impact a significant number of
devices.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441274
Original-Revision: 16bbce4a3f841513393801aa09b211ebfc39c47d
GitOrigin-RevId: 9b965c54ec9fe589021cb62cd2b4768a065e24fd
Change-Id: I2cabc4576e07230948d656062c44c8bb42629f00