[roll] Roll fuchsia [fuchsia-controller][honeydew] tracing refactor.
This refactors the tracing affordance to use fuchsia-controller more
idiomatically. This avoids the use of threads (which is likely to cause
undefined behavior with fuchsia-controller code, as it assumes a
single Python thread).

This also uses the async adapter code to clean some lines up that were
previously using `asyncio.run()`.

This also updates some of the fuchsia-controller documentation to
clarify some of the potential pitfalls in the existing libraries.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440728
Original-Revision: a357f3cf8311490a18861de92dd73e87b6e48782
GitOrigin-RevId: 59810565dd0e18fb8cc48c2465b170c775f09619
Change-Id: If50315c3da2541d58c33ac84117056ad3ed03672