[roll] Roll fuchsia [ffx][fuchsia-controller] Adds FcStatus error type for FDomain.
This is the new exception used for describing FDomain based errors in
fuchsia-controller. This isn't in use in this change, but is employed in a
follow-up. This just contains a fair amount of boilerplate.

NO_IFTTT=Does not affect fuchsia-controller ABI.
Original-Bug: 465842888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440830
Original-Revision: 36822de106ca02a11e5a8adcf5370701654a1d2e
GitOrigin-RevId: fdb20a5031f29968b155b9a55b058f3e7b31bdcf
Change-Id: Ied5a27344c3a7b29dd3f28beeb63ff2ffa0fc9b2