[roll] Roll fuchsia [log] Do not use a stack based buffer for sending log message
This buffer is responsible for 4Mo of permanent used memory in Starnix
when userspace logs to kmsg.
Moving this on the heap ensure this memory is reused using scudo buffer
instead.

This can probably be further improved by not initializing this memory to
0.

Original-Bug: 466294903

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441148
Original-Revision: f4d53a803a322e3070be31619f51dc2324d03d7f
GitOrigin-RevId: 1a8f727c75af373b2c322d21374794a6ffa07137
Change-Id: Ib194ac47dd3b9c25e4e92c19a7a22c7d5388efca