[roll] Roll fuchsia [cog] run git-polyfill inside FUCHSIA_DIR
This change makes it so the git-polyfill executes from within the
fuchsia directory. This is required to satisfy the constraints of the
hermetic-env. We have also passed along the original invoker directory
so it can be used in the commands if needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441282
Original-Revision: 6978a018cbf96dddc72253c43308ce1aae58ec58
GitOrigin-RevId: 4ed6601a19f47c0ab970ecd600fabe7608e97bb6
Change-Id: I5265023908488f2e4f2bb83716ad87efa0c5f6ec