[roll] Roll fuchsia [packet] Introduce SplitByteSliceBufferView
Expose an implementation of `BufferView<B>` on top of a SplitByteSlice.
This type will be used in a subsequent packet-formats change to parse
TcpOptions.

While here, fix a latent bug in the `SplitSliceBufferView::take_back`
implementation. Fortunately this code path was unused so there should
not be any behavior changes, but it removes a trap for future users.

Original-Bug: 462085360
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1442308
Original-Revision: f0f05f153988e9eab0c475cca9bc4f1e949b2310
GitOrigin-RevId: edb4cc3106ebef8be6c8eb18e9dbf9069388ab62
Change-Id: I201321d7baebb361b6cdc88c28b78d70f9709df1