[roll] Roll fuchsia [netstack3] Improve route table `Debug` impl
The old `Debug` impl formats the pointer address of the table, which
is not very useful. This change improves the impl to output the
bindings ID, which is a numeric value.

Also change the bindings_id of a Core RoutingTableId to use an opaque
ID type instead of `u32`. This way Core has no obligation to verify
the given Bindings ID is unique and it is up to the Bindings to ensure
this.

Original-Fixed: 462746973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438304
Original-Revision: 77d13494594add2377b66c869ab62d6a35b23145
GitOrigin-RevId: be82544075e3dda661d86a6bb72956fef60714c8
Change-Id: I728535e9d5bab5be495da512dab2c3beb508a5a0