[roll] Roll fuchsia [fxfs] Fix race while force-locking in flush
If a Crypt error occurs while flushing a volume, we still want flushing
to proceed, so we force-lock the volume and retry.

There was a race condition which arose if this happened at the same time
that the volume was being unmounted, which could cause force-lock to
fail.

Fix this by making force-lock tolerant of the volume already being
locked.

Original-Fixed: 457763372
Multiply: fxfs-platform-test.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441513
Original-Revision: 9fd9e27d9f4a48579406335961562132927ee990
GitOrigin-RevId: a414e690b1d4836cd5dc3ec58ecc80036ea8be81
Change-Id: I3a0a1a9b4685bea122096bd15824711249c7fc62