[roll] Roll fuchsia [cml] rewrite spanned -> context
yet another redo of my still not living code.
the inspiration was because the current merging
algorithm raises validation errors which my Spanned*
types cannot effectively catch.

The Spanned* types are replaced by Parsed* and Context*.
The Parsed* of each type is an intermediate step added
for parsing json. The Context* naming is the final form
that validation interacts with and uses to raise errors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440408
Original-Revision: 8807278904977aaea6b751d6035997da30915e5d
GitOrigin-RevId: cc744c2e32aa1ff74c7380922e988018a88b6732
Change-Id: Id8237dff9a84991a248d091f511d17f1777ff495