[roll] Roll fuchsia [assembly][vbmeta] Support rollback indices
This change adds support for constructing a VBMeta with a specific
rollback index. We do not need this for Fuchsia VBMeta (and have adopted
no release convention around it), but we need to have a non-zero index
for VBMeta that we construct for the Android pVM context, even when
non-rollback-protected contexts. When we eventually add assembly logic
for such VBMeta, we'll just simply specify a 1 for now.

Original-Bug: 449201661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425320
Original-Revision: 3b31e5b7fbabbe5f8aac773f0110c1b2c9a256ff
GitOrigin-RevId: 33c04f3239cb9d090b40f3ffe26fb6b9622aaf80
Change-Id: I243576fb75c22cddc9855e20368d32532895ac3b